Retrieve song lyrics for an album

Obtain the lyrics to an album in a tidy format.

genius_album(artist = NULL, album = NULL, info = "simple")

Arguments

artist
The quoted name of the artist. Spelling matters, capitalization does not.

album
The quoted name of the album Spelling matters, capitalization does not.

info
Return extra information about each song. Default `"simple"` returns `title`, `track_n`, and `text`. Set `info = "artist"` for artist and track title. See args to `genius_lyrics()`.

Examples

genius_album(artist = "Petal", album = "Comfort EP")

#> # A tibble: 81 x 4
#>    track_title track_n lyric                                          line
#>    <chr>         <int> <chr>                                         <int>
#>  1 Comfort           1 And I see the way you look                        1
#>  2 Comfort           1 When you think I'm not looking forward            2
#>  3 Comfort           1 That I'm lying in a trap of scared assumption     3
#>  4 Comfort           1 It's a certain kind of sickness                   4
#>  5 Comfort           1 A wrong turn of conversation                      5
#>  6 Comfort           1 A kiss on the trembling hand of our love          6
#>  7 Comfort           1 And you could barely drive when I said            7
#>  8 Comfort           1 "\"I don't fucking care anymore"                  8
#>  9 Comfort           1 I don't see the point of lying for                9
#> 10 Comfort           1 "What I am only tearing apart.\""                10
#> # ... with 71 more rows

genius_album(artist = "Fit For A King", album = "Deathgrip")

#> # A tibble: 404 x 4
#>    track_title         track_n lyric                                line
#>    <chr>                 <int> <chr>                               <int>
#>  1 The End's Beginning       1 <NA>                                   NA
#>  2 Pissed Off                2 Everyday is another war                 1
#>  3 Pissed Off                2 We live and die in a hopeless world     2
#>  4 Pissed Off                2 Laying waste to the innocent            3
#>  5 Pissed Off                2 With no regret, I'm sick of it          4
#>  6 Pissed Off                2 We create, we destroy                   5
#>  7 Pissed Off                2 The descendants blade                   6
#>  8 Pissed Off                2 Will pierce the flesh                   7
#>  9 Pissed Off                2 We attack and deploy                    8
#> 10 Pissed Off                2 We are slaves to nothing                9
#> # ... with 394 more rows