Retrieve song lyrics from Genius.com

Retrieve the lyrics of a song with supplied artist and song name.

genius_lyrics(artist = NULL, song = NULL, info = "title")

Arguments

artist
The quoted name of the artist. Spelling matters, capitalization does not.

song
The quoted name of the song. Spelling matters, capitalization does not.

info
Default `"title"`, returns the track title. Set to `"simple"` for only lyrics, `"artist"` for the lyrics and artist, or `"all"` to return the lyrics, artist, and title.

Examples

genius_lyrics(artist = "Margaret Glaspy", song = "Memory Street")

#> # A tibble: 27 x 3
#>    track_title   lyric                                    line
#>    <chr>         <chr>                                   <int>
#>  1 Memory Street Ring the alarm                              1
#>  2 Memory Street I'm on memory street                        2
#>  3 Memory Street With him on my arm                          3
#>  4 Memory Street And my feet on the dash of that car         4
#>  5 Memory Street I don't dare                                5
#>  6 Memory Street Walk down memory street                     6
#>  7 Memory Street Why remember                                7
#>  8 Memory Street All the times I took forever to forget?     8
#>  9 Memory Street Call the guards                             9
#> 10 Memory Street I'm at the gates                           10
#> # ... with 17 more rows

genius_lyrics(artist = "Kendrick Lamar", song = "Money Trees")

#> # A tibble: 123 x 3
#>    track_title lyric                                                  line
#>    <chr>       <chr>                                                 <int>
#>  1 Money Trees Me and my niggas tryna get it, ya bish                    1
#>  2 Money Trees Hit the house lick: tell me, is you wit' it, ya bish?     2
#>  3 Money Trees Home invasion was persuasive                              3
#>  4 Money Trees From nine to five I know it's vacant, ya bish             4
#>  5 Money Trees Dreams of livin' life like rappers do                     5
#>  6 Money Trees Back when condom wrappers wasn't cool                     6
#>  7 Money Trees I fucked Sherane and went to tell my bros                 7
#>  8 Money Trees "Then Usher Raymond \"Let It Burn\" came on"              8
#>  9 Money Trees Hot sauce all in our Top Ramen, ya bish                   9
#> 10 Money Trees Park the car, then we start rhymin', ya bish             10
#> # ... with 113 more rows

genius_lyrics("JMSN", "Drinkin'")

#> # A tibble: 87 x 3
#>    track_title lyric                                     line
#>    <chr>       <chr>                                    <int>
#>  1 Drinkin'    I've been drinkin'                           1
#>  2 Drinkin'    "I've been drinkin', "                       2
#>  3 Drinkin'    I'm drinkin' all the time                    3
#>  4 Drinkin'    All the time, every night                    4
#>  5 Drinkin'    I've been thinkin'                           5
#>  6 Drinkin'    Tell me what you're thinkin' 'bout           6
#>  7 Drinkin'    Thinkin' every night, I don't feel right     7
#>  8 Drinkin'    Don't feel right                             8
#>  9 Drinkin'    Some might think I've lost my mind           9
#> 10 Drinkin'    You've lost your mind                       10
#> # ... with 77 more rows