Retrieve artist discography with song lyrics and audio info

Retrieve the entire discography of an artist with the lyrics of each song and the associated audio information. Returns the song data as a nested tibble (see tidyr::nest). This way we can easily see each album, artist, and song title before expanding our data.

get_album_data(
  artist,
  albums = character(),
  authorization = get_spotify_access_token()
)

Arguments

artist
The quoted name of the artist. Spelling matters, capitalization does not.

albums
A character vector of album names. Spelling matters, capitalization does not

authorization
Authorization token for Spotify web API. Defaults to get_spotify_access_token()

Value

A nested tibble. See tidyr::nest.

Examples

# \donttest{
get_album_data(artist = "Wild child",
               albums = "Expectations")

#> Warning: `html_session()` was deprecated in rvest 1.0.0.
#> Please use `session()` instead.

#> Joining, by = c("album_name", "track_n", "track_url")

#> Warning: All elements of `...` must be named.
#> Did you want `data = c(line, lyric)`?

#> # A tibble: 36 x 41
#>    artist_name artist_id    album_id    album_type album_images album_release_d~
#>    <chr>       <chr>        <chr>       <chr>      <list>       <chr>           
#>  1 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  2 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  3 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  4 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  5 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  6 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  7 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  8 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#>  9 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#> 10 Wild Child  1xLMexpeeTK~ 5dab4s1Urw~ album      <df [3 x 3]> 2018-02-09      
#> # ... with 26 more rows, and 35 more variables: album_release_year <dbl>,
#> #   album_release_date_precision <chr>, danceability <dbl>, energy <dbl>,
#> #   key <int>, loudness <dbl>, mode <int>, speechiness <dbl>,
#> #   acousticness <dbl>, instrumentalness <dbl>, liveness <dbl>, valence <dbl>,
#> #   tempo <dbl>, track_id <chr>, analysis_url <chr>, time_signature <int>,
#> #   artists <list>, available_markets <list>, disc_number <int>,
#> #   duration_ms <int>, explicit <lgl>, track_href <chr>, is_local <lgl>,
#> #   track_name <chr>, track_preview_url <chr>, track_number <int>, type <chr>,
#> #   track_uri <chr>, external_urls.spotify <chr>, album_name <chr>,
#> #   key_name <chr>, mode_name <chr>, key_mode <chr>, track_n <int>,
#> #   lyrics <list>

# }