Get Spotify catalog information about an album’s tracks. Optional parameters can be used to limit the number of tracks returned.

Get Spotify catalog information about an album’s tracks. Optional parameters can be used to limit the number of tracks returned.

get_album_tracks(id, limit = 20, offset = 0, market = NULL,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE)

Arguments

id
Required. The Spotify ID for the album.

limit
Optional.
Maximum number of results to return.
Default: 20
Minimum: 1
Maximum: 50
Note: The limit is applied within each type, not on the total response.
For example, if the limit value is 3 and the type is c("artist", "album"), the response contains 3 artists and 3 albums.

offset
Optional.
The index of the first album to return.
Default: 0 (the first album).
Maximum offset (including limit): 10,000.
Use with limit to get the next page of albums.

market
Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token". Provide this parameter if you want to apply Track Relinking

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing album data. See the official API documentation for more information.