Get Audio Features For Artists' Discography

Get audio feature information for all or part of an artists' discography.

get_artist_audio_features(
  artist = NULL,
  include_groups = "album",
  return_closest_artist = TRUE,
  dedupe_albums = TRUE,
  market = NULL,
  authorization = get_spotify_access_token()
)

Arguments

artist

Required. String of either an artist name or an artist Spotify ID. If an artist name is provided, search_spotify() will be used to find a Spotify ID matching the name provided.

include_groups

Optional. A character vector of keywords that will be used to filter the response. Defaults to "album". Valid values are:
"album"
"single"
"appears_on"
"compilation"
For example: include_groups = c("album", "single")

return_closest_artist

Optional. Boolean.

dedupe_albums

Optional. Logical, boolean parameter, defaults to TRUE.

market

Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token".
Supply this parameter to limit the response to one particular geographical market. For example, for albums available in Sweden: market = "SE".
If not given, results will be returned for all markets and you are likely to get duplicate results per album, one for each market in which the album is available!

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

Returns a data frame of results containing track audio features data. See the Spotify Web API documentation for more information.

See also