Get Spotify catalog information about an artist’s top tracks by country.

Get Spotify catalog information about an artist’s top tracks by country.

get_artist_top_tracks(
  id,
  market = "US",
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

id
The Spotify ID for the artist.

market
Required.
An ISO 3166-1 alpha-2 country code or the string "from_token". Defaults to "US".

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing artist data. See https://developer.spotify.com/documentation/web-api/reference/artists/get-several-artists/ for more information.

See also