Get a list of Spotify playlists tagged with a particular category

Get a list of Spotify playlists tagged with a particular category

get_category_playlists(
  category_id = "party",
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

category_id
Required. The Spotify ID for the category.

country
Optional. A country: an ISO 3166-1 alpha-2 country code.

limit
Optional. The maximum number of items to return. Defaults to 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Defaults to 0, the first object. Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing category playlists. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
get_category_playlists('party', country = 'BR')

#>    collaborative
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11         FALSE
#> 12         FALSE
#> 13         FALSE
#> 14         FALSE
#> 15         FALSE
#> 16         FALSE
#> 17         FALSE
#> 18         FALSE
#> 19         FALSE
#> 20         FALSE
#>                                                                                                                                                                                 description
#> 1                                                                                                                        Não espere até sexta-feira à noite! Aperte o play e partiu sextar.
#> 2                                                                                                         Tem Sertanejo! Tem Funk! Tem Pagode! A melhor trilha pra embalar o seu churrasco.
#> 3                                                                                                       Conforto para beber gin tônica, mojito ou caipirinha sem precisar levantar do sofá.
#> 4                                                                                                                        O batidão dos funks mais bombados pra agitar a pista do seu baile!
#> 5  O funk tá diferente, meu sócio. Vem pro som do bololô. Foto: Mc Davi, MC Marks, MC Ryan SP, Mc Kevin, Mc Don Juan, MC Hariel, MC G15, Mc Phe Cachorrera, Gaab, MC Menor da VG, MC Kapela
#> 6                                                                                            Queeeima paredão! Pra forrozeiro sofrer dançando e dançar sofrendo, bora. Foto: Mari Fernandez
#> 7                                                                                                                Entre na onda com hits tropicais e remixes para curtir a praia com amigos.
#> 8                                                                                                                      Pôr do sol com pé na areia e aquela resenha com violão e muita rima.
#> 9                                                                                                      Para curtir aquela noite sertaneja puro modão. Foto: Thiago Brava, Israel & Rodolffo
#> 10                                                                                                                      Grandes hinos da música brasileira para animar sua pista de dança! 
#> 11                                                                                                                             Calor, amigos e muita música brasileira pra animar a galera!
#> 12                                                                                                                Grandes sucessos do samba rock pra arrastar o sofá e quebrar o bracinho. 
#> 13                                                                                                                                                    The biggest party hits of the 1980s. 
#> 14                                                                                                                                    <a href="spotify:genre:edm_dance">Move</a> your feet!
#> 15                                                                                           Uma playlist com músicas para você que quer encontrar seu lugar nesse mundo, feita por Ananda.
#> 16                                                                                                                                                        The ultimate rock party playlist!
#> 17                                                                                                      Já pode beber? A melhor trilha para embalar o seu momento pós-trabalho de cada dia!
#> 18                                                                                                                                                        It's all about the girls tonight.
#> 19                                                                                                                             The hip-hop playlist that's a whole mood. Cover: Flo Milli\n
#> 20                                                                                                                                    The biggest party hits of the 2000s. Cover: Shakira. 
#>                                                           href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWX9JsANXY37r
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DX6xnkAwJX7tn
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DX8vpLK1FoEw3
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DWZ72qOlbizxi
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DX0Uv9tZ47pWo
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DX6GwdWRQMQpq
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DX7e8TjkFNKWH
#>                        id
#> 1  37i9dQZF1DX8mBRYewE6or
#> 2  37i9dQZF1DXb1oSegSL8id
#> 3  37i9dQZF1DWX9JsANXY37r
#> 4  37i9dQZF1DWWmaszSfZpom
#> 5  37i9dQZF1DX4MdXmAY6EDq
#> 6  37i9dQZF1DXdXbKzsf80lE
#> 7  37i9dQZF1DX41UmHm2vtav
#> 8  37i9dQZF1DWXvzyniz2fv4
#> 9  37i9dQZF1DX5GJ0BQvLMjM
#> 10 37i9dQZF1DX1oTjVZnUYT3
#> 11 37i9dQZF1DXdWpx7KRCSbJ
#> 12 37i9dQZF1DX94Xc0f4tUlY
#> 13 37i9dQZF1DX6xnkAwJX7tn
#> 14 37i9dQZF1DXaXB8fQg7xif
#> 15 37i9dQZF1DX8vpLK1FoEw3
#> 16 37i9dQZF1DX8FwnYE6PRvL
#> 17 37i9dQZF1DWZ72qOlbizxi
#> 18 37i9dQZF1DX0Uv9tZ47pWo
#> 19 37i9dQZF1DX6GwdWRQMQpq
#> 20 37i9dQZF1DX7e8TjkFNKWH
#>                                                                      images
#> 1  NA, https://i.scdn.co/image/ab67706f0000000346878f5872e6fdea9686bd31, NA
#> 2  NA, https://i.scdn.co/image/ab67706f000000032f3b2073765b80193eb26ef2, NA
#> 3  NA, https://i.scdn.co/image/ab67706f00000003fbabce11ab019931ca40c588, NA
#> 4  NA, https://i.scdn.co/image/ab67706f0000000330b91682f0604c3a3d071904, NA
#> 5  NA, https://i.scdn.co/image/ab67706f00000003e1130cb04e6615d18143d012, NA
#> 6  NA, https://i.scdn.co/image/ab67706f000000039ddb6e3c827a45b3c47ffe8d, NA
#> 7  NA, https://i.scdn.co/image/ab67706f00000003bcf7685d4e9d6d0ccb44bfb2, NA
#> 8  NA, https://i.scdn.co/image/ab67706f00000003ee42a542a4c259708eff50c1, NA
#> 9  NA, https://i.scdn.co/image/ab67706f0000000315412c27faaa68ca08105f26, NA
#> 10 NA, https://i.scdn.co/image/ab67706f00000003c470446f590f6a8b8a55eaf8, NA
#> 11 NA, https://i.scdn.co/image/ab67706f000000033b8dc8bf62a876d30695ec1d, NA
#> 12 NA, https://i.scdn.co/image/ab67706f000000036d84d98d614876aa51e67fbe, NA
#> 13 NA, https://i.scdn.co/image/ab67706f00000003b71bec7748df963faa09f624, NA
#> 14 NA, https://i.scdn.co/image/ab67706f00000003470dd505fcf08e4693db9b24, NA
#> 15 NA, https://i.scdn.co/image/ab67706f00000003058d475cbc127c2f0c712490, NA
#> 16 NA, https://i.scdn.co/image/ab67706f00000003feb66cd1f6632ecfa89b3e8c, NA
#> 17 NA, https://i.scdn.co/image/ab67706f00000003f33dccf0dbb3500f622798db, NA
#> 18 NA, https://i.scdn.co/image/ab67706f0000000324bf380143d949f9bbccc187, NA
#> 19 NA, https://i.scdn.co/image/ab67706f000000033e0cf0fc5e314fa6eaa9e00d, NA
#> 20 NA, https://i.scdn.co/image/ab67706f0000000329a3de2418e65574d461618a, NA
#>                 name primary_color public
#> 1             Sextou            NA     NA
#> 2    Sol e Churrasco            NA     NA
#> 3     Drinks no Sofá            NA     NA
#> 4      Segue o Baile            NA     NA
#> 5            Revoada            NA     NA
#> 6      Bora arrochar            NA     NA
#> 7     Festa na Praia            NA     NA
#> 8   Resenha na Praia            NA     NA
#> 9    Farra Sertaneja            NA     NA
#> 10  Festa Brasileira            NA     NA
#> 11  Festa na Piscina            NA     NA
#> 12        Sambalanço            NA     NA
#> 13         80s Party            NA     NA
#> 14       Dance Party            NA     NA
#> 15 Lista VIP: Ananda            NA     NA
#> 16        Rock Party            NA     NA
#> 17        Happy Hour            NA     NA
#> 18      Girls' Night            NA     NA
#> 19    Feelin' Myself            NA     NA
#> 20         00s Party            NA     NA
#>                                                             snapshot_id
#> 1  MTYzNTg3MTUzMywwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 2  MTYzNTg3MTUyNSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 3  MTYzNTg3MTUzOCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 4  MTYzNTg3MTU0NCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 5  MTYzNTQ3NjUyMCwwMDAwMDAzYjAwMDAwMTdjYzlmZjNjYmMwMDAwMDE3YzgwN2RhYTU1
#> 6  MTYzNTg3MTU1NCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 7  MTYzNTg3MTUxNywwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 8  MTYzNDI2Njk4MCwwMDAwMDA2ZDAwMDAwMTdjODFlNzFlZmIwMDAwMDE3MmYxYzU1MTVh
#> 9  MTYzNTQ3NjQwMCwwMDAwMDQwZDAwMDAwMTdjYzlmZDY4MmIwMDAwMDE3YzgwZWExY2Iz
#> 10 MTYzNTg3MTU0OCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 11 MTYzNTg3MTU2NCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 12 MTYzNTQ3NjQ2MCwwMDAwMDAxYjAwMDAwMTdjYzlmZTUyMzAwMDAwMDE3ODg1NTkxMzBi
#> 13 MTYzNTg3MTU2NCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 14 MTYzNTg3MTUyNSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 15 MTYzNTQ3NjQwMCwwMDAwMDQ0YTAwMDAwMTdjYzlmZDY4MTgwMDAwMDE3Y2M4NDU0YmE2
#> 16 MTYxOTEyOTEwOCwwMDAwMDAzOTAwMDAwMTc4ZmI5ZDg3MDQwMDAwMDE2ZDE1NTBjOWZh
#> 17 MTYzNTg3MTUzOSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 18 MTYzNTg3MTU1OSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 19 MTYzNTU0Mjg2NywwMDAwMDQ1ZTAwMDAwMTdjY2RmMzlmMjcwMDAwMDE3Y2M4MjI1ZmYy
#> 20 MTYzNTg3MTUyNywwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#>        type                                     uri
#> 1  playlist spotify:playlist:37i9dQZF1DX8mBRYewE6or
#> 2  playlist spotify:playlist:37i9dQZF1DXb1oSegSL8id
#> 3  playlist spotify:playlist:37i9dQZF1DWX9JsANXY37r
#> 4  playlist spotify:playlist:37i9dQZF1DWWmaszSfZpom
#> 5  playlist spotify:playlist:37i9dQZF1DX4MdXmAY6EDq
#> 6  playlist spotify:playlist:37i9dQZF1DXdXbKzsf80lE
#> 7  playlist spotify:playlist:37i9dQZF1DX41UmHm2vtav
#> 8  playlist spotify:playlist:37i9dQZF1DWXvzyniz2fv4
#> 9  playlist spotify:playlist:37i9dQZF1DX5GJ0BQvLMjM
#> 10 playlist spotify:playlist:37i9dQZF1DX1oTjVZnUYT3
#> 11 playlist spotify:playlist:37i9dQZF1DXdWpx7KRCSbJ
#> 12 playlist spotify:playlist:37i9dQZF1DX94Xc0f4tUlY
#> 13 playlist spotify:playlist:37i9dQZF1DX6xnkAwJX7tn
#> 14 playlist spotify:playlist:37i9dQZF1DXaXB8fQg7xif
#> 15 playlist spotify:playlist:37i9dQZF1DX8vpLK1FoEw3
#> 16 playlist spotify:playlist:37i9dQZF1DX8FwnYE6PRvL
#> 17 playlist spotify:playlist:37i9dQZF1DWZ72qOlbizxi
#> 18 playlist spotify:playlist:37i9dQZF1DX0Uv9tZ47pWo
#> 19 playlist spotify:playlist:37i9dQZF1DX6GwdWRQMQpq
#> 20 playlist spotify:playlist:37i9dQZF1DX7e8TjkFNKWH
#>                                       external_urls.spotify owner.display_name
#> 1  https://open.spotify.com/playlist/37i9dQZF1DX8mBRYewE6or            Spotify
#> 2  https://open.spotify.com/playlist/37i9dQZF1DXb1oSegSL8id            Spotify
#> 3  https://open.spotify.com/playlist/37i9dQZF1DWX9JsANXY37r            Spotify
#> 4  https://open.spotify.com/playlist/37i9dQZF1DWWmaszSfZpom            Spotify
#> 5  https://open.spotify.com/playlist/37i9dQZF1DX4MdXmAY6EDq            Spotify
#> 6  https://open.spotify.com/playlist/37i9dQZF1DXdXbKzsf80lE            Spotify
#> 7  https://open.spotify.com/playlist/37i9dQZF1DX41UmHm2vtav            Spotify
#> 8  https://open.spotify.com/playlist/37i9dQZF1DWXvzyniz2fv4            Spotify
#> 9  https://open.spotify.com/playlist/37i9dQZF1DX5GJ0BQvLMjM            Spotify
#> 10 https://open.spotify.com/playlist/37i9dQZF1DX1oTjVZnUYT3            Spotify
#> 11 https://open.spotify.com/playlist/37i9dQZF1DXdWpx7KRCSbJ            Spotify
#> 12 https://open.spotify.com/playlist/37i9dQZF1DX94Xc0f4tUlY            Spotify
#> 13 https://open.spotify.com/playlist/37i9dQZF1DX6xnkAwJX7tn            Spotify
#> 14 https://open.spotify.com/playlist/37i9dQZF1DXaXB8fQg7xif            Spotify
#> 15 https://open.spotify.com/playlist/37i9dQZF1DX8vpLK1FoEw3            Spotify
#> 16 https://open.spotify.com/playlist/37i9dQZF1DX8FwnYE6PRvL            Spotify
#> 17 https://open.spotify.com/playlist/37i9dQZF1DWZ72qOlbizxi            Spotify
#> 18 https://open.spotify.com/playlist/37i9dQZF1DX0Uv9tZ47pWo            Spotify
#> 19 https://open.spotify.com/playlist/37i9dQZF1DX6GwdWRQMQpq            Spotify
#> 20 https://open.spotify.com/playlist/37i9dQZF1DX7e8TjkFNKWH            Spotify
#>                                  owner.href owner.id owner.type
#> 1  https://api.spotify.com/v1/users/spotify  spotify       user
#> 2  https://api.spotify.com/v1/users/spotify  spotify       user
#> 3  https://api.spotify.com/v1/users/spotify  spotify       user
#> 4  https://api.spotify.com/v1/users/spotify  spotify       user
#> 5  https://api.spotify.com/v1/users/spotify  spotify       user
#> 6  https://api.spotify.com/v1/users/spotify  spotify       user
#> 7  https://api.spotify.com/v1/users/spotify  spotify       user
#> 8  https://api.spotify.com/v1/users/spotify  spotify       user
#> 9  https://api.spotify.com/v1/users/spotify  spotify       user
#> 10 https://api.spotify.com/v1/users/spotify  spotify       user
#> 11 https://api.spotify.com/v1/users/spotify  spotify       user
#> 12 https://api.spotify.com/v1/users/spotify  spotify       user
#> 13 https://api.spotify.com/v1/users/spotify  spotify       user
#> 14 https://api.spotify.com/v1/users/spotify  spotify       user
#> 15 https://api.spotify.com/v1/users/spotify  spotify       user
#> 16 https://api.spotify.com/v1/users/spotify  spotify       user
#> 17 https://api.spotify.com/v1/users/spotify  spotify       user
#> 18 https://api.spotify.com/v1/users/spotify  spotify       user
#> 19 https://api.spotify.com/v1/users/spotify  spotify       user
#> 20 https://api.spotify.com/v1/users/spotify  spotify       user
#>               owner.uri           owner.external_urls.spotify
#> 1  spotify:user:spotify https://open.spotify.com/user/spotify
#> 2  spotify:user:spotify https://open.spotify.com/user/spotify
#> 3  spotify:user:spotify https://open.spotify.com/user/spotify
#> 4  spotify:user:spotify https://open.spotify.com/user/spotify
#> 5  spotify:user:spotify https://open.spotify.com/user/spotify
#> 6  spotify:user:spotify https://open.spotify.com/user/spotify
#> 7  spotify:user:spotify https://open.spotify.com/user/spotify
#> 8  spotify:user:spotify https://open.spotify.com/user/spotify
#> 9  spotify:user:spotify https://open.spotify.com/user/spotify
#> 10 spotify:user:spotify https://open.spotify.com/user/spotify
#> 11 spotify:user:spotify https://open.spotify.com/user/spotify
#> 12 spotify:user:spotify https://open.spotify.com/user/spotify
#> 13 spotify:user:spotify https://open.spotify.com/user/spotify
#> 14 spotify:user:spotify https://open.spotify.com/user/spotify
#> 15 spotify:user:spotify https://open.spotify.com/user/spotify
#> 16 spotify:user:spotify https://open.spotify.com/user/spotify
#> 17 spotify:user:spotify https://open.spotify.com/user/spotify
#> 18 spotify:user:spotify https://open.spotify.com/user/spotify
#> 19 spotify:user:spotify https://open.spotify.com/user/spotify
#> 20 spotify:user:spotify https://open.spotify.com/user/spotify
#>                                                           tracks.href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or/tracks
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id/tracks
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWX9JsANXY37r/tracks
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom/tracks
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq/tracks
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE/tracks
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav/tracks
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4/tracks
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM/tracks
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3/tracks
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ/tracks
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY/tracks
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DX6xnkAwJX7tn/tracks
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif/tracks
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DX8vpLK1FoEw3/tracks
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL/tracks
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DWZ72qOlbizxi/tracks
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DX0Uv9tZ47pWo/tracks
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DX6GwdWRQMQpq/tracks
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DX7e8TjkFNKWH/tracks
#>    tracks.total
#> 1            75
#> 2           100
#> 3            60
#> 4            70
#> 5            50
#> 6           100
#> 7            70
#> 8            65
#> 9           100
#> 10           57
#> 11           65
#> 12           50
#> 13          150
#> 14          125
#> 15           50
#> 16          145
#> 17           75
#> 18           80
#> 19           50
#> 20          150

# }