Get a list of Spotify featured playlists (shown, for example, on a Spotify player’s ‘Browse’ tab).

Get a list of Spotify featured playlists (shown, for example, on a Spotify player’s ‘Browse’ tab).

get_featured_playlists(
  locale = NULL,
  country = NULL,
  timestamp = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

locale
Optional. The desired language, consisting of an ISO 639-1 language code and an ISO 3166-1 alpha-2 country code, joined by an underscore. For example: es_MX, meaning "Spanish (Mexico)". Provide this parameter if you want the category strings returned in a particular language. Note that, if locale is not supplied, or if the specified language is not available, the category strings returned will be in the Spotify default language (American English). The locale parameter, combined with the country parameter, may give odd results if not carefully matched. For example country=SE&locale=de_DE will return a list of categories relevant to Sweden but as German language strings.

country
Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

timestamp
Optional. A timestamp in ISO 8601 format: yyyy-MM-ddTHH:mm:ss. Use this parameter to specify the user’s local time to get results tailored for that specific date and time in the day. If not provided, the response defaults to the current UTC time. Example: “2014-10-23T09:00:00” for a user whose local time is 9AM. If there were no featured playlists (or there is no data) at the specified time, the response will revert to the current UTC time.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing featured playlists. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
## Get new Swedish music
get_featured_playlists(country = 'SE')

#>    collaborative
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11         FALSE
#> 12         FALSE
#>                                                                           description
#> 1                                  Äntligen fredag! Ny musik! Happy New Music Friday!
#> 2                                       It's Hits - bästa och största hitsen just nu.
#> 3                          Ricky Rich och Blizzy återförenas i Studio 100. Lyssna nu!
#> 4                                            De senaste hitsen du behöver ha koll på.
#> 5                                                 Akustiskt soundtrack till sommaren.
#> 6                                         Det bästa från svensk & internationell pop.
#> 7                   Hör guldkornen från indievärlden just nu. Cover: Linn Koch-Emmery
#> 8                                        Här hör du de bästa svenska låtarna just nu.
#> 9  Våra bästa lokala stjärnor hela vägen från Kalix till Höör. Cover: Lars Winnerbäck
#> 10                                                             Låtar som är On Point.
#> 11                             Aktuell musik med etablerade och intressanta artister.
#> 12                                                    Svenska visor från förr och nu.
#>                                                           href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DXcecv7ESbOPu
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DWTMYgB8TqtmR
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWXfgo3OOonqa
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DXc4CrXgl8cum
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DWWGRfB2jBjnV
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DX7FV7CCq9byu
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DWVwRuJjs9A3S
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DX0ekBHz1NmKS
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DWTmWCffxdqas
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX9jjafmzFEux
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DX9kaiEDpvmoR
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DXbknicFbzYx6
#>                        id
#> 1  37i9dQZF1DXcecv7ESbOPu
#> 2  37i9dQZF1DWTMYgB8TqtmR
#> 3  37i9dQZF1DWXfgo3OOonqa
#> 4  37i9dQZF1DXc4CrXgl8cum
#> 5  37i9dQZF1DWWGRfB2jBjnV
#> 6  37i9dQZF1DX7FV7CCq9byu
#> 7  37i9dQZF1DWVwRuJjs9A3S
#> 8  37i9dQZF1DX0ekBHz1NmKS
#> 9  37i9dQZF1DWTmWCffxdqas
#> 10 37i9dQZF1DX9jjafmzFEux
#> 11 37i9dQZF1DX9kaiEDpvmoR
#> 12 37i9dQZF1DXbknicFbzYx6
#>                                                                      images
#> 1  NA, https://i.scdn.co/image/ab67706f0000000357d3303a4e079b77c0eac4f0, NA
#> 2  NA, https://i.scdn.co/image/ab67706f00000003919243ea2d9c07a9469d5a91, NA
#> 3  NA, https://i.scdn.co/image/ab67706f00000003bb5fd9b8c70b6cf891a32b69, NA
#> 4  NA, https://i.scdn.co/image/ab67706f00000003810ab08ceb5bd9b0f573fa2d, NA
#> 5  NA, https://i.scdn.co/image/ab67706f00000003fe8b30220314a252351f1281, NA
#> 6  NA, https://i.scdn.co/image/ab67706f000000038263753d219cd7a51bdc49a5, NA
#> 7  NA, https://i.scdn.co/image/ab67706f00000003ad9e874972627fb2ecaf9fce, NA
#> 8  NA, https://i.scdn.co/image/ab67706f0000000340c523fe54ca813cacc2edc8, NA
#> 9  NA, https://i.scdn.co/image/ab67706f00000003b3706762de7098797567657e, NA
#> 10 NA, https://i.scdn.co/image/ab67706f000000033e2583d069841fae66ae47d6, NA
#> 11 NA, https://i.scdn.co/image/ab67706f000000035678172609436a286a6e8a54, NA
#> 12 NA, https://i.scdn.co/image/ab67706f000000033e97a2abad7762e55e444210, NA
#>                       name primary_color public
#> 1  New Music Friday Sweden            NA     NA
#> 2         It's Hits Sweden            NA     NA
#> 3                      100            NA     NA
#> 4   Get Your Hits Together            NA     NA
#> 5          Akustisk Sommar            NA     NA
#> 6                   Ny Pop            NA     NA
#> 7         Indie Highlights            NA     NA
#> 8           Made in Sweden            NA     NA
#> 9     Från norr till söder            NA     NA
#> 10                on point            NA     NA
#> 11                Aktuellt            NA     NA
#> 12          Visa och vemod            NA     NA
#>                                                             snapshot_id
#> 1  MTYyMzM2MjQwMCwwMDAwMDMwNTAwMDAwMTc5ZjdmMDZmOTgwMDAwMDE3OWY1NTA5ZmEz
#> 2  MTYyMzQwMTI4NiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 3  MTYyMzQwMTI5NCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 4  MTYyMzQwMTMyNiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 5  MTYyMzQwMTI4NiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 6  MTYyMzM2MjQwMCwwMDAwMDQwMDAwMDAwMTc5ZjdmMDZmYWQwMDAwMDE3OWY3OThjMjhk
#> 7  MTYyMzQwMTI4NiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 8  MTYyMzQwMTI4NiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 9  MTYyMzQwMTMwMSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 10 MTYyMzQwMTI4OCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 11 MTYyMzQwMTI4NSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 12 MTYyMzQwMTI4NSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#>        type                                     uri
#> 1  playlist spotify:playlist:37i9dQZF1DXcecv7ESbOPu
#> 2  playlist spotify:playlist:37i9dQZF1DWTMYgB8TqtmR
#> 3  playlist spotify:playlist:37i9dQZF1DWXfgo3OOonqa
#> 4  playlist spotify:playlist:37i9dQZF1DXc4CrXgl8cum
#> 5  playlist spotify:playlist:37i9dQZF1DWWGRfB2jBjnV
#> 6  playlist spotify:playlist:37i9dQZF1DX7FV7CCq9byu
#> 7  playlist spotify:playlist:37i9dQZF1DWVwRuJjs9A3S
#> 8  playlist spotify:playlist:37i9dQZF1DX0ekBHz1NmKS
#> 9  playlist spotify:playlist:37i9dQZF1DWTmWCffxdqas
#> 10 playlist spotify:playlist:37i9dQZF1DX9jjafmzFEux
#> 11 playlist spotify:playlist:37i9dQZF1DX9kaiEDpvmoR
#> 12 playlist spotify:playlist:37i9dQZF1DXbknicFbzYx6
#>                                       external_urls.spotify owner.display_name
#> 1  https://open.spotify.com/playlist/37i9dQZF1DXcecv7ESbOPu            Spotify
#> 2  https://open.spotify.com/playlist/37i9dQZF1DWTMYgB8TqtmR            Spotify
#> 3  https://open.spotify.com/playlist/37i9dQZF1DWXfgo3OOonqa            Spotify
#> 4  https://open.spotify.com/playlist/37i9dQZF1DXc4CrXgl8cum            Spotify
#> 5  https://open.spotify.com/playlist/37i9dQZF1DWWGRfB2jBjnV            Spotify
#> 6  https://open.spotify.com/playlist/37i9dQZF1DX7FV7CCq9byu            Spotify
#> 7  https://open.spotify.com/playlist/37i9dQZF1DWVwRuJjs9A3S            Spotify
#> 8  https://open.spotify.com/playlist/37i9dQZF1DX0ekBHz1NmKS            Spotify
#> 9  https://open.spotify.com/playlist/37i9dQZF1DWTmWCffxdqas            Spotify
#> 10 https://open.spotify.com/playlist/37i9dQZF1DX9jjafmzFEux            Spotify
#> 11 https://open.spotify.com/playlist/37i9dQZF1DX9kaiEDpvmoR            Spotify
#> 12 https://open.spotify.com/playlist/37i9dQZF1DXbknicFbzYx6            Spotify
#>                                  owner.href owner.id owner.type
#> 1  https://api.spotify.com/v1/users/spotify  spotify       user
#> 2  https://api.spotify.com/v1/users/spotify  spotify       user
#> 3  https://api.spotify.com/v1/users/spotify  spotify       user
#> 4  https://api.spotify.com/v1/users/spotify  spotify       user
#> 5  https://api.spotify.com/v1/users/spotify  spotify       user
#> 6  https://api.spotify.com/v1/users/spotify  spotify       user
#> 7  https://api.spotify.com/v1/users/spotify  spotify       user
#> 8  https://api.spotify.com/v1/users/spotify  spotify       user
#> 9  https://api.spotify.com/v1/users/spotify  spotify       user
#> 10 https://api.spotify.com/v1/users/spotify  spotify       user
#> 11 https://api.spotify.com/v1/users/spotify  spotify       user
#> 12 https://api.spotify.com/v1/users/spotify  spotify       user
#>               owner.uri           owner.external_urls.spotify
#> 1  spotify:user:spotify https://open.spotify.com/user/spotify
#> 2  spotify:user:spotify https://open.spotify.com/user/spotify
#> 3  spotify:user:spotify https://open.spotify.com/user/spotify
#> 4  spotify:user:spotify https://open.spotify.com/user/spotify
#> 5  spotify:user:spotify https://open.spotify.com/user/spotify
#> 6  spotify:user:spotify https://open.spotify.com/user/spotify
#> 7  spotify:user:spotify https://open.spotify.com/user/spotify
#> 8  spotify:user:spotify https://open.spotify.com/user/spotify
#> 9  spotify:user:spotify https://open.spotify.com/user/spotify
#> 10 spotify:user:spotify https://open.spotify.com/user/spotify
#> 11 spotify:user:spotify https://open.spotify.com/user/spotify
#> 12 spotify:user:spotify https://open.spotify.com/user/spotify
#>                                                           tracks.href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DXcecv7ESbOPu/tracks
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DWTMYgB8TqtmR/tracks
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWXfgo3OOonqa/tracks
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DXc4CrXgl8cum/tracks
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DWWGRfB2jBjnV/tracks
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DX7FV7CCq9byu/tracks
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DWVwRuJjs9A3S/tracks
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DX0ekBHz1NmKS/tracks
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DWTmWCffxdqas/tracks
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX9jjafmzFEux/tracks
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DX9kaiEDpvmoR/tracks
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DXbknicFbzYx6/tracks
#>    tracks.total
#> 1           105
#> 2            50
#> 3            85
#> 4            80
#> 5            75
#> 6            70
#> 7            80
#> 8           100
#> 9            60
#> 10           75
#> 11           70
#> 12           55

# }