Search for Artists by Genre

Get Spotify Catalog information about artists belonging to a given genre.

get_genre_artists(
  genre = character(),
  market = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)

Arguments

genre
Required.
String of genre name to search for
For example: genre = "wonky".

market
Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token".
If a country code is specified, only artists with content that is playable in that market is returned.
Note:
- If market is set to "from_token", and a valid access token is specified in the request header, only content playable in the country associated with the user account is returned.
- Users can view the country that is associated with their account in the account settings. A user must grant access to the user-read-private scope prior to when the access token is issued.

limit
Optional.
Maximum number of results to return.
Default: 20
Minimum: 1
Maximum: 50

offset
Optional.
The index of the first result to return.
Default: 0 (the first result).
Maximum offset (including limit): 10,000.
Use with limit to get the next page of search results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

A data frame of the artists belonging to the genre with data and metadata about the artists in a tibble.

See also

Examples

# \donttest{
get_genre_artists('wonky')

#> # A tibble: 20 x 12
#>    genres     href   id     images name  popularity type  uri   external_urls.s~
#>    <list>     <chr>  <chr>  <list> <chr>      <int> <chr> <chr> <chr>           
#>  1 <chr [6]>  https~ 17Zu0~ <df [~ Knxw~         62 arti~ spot~ https://open.sp~
#>  2 <chr [12]> https~ 29XOe~ <df [~ Flyi~         63 arti~ spot~ https://open.sp~
#>  3 <chr [7]>  https~ 3VwKS~ <df [~ TOKi~         59 arti~ spot~ https://open.sp~
#>  4 <chr [2]>  https~ 1PmVy~ <df [~ Lorn          57 arti~ spot~ https://open.sp~
#>  5 <chr [8]>  https~ 3NUtp~ <df [~ Moun~         57 arti~ spot~ https://open.sp~
#>  6 <chr [8]>  https~ 6y80I~ <df [~ Shlo~         52 arti~ spot~ https://open.sp~
#>  7 <chr [6]>  https~ 0IVap~ <df [~ Nosa~         51 arti~ spot~ https://open.sp~
#>  8 <chr [9]>  https~ 4GcpB~ <df [~ Mnds~         52 arti~ spot~ https://open.sp~
#>  9 <chr [3]>  https~ 1AEcp~ <df [~ DaM-~         51 arti~ spot~ https://open.sp~
#> 10 <chr [5]>  https~ 2AR42~ <df [~ Floa~         54 arti~ spot~ https://open.sp~
#> 11 <chr [10]> https~ 5wZOr~ <df [~ Lone          53 arti~ spot~ https://open.sp~
#> 12 <chr [3]>  https~ 2DcAd~ <df [~ Dirt~         47 arti~ spot~ https://open.sp~
#> 13 <chr [6]>  https~ 0uCCB~ <df [~ Buri~         54 arti~ spot~ https://open.sp~
#> 14 <chr [4]>  https~ 6e7BQ~ <df [~ Karr~         50 arti~ spot~ https://open.sp~
#> 15 <chr [4]>  https~ 2L2un~ <df [~ Teebs         48 arti~ spot~ https://open.sp~
#> 16 <chr [2]>  https~ 1MSxO~ <df [~ Phot~         47 arti~ spot~ https://open.sp~
#> 17 <chr [12]> https~ 06xa1~ <df [~ Mach~         51 arti~ spot~ https://open.sp~
#> 18 <chr [3]>  https~ 3i7IU~ <df [~ Jonw~         47 arti~ spot~ https://open.sp~
#> 19 <chr [1]>  https~ 1iqXh~ <df [~ xxyy~         45 arti~ spot~ https://open.sp~
#> 20 <chr [2]>  https~ 47V6n~ <df [~ Robo~         51 arti~ spot~ https://open.sp~
#> # ... with 3 more variables: followers.href <lgl>, followers.total <int>,
#> #   genre <chr>

# }