Search for artists by label

Get Spotify Catalog information about artists belonging to a given label.

get_label_artists(
  label = character(),
  market = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)

Arguments

label

Required.
String of label name to search for
For example: label = "brainfeeder".

market

Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token".
If a country code is specified, only artists with content that is playable in that market is returned.
Note:
- If market is set to "from_token", and a valid access token is specified in the request header, only content playable in the country associated with the user account is returned.
- Users can view the country that is associated with their account in the account settings. A user must grant access to the user-read-private scope prior to when the access token is issued.

limit

Optional.
Maximum number of results to return.
Default: 20
Minimum: 1
Maximum: 50

offset

Optional.
The index of the first result to return.
Default: 0 (the first result).
Maximum offset (including limit): 10,000.
Use with limit to get the next page of search results.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

A data frame with the label information of the artists.

Examples

# \donttest{
get_label_artists('brainfeeder')
#> $q
#> [1] "label:\"brainfeeder\""
#> 
#> $type
#> [1] "artist"
#> 
#> $market
#> NULL
#> 
#> $limit
#> [1] 20
#> 
#> $offset
#> [1] 0
#> 
#> $access_token
#> [1] "BQDeoPHIxUHsKiX6-8fSpXFwh9UJkfzwv7IdIiM0afjFqHBoQXIB6d6Nlf3kL_Mgj060SlJ69rhr_57iB8_ev3KZmbovkkkEwxe2xHwi2l35suiFLCY"
#> 
#> # A tibble: 20 × 12
#>    genres href    id    images name  popul…¹ type  uri   exter…² follo…³ follo…⁴
#>    <list> <chr>   <chr> <list> <chr>   <int> <chr> <chr> <chr>   <lgl>     <int>
#>  1 <chr>  https:… 4frX… <df>   Thun…      73 arti… spot… https:… NA      1019102
#>  2 <chr>  https:… 1ybI… <df>   Smino      71 arti… spot… https:… NA       728920
#>  3 <chr>  https:… 3Y3x… <df>   Kenn…      68 arti… spot… https:… NA       913060
#>  4 <chr>  https:… 43Jl… <df>   Hiat…      56 arti… spot… https:… NA       523106
#>  5 <chr>  https:… 29XO… <df>   Flyi…      57 arti… spot… https:… NA       710889
#>  6 <chr>  https:… 5cM1… <df>   Robe…      61 arti… spot… https:… NA       370949
#>  7 <chr>  https:… 6uIs… <df>   Loui…      51 arti… spot… https:… NA       108588
#>  8 <chr>  https:… 1PmV… <df>   Lorn       54 arti… spot… https:… NA       269578
#>  9 <chr>  https:… 6HQY… <df>   Kama…      55 arti… spot… https:… NA       409067
#> 10 <chr>  https:… 6EHS… <df>   Sala…      48 arti… spot… https:… NA        26096
#> 11 <chr>  https:… 65dG… <df>   BADB…      63 arti… spot… https:… NA       641262
#> 12 <chr>  https:… 2L2u… <df>   Teebs      50 arti… spot… https:… NA        95668
#> 13 <chr>  https:… 2Njf… <df>   Aziz…      51 arti… spot… https:… NA       146048
#> 14 <chr>  https:… 0NcP… <df>   Guap…      57 arti… spot… https:… NA       108889
#> 15 <chr>  https:… 46Ce… <df>   Lapa…      45 arti… spot… https:… NA       113068
#> 16 <chr>  https:… 6jlW… <df>   Emil…      49 arti… spot… https:… NA       179549
#> 17 <chr>  https:… 3VwK… <df>   TOKi…      55 arti… spot… https:… NA       289191
#> 18 <chr>  https:… 24hJ… <df>   Mich…      55 arti… spot… https:… NA       428526
#> 19 <chr>  https:… 1R84… <df>   Pand…      52 arti… spot… https:… NA       196310
#> 20 <chr>  https:… 1tP3… <df>   Arth…      47 arti… spot… https:… NA        42240
#> # … with 1 more variable: label <chr>, and abbreviated variable names
#> #   ¹​popularity, ²​external_urls.spotify, ³​followers.href, ⁴​followers.total
# }