Get new releases

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country
Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases.
See https://developer.spotify.com/documentation/web-api/reference/#category-browse for more information.

Examples

# \donttest{
## Get new Swedish music
get_new_releases(country = 'SE', limit = 5)

#>   album_type
#> 1     single
#> 2     single
#> 3     single
#> 4     single
#> 5     single
#>                                                                                                                                                                                                      artists
#> 1    https://api.spotify.com/v1/artists/2VWX6ZalDVTp1KMGN6JDSl, 2VWX6ZalDVTp1KMGN6JDSl, Siw Malmkvist, artist, spotify:artist:2VWX6ZalDVTp1KMGN6JDSl, https://open.spotify.com/artist/2VWX6ZalDVTp1KMGN6JDSl
#> 2     https://api.spotify.com/v1/artists/3fA5zBQWEa0uzYo43SvXaj, 3fA5zBQWEa0uzYo43SvXaj, Peter Jöback, artist, spotify:artist:3fA5zBQWEa0uzYo43SvXaj, https://open.spotify.com/artist/3fA5zBQWEa0uzYo43SvXaj
#> 3          https://api.spotify.com/v1/artists/59E9dYtwLTUAGuAppwFKAW, 59E9dYtwLTUAGuAppwFKAW, Cherrie, artist, spotify:artist:59E9dYtwLTUAGuAppwFKAW, https://open.spotify.com/artist/59E9dYtwLTUAGuAppwFKAW
#> 4 https://api.spotify.com/v1/artists/2DWNa2o3BbsIHeBq7Iu6E7, 2DWNa2o3BbsIHeBq7Iu6E7, Thomas Stenström, artist, spotify:artist:2DWNa2o3BbsIHeBq7Iu6E7, https://open.spotify.com/artist/2DWNa2o3BbsIHeBq7Iu6E7
#> 5     https://api.spotify.com/v1/artists/0tOdy7SqogILIjzGU2hSr5, 0tOdy7SqogILIjzGU2hSr5, Melissa Horn, artist, spotify:artist:0tOdy7SqogILIjzGU2hSr5, https://open.spotify.com/artist/0tOdy7SqogILIjzGU2hSr5
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 3 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 4 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 5 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                       href
#> 1 https://api.spotify.com/v1/albums/6xhAkcrAHgFTjMZs4bP8YY
#> 2 https://api.spotify.com/v1/albums/6CQpZRZnlKvzNgc3f6rt1c
#> 3 https://api.spotify.com/v1/albums/4zcd1yACKoc6R3fAi8NhCn
#> 4 https://api.spotify.com/v1/albums/5BUAq7X644KHxL5cGWIUVH
#> 5 https://api.spotify.com/v1/albums/7veo0FwLrcfOOPVverDM4t
#>                       id
#> 1 6xhAkcrAHgFTjMZs4bP8YY
#> 2 6CQpZRZnlKvzNgc3f6rt1c
#> 3 4zcd1yACKoc6R3fAi8NhCn
#> 4 5BUAq7X644KHxL5cGWIUVH
#> 5 7veo0FwLrcfOOPVverDM4t
#>                                                                                                                                                                                                                             images
#> 1 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273959047b65bbc5f8c56d2de6b, https://i.scdn.co/image/ab67616d00001e02959047b65bbc5f8c56d2de6b, https://i.scdn.co/image/ab67616d00004851959047b65bbc5f8c56d2de6b, 640, 300, 64
#> 2 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27380ce1d930a7da1b0be2b7c84, https://i.scdn.co/image/ab67616d00001e0280ce1d930a7da1b0be2b7c84, https://i.scdn.co/image/ab67616d0000485180ce1d930a7da1b0be2b7c84, 640, 300, 64
#> 3 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2738b0adddd70364891b348c494, https://i.scdn.co/image/ab67616d00001e028b0adddd70364891b348c494, https://i.scdn.co/image/ab67616d000048518b0adddd70364891b348c494, 640, 300, 64
#> 4 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273813d1fbc41aa24adbf08fd41, https://i.scdn.co/image/ab67616d00001e02813d1fbc41aa24adbf08fd41, https://i.scdn.co/image/ab67616d00004851813d1fbc41aa24adbf08fd41, 640, 300, 64
#> 5 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b6e2f9ac92b2165c88527ea0, https://i.scdn.co/image/ab67616d00001e02b6e2f9ac92b2165c88527ea0, https://i.scdn.co/image/ab67616d00004851b6e2f9ac92b2165c88527ea0, 640, 300, 64
#>                        name release_date release_date_precision total_tracks
#> 1 Smek mig mjukt i ansiktet   2021-10-30                    day            1
#> 2               Falla fritt   2021-10-30                    day            1
#> 3    Mamma är lik sin mamma   2021-10-30                    day            1
#> 4              Sweet Jackie   2021-10-30                    day            1
#> 5               Lämna honom   2021-10-30                    day            1
#>    type                                  uri
#> 1 album spotify:album:6xhAkcrAHgFTjMZs4bP8YY
#> 2 album spotify:album:6CQpZRZnlKvzNgc3f6rt1c
#> 3 album spotify:album:4zcd1yACKoc6R3fAi8NhCn
#> 4 album spotify:album:5BUAq7X644KHxL5cGWIUVH
#> 5 album spotify:album:7veo0FwLrcfOOPVverDM4t
#>                                   external_urls.spotify
#> 1 https://open.spotify.com/album/6xhAkcrAHgFTjMZs4bP8YY
#> 2 https://open.spotify.com/album/6CQpZRZnlKvzNgc3f6rt1c
#> 3 https://open.spotify.com/album/4zcd1yACKoc6R3fAi8NhCn
#> 4 https://open.spotify.com/album/5BUAq7X644KHxL5cGWIUVH
#> 5 https://open.spotify.com/album/7veo0FwLrcfOOPVverDM4t

# }