Get features and popularity for all of a given set of playlists on Spotify

This function returns the popularity and audio features for every song for a given set of playlists on Spotify

get_playlist_audio_features(username, playlist_uris,
  authorization = get_spotify_access_token())

Arguments

username
String of Spotify username. Can be found on the Spotify app.

playlist_uris
Character vector of Spotify playlist uris. Can be found within the Spotify App

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details. Defaults to spotifyr::get_spotify_access_token()

Examples

# NOT RUN {
playlist_username <- 'spotify'
playlist_uris <- c('37i9dQZF1E9T1oFsQFg98K', '37i9dQZF1CyQNOI21QVf3p')
playlist_audio_features <- get_playlist_audio_features(playlist_username, playlist_uris)
# }