Get playlists

This function returns a dataframe of playlists for a given Spotify username and a character vector of playlist uris

get_playlists(username, playlist_uris,
  access_token = get_spotify_access_token())

Arguments

username
String of Spotify username. Can be found within the Spotify App

playlist_uris
Character vector of Spotify playlist uris associated with the given username. Can be found within the Spotify App

access_token
Spotify Web API token. Defaults to spotifyr::get_spotify_access_token().

Examples

# NOT RUN {
playlist_username <- 'spotify'
playlist_uris <- c('37i9dQZF1E9T1oFsQFg98K', '37i9dQZF1CyQNOI21QVf3p')
my_playlist_audio_features <- get_playlist_audio_features('spotify', playlist_uris)
# }