Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit
Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market
Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists
A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres
A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks
A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms
Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key
Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode
Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity
Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo
Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature
Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms
Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key
Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode
Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity
Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo
Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature
Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms
Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key
Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode
Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity
Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo
Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature
Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response
Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

# \donttest{
get_recommendations(market = 'SE', seed_genres = 'rock')

#>                                                                                                                                                                                                                                                                                                                                                                                                    artists
#> 1                                                                                                                                                                                                    https://api.spotify.com/v1/artists/6mdiAmATAx73kdxrNrnlao, 6mdiAmATAx73kdxrNrnlao, Iron Maiden, artist, spotify:artist:6mdiAmATAx73kdxrNrnlao, https://open.spotify.com/artist/6mdiAmATAx73kdxrNrnlao
#> 2                                                                                                                                                                                                         https://api.spotify.com/v1/artists/4KWTAlx2RvbpseOGMEmROg, 4KWTAlx2RvbpseOGMEmROg, R.E.M., artist, spotify:artist:4KWTAlx2RvbpseOGMEmROg, https://open.spotify.com/artist/4KWTAlx2RvbpseOGMEmROg
#> 3                                                                                                                                                                                                       https://api.spotify.com/v1/artists/7EQ0qTo7fWT7DPxmxtSYEc, 7EQ0qTo7fWT7DPxmxtSYEc, Bastille, artist, spotify:artist:7EQ0qTo7fWT7DPxmxtSYEc, https://open.spotify.com/artist/7EQ0qTo7fWT7DPxmxtSYEc
#> 4                                                                                                                                                                                                       https://api.spotify.com/v1/artists/4vpDg7Y7fU982Ds30zawDA, 4vpDg7Y7fU982Ds30zawDA, The Band, artist, spotify:artist:4vpDg7Y7fU982Ds30zawDA, https://open.spotify.com/artist/4vpDg7Y7fU982Ds30zawDA
#> 5                                                                                                                                                                                                        https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, 6GI52t8N5F02MxU0g5U69P, Santana, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P
#> 6                                                                                                                                                                                                        https://api.spotify.com/v1/artists/6B5c4sch27tWHAGdarpPaW, 6B5c4sch27tWHAGdarpPaW, Seether, artist, spotify:artist:6B5c4sch27tWHAGdarpPaW, https://open.spotify.com/artist/6B5c4sch27tWHAGdarpPaW
#> 7                                                                                                                                                                                                         https://api.spotify.com/v1/artists/2HOMW1liAxI3URnEq95SMS, 2HOMW1liAxI3URnEq95SMS, MISSIO, artist, spotify:artist:2HOMW1liAxI3URnEq95SMS, https://open.spotify.com/artist/2HOMW1liAxI3URnEq95SMS
#> 8                                                                                                                                                                                                https://api.spotify.com/v1/artists/4NpFxQe2UvRCAjto3JqlSl, 4NpFxQe2UvRCAjto3JqlSl, Greta Van Fleet, artist, spotify:artist:4NpFxQe2UvRCAjto3JqlSl, https://open.spotify.com/artist/4NpFxQe2UvRCAjto3JqlSl
#> 9                                                                                                                                                                                                 https://api.spotify.com/v1/artists/0vEsuISMWAKNctLlUAhSZC, 0vEsuISMWAKNctLlUAhSZC, Counting Crows, artist, spotify:artist:0vEsuISMWAKNctLlUAhSZC, https://open.spotify.com/artist/0vEsuISMWAKNctLlUAhSZC
#> 10                                                                                                                                                                                                  https://api.spotify.com/v1/artists/4DWX7u8BV0vZIQSpJQQDWU, 4DWX7u8BV0vZIQSpJQQDWU, Alter Bridge, artist, spotify:artist:4DWX7u8BV0vZIQSpJQQDWU, https://open.spotify.com/artist/4DWX7u8BV0vZIQSpJQQDWU
#> 11                                                                                                                                                                                             https://api.spotify.com/v1/artists/0nmQIMXWTXfhgOBdNzhGOs, 0nmQIMXWTXfhgOBdNzhGOs, Avenged Sevenfold, artist, spotify:artist:0nmQIMXWTXfhgOBdNzhGOs, https://open.spotify.com/artist/0nmQIMXWTXfhgOBdNzhGOs
#> 12                                                                                                                                                                                                   https://api.spotify.com/v1/artists/6XyY86QOPPrYVGvF9ch6wz, 6XyY86QOPPrYVGvF9ch6wz, Linkin Park, artist, spotify:artist:6XyY86QOPPrYVGvF9ch6wz, https://open.spotify.com/artist/6XyY86QOPPrYVGvF9ch6wz
#> 13                                                                                                                                                                                           https://api.spotify.com/v1/artists/20JZFwl6HVl6yg8a4H3ZqK, 20JZFwl6HVl6yg8a4H3ZqK, Panic! At The Disco, artist, spotify:artist:20JZFwl6HVl6yg8a4H3ZqK, https://open.spotify.com/artist/20JZFwl6HVl6yg8a4H3ZqK
#> 14                                                                                                                                                                                                        https://api.spotify.com/v1/artists/0ECwFtbIWEVNwjlrfc6xoL, 0ECwFtbIWEVNwjlrfc6xoL, Eagles, artist, spotify:artist:0ECwFtbIWEVNwjlrfc6xoL, https://open.spotify.com/artist/0ECwFtbIWEVNwjlrfc6xoL
#> 15                                                                                                                                                                                                   https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 16 https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, https://api.spotify.com/v1/artists/3aBkeBhwadnWMWoVJ2CxJC, 6GI52t8N5F02MxU0g5U69P, 3aBkeBhwadnWMWoVJ2CxJC, Santana, Rob Thomas, artist, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, spotify:artist:3aBkeBhwadnWMWoVJ2CxJC, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/3aBkeBhwadnWMWoVJ2CxJC
#> 17                                                                                                                                                                                                        https://api.spotify.com/v1/artists/0qT79UgT5tY4yudH9VfsdT, 0qT79UgT5tY4yudH9VfsdT, Sum 41, artist, spotify:artist:0qT79UgT5tY4yudH9VfsdT, https://open.spotify.com/artist/0qT79UgT5tY4yudH9VfsdT
#> 18                                                                                                                                                                                                 https://api.spotify.com/v1/artists/08GQAI4eElDnROBrJRGE0X, 08GQAI4eElDnROBrJRGE0X, Fleetwood Mac, artist, spotify:artist:08GQAI4eElDnROBrJRGE0X, https://open.spotify.com/artist/08GQAI4eElDnROBrJRGE0X
#> 19                                                                                                                                                                                             https://api.spotify.com/v1/artists/2sil8z5kiy4r76CRTXxBCA, 2sil8z5kiy4r76CRTXxBCA, The Goo Goo Dolls, artist, spotify:artist:2sil8z5kiy4r76CRTXxBCA, https://open.spotify.com/artist/2sil8z5kiy4r76CRTXxBCA
#> 20                                                                                                                                                                                                      https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#>    disc_number duration_ms explicit
#> 1            1      233506    FALSE
#> 2            1      246933    FALSE
#> 3            1      201799    FALSE
#> 4            1      274493    FALSE
#> 5            1      196466    FALSE
#> 6            1      244346    FALSE
#> 7            1      215200    FALSE
#> 8            1      259893    FALSE
#> 9            1      272293    FALSE
#> 10           1      259959    FALSE
#> 11           1      298531    FALSE
#> 12           1      216281    FALSE
#> 13           1      186600    FALSE
#> 14           1      391376    FALSE
#> 15           1      366136    FALSE
#> 16           1      294986    FALSE
#> 17           1      207093    FALSE
#> 18           1      270213    FALSE
#> 19           1      289533    FALSE
#> 20           1      307879    FALSE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/4Zc7TCHzuNwL0AFBlyLdyr
#> 2  https://api.spotify.com/v1/tracks/2oSpQ7QtIKTNFfA08Cy0ku
#> 3  https://api.spotify.com/v1/tracks/1etiUDkISHELzQGMY79ryt
#> 4  https://api.spotify.com/v1/tracks/0P7DoyGrr4Wp9w5TotEtUC
#> 5  https://api.spotify.com/v1/tracks/4YMQXzscifAREG0a7KNGhB
#> 6  https://api.spotify.com/v1/tracks/3XHnaUAFX8PfNMW1hHC75N
#> 7  https://api.spotify.com/v1/tracks/1RgvTl4vQN4qECtnO1guEp
#> 8  https://api.spotify.com/v1/tracks/32BEasyNQgqXQJKoVa0fUw
#> 9  https://api.spotify.com/v1/tracks/5DiXcVovI0FcY2s0icWWUu
#> 10 https://api.spotify.com/v1/tracks/3UULkHdmLDqBDPmBYltoRE
#> 11 https://api.spotify.com/v1/tracks/6Gcv1c96EIihHdJeERkpc8
#> 12 https://api.spotify.com/v1/tracks/3FQCJI2t5LTbsRPfYVBSVB
#> 13 https://api.spotify.com/v1/tracks/0wdV94AXGRphrbkDWLNzVC
#> 14 https://api.spotify.com/v1/tracks/40riOy7x9W7GXjyGp4pjAv
#> 15 https://api.spotify.com/v1/tracks/3uMmllZo1AfoEnVT4ENCD3
#> 16 https://api.spotify.com/v1/tracks/0n2SEXB2qoRQg171q7XqeW
#> 17 https://api.spotify.com/v1/tracks/1HNE2PX70ztbEl6MLxrpNL
#> 18 https://api.spotify.com/v1/tracks/5e9TFTbltYBg2xThimr0rU
#> 19 https://api.spotify.com/v1/tracks/6Qyc6fS4DsZjB2mRW9DsQs
#> 20 https://api.spotify.com/v1/tracks/0BCPKOYdS2jbQ8iyB56Zns
#>                        id is_local is_playable
#> 1  4Zc7TCHzuNwL0AFBlyLdyr    FALSE        TRUE
#> 2  2oSpQ7QtIKTNFfA08Cy0ku    FALSE        TRUE
#> 3  1etiUDkISHELzQGMY79ryt    FALSE        TRUE
#> 4  0P7DoyGrr4Wp9w5TotEtUC    FALSE        TRUE
#> 5  4YMQXzscifAREG0a7KNGhB    FALSE        TRUE
#> 6  3XHnaUAFX8PfNMW1hHC75N    FALSE        TRUE
#> 7  1RgvTl4vQN4qECtnO1guEp    FALSE        TRUE
#> 8  32BEasyNQgqXQJKoVa0fUw    FALSE        TRUE
#> 9  5DiXcVovI0FcY2s0icWWUu    FALSE        TRUE
#> 10 3UULkHdmLDqBDPmBYltoRE    FALSE        TRUE
#> 11 6Gcv1c96EIihHdJeERkpc8    FALSE        TRUE
#> 12 3FQCJI2t5LTbsRPfYVBSVB    FALSE        TRUE
#> 13 0wdV94AXGRphrbkDWLNzVC    FALSE        TRUE
#> 14 40riOy7x9W7GXjyGp4pjAv    FALSE        TRUE
#> 15 3uMmllZo1AfoEnVT4ENCD3    FALSE        TRUE
#> 16 0n2SEXB2qoRQg171q7XqeW    FALSE        TRUE
#> 17 1HNE2PX70ztbEl6MLxrpNL    FALSE        TRUE
#> 18 5e9TFTbltYBg2xThimr0rU    FALSE        TRUE
#> 19 6Qyc6fS4DsZjB2mRW9DsQs    FALSE        TRUE
#> 20 0BCPKOYdS2jbQ8iyB56Zns    FALSE        TRUE
#>                                                         name popularity
#> 1                           Run to the Hills - 2015 Remaster         71
#> 2  It's The End Of The World As We Know It (And I Feel Fine)         70
#> 3                                      Quarter Past Midnight         59
#> 4                               The Weight - Remastered 2000         73
#> 5                         Black Magic Woman - Single Version         69
#> 6                                         Betray And Degrade         54
#> 7                                             Middle Fingers         49
#> 8                                         Black Smoke Rising         70
#> 9                                                  Mr. Jones         77
#> 10                                                Metalingus         66
#> 11                                                      Dose         39
#> 12                                           Battle Symphony         64
#> 13                                               Roaring 20s         62
#> 14                          Hotel California - 2013 Remaster         83
#> 15                            Highway Star - Remastered 2012         67
#> 16                                 Smooth (feat. Rob Thomas)         69
#> 17                                               In Too Deep         75
#> 18                                 The Chain - 2004 Remaster         81
#> 19                                                      Iris         80
#> 20                                                    Clocks         78
#>                                                                                                    preview_url
#> 1  https://p.scdn.co/mp3-preview/66a6b7102df35ce9e83d29f8f5092ad529eaa14e?cid=125a10da5fa94687b3d0913168bc3698
#> 2                                                                                                         <NA>
#> 3                                                                                                         <NA>
#> 4                                                                                                         <NA>
#> 5  https://p.scdn.co/mp3-preview/09ba9494a44c24340443d4a7069a80240fec7066?cid=125a10da5fa94687b3d0913168bc3698
#> 6                                                                                                         <NA>
#> 7  https://p.scdn.co/mp3-preview/9289bf9795c01feda4ecb9b4e18c9cb1a6c0385b?cid=125a10da5fa94687b3d0913168bc3698
#> 8                                                                                                         <NA>
#> 9                                                                                                         <NA>
#> 10                                                                                                        <NA>
#> 11                                                                                                        <NA>
#> 12 https://p.scdn.co/mp3-preview/0d5e44827994c1530c93331aacbd774def9f0511?cid=125a10da5fa94687b3d0913168bc3698
#> 13 https://p.scdn.co/mp3-preview/2ee94c5fed3c87e01b7d5cb97cb2739c9091e426?cid=125a10da5fa94687b3d0913168bc3698
#> 14 https://p.scdn.co/mp3-preview/50e82c99c20ffa4223e82250605bbd8500cb3928?cid=125a10da5fa94687b3d0913168bc3698
#> 15                                                                                                        <NA>
#> 16 https://p.scdn.co/mp3-preview/9f977ebb2e695502c8b1d3bd678ef706c9ea4ccb?cid=125a10da5fa94687b3d0913168bc3698
#> 17                                                                                                        <NA>
#> 18 https://p.scdn.co/mp3-preview/39267f043904cef22e6652c0ce761114321953fd?cid=125a10da5fa94687b3d0913168bc3698
#> 19 https://p.scdn.co/mp3-preview/0e87ade9fa869e31b3c32115c2c10ee8fa522c81?cid=125a10da5fa94687b3d0913168bc3698
#> 20 https://p.scdn.co/mp3-preview/77d0dd7e02ca1de794143b6034d86df6243328b9?cid=125a10da5fa94687b3d0913168bc3698
#>    track_number  type                                  uri album.album_type
#> 1             6 track spotify:track:4Zc7TCHzuNwL0AFBlyLdyr            ALBUM
#> 2             6 track spotify:track:2oSpQ7QtIKTNFfA08Cy0ku            ALBUM
#> 3             1 track spotify:track:1etiUDkISHELzQGMY79ryt           SINGLE
#> 4             5 track spotify:track:0P7DoyGrr4Wp9w5TotEtUC            ALBUM
#> 5             6 track spotify:track:4YMQXzscifAREG0a7KNGhB      COMPILATION
#> 6             2 track spotify:track:3XHnaUAFX8PfNMW1hHC75N            ALBUM
#> 7             3 track spotify:track:1RgvTl4vQN4qECtnO1guEp            ALBUM
#> 8             8 track spotify:track:32BEasyNQgqXQJKoVa0fUw            ALBUM
#> 9             3 track spotify:track:5DiXcVovI0FcY2s0icWWUu            ALBUM
#> 10            5 track spotify:track:3UULkHdmLDqBDPmBYltoRE            ALBUM
#> 11           12 track spotify:track:6Gcv1c96EIihHdJeERkpc8            ALBUM
#> 12            4 track spotify:track:3FQCJI2t5LTbsRPfYVBSVB            ALBUM
#> 13            5 track spotify:track:0wdV94AXGRphrbkDWLNzVC            ALBUM
#> 14            1 track spotify:track:40riOy7x9W7GXjyGp4pjAv            ALBUM
#> 15            1 track spotify:track:3uMmllZo1AfoEnVT4ENCD3            ALBUM
#> 16            5 track spotify:track:0n2SEXB2qoRQg171q7XqeW            ALBUM
#> 17            7 track spotify:track:1HNE2PX70ztbEl6MLxrpNL            ALBUM
#> 18            7 track spotify:track:5e9TFTbltYBg2xThimr0rU            ALBUM
#> 19           11 track spotify:track:6Qyc6fS4DsZjB2mRW9DsQs            ALBUM
#> 20            5 track spotify:track:0BCPKOYdS2jbQ8iyB56Zns            ALBUM
#>                                                                                                                                                                                                    album.artists
#> 1          https://api.spotify.com/v1/artists/6mdiAmATAx73kdxrNrnlao, 6mdiAmATAx73kdxrNrnlao, Iron Maiden, artist, spotify:artist:6mdiAmATAx73kdxrNrnlao, https://open.spotify.com/artist/6mdiAmATAx73kdxrNrnlao
#> 2               https://api.spotify.com/v1/artists/4KWTAlx2RvbpseOGMEmROg, 4KWTAlx2RvbpseOGMEmROg, R.E.M., artist, spotify:artist:4KWTAlx2RvbpseOGMEmROg, https://open.spotify.com/artist/4KWTAlx2RvbpseOGMEmROg
#> 3             https://api.spotify.com/v1/artists/7EQ0qTo7fWT7DPxmxtSYEc, 7EQ0qTo7fWT7DPxmxtSYEc, Bastille, artist, spotify:artist:7EQ0qTo7fWT7DPxmxtSYEc, https://open.spotify.com/artist/7EQ0qTo7fWT7DPxmxtSYEc
#> 4             https://api.spotify.com/v1/artists/4vpDg7Y7fU982Ds30zawDA, 4vpDg7Y7fU982Ds30zawDA, The Band, artist, spotify:artist:4vpDg7Y7fU982Ds30zawDA, https://open.spotify.com/artist/4vpDg7Y7fU982Ds30zawDA
#> 5              https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, 6GI52t8N5F02MxU0g5U69P, Santana, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P
#> 6              https://api.spotify.com/v1/artists/6B5c4sch27tWHAGdarpPaW, 6B5c4sch27tWHAGdarpPaW, Seether, artist, spotify:artist:6B5c4sch27tWHAGdarpPaW, https://open.spotify.com/artist/6B5c4sch27tWHAGdarpPaW
#> 7               https://api.spotify.com/v1/artists/2HOMW1liAxI3URnEq95SMS, 2HOMW1liAxI3URnEq95SMS, MISSIO, artist, spotify:artist:2HOMW1liAxI3URnEq95SMS, https://open.spotify.com/artist/2HOMW1liAxI3URnEq95SMS
#> 8      https://api.spotify.com/v1/artists/4NpFxQe2UvRCAjto3JqlSl, 4NpFxQe2UvRCAjto3JqlSl, Greta Van Fleet, artist, spotify:artist:4NpFxQe2UvRCAjto3JqlSl, https://open.spotify.com/artist/4NpFxQe2UvRCAjto3JqlSl
#> 9       https://api.spotify.com/v1/artists/0vEsuISMWAKNctLlUAhSZC, 0vEsuISMWAKNctLlUAhSZC, Counting Crows, artist, spotify:artist:0vEsuISMWAKNctLlUAhSZC, https://open.spotify.com/artist/0vEsuISMWAKNctLlUAhSZC
#> 10        https://api.spotify.com/v1/artists/4DWX7u8BV0vZIQSpJQQDWU, 4DWX7u8BV0vZIQSpJQQDWU, Alter Bridge, artist, spotify:artist:4DWX7u8BV0vZIQSpJQQDWU, https://open.spotify.com/artist/4DWX7u8BV0vZIQSpJQQDWU
#> 11   https://api.spotify.com/v1/artists/0nmQIMXWTXfhgOBdNzhGOs, 0nmQIMXWTXfhgOBdNzhGOs, Avenged Sevenfold, artist, spotify:artist:0nmQIMXWTXfhgOBdNzhGOs, https://open.spotify.com/artist/0nmQIMXWTXfhgOBdNzhGOs
#> 12         https://api.spotify.com/v1/artists/6XyY86QOPPrYVGvF9ch6wz, 6XyY86QOPPrYVGvF9ch6wz, Linkin Park, artist, spotify:artist:6XyY86QOPPrYVGvF9ch6wz, https://open.spotify.com/artist/6XyY86QOPPrYVGvF9ch6wz
#> 13 https://api.spotify.com/v1/artists/20JZFwl6HVl6yg8a4H3ZqK, 20JZFwl6HVl6yg8a4H3ZqK, Panic! At The Disco, artist, spotify:artist:20JZFwl6HVl6yg8a4H3ZqK, https://open.spotify.com/artist/20JZFwl6HVl6yg8a4H3ZqK
#> 14              https://api.spotify.com/v1/artists/0ECwFtbIWEVNwjlrfc6xoL, 0ECwFtbIWEVNwjlrfc6xoL, Eagles, artist, spotify:artist:0ECwFtbIWEVNwjlrfc6xoL, https://open.spotify.com/artist/0ECwFtbIWEVNwjlrfc6xoL
#> 15         https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 16             https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, 6GI52t8N5F02MxU0g5U69P, Santana, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P
#> 17              https://api.spotify.com/v1/artists/0qT79UgT5tY4yudH9VfsdT, 0qT79UgT5tY4yudH9VfsdT, Sum 41, artist, spotify:artist:0qT79UgT5tY4yudH9VfsdT, https://open.spotify.com/artist/0qT79UgT5tY4yudH9VfsdT
#> 18       https://api.spotify.com/v1/artists/08GQAI4eElDnROBrJRGE0X, 08GQAI4eElDnROBrJRGE0X, Fleetwood Mac, artist, spotify:artist:08GQAI4eElDnROBrJRGE0X, https://open.spotify.com/artist/08GQAI4eElDnROBrJRGE0X
#> 19   https://api.spotify.com/v1/artists/2sil8z5kiy4r76CRTXxBCA, 2sil8z5kiy4r76CRTXxBCA, The Goo Goo Dolls, artist, spotify:artist:2sil8z5kiy4r76CRTXxBCA, https://open.spotify.com/artist/2sil8z5kiy4r76CRTXxBCA
#> 20            https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/5S3gls8Kjn8KVmqlIDEBbO
#> 2  https://api.spotify.com/v1/albums/6gMv3MgFlieOM6Uz5GZBzy
#> 3  https://api.spotify.com/v1/albums/70p5NvBOkvaxU1UeIxhE1Z
#> 4  https://api.spotify.com/v1/albums/0ky5kdvfPxSmSpj03hpSAE
#> 5  https://api.spotify.com/v1/albums/1YrdGK2y2Edq2R5gFja2ai
#> 6  https://api.spotify.com/v1/albums/2CbMgWfTfT4n80PjrDkbbt
#> 7  https://api.spotify.com/v1/albums/08iSi4GIyoDt2zSvNyru7M
#> 8  https://api.spotify.com/v1/albums/6uSnHSIBGKUiW1uKQLYZ7w
#> 9  https://api.spotify.com/v1/albums/4nKfZbCALT9H9LfedtDwnZ
#> 10 https://api.spotify.com/v1/albums/6hqietEuZCikevdO2oGTcx
#> 11 https://api.spotify.com/v1/albums/0OJ2cB135AqvHEtfXifM5D
#> 12 https://api.spotify.com/v1/albums/5Eevxp2BCbWq25ZdiXRwYd
#> 13 https://api.spotify.com/v1/albums/6ApYSpXF8GxZAgBTHDzYge
#> 14 https://api.spotify.com/v1/albums/2widuo17g5CEC66IbzveRu
#> 15 https://api.spotify.com/v1/albums/32NQ56VZDTXSH3SMv4XSGN
#> 16 https://api.spotify.com/v1/albums/10aiDpdFGyfCFEcqpx6XTq
#> 17 https://api.spotify.com/v1/albums/2UCWsnmZEVg9HhnMeKTsim
#> 18 https://api.spotify.com/v1/albums/0BwWUstDMUbgq2NYONRqlu
#> 19 https://api.spotify.com/v1/albums/4UMjBXcRqIgMZ1XumU2x5T
#> 20 https://api.spotify.com/v1/albums/0RHX9XECH8IVI3LNgWDpmQ
#>                  album.id
#> 1  5S3gls8Kjn8KVmqlIDEBbO
#> 2  6gMv3MgFlieOM6Uz5GZBzy
#> 3  70p5NvBOkvaxU1UeIxhE1Z
#> 4  0ky5kdvfPxSmSpj03hpSAE
#> 5  1YrdGK2y2Edq2R5gFja2ai
#> 6  2CbMgWfTfT4n80PjrDkbbt
#> 7  08iSi4GIyoDt2zSvNyru7M
#> 8  6uSnHSIBGKUiW1uKQLYZ7w
#> 9  4nKfZbCALT9H9LfedtDwnZ
#> 10 6hqietEuZCikevdO2oGTcx
#> 11 0OJ2cB135AqvHEtfXifM5D
#> 12 5Eevxp2BCbWq25ZdiXRwYd
#> 13 6ApYSpXF8GxZAgBTHDzYge
#> 14 2widuo17g5CEC66IbzveRu
#> 15 32NQ56VZDTXSH3SMv4XSGN
#> 16 10aiDpdFGyfCFEcqpx6XTq
#> 17 2UCWsnmZEVg9HhnMeKTsim
#> 18 0BwWUstDMUbgq2NYONRqlu
#> 19 4UMjBXcRqIgMZ1XumU2x5T
#> 20 0RHX9XECH8IVI3LNgWDpmQ
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735c29a88ba5341ca428f0c322, https://i.scdn.co/image/ab67616d00001e025c29a88ba5341ca428f0c322, https://i.scdn.co/image/ab67616d000048515c29a88ba5341ca428f0c322, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a7c7f94cfccd5ab6de233916, https://i.scdn.co/image/ab67616d00001e02a7c7f94cfccd5ab6de233916, https://i.scdn.co/image/ab67616d00004851a7c7f94cfccd5ab6de233916, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273137d133f1c79c4965eace78c, https://i.scdn.co/image/ab67616d00001e02137d133f1c79c4965eace78c, https://i.scdn.co/image/ab67616d00004851137d133f1c79c4965eace78c, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27320bbceac6950d3fe13fa13c3, https://i.scdn.co/image/ab67616d00001e0220bbceac6950d3fe13fa13c3, https://i.scdn.co/image/ab67616d0000485120bbceac6950d3fe13fa13c3, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273930858249e7abbb9b5317929, https://i.scdn.co/image/ab67616d00001e02930858249e7abbb9b5317929, https://i.scdn.co/image/ab67616d00004851930858249e7abbb9b5317929, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f60fd6ac5c9c508d1f636fdb, https://i.scdn.co/image/ab67616d00001e02f60fd6ac5c9c508d1f636fdb, https://i.scdn.co/image/ab67616d00004851f60fd6ac5c9c508d1f636fdb, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273eb3a6d9c69ee04f74b4250d4, https://i.scdn.co/image/ab67616d00001e02eb3a6d9c69ee04f74b4250d4, https://i.scdn.co/image/ab67616d00004851eb3a6d9c69ee04f74b4250d4, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27357b7f789d328c205b4d15893, https://i.scdn.co/image/ab67616d00001e0257b7f789d328c205b4d15893, https://i.scdn.co/image/ab67616d0000485157b7f789d328c205b4d15893, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737e09670f90cd47b3fb9a23e0, https://i.scdn.co/image/ab67616d00001e027e09670f90cd47b3fb9a23e0, https://i.scdn.co/image/ab67616d000048517e09670f90cd47b3fb9a23e0, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273bc7ddb77993dd1d8d19c22a2, https://i.scdn.co/image/ab67616d00001e02bc7ddb77993dd1d8d19c22a2, https://i.scdn.co/image/ab67616d00004851bc7ddb77993dd1d8d19c22a2, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734843e0f824a00334e811279c, https://i.scdn.co/image/ab67616d00001e024843e0f824a00334e811279c, https://i.scdn.co/image/ab67616d000048514843e0f824a00334e811279c, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273145e1cf12538f5666511237d, https://i.scdn.co/image/ab67616d00001e02145e1cf12538f5666511237d, https://i.scdn.co/image/ab67616d00004851145e1cf12538f5666511237d, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c5148520a59be191eea16989, https://i.scdn.co/image/ab67616d00001e02c5148520a59be191eea16989, https://i.scdn.co/image/ab67616d00004851c5148520a59be191eea16989, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734637341b9f507521afa9a778, https://i.scdn.co/image/ab67616d00001e024637341b9f507521afa9a778, https://i.scdn.co/image/ab67616d000048514637341b9f507521afa9a778, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00001e02a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00004851a729c9c3dec04b99d889c66f, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27347eb3ea5a92904c19e102e54, https://i.scdn.co/image/ab67616d00001e0247eb3ea5a92904c19e102e54, https://i.scdn.co/image/ab67616d0000485147eb3ea5a92904c19e102e54, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2739a482180e6a306229bff49dc, https://i.scdn.co/image/ab67616d00001e029a482180e6a306229bff49dc, https://i.scdn.co/image/ab67616d000048519a482180e6a306229bff49dc, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e52a59a28efa4773dd2bfe1b, https://i.scdn.co/image/ab67616d00001e02e52a59a28efa4773dd2bfe1b, https://i.scdn.co/image/ab67616d00004851e52a59a28efa4773dd2bfe1b, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273eda9478c39a21e1cdc6609ca, https://i.scdn.co/image/ab67616d00001e02eda9478c39a21e1cdc6609ca, https://i.scdn.co/image/ab67616d00004851eda9478c39a21e1cdc6609ca, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27390afd8e4ec6d787114ed6c40, https://i.scdn.co/image/ab67616d00001e0290afd8e4ec6d787114ed6c40, https://i.scdn.co/image/ab67616d0000485190afd8e4ec6d787114ed6c40, 640, 300, 64
#>                                 album.name album.release_date
#> 1  The Number of the Beast (2015 Remaster)               1982
#> 2                  Document (R.E.M. No. 5)         1987-09-01
#> 3                    Quarter Past Midnight         2018-05-09
#> 4         Music From Big Pink (Remastered)         1968-07-01
#> 5                  Santana's Greatest Hits               1969
#> 6       Poison The Parish (Deluxe Edition)         2017-05-12
#> 7                                    Loner         2017-05-19
#> 8                           From The Fires         2017-11-10
#> 9              August And Everything After         1993-01-01
#> 10                         One Day Remains         2004-08-10
#> 11              The Stage (Deluxe Edition)         2017-12-22
#> 12                          One More Light         2017-05-19
#> 13                     Pray for the Wicked         2018-06-22
#> 14        Hotel California (2013 Remaster)         1976-12-08
#> 15               Machine Head (Remastered)         1972-03-25
#> 16               Supernatural (Remastered)         1999-06-15
#> 17                   All Killer, No Filler         2001-01-01
#> 18                  Rumours (Super Deluxe)         1977-02-04
#> 19                       Dizzy up the Girl         1998-09-11
#> 20             A Rush of Blood to the Head         2002-08-08
#>    album.release_date_precision album.total_tracks album.type
#> 1                          year                  8      album
#> 2                           day                 11      album
#> 3                           day                  1      album
#> 4                           day                 11      album
#> 5                          year                 10      album
#> 6                           day                 15      album
#> 7                           day                 11      album
#> 8                           day                  8      album
#> 9                           day                 11      album
#> 10                          day                 11      album
#> 11                          day                 22      album
#> 12                          day                 10      album
#> 13                          day                 11      album
#> 14                          day                  9      album
#> 15                          day                  8      album
#> 16                          day                 14      album
#> 17                          day                 13      album
#> 18                          day                 58      album
#> 19                          day                 13      album
#> 20                          day                 11      album
#>                               album.uri
#> 1  spotify:album:5S3gls8Kjn8KVmqlIDEBbO
#> 2  spotify:album:6gMv3MgFlieOM6Uz5GZBzy
#> 3  spotify:album:70p5NvBOkvaxU1UeIxhE1Z
#> 4  spotify:album:0ky5kdvfPxSmSpj03hpSAE
#> 5  spotify:album:1YrdGK2y2Edq2R5gFja2ai
#> 6  spotify:album:2CbMgWfTfT4n80PjrDkbbt
#> 7  spotify:album:08iSi4GIyoDt2zSvNyru7M
#> 8  spotify:album:6uSnHSIBGKUiW1uKQLYZ7w
#> 9  spotify:album:4nKfZbCALT9H9LfedtDwnZ
#> 10 spotify:album:6hqietEuZCikevdO2oGTcx
#> 11 spotify:album:0OJ2cB135AqvHEtfXifM5D
#> 12 spotify:album:5Eevxp2BCbWq25ZdiXRwYd
#> 13 spotify:album:6ApYSpXF8GxZAgBTHDzYge
#> 14 spotify:album:2widuo17g5CEC66IbzveRu
#> 15 spotify:album:32NQ56VZDTXSH3SMv4XSGN
#> 16 spotify:album:10aiDpdFGyfCFEcqpx6XTq
#> 17 spotify:album:2UCWsnmZEVg9HhnMeKTsim
#> 18 spotify:album:0BwWUstDMUbgq2NYONRqlu
#> 19 spotify:album:4UMjBXcRqIgMZ1XumU2x5T
#> 20 spotify:album:0RHX9XECH8IVI3LNgWDpmQ
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/5S3gls8Kjn8KVmqlIDEBbO      GBCHB1800025
#> 2  https://open.spotify.com/album/6gMv3MgFlieOM6Uz5GZBzy      USIR38700037
#> 3  https://open.spotify.com/album/70p5NvBOkvaxU1UeIxhE1Z      GBUM71801713
#> 4  https://open.spotify.com/album/0ky5kdvfPxSmSpj03hpSAE      USCA20000789
#> 5  https://open.spotify.com/album/1YrdGK2y2Edq2R5gFja2ai      USSM17000643
#> 6  https://open.spotify.com/album/2CbMgWfTfT4n80PjrDkbbt      USC4R1700002
#> 7  https://open.spotify.com/album/08iSi4GIyoDt2zSvNyru7M      USRC11700139
#> 8  https://open.spotify.com/album/6uSnHSIBGKUiW1uKQLYZ7w      USUM71703681
#> 9  https://open.spotify.com/album/4nKfZbCALT9H9LfedtDwnZ      USIR10000287
#> 10 https://open.spotify.com/album/6hqietEuZCikevdO2oGTcx      USWU30400101
#> 11 https://open.spotify.com/album/0OJ2cB135AqvHEtfXifM5D      USUG11700613
#> 12 https://open.spotify.com/album/5Eevxp2BCbWq25ZdiXRwYd      USWB11700221
#> 13 https://open.spotify.com/album/6ApYSpXF8GxZAgBTHDzYge      USAT21801175
#> 14 https://open.spotify.com/album/2widuo17g5CEC66IbzveRu      USEE11300353
#> 15 https://open.spotify.com/album/32NQ56VZDTXSH3SMv4XSGN      GBDXG1200002
#> 16 https://open.spotify.com/album/10aiDpdFGyfCFEcqpx6XTq      USAR19900033
#> 17 https://open.spotify.com/album/2UCWsnmZEVg9HhnMeKTsim      USIR20110114
#> 18 https://open.spotify.com/album/0BwWUstDMUbgq2NYONRqlu      USWB11301116
#> 19 https://open.spotify.com/album/4UMjBXcRqIgMZ1XumU2x5T      USWB19800160
#> 20 https://open.spotify.com/album/0RHX9XECH8IVI3LNgWDpmQ      GBAYE0200771
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/4Zc7TCHzuNwL0AFBlyLdyr
#> 2  https://open.spotify.com/track/2oSpQ7QtIKTNFfA08Cy0ku
#> 3  https://open.spotify.com/track/1etiUDkISHELzQGMY79ryt
#> 4  https://open.spotify.com/track/0P7DoyGrr4Wp9w5TotEtUC
#> 5  https://open.spotify.com/track/4YMQXzscifAREG0a7KNGhB
#> 6  https://open.spotify.com/track/3XHnaUAFX8PfNMW1hHC75N
#> 7  https://open.spotify.com/track/1RgvTl4vQN4qECtnO1guEp
#> 8  https://open.spotify.com/track/32BEasyNQgqXQJKoVa0fUw
#> 9  https://open.spotify.com/track/5DiXcVovI0FcY2s0icWWUu
#> 10 https://open.spotify.com/track/3UULkHdmLDqBDPmBYltoRE
#> 11 https://open.spotify.com/track/6Gcv1c96EIihHdJeERkpc8
#> 12 https://open.spotify.com/track/3FQCJI2t5LTbsRPfYVBSVB
#> 13 https://open.spotify.com/track/0wdV94AXGRphrbkDWLNzVC
#> 14 https://open.spotify.com/track/40riOy7x9W7GXjyGp4pjAv
#> 15 https://open.spotify.com/track/3uMmllZo1AfoEnVT4ENCD3
#> 16 https://open.spotify.com/track/0n2SEXB2qoRQg171q7XqeW
#> 17 https://open.spotify.com/track/1HNE2PX70ztbEl6MLxrpNL
#> 18 https://open.spotify.com/track/5e9TFTbltYBg2xThimr0rU
#> 19 https://open.spotify.com/track/6Qyc6fS4DsZjB2mRW9DsQs
#> 20 https://open.spotify.com/track/0BCPKOYdS2jbQ8iyB56Zns
#>                                            linked_from.href
#> 1  https://api.spotify.com/v1/tracks/2JS1iE5A5RHvUPH5Zl9jlF
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4                                                      <NA>
#> 5                                                      <NA>
#> 6                                                      <NA>
#> 7                                                      <NA>
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#> 11                                                     <NA>
#> 12                                                     <NA>
#> 13                                                     <NA>
#> 14                                                     <NA>
#> 15                                                     <NA>
#> 16                                                     <NA>
#> 17                                                     <NA>
#> 18                                                     <NA>
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1  2JS1iE5A5RHvUPH5Zl9jlF            track spotify:track:2JS1iE5A5RHvUPH5Zl9jlF
#> 2                    <NA>             <NA>                                 <NA>
#> 3                    <NA>             <NA>                                 <NA>
#> 4                    <NA>             <NA>                                 <NA>
#> 5                    <NA>             <NA>                                 <NA>
#> 6                    <NA>             <NA>                                 <NA>
#> 7                    <NA>             <NA>                                 <NA>
#> 8                    <NA>             <NA>                                 <NA>
#> 9                    <NA>             <NA>                                 <NA>
#> 10                   <NA>             <NA>                                 <NA>
#> 11                   <NA>             <NA>                                 <NA>
#> 12                   <NA>             <NA>                                 <NA>
#> 13                   <NA>             <NA>                                 <NA>
#> 14                   <NA>             <NA>                                 <NA>
#> 15                   <NA>             <NA>                                 <NA>
#> 16                   <NA>             <NA>                                 <NA>
#> 17                   <NA>             <NA>                                 <NA>
#> 18                   <NA>             <NA>                                 <NA>
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1  https://open.spotify.com/track/2JS1iE5A5RHvUPH5Zl9jlF
#> 2                                                   <NA>
#> 3                                                   <NA>
#> 4                                                   <NA>
#> 5                                                   <NA>
#> 6                                                   <NA>
#> 7                                                   <NA>
#> 8                                                   <NA>
#> 9                                                   <NA>
#> 10                                                  <NA>
#> 11                                                  <NA>
#> 12                                                  <NA>
#> 13                                                  <NA>
#> 14                                                  <NA>
#> 15                                                  <NA>
#> 16                                                  <NA>
#> 17                                                  <NA>
#> 18                                                  <NA>
#> 19                                                  <NA>
#> 20                                                  <NA>

# }