Get Recommendations for Unlimited Vector of Track IDs

Get recommendations for a submitted vector of track IDs, with no limit on the number of seed tracks

get_recommendations_all(track_ids, valence = NULL)

Arguments

track_ids
A vector containing the IDs of the tracks you'd like recommendations for

valence
The target valence for the recommendations

Value

Returns a data frame containing binded recommendations from the Spotify API in a single data frame.

Details

This is a wrapper for the get_recommendations function, which provides a workaround for the limit of 5 seed tracks per recommendation call. The function splits a supplied vector of track IDs into subsets of length 5, then applies a get_recommendations call, 5 tracks at a time. This should generate a data frame of recommended tracks, with length equal to the supplied vector of track ids.

See also

Examples

# \donttest{
get_recommendations_all(
   track_ids = c("5VIpLopHgolKcSSj7JPCMA", "3QRGYDFFUVb4qneE4DX1gR")
   )

#> Request failed [400]. Retrying in 1 seconds...

#> Request failed [400]. Retrying in 1 seconds...

#> Error in get_spotify_access_token(): 

# }