Get Spotify catalog information for a single track identified by its unique Spotify ID.

Get Spotify catalog information for a single track identified by its unique Spotify ID.

get_track(id, market = NULL, authorization = get_spotify_access_token())

Arguments

id

The Spotify ID for the track.

market

Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token". Provide this parameter if you want to apply Track Relinking

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

Returns a data frame of results containing track data. See https://developer.spotify.com/documentation/web-api/reference/tracks/get-several-tracks/ for more information.

See also