Get audio features of tracks

Get audio feature information for up to 100 tracks identified by their unique Spotify IDs.

get_track_audio_features(ids, authorization = get_spotify_access_token())

Arguments

ids
Required. A comma-separated list of the Spotify IDs of the tracks. Maximum: 100 IDs.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

Returns a data frame of results containing track audio features data. See https://developer.spotify.com/documentation/web-api/reference/tracks/get-several-audio-features/ for more information.

See also