Skips to previous track in the user’s queue.

Skips to previous track in the user’s queue.

skip_my_playback_previous(device_id = NULL,
  authorization = get_spotify_authorization_code())

Arguments

device_id
Optional. The id of the device this command is targeting. If not supplied, the user’s currently active device is the target.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details. Defaults to spotifyr::get_spotify_access_token(). The access token must have been issued on behalf of the current user.
The access token must have the user-modify-playback-state scope authorized in order to control playback.