spotifyr package

A Quick and Easy Wrapper for Pulling Track Audio Features from Spotify's Web API in Bulk

Details

See the README on GitHub

artist functions

Retrieve information related to artists. get_artist: Get the data of a single artist.
get_artists: Get the data of multiple artist.
get_related_artists: Get that of related artist to an original artist.
get_artist_albums: Get artists who appear on an album.

track functions

Retrieve information related to individual song recordings (songs, concerts). get_track
get_tracks
get_playlist_tracks

album functions

Retrieve information about albums. See also get_album_tracks.

playlist functions

Work with playlists.
add_tracks_to_playlist
change_playlist_details
get_user_playlists

player functions

Interact with the user's devices and players.

personalization functions

get_my_top_artists_or_tracks

musicology functions

Functions related to the contents of the music.
get_track_audio_analysis
get_artist_audio_features
get_playlist_audio_features
get_user_audio_features
get_genre_artists

search functions

Search for an artist, song or other keyword.
The main function is search_spotify.

authentication functions

Helper functions to provide authentication function for requests. Never reveal in public documents, markdown files the returned values of these functions:
get_spotify_authorization_code
get_spotify_access_token.