Skips to previous track in the user’s queue.

Skips to previous track in the user’s queue.

start_my_playback(
  device_id = NULL,
  context_uri = NULL,
  uris = NULL,
  offset = NULL,
  position_ms = NULL,
  authorization = get_spotify_authorization_code()
)

Arguments

device_id

Optional. The id of the device this command is targeting. If not supplied, the user’s currently active device is the target.

context_uri

Optional. String of the Spotify URI of the context to play. Valid contexts are albums, artists, playlists. Example context_uri = "spotify:album:1Je1IMUlBXcx1Fz0WE7oPT".

uris

Optional. A character vector of the Spotify track URIs to play. For example: "uris": c("spotify:track:4iV5W9uYEdYUVa79Axb7Rh", "spotify:track:1301WleyT98MSxVHPZCA6M").

offset

Optional. A named list indicating from where the context playback should start. Only available when context_uri corresponds to an album or playlist object, or when the uris parameter is used.
"position" is zero based and can't be negative. Example: "offset" = list("position" = 5).
"uri" is a string representing the uri of the item to start at. Example: "offset" = list("uri" = "spotify:track:1301WleyT98MSxVHPZCA6M").

position_ms

Optional. Integer indicating from what position to start playback. Must be a positive number. Passing in a position that is greater than the length of the track will cause the player to start playing the next song.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details. Defaults to spotifyr::get_spotify_access_token(). The access token must have been issued on behalf of the current user.
The access token must have the user-modify-playback-state scope authorized in order to control playback.

Value

The response from the Spotify Web API on the POST request, with a meaningful error message if the operation was not successful.

See also