Check if Users Follow a Playlist

Check if particular users are following a particular playlist.

check_users_following(
  playlist_id,
  ids,
  authorization = get_spotify_authorization_code()
)

Arguments

playlist_id

Required. The Spotify ID of the playlist.

ids

Required. Spotify User IDs; the ids of the users that you want to check to see if they follow the playlist. Maximum: 5 ids.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization Guide for more details.
Following a playlist can be done publicly or privately. Checking if a user publicly follows a playlist doesn’t require any scopes; if the user is publicly following the playlist, this endpoint returns TRUE.
Checking if the user is privately following a playlist is only possible for the current user when that user has granted access to the playlist-read-private scope. See Using Scopes.

Value

A data frame containing three columns,

"user_id" for the particular users,

"playlist_id" for the particular playlists,

"is_following" as a logical, boolean variable.

See also