Retrieve artist discography with song lyrics and audio info

Retrieve the entire discography of an artist with the lyrics of each song and the associated audio information. Returns the song data as a nested tibble. This way we can easily see each album, artist, and song title before expanding our data.

get_album_data(artist, albums = character())

Arguments

artist
The quoted name of the artist. Spelling matters, capitalization does not.

albums
A character vector of album names. Spelling matters, capitalization does not

Examples

get_album_data("Wild child", "Expectations")

#> # A tibble: 12 x 19
#>    album_name  track_name track_uri  track_n danceability energy key   loudness
#>    <chr>       <chr>      <chr>        <int>        <dbl>  <dbl> <chr>    <dbl>
#>  1 Expectatio… Alex       3SgSloOID…       1        0.622 0.720  A       - 8.59
#>  2 Expectatio… Eggshells  6n22ZSLAR…       2        0.408 0.369  F       -10.2 
#>  3 Expectatio… Back & Fo… 3FpsbToUN…       3        0.749 0.694  F       - 7.20
#>  4 Expectatio… Think It … 7jyEUpROt…       4        0.783 0.419  D       - 8.52
#>  5 Expectatio… Follow Me  1jCfkIVUw…       5        0.648 0.369  C       -13.6 
#>  6 Expectatio… Expectati… 6l6vgauYD…       6        0.689 0.643  B       - 6.50
#>  7 Expectatio… Sinking S… 6QzpfDKRj…       7        0.514 0.180  E       -19.3 
#>  8 Expectatio… My Town    1WCB6YXOU…       8        0.543 0.339  A       - 8.00
#>  9 Expectatio… The One    1qLVOZKRL…       9        0.563 0.0722 A       -20.7 
#> 10 Expectatio… Break You… 4fpse93GT…      10        0.709 0.511  F       -10.5 
#> 11 Expectatio… Leave It … 5KYmcAfpQ…      11        0.637 0.446  C       - 9.56
#> 12 Expectatio… Goodbye G… 5oMReIKx6…      12        0.466 0.461  C       - 5.18
#> # ... with 11 more variables: mode <chr>, speechiness <dbl>,
#> #   acousticness <dbl>, instrumentalness <dbl>, liveness <dbl>, valence <dbl>,
#> #   tempo <dbl>, duration_ms <dbl>, time_signature <dbl>, key_mode <chr>,
#> #   lyrics <list>