Get a list of Spotify playlists tagged with a particular category

Get a list of Spotify playlists tagged with a particular category

get_category_playlists(
  category_id = "party",
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

category_id

Required. The Spotify ID for the category.

country

Optional. A country: an ISO 3166-1 alpha-2 country code.

limit

Optional. The maximum number of items to return. Defaults to 20. Minimum: 1. Maximum: 50.

offset

Optional. The index of the first item to return. Defaults to 0, the first object. Use with limit to get the next set of items.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info

Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing category playlists. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
get_category_playlists('party', country = 'BR')
#>    collaborative
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11         FALSE
#> 12         FALSE
#> 13         FALSE
#> 14         FALSE
#> 15         FALSE
#> 16         FALSE
#> 17         FALSE
#> 18         FALSE
#> 19         FALSE
#> 20         FALSE
#>                                                                                                                                           description
#> 1  O funk tá diferente, meu sócio. Vem pro som do bololô. Foto: Mc IG, Mc Kadu, Mc Ryan SP, Trap Laudo, Mc PH, MC GP, MC Luki, Mc Kanhoto, Menor Will
#> 2                                                           Tem Sertanejo! Tem Funk! Tem Pagode! A melhor trilha pra embalar o seu churrasco do jogo.
#> 3                                                                                  Pra que esperar até sexta-feira a noite?? Aperte o play e SEXTOU. 
#> 4                                                                           Queeeima paredão! Pra forrozeiro sofrer dançando e dançar sofrendo, bora.
#> 5                                                                          Entre na onda com hits tropicais e remixes para curtir a praia com amigos.
#> 6                                                                                                                                         lista na dm
#> 7                                                                                Pôr do sol com pé na areia e aquela resenha com violão e muita rima.
#> 8                                                                                                                   The ultimate rock party playlist!
#> 9                                                                  O batidão dos funks mais bombados pra agitar a pista do seu baile! Foto: Mc Jacaré
#> 10                                                                                                            Pronto pra pool party? Vem com a gente!
#> 11                                                                                Grandes hinos da música brasileira para animar sua pista de dança! 
#> 12                                                                                              The biggest party hits of the 2000s. Cover: Shakira. 
#> 13                                                                                               The biggest party hits of the 2010s. Cover: Flo Rida
#> 14                                              playlist pra viajar e se perder em algum lugar entre o brasil e marte. vou te mostrar meu universo ;)
#> 15                                                                                             Para curtir aquela festa sertaneja puro modão do jogo!
#> 16                                                                                                  The biggest party hits of the 1990s. Cover: Aqua.
#> 17                                                                          Grandes sucessos do samba rock pra arrastar o sofá e quebrar o bracinho. 
#> 18                                                                                          Here it is, the playlist for the ultimate throwback bash!
#> 19                                                                                Reggaeton hits from Anuel AA, Rauw Alejandro, KAROL G and more.🔥🥵
#> 20                                                                                              <a href="spotify:genre:edm_dance">Move</a> your feet!
#>                                                           href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DX9mwdmcCXQM1
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX7e8TjkFNKWH
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DWWylYLMvjuRG
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DX8vpLK1FoEw3
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DXdo6A3mWpdWx
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DX7F6T2n2fegs
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DWY7IeIP1cdjF
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif
#>                        id
#> 1  37i9dQZF1DX4MdXmAY6EDq
#> 2  37i9dQZF1DXb1oSegSL8id
#> 3  37i9dQZF1DX8mBRYewE6or
#> 4  37i9dQZF1DXdXbKzsf80lE
#> 5  37i9dQZF1DX41UmHm2vtav
#> 6  37i9dQZF1DX9mwdmcCXQM1
#> 7  37i9dQZF1DWXvzyniz2fv4
#> 8  37i9dQZF1DX8FwnYE6PRvL
#> 9  37i9dQZF1DWWmaszSfZpom
#> 10 37i9dQZF1DXdWpx7KRCSbJ
#> 11 37i9dQZF1DX1oTjVZnUYT3
#> 12 37i9dQZF1DX7e8TjkFNKWH
#> 13 37i9dQZF1DWWylYLMvjuRG
#> 14 37i9dQZF1DX8vpLK1FoEw3
#> 15 37i9dQZF1DX5GJ0BQvLMjM
#> 16 37i9dQZF1DXdo6A3mWpdWx
#> 17 37i9dQZF1DX94Xc0f4tUlY
#> 18 37i9dQZF1DX7F6T2n2fegs
#> 19 37i9dQZF1DWY7IeIP1cdjF
#> 20 37i9dQZF1DXaXB8fQg7xif
#>                                                                      images
#> 1  NA, https://i.scdn.co/image/ab67706f000000037cf885a0034db29a9d66dec8, NA
#> 2  NA, https://i.scdn.co/image/ab67706f00000003878b6aa265f0147f251302de, NA
#> 3  NA, https://i.scdn.co/image/ab67706f00000003e138087c0972c43645489dd9, NA
#> 4  NA, https://i.scdn.co/image/ab67706f000000039ddb6e3c827a45b3c47ffe8d, NA
#> 5  NA, https://i.scdn.co/image/ab67706f00000003d7053e74da84ee2575ad645d, NA
#> 6  NA, https://i.scdn.co/image/ab67706f0000000394bbeda3ad6c7483b86a5775, NA
#> 7  NA, https://i.scdn.co/image/ab67706f0000000346cd6c19d9e17d1f9b044510, NA
#> 8  NA, https://i.scdn.co/image/ab67706f000000033c52ff5ce471594a9d6b848e, NA
#> 9  NA, https://i.scdn.co/image/ab67706f0000000328ddd529e3eded4b3949c73f, NA
#> 10 NA, https://i.scdn.co/image/ab67706f0000000370ea8f47d22d482cc6b3fecb, NA
#> 11 NA, https://i.scdn.co/image/ab67706f00000003c470446f590f6a8b8a55eaf8, NA
#> 12 NA, https://i.scdn.co/image/ab67706f00000003128b6654e84ef4cffba70489, NA
#> 13 NA, https://i.scdn.co/image/ab67706f000000030b3215962492991ff8354ee5, NA
#> 14 NA, https://i.scdn.co/image/ab67706f0000000381603a8e908e99dc7d8a57d9, NA
#> 15 NA, https://i.scdn.co/image/ab67706f00000003d83418c3a78f3faa5f097c17, NA
#> 16 NA, https://i.scdn.co/image/ab67706f00000003a9a7853641d94519d81f1a3c, NA
#> 17 NA, https://i.scdn.co/image/ab67706f000000036d84d98d614876aa51e67fbe, NA
#> 18 NA, https://i.scdn.co/image/ab67706f00000003370da67d71f7059f9b0a24a9, NA
#> 19 NA, https://i.scdn.co/image/ab67706f00000003d42bbcff2d07404bbe3b6050, NA
#> 20 NA, https://i.scdn.co/image/ab67706f00000003310432fef1cad89faa1ccc1b, NA
#>                                        name primary_color public
#> 1                                   Revoada            NA     NA
#> 2                       Churrasco e Futebol            NA     NA
#> 3                                    Sextou            NA     NA
#> 4                             Bora arrochar            NA     NA
#> 5                            Festa na Praia            NA     NA
#> 6                          minha party pfta            NA     NA
#> 7                          Resenha na Praia            NA     NA
#> 8                                Rock Party            NA     NA
#> 9                             Segue o Baile            NA     NA
#> 10                         Festa na Piscina            NA     NA
#> 11                         Festa Brasileira            NA     NA
#> 12                         Party Hits 2000s            NA     NA
#> 13                         Party Hits 2010s            NA     NA
#> 14 Nave Espacial - #MeuSpotify da Giulia Be            NA     NA
#> 15                      Farra, Bola e Modão            NA     NA
#> 16                                90s Party            NA     NA
#> 17                               Sambalanço            NA     NA
#> 18                          Throwback Party            NA     NA
#> 19                          Baila Reggaeton            NA     NA
#> 20                              Dance Party            NA     NA
#>                                                             snapshot_id
#> 1  MTY3MDU1NDk4MCwwMDAwMDAwMDhmYmQyY2MxZjJiYmRjOGRiYzBiNThlOTcwYTI3YjM3
#> 2  MTY3MDM0NTEyOSwwMDAwMDAwMDQxZTQzYWNlNDVlMWMwNDIzODk2MTE0OWI3ZWNkYWVj
#> 3  MTY2OTk1MDAwMCwwMDAwMDAwMDc2ZjZmZDY2OTkxNzcxMTM5M2U1ZTNkZGM5OWE2YzA2
#> 4  MTY3MDM0Mzc4MSwwMDAwMDAwMGQ4ZjY4MDU4MDllNjZmZTgxYjgxZmE1NzEwMjQyNGNi
#> 5  MTY3MDU1NDk4MCwwMDAwMDAwMDg2ODIwNjllNjg4MzYyOGNhZDE0ZWMwN2I3ZDE3NWRk
#> 6  MTY1OTU1MTgyOCwwMDAwMDAwMDgwOTg3M2FiN2JjZTUyZTk5NzRjNTBhMWJmZGU4YzUz
#> 7  MTY2OTk1MDA2MCwwMDAwMDAwMGFmMTI1MDRkYjU0ZjRlZDcwNjQ1ZTIwMzM0NzNkZGZh
#> 8  MTY2OTMwMzAwMSwwMDAwMDAwMGEzZDk2Yjc5ZGZlNTlmZDAxYjU1ZGU2MzMyODBjYzlk
#> 9  MTY2NTc2OTI0MCwwMDAwMDAwMDE4YWYwMzY0ZjU2ZWRlNTExYWNkYzlhYjU0YzlhMjU5
#> 10 MTY1MzkxNzg2OCwwMDAwMDAwMDAxZTBjZTY3MDg0MWJlZmM5MGVhNmU2YzU5ODc5MmYx
#> 11 MTYzNzg5NjczMiwwMDAwMDAwMDBmODhjZmQwNzI3MmM2YzM4ODgxOTBjMDUyMjVhMmM3
#> 12 MTY2MTc4MzI0MCwwMDAwMDAwMGVjMmE1NmUyY2MyZWRjZTBmMzg4NzkwY2I0ODBjZmY2
#> 13 MTY2MTc4MzE5MSwwMDAwMDAwMGJmODIyYjk1OTU4NDNiNDgyM2Q3NzZjY2E5NjI1NDlj
#> 14 MTY3MDM1NDc0NywwMDAwMDAwMDExYzExMTZhMWE2Y2FlOWMwODBmMDIxMDRhMzhjMTBh
#> 15 MTY2Njk2Mjg5MCwwMDAwMDAwMGY5ZmY0ZDc1YzE4ZjRiOWVmZmMyZTYyY2ZhYjgzNTBj
#> 16 MTY2NTQ0NjQwMCwwMDAwMDAwMDcxNmI3OThjZDYzYjIwOTZhYzI4ODc0OTc0YWQ0ODBk
#> 17 MTY2NTcxNjQ2MCwwMDAwMDAwMDk2NjA2MmVjYjA3YjQ2NmNkNzcyNDMxZWY1YzFkZWNh
#> 18 MTY3MDkzOTM4MSwwMDAwMDAwMDZmNDFiMDc2ZTRlN2MwMWEwODNhYTM0OTM2OWQwNzMz
#> 19 MTY3MDg2OTY0NywwMDAwMDAwMGIxNmFkNjFiNzQ4NTU4ZGQ4M2Q0OTAwYTU2NmY2NTcw
#> 20 MTY2OTk1NzIwMCwwMDAwMDAwMGYzYjhlZDBhYzUwZjhiNzE1MWIyOGUyOGFhMzQ1YzA2
#>        type                                     uri
#> 1  playlist spotify:playlist:37i9dQZF1DX4MdXmAY6EDq
#> 2  playlist spotify:playlist:37i9dQZF1DXb1oSegSL8id
#> 3  playlist spotify:playlist:37i9dQZF1DX8mBRYewE6or
#> 4  playlist spotify:playlist:37i9dQZF1DXdXbKzsf80lE
#> 5  playlist spotify:playlist:37i9dQZF1DX41UmHm2vtav
#> 6  playlist spotify:playlist:37i9dQZF1DX9mwdmcCXQM1
#> 7  playlist spotify:playlist:37i9dQZF1DWXvzyniz2fv4
#> 8  playlist spotify:playlist:37i9dQZF1DX8FwnYE6PRvL
#> 9  playlist spotify:playlist:37i9dQZF1DWWmaszSfZpom
#> 10 playlist spotify:playlist:37i9dQZF1DXdWpx7KRCSbJ
#> 11 playlist spotify:playlist:37i9dQZF1DX1oTjVZnUYT3
#> 12 playlist spotify:playlist:37i9dQZF1DX7e8TjkFNKWH
#> 13 playlist spotify:playlist:37i9dQZF1DWWylYLMvjuRG
#> 14 playlist spotify:playlist:37i9dQZF1DX8vpLK1FoEw3
#> 15 playlist spotify:playlist:37i9dQZF1DX5GJ0BQvLMjM
#> 16 playlist spotify:playlist:37i9dQZF1DXdo6A3mWpdWx
#> 17 playlist spotify:playlist:37i9dQZF1DX94Xc0f4tUlY
#> 18 playlist spotify:playlist:37i9dQZF1DX7F6T2n2fegs
#> 19 playlist spotify:playlist:37i9dQZF1DWY7IeIP1cdjF
#> 20 playlist spotify:playlist:37i9dQZF1DXaXB8fQg7xif
#>                                       external_urls.spotify owner.display_name
#> 1  https://open.spotify.com/playlist/37i9dQZF1DX4MdXmAY6EDq            Spotify
#> 2  https://open.spotify.com/playlist/37i9dQZF1DXb1oSegSL8id            Spotify
#> 3  https://open.spotify.com/playlist/37i9dQZF1DX8mBRYewE6or            Spotify
#> 4  https://open.spotify.com/playlist/37i9dQZF1DXdXbKzsf80lE            Spotify
#> 5  https://open.spotify.com/playlist/37i9dQZF1DX41UmHm2vtav            Spotify
#> 6  https://open.spotify.com/playlist/37i9dQZF1DX9mwdmcCXQM1            Spotify
#> 7  https://open.spotify.com/playlist/37i9dQZF1DWXvzyniz2fv4            Spotify
#> 8  https://open.spotify.com/playlist/37i9dQZF1DX8FwnYE6PRvL            Spotify
#> 9  https://open.spotify.com/playlist/37i9dQZF1DWWmaszSfZpom            Spotify
#> 10 https://open.spotify.com/playlist/37i9dQZF1DXdWpx7KRCSbJ            Spotify
#> 11 https://open.spotify.com/playlist/37i9dQZF1DX1oTjVZnUYT3            Spotify
#> 12 https://open.spotify.com/playlist/37i9dQZF1DX7e8TjkFNKWH            Spotify
#> 13 https://open.spotify.com/playlist/37i9dQZF1DWWylYLMvjuRG            Spotify
#> 14 https://open.spotify.com/playlist/37i9dQZF1DX8vpLK1FoEw3            Spotify
#> 15 https://open.spotify.com/playlist/37i9dQZF1DX5GJ0BQvLMjM            Spotify
#> 16 https://open.spotify.com/playlist/37i9dQZF1DXdo6A3mWpdWx            Spotify
#> 17 https://open.spotify.com/playlist/37i9dQZF1DX94Xc0f4tUlY            Spotify
#> 18 https://open.spotify.com/playlist/37i9dQZF1DX7F6T2n2fegs            Spotify
#> 19 https://open.spotify.com/playlist/37i9dQZF1DWY7IeIP1cdjF            Spotify
#> 20 https://open.spotify.com/playlist/37i9dQZF1DXaXB8fQg7xif            Spotify
#>                                  owner.href owner.id owner.type
#> 1  https://api.spotify.com/v1/users/spotify  spotify       user
#> 2  https://api.spotify.com/v1/users/spotify  spotify       user
#> 3  https://api.spotify.com/v1/users/spotify  spotify       user
#> 4  https://api.spotify.com/v1/users/spotify  spotify       user
#> 5  https://api.spotify.com/v1/users/spotify  spotify       user
#> 6  https://api.spotify.com/v1/users/spotify  spotify       user
#> 7  https://api.spotify.com/v1/users/spotify  spotify       user
#> 8  https://api.spotify.com/v1/users/spotify  spotify       user
#> 9  https://api.spotify.com/v1/users/spotify  spotify       user
#> 10 https://api.spotify.com/v1/users/spotify  spotify       user
#> 11 https://api.spotify.com/v1/users/spotify  spotify       user
#> 12 https://api.spotify.com/v1/users/spotify  spotify       user
#> 13 https://api.spotify.com/v1/users/spotify  spotify       user
#> 14 https://api.spotify.com/v1/users/spotify  spotify       user
#> 15 https://api.spotify.com/v1/users/spotify  spotify       user
#> 16 https://api.spotify.com/v1/users/spotify  spotify       user
#> 17 https://api.spotify.com/v1/users/spotify  spotify       user
#> 18 https://api.spotify.com/v1/users/spotify  spotify       user
#> 19 https://api.spotify.com/v1/users/spotify  spotify       user
#> 20 https://api.spotify.com/v1/users/spotify  spotify       user
#>               owner.uri           owner.external_urls.spotify
#> 1  spotify:user:spotify https://open.spotify.com/user/spotify
#> 2  spotify:user:spotify https://open.spotify.com/user/spotify
#> 3  spotify:user:spotify https://open.spotify.com/user/spotify
#> 4  spotify:user:spotify https://open.spotify.com/user/spotify
#> 5  spotify:user:spotify https://open.spotify.com/user/spotify
#> 6  spotify:user:spotify https://open.spotify.com/user/spotify
#> 7  spotify:user:spotify https://open.spotify.com/user/spotify
#> 8  spotify:user:spotify https://open.spotify.com/user/spotify
#> 9  spotify:user:spotify https://open.spotify.com/user/spotify
#> 10 spotify:user:spotify https://open.spotify.com/user/spotify
#> 11 spotify:user:spotify https://open.spotify.com/user/spotify
#> 12 spotify:user:spotify https://open.spotify.com/user/spotify
#> 13 spotify:user:spotify https://open.spotify.com/user/spotify
#> 14 spotify:user:spotify https://open.spotify.com/user/spotify
#> 15 spotify:user:spotify https://open.spotify.com/user/spotify
#> 16 spotify:user:spotify https://open.spotify.com/user/spotify
#> 17 spotify:user:spotify https://open.spotify.com/user/spotify
#> 18 spotify:user:spotify https://open.spotify.com/user/spotify
#> 19 spotify:user:spotify https://open.spotify.com/user/spotify
#> 20 spotify:user:spotify https://open.spotify.com/user/spotify
#>                                                           tracks.href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq/tracks
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id/tracks
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or/tracks
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE/tracks
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav/tracks
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DX9mwdmcCXQM1/tracks
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4/tracks
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL/tracks
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom/tracks
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ/tracks
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3/tracks
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX7e8TjkFNKWH/tracks
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DWWylYLMvjuRG/tracks
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DX8vpLK1FoEw3/tracks
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM/tracks
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DXdo6A3mWpdWx/tracks
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY/tracks
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DX7F6T2n2fegs/tracks
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DWY7IeIP1cdjF/tracks
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif/tracks
#>    tracks.total
#> 1            50
#> 2           100
#> 3            75
#> 4           100
#> 5            70
#> 6            65
#> 7            68
#> 8           100
#> 9            75
#> 10           70
#> 11           57
#> 12          150
#> 13          150
#> 14           61
#> 15          102
#> 16          150
#> 17           50
#> 18          150
#> 19           50
#> 20          150
# }