Get a list of Spotify playlists tagged with a particular category.

Get a list of Spotify playlists tagged with a particular category.

get_category_playlists(
  category_id,
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

category_id
Required. The Spotify ID for the category.

country
Optional. A country: an ISO 3166-1 alpha-2 country code.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing category playlists. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
## Get Brazilian party playlists
get_category_playlists('party', country = 'BR')

#>    collaborative
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11         FALSE
#> 12         FALSE
#> 13         FALSE
#> 14         FALSE
#> 15         FALSE
#> 16         FALSE
#> 17         FALSE
#> 18         FALSE
#> 19         FALSE
#> 20         FALSE
#>                                                                                            description
#> 1                                           Chegou o grande dia, aperte o play e partiu fim de semana!
#> 2                    Tem Sertanejo! Tem Funk! Tem Pagode! A melhor trilha pra embalar o seu churrasco.
#> 3                  Conforto para beber gin tônica, mojito ou caipirinha sem precisar levantar do sofá.
#> 4                                   O batidão dos funks mais bombados pra agitar a pista do seu baile!
#> 5  O funk tá diferente, meu sócio. Vem pro som do bololô. Foto: Mc Magal, Mc Piedro, Mc Lele JP, DJ GM
#> 6                            Queeeima paredão! Pra forrozeiro sofrer dançando e dançar sofrendo, bora.
#> 7                           Entre na onda com hits tropicais e remixes para curtir a praia com amigos.
#> 8                                 Pôr do sol com pé na areia e aquela resenha com violão e muita rima.
#> 9                                                                    It's all about the girls tonight.
#> 10                          Para curtir aquela noite sertaneja puro modão. Foto:  João Bosco e Gabriel
#> 11             Já pode beber? A melhor trilha sonora para embalar a festinha pós-trabalho de cada dia!
#> 12                                  Grandes hinos da música brasileira para animar sua pista de dança!
#> 13                                               <a href="spotify:genre:edm_dance">Move</a> your feet!
#> 14                                                                   The ultimate rock party playlist!
#> 15                                        Calor, amigos e muita música brasileira pra animar a galera!
#> 16            The hip-hop playlist that's a whole mood. Cover: Megan Thee Stallion; Art by Laci Jordan
#> 17                                    Um pouco de MPB, Samba e Axé para alegrar sua carne de carnaval.
#> 18                                                                                         Mode: Turnt
#> 19                                                     Pista boa é aquela pra bater cabelo e se jogar.
#> 20                           Grandes sucessos do samba rock pra arrastar o sofá e quebrar o bracinho. 
#>                                                           href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWX9JsANXY37r
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DX0Uv9tZ47pWo
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DWZ72qOlbizxi
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DX6GwdWRQMQpq
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DX9udUjwxj8v9
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DWY4xHQp97fN6
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DXcjvYvHoIYE6
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY
#>                        id
#> 1  37i9dQZF1DX8mBRYewE6or
#> 2  37i9dQZF1DXb1oSegSL8id
#> 3  37i9dQZF1DWX9JsANXY37r
#> 4  37i9dQZF1DWWmaszSfZpom
#> 5  37i9dQZF1DX4MdXmAY6EDq
#> 6  37i9dQZF1DXdXbKzsf80lE
#> 7  37i9dQZF1DX41UmHm2vtav
#> 8  37i9dQZF1DWXvzyniz2fv4
#> 9  37i9dQZF1DX0Uv9tZ47pWo
#> 10 37i9dQZF1DX5GJ0BQvLMjM
#> 11 37i9dQZF1DWZ72qOlbizxi
#> 12 37i9dQZF1DX1oTjVZnUYT3
#> 13 37i9dQZF1DXaXB8fQg7xif
#> 14 37i9dQZF1DX8FwnYE6PRvL
#> 15 37i9dQZF1DXdWpx7KRCSbJ
#> 16 37i9dQZF1DX6GwdWRQMQpq
#> 17 37i9dQZF1DX9udUjwxj8v9
#> 18 37i9dQZF1DWY4xHQp97fN6
#> 19 37i9dQZF1DXcjvYvHoIYE6
#> 20 37i9dQZF1DX94Xc0f4tUlY
#>                                                                      images
#> 1  NA, https://i.scdn.co/image/ab67706f00000003206a95fa5badbe1d33b65e14, NA
#> 2  NA, https://i.scdn.co/image/ab67706f000000032f3b2073765b80193eb26ef2, NA
#> 3  NA, https://i.scdn.co/image/ab67706f00000003fbabce11ab019931ca40c588, NA
#> 4  NA, https://i.scdn.co/image/ab67706f0000000330b91682f0604c3a3d071904, NA
#> 5  NA, https://i.scdn.co/image/ab67706f0000000310d0c93d20d37a9607ea1f25, NA
#> 6  NA, https://i.scdn.co/image/ab67706f00000003dcb303df890bfc5ba3e1906d, NA
#> 7  NA, https://i.scdn.co/image/ab67706f00000003bcf7685d4e9d6d0ccb44bfb2, NA
#> 8  NA, https://i.scdn.co/image/ab67706f00000003ee42a542a4c259708eff50c1, NA
#> 9  NA, https://i.scdn.co/image/ab67706f00000003955a83cb4667cf205fdf58eb, NA
#> 10 NA, https://i.scdn.co/image/ab67706f000000036ebdfeaeea4aa5c38d38a25b, NA
#> 11 NA, https://i.scdn.co/image/ab67706f00000003ca391c001dd460eb50815f99, NA
#> 12 NA, https://i.scdn.co/image/ab67706f0000000338930636b13ee4d2e1ee2e4e, NA
#> 13 NA, https://i.scdn.co/image/ab67706f00000003470dd505fcf08e4693db9b24, NA
#> 14 NA, https://i.scdn.co/image/ab67706f00000003feb66cd1f6632ecfa89b3e8c, NA
#> 15 NA, https://i.scdn.co/image/ab67706f000000033b8dc8bf62a876d30695ec1d, NA
#> 16 NA, https://i.scdn.co/image/ab67706f00000003bf5b729a4e375ebb081f82b9, NA
#> 17 NA, https://i.scdn.co/image/ab67706f00000003076d372cae48aa2772b70705, NA
#> 18 NA, https://i.scdn.co/image/ab67706f000000030aa2cf87a0674e630eecb75f, NA
#> 19 NA, https://i.scdn.co/image/ab67706f000000036c7f1e4ccf8ba2ee554cf5e0, NA
#> 20 NA, https://i.scdn.co/image/ab67706f000000036d84d98d614876aa51e67fbe, NA
#>                name primary_color public
#> 1             Sexta            NA     NA
#> 2   Sol e Churrasco            NA     NA
#> 3    Drinks no Sofá            NA     NA
#> 4     Segue o Baile            NA     NA
#> 5           Revoada            NA     NA
#> 6     Bora arrochar            NA     NA
#> 7    Festa na Praia            NA     NA
#> 8  Resenha na Praia            NA     NA
#> 9      Girls' Night            NA     NA
#> 10  Farra Sertaneja            NA     NA
#> 11       Happy Hour            NA     NA
#> 12 Festa Brasileira            NA     NA
#> 13      Dance Party            NA     NA
#> 14       Rock Party            NA     NA
#> 15 Festa na Piscina            NA     NA
#> 16   Feelin' Myself            NA     NA
#> 17     Carnavalizar            NA     NA
#> 18        Get Turnt            NA     NA
#> 19            Fervo            NA     NA
#> 20       Sambalanço            NA     NA
#>                                                             snapshot_id
#> 1  MTYyMzQwMTIzMywwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 2  MTYyMzQwMTI3MSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 3  MTYyMzQwMTI0NiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 4  MTYyMzI2NzE2MSwwMDAwMDBiODAwMDAwMTc5ZjI0MzM0ZTcwMDAwMDE3MGFmM2M0YWQy
#> 5  MTYyMzM4MDQ2MCwwMDAwMDAyNTAwMDAwMTc5ZjkwNDAyNWUwMDAwMDE3OTFmNjlkNTI3
#> 6  MTYyMzQwMTI3MCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 7  MTYyMzQwMTIzOSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 8  MTYyMjgzODQ0NSwwMDAwMDA2MzAwMDAwMTc5ZDhiNTg2M2UwMDAwMDE3MmYxYzU1MTVh
#> 9  MTYyMzQwMTIyNywwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 10 MTYyMzM4MDQ2MCwwMDAwMDNmYzAwMDAwMTc5ZjkwNDAyNWUwMDAwMDE3OGIzY2E1YjQ5
#> 11 MTYyMzQwMTIzOSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 12 MTYyMzQwMTI4MCwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 13 MTYyMzQwMTIyNiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 14 MTYxOTEyOTEwOCwwMDAwMDAzOTAwMDAwMTc4ZmI5ZDg3MDQwMDAwMDE2ZDE1NTBjOWZh
#> 15 MTYyMzQwMTI2MiwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 16 MTYyMzM4NDAwMCwwMDAwMDQzZTAwMDAwMTc5ZjkzYTA2ZTMwMDAwMDE3OWYxYjZlYmM0
#> 17 MTYyMzQwMTI2MSwwMDAwMDAwMGQ0MWQ4Y2Q5OGYwMGIyMDRlOTgwMDk5OGVjZjg0Mjdl
#> 18 MTYyMzM4NDAwMCwwMDAwMDRhOTAwMDAwMTc5ZjkzYTA2Y2UwMDAwMDE3OWYxYmMwZmMx
#> 19 MTYxNjEyMjg2MCwwMDAwMDAwYTAwMDAwMTc4NDg2ZGQyN2YwMDAwMDE3Mzk1YzhjYjVm
#> 20 MTYyMzM4MDQ2MCwwMDAwMDAwYzAwMDAwMTc5ZjkwNDAyNTQwMDAwMDE3ODg1NTkxMzBi
#>        type                                     uri
#> 1  playlist spotify:playlist:37i9dQZF1DX8mBRYewE6or
#> 2  playlist spotify:playlist:37i9dQZF1DXb1oSegSL8id
#> 3  playlist spotify:playlist:37i9dQZF1DWX9JsANXY37r
#> 4  playlist spotify:playlist:37i9dQZF1DWWmaszSfZpom
#> 5  playlist spotify:playlist:37i9dQZF1DX4MdXmAY6EDq
#> 6  playlist spotify:playlist:37i9dQZF1DXdXbKzsf80lE
#> 7  playlist spotify:playlist:37i9dQZF1DX41UmHm2vtav
#> 8  playlist spotify:playlist:37i9dQZF1DWXvzyniz2fv4
#> 9  playlist spotify:playlist:37i9dQZF1DX0Uv9tZ47pWo
#> 10 playlist spotify:playlist:37i9dQZF1DX5GJ0BQvLMjM
#> 11 playlist spotify:playlist:37i9dQZF1DWZ72qOlbizxi
#> 12 playlist spotify:playlist:37i9dQZF1DX1oTjVZnUYT3
#> 13 playlist spotify:playlist:37i9dQZF1DXaXB8fQg7xif
#> 14 playlist spotify:playlist:37i9dQZF1DX8FwnYE6PRvL
#> 15 playlist spotify:playlist:37i9dQZF1DXdWpx7KRCSbJ
#> 16 playlist spotify:playlist:37i9dQZF1DX6GwdWRQMQpq
#> 17 playlist spotify:playlist:37i9dQZF1DX9udUjwxj8v9
#> 18 playlist spotify:playlist:37i9dQZF1DWY4xHQp97fN6
#> 19 playlist spotify:playlist:37i9dQZF1DXcjvYvHoIYE6
#> 20 playlist spotify:playlist:37i9dQZF1DX94Xc0f4tUlY
#>                                       external_urls.spotify owner.display_name
#> 1  https://open.spotify.com/playlist/37i9dQZF1DX8mBRYewE6or            Spotify
#> 2  https://open.spotify.com/playlist/37i9dQZF1DXb1oSegSL8id            Spotify
#> 3  https://open.spotify.com/playlist/37i9dQZF1DWX9JsANXY37r            Spotify
#> 4  https://open.spotify.com/playlist/37i9dQZF1DWWmaszSfZpom            Spotify
#> 5  https://open.spotify.com/playlist/37i9dQZF1DX4MdXmAY6EDq            Spotify
#> 6  https://open.spotify.com/playlist/37i9dQZF1DXdXbKzsf80lE            Spotify
#> 7  https://open.spotify.com/playlist/37i9dQZF1DX41UmHm2vtav            Spotify
#> 8  https://open.spotify.com/playlist/37i9dQZF1DWXvzyniz2fv4            Spotify
#> 9  https://open.spotify.com/playlist/37i9dQZF1DX0Uv9tZ47pWo            Spotify
#> 10 https://open.spotify.com/playlist/37i9dQZF1DX5GJ0BQvLMjM            Spotify
#> 11 https://open.spotify.com/playlist/37i9dQZF1DWZ72qOlbizxi            Spotify
#> 12 https://open.spotify.com/playlist/37i9dQZF1DX1oTjVZnUYT3            Spotify
#> 13 https://open.spotify.com/playlist/37i9dQZF1DXaXB8fQg7xif            Spotify
#> 14 https://open.spotify.com/playlist/37i9dQZF1DX8FwnYE6PRvL            Spotify
#> 15 https://open.spotify.com/playlist/37i9dQZF1DXdWpx7KRCSbJ            Spotify
#> 16 https://open.spotify.com/playlist/37i9dQZF1DX6GwdWRQMQpq            Spotify
#> 17 https://open.spotify.com/playlist/37i9dQZF1DX9udUjwxj8v9            Spotify
#> 18 https://open.spotify.com/playlist/37i9dQZF1DWY4xHQp97fN6            Spotify
#> 19 https://open.spotify.com/playlist/37i9dQZF1DXcjvYvHoIYE6            Spotify
#> 20 https://open.spotify.com/playlist/37i9dQZF1DX94Xc0f4tUlY            Spotify
#>                                  owner.href owner.id owner.type
#> 1  https://api.spotify.com/v1/users/spotify  spotify       user
#> 2  https://api.spotify.com/v1/users/spotify  spotify       user
#> 3  https://api.spotify.com/v1/users/spotify  spotify       user
#> 4  https://api.spotify.com/v1/users/spotify  spotify       user
#> 5  https://api.spotify.com/v1/users/spotify  spotify       user
#> 6  https://api.spotify.com/v1/users/spotify  spotify       user
#> 7  https://api.spotify.com/v1/users/spotify  spotify       user
#> 8  https://api.spotify.com/v1/users/spotify  spotify       user
#> 9  https://api.spotify.com/v1/users/spotify  spotify       user
#> 10 https://api.spotify.com/v1/users/spotify  spotify       user
#> 11 https://api.spotify.com/v1/users/spotify  spotify       user
#> 12 https://api.spotify.com/v1/users/spotify  spotify       user
#> 13 https://api.spotify.com/v1/users/spotify  spotify       user
#> 14 https://api.spotify.com/v1/users/spotify  spotify       user
#> 15 https://api.spotify.com/v1/users/spotify  spotify       user
#> 16 https://api.spotify.com/v1/users/spotify  spotify       user
#> 17 https://api.spotify.com/v1/users/spotify  spotify       user
#> 18 https://api.spotify.com/v1/users/spotify  spotify       user
#> 19 https://api.spotify.com/v1/users/spotify  spotify       user
#> 20 https://api.spotify.com/v1/users/spotify  spotify       user
#>               owner.uri           owner.external_urls.spotify
#> 1  spotify:user:spotify https://open.spotify.com/user/spotify
#> 2  spotify:user:spotify https://open.spotify.com/user/spotify
#> 3  spotify:user:spotify https://open.spotify.com/user/spotify
#> 4  spotify:user:spotify https://open.spotify.com/user/spotify
#> 5  spotify:user:spotify https://open.spotify.com/user/spotify
#> 6  spotify:user:spotify https://open.spotify.com/user/spotify
#> 7  spotify:user:spotify https://open.spotify.com/user/spotify
#> 8  spotify:user:spotify https://open.spotify.com/user/spotify
#> 9  spotify:user:spotify https://open.spotify.com/user/spotify
#> 10 spotify:user:spotify https://open.spotify.com/user/spotify
#> 11 spotify:user:spotify https://open.spotify.com/user/spotify
#> 12 spotify:user:spotify https://open.spotify.com/user/spotify
#> 13 spotify:user:spotify https://open.spotify.com/user/spotify
#> 14 spotify:user:spotify https://open.spotify.com/user/spotify
#> 15 spotify:user:spotify https://open.spotify.com/user/spotify
#> 16 spotify:user:spotify https://open.spotify.com/user/spotify
#> 17 spotify:user:spotify https://open.spotify.com/user/spotify
#> 18 spotify:user:spotify https://open.spotify.com/user/spotify
#> 19 spotify:user:spotify https://open.spotify.com/user/spotify
#> 20 spotify:user:spotify https://open.spotify.com/user/spotify
#>                                                           tracks.href
#> 1  https://api.spotify.com/v1/playlists/37i9dQZF1DX8mBRYewE6or/tracks
#> 2  https://api.spotify.com/v1/playlists/37i9dQZF1DXb1oSegSL8id/tracks
#> 3  https://api.spotify.com/v1/playlists/37i9dQZF1DWX9JsANXY37r/tracks
#> 4  https://api.spotify.com/v1/playlists/37i9dQZF1DWWmaszSfZpom/tracks
#> 5  https://api.spotify.com/v1/playlists/37i9dQZF1DX4MdXmAY6EDq/tracks
#> 6  https://api.spotify.com/v1/playlists/37i9dQZF1DXdXbKzsf80lE/tracks
#> 7  https://api.spotify.com/v1/playlists/37i9dQZF1DX41UmHm2vtav/tracks
#> 8  https://api.spotify.com/v1/playlists/37i9dQZF1DWXvzyniz2fv4/tracks
#> 9  https://api.spotify.com/v1/playlists/37i9dQZF1DX0Uv9tZ47pWo/tracks
#> 10 https://api.spotify.com/v1/playlists/37i9dQZF1DX5GJ0BQvLMjM/tracks
#> 11 https://api.spotify.com/v1/playlists/37i9dQZF1DWZ72qOlbizxi/tracks
#> 12 https://api.spotify.com/v1/playlists/37i9dQZF1DX1oTjVZnUYT3/tracks
#> 13 https://api.spotify.com/v1/playlists/37i9dQZF1DXaXB8fQg7xif/tracks
#> 14 https://api.spotify.com/v1/playlists/37i9dQZF1DX8FwnYE6PRvL/tracks
#> 15 https://api.spotify.com/v1/playlists/37i9dQZF1DXdWpx7KRCSbJ/tracks
#> 16 https://api.spotify.com/v1/playlists/37i9dQZF1DX6GwdWRQMQpq/tracks
#> 17 https://api.spotify.com/v1/playlists/37i9dQZF1DX9udUjwxj8v9/tracks
#> 18 https://api.spotify.com/v1/playlists/37i9dQZF1DWY4xHQp97fN6/tracks
#> 19 https://api.spotify.com/v1/playlists/37i9dQZF1DXcjvYvHoIYE6/tracks
#> 20 https://api.spotify.com/v1/playlists/37i9dQZF1DX94Xc0f4tUlY/tracks
#>    tracks.total
#> 1            62
#> 2           100
#> 3            60
#> 4            71
#> 5            50
#> 6           100
#> 7            80
#> 8            68
#> 9            80
#> 10          100
#> 11           75
#> 12           57
#> 13          100
#> 14          145
#> 15           65
#> 16           50
#> 17           50
#> 18          100
#> 19           56
#> 20           51

# }