Get list of Spotify featured playlists

Get a list of Spotify featured playlists (as shown, for example, on a Spotify player’s ‘Browse’ tab)

get_featured_playlists(
  locale = NULL,
  country = NULL,
  timestamp = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

locale

Optional. The desired language, consisting of an ISO 639-1 language code and an ISO 3166-1 alpha-2 country code, joined by an underscore. For example: es_MX, meaning "Spanish (Mexico)". Provide this parameter if you want the category strings returned in a particular language. Note that, if locale is not supplied, or if the specified language is not available, the category strings returned will be in the Spotify default language (American English). The locale parameter, combined with the country parameter, may give odd results if not carefully matched. For example country=SE&locale=de_DE will return a list of categories relevant to Sweden but as German language strings.

country

Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

timestamp

Optional. A timestamp in ISO 8601 format: yyyy-MM-ddTHH:mm:ss. Use this parameter to specify the user’s local time to get results tailored for that specific date and time in the day. If not provided, the response defaults to the current UTC time. Example: “2014-10-23T09:00:00” for a user whose local time is 9AM. If there were no featured playlists (or there is no data) at the specified time, the response will revert to the current UTC time.

limit

Optional. The maximum number of items to return. Default to 20. Minimum: 1. Maximum: 50.

offset

Optional. The index of the first item to return. Defaults to 0, i.e., the first object. Use with limit to get the next set of items.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info

Optional. Boolean indicating whether to include full result, with meta information such as "total", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing featured playlists.

See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{

## Get Flemish-Dutch playlists from Belgium:

get_featured_playlists(
   country = 'BE',
   locale = "nl_BE" )
#>   collaborative
#> 1         FALSE
#> 2         FALSE
#> 3         FALSE
#> 4         FALSE
#> 5         FALSE
#> 6         FALSE
#> 7         FALSE
#> 8         FALSE
#> 9         FALSE
#>                                                                                      description
#> 1                              50 Hits van dit moment - 50 hits du moment. Cover: Meghan Trainor
#> 2              Restez au courant des nouveaux titres et des tubes d'ici et ailleurs. Cover: Nour
#> 3                                                    Voel je goed met deze tijdloze Happy Tunes!
#> 4                      Luister naar muziek uit België - Listen music made in Belgium. #playlocal
#> 5                   Le meilleur de la musique d'hier et d'aujourd'hui pour une journée parfaite.
#> 6                                       DIKKE full album in Code 32! The best BE rap in Code 32.
#> 7               Indie, Alternative en Rock met een randje in één playlist. Cover: Orange Skyline
#> 8                                            De grootste dance hits van het moment! Cover: Jengi
#> 9 Meezingen met de beste Vlaamse hits, levensliederen en Vlaams Populair. Cover: Pommelien Thijs
#>                                                          href
#> 1 https://api.spotify.com/v1/playlists/37i9dQZF1DX1kfybUJZB6S
#> 2 https://api.spotify.com/v1/playlists/37i9dQZF1DWSq1nQilUgG8
#> 3 https://api.spotify.com/v1/playlists/37i9dQZF1DX9u7XXOp0l5L
#> 4 https://api.spotify.com/v1/playlists/37i9dQZF1DX3S2ONCFIYHU
#> 5 https://api.spotify.com/v1/playlists/37i9dQZF1DXdrln2UyZD7F
#> 6 https://api.spotify.com/v1/playlists/37i9dQZF1DXadhN5awWZtL
#> 7 https://api.spotify.com/v1/playlists/37i9dQZF1DX6MUrG3NBYtM
#> 8 https://api.spotify.com/v1/playlists/37i9dQZF1DWTwCImwcYjDL
#> 9 https://api.spotify.com/v1/playlists/37i9dQZF1DX4UF20jxFQBv
#>                       id
#> 1 37i9dQZF1DX1kfybUJZB6S
#> 2 37i9dQZF1DWSq1nQilUgG8
#> 3 37i9dQZF1DX9u7XXOp0l5L
#> 4 37i9dQZF1DX3S2ONCFIYHU
#> 5 37i9dQZF1DXdrln2UyZD7F
#> 6 37i9dQZF1DXadhN5awWZtL
#> 7 37i9dQZF1DX6MUrG3NBYtM
#> 8 37i9dQZF1DWTwCImwcYjDL
#> 9 37i9dQZF1DX4UF20jxFQBv
#>                                                                     images
#> 1 NA, https://i.scdn.co/image/ab67706f00000003fdb580e113383783b674d3d2, NA
#> 2 NA, https://i.scdn.co/image/ab67706f00000003af80586535c6916023ea2c69, NA
#> 3 NA, https://i.scdn.co/image/ab67706f0000000380e48e10df93b4f887cb969d, NA
#> 4 NA, https://i.scdn.co/image/ab67706f0000000374ef3354f9509a43988a935a, NA
#> 5 NA, https://i.scdn.co/image/ab67706f00000003ed570b97eecf4d3d33e49b9b, NA
#> 6 NA, https://i.scdn.co/image/ab67706f00000003a9cbde5ae99dae8a4c199a98, NA
#> 7 NA, https://i.scdn.co/image/ab67706f0000000359e296d727972950f33a771e, NA
#> 8 NA, https://i.scdn.co/image/ab67706f0000000373b1aaa0d39ca5e41a39aa3e, NA
#> 9 NA, https://i.scdn.co/image/ab67706f00000003721f1c2543cb0cd4184d6915, NA
#>                       name primary_color public
#> 1              Hot Hits BE            NA     NA
#> 2           Restez Branché            NA     NA
#> 3              Happy Tunes            NA     NA
#> 4         Made In  Belgium            NA     NA
#> 5         La vie est belle            NA     NA
#> 6                  Code 32            NA     NA
#> 7 Alternative All Lined Up            NA     NA
#> 8                360 Dance            NA     NA
#> 9        Vlaanderen Boven!            NA     NA
#>                                                            snapshot_id     type
#> 1 MTY3MDU0MDQwMCwwMDAwMDAwMDBkZTllMGVkNWJkOTFmMzg2MGY3NjBiYzA5NWM2OWY3 playlist
#> 2 MTY3MDU0MDQwMCwwMDAwMDAwMDhlNzIzMzJhOTQ1YWM5OTIwODI4YzU0MmI3MDIzYjU4 playlist
#> 3 MTY3MDQyODQyOSwwMDAwMDAwMGRkZDNhZjU5NTNkMGI2ZDBjZjA4MWMwMWViNDRjN2Yw playlist
#> 4 MTY3MDU5MjI1OSwwMDAwMDAwMDljMWVjNzYwZjc0NjgwNTkwZjJlODM3ODkxNWQ2Y2Yx playlist
#> 5 MTY3MDM5OTgzNywwMDAwMDAwMDIzM2NlZDA0OTBhMmE4NGZmMDdkMjc1OGJiMzM4OGVk playlist
#> 6 MTY3MDU0MDQ2MCwwMDAwMDAwMDhlM2E2NDdhZTFiN2MwYjc3MTFmYTA0ZmVmNzNhNjYx playlist
#> 7 MTY2NzUxNjQwMCwwMDAwMDAwMDFmNzUzNjA3OTUxNjY4YTY5NzQxMDY3NTFiYTE2MDQ3 playlist
#> 8 MTY3MDkzMjk1MSwwMDAwMDAwMDhiZTdkYThlOTBhNDQyNzA1N2I4NTg5NGFiYmJjMzBi playlist
#> 9 MTY3MDMxNTU4MiwwMDAwMDAwMDljMmM5NzA4ZDlmM2ZiOTljMGVlNzExOWY4ODZmYThi playlist
#>                                       uri
#> 1 spotify:playlist:37i9dQZF1DX1kfybUJZB6S
#> 2 spotify:playlist:37i9dQZF1DWSq1nQilUgG8
#> 3 spotify:playlist:37i9dQZF1DX9u7XXOp0l5L
#> 4 spotify:playlist:37i9dQZF1DX3S2ONCFIYHU
#> 5 spotify:playlist:37i9dQZF1DXdrln2UyZD7F
#> 6 spotify:playlist:37i9dQZF1DXadhN5awWZtL
#> 7 spotify:playlist:37i9dQZF1DX6MUrG3NBYtM
#> 8 spotify:playlist:37i9dQZF1DWTwCImwcYjDL
#> 9 spotify:playlist:37i9dQZF1DX4UF20jxFQBv
#>                                      external_urls.spotify owner.display_name
#> 1 https://open.spotify.com/playlist/37i9dQZF1DX1kfybUJZB6S            Spotify
#> 2 https://open.spotify.com/playlist/37i9dQZF1DWSq1nQilUgG8            Spotify
#> 3 https://open.spotify.com/playlist/37i9dQZF1DX9u7XXOp0l5L            Spotify
#> 4 https://open.spotify.com/playlist/37i9dQZF1DX3S2ONCFIYHU            Spotify
#> 5 https://open.spotify.com/playlist/37i9dQZF1DXdrln2UyZD7F            Spotify
#> 6 https://open.spotify.com/playlist/37i9dQZF1DXadhN5awWZtL            Spotify
#> 7 https://open.spotify.com/playlist/37i9dQZF1DX6MUrG3NBYtM            Spotify
#> 8 https://open.spotify.com/playlist/37i9dQZF1DWTwCImwcYjDL            Spotify
#> 9 https://open.spotify.com/playlist/37i9dQZF1DX4UF20jxFQBv            Spotify
#>                                 owner.href owner.id owner.type
#> 1 https://api.spotify.com/v1/users/spotify  spotify       user
#> 2 https://api.spotify.com/v1/users/spotify  spotify       user
#> 3 https://api.spotify.com/v1/users/spotify  spotify       user
#> 4 https://api.spotify.com/v1/users/spotify  spotify       user
#> 5 https://api.spotify.com/v1/users/spotify  spotify       user
#> 6 https://api.spotify.com/v1/users/spotify  spotify       user
#> 7 https://api.spotify.com/v1/users/spotify  spotify       user
#> 8 https://api.spotify.com/v1/users/spotify  spotify       user
#> 9 https://api.spotify.com/v1/users/spotify  spotify       user
#>              owner.uri           owner.external_urls.spotify
#> 1 spotify:user:spotify https://open.spotify.com/user/spotify
#> 2 spotify:user:spotify https://open.spotify.com/user/spotify
#> 3 spotify:user:spotify https://open.spotify.com/user/spotify
#> 4 spotify:user:spotify https://open.spotify.com/user/spotify
#> 5 spotify:user:spotify https://open.spotify.com/user/spotify
#> 6 spotify:user:spotify https://open.spotify.com/user/spotify
#> 7 spotify:user:spotify https://open.spotify.com/user/spotify
#> 8 spotify:user:spotify https://open.spotify.com/user/spotify
#> 9 spotify:user:spotify https://open.spotify.com/user/spotify
#>                                                          tracks.href
#> 1 https://api.spotify.com/v1/playlists/37i9dQZF1DX1kfybUJZB6S/tracks
#> 2 https://api.spotify.com/v1/playlists/37i9dQZF1DWSq1nQilUgG8/tracks
#> 3 https://api.spotify.com/v1/playlists/37i9dQZF1DX9u7XXOp0l5L/tracks
#> 4 https://api.spotify.com/v1/playlists/37i9dQZF1DX3S2ONCFIYHU/tracks
#> 5 https://api.spotify.com/v1/playlists/37i9dQZF1DXdrln2UyZD7F/tracks
#> 6 https://api.spotify.com/v1/playlists/37i9dQZF1DXadhN5awWZtL/tracks
#> 7 https://api.spotify.com/v1/playlists/37i9dQZF1DX6MUrG3NBYtM/tracks
#> 8 https://api.spotify.com/v1/playlists/37i9dQZF1DWTwCImwcYjDL/tracks
#> 9 https://api.spotify.com/v1/playlists/37i9dQZF1DX4UF20jxFQBv/tracks
#>   tracks.total
#> 1           50
#> 2           92
#> 3          100
#> 4           90
#> 5           60
#> 6           69
#> 7          100
#> 8           60
#> 9           75
# }