Search for artists by label

Get Spotify Catalog information about artists belonging to a given label.

get_label_artists(
  label = character(),
  market = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)

Arguments

label
Required.
String of label name to search for
For example: label = "brainfeeder".

market
Optional.
An ISO 3166-1 alpha-2 country code or the string "from_token".
If a country code is specified, only artists with content that is playable in that market is returned.
Note:
- If market is set to "from_token", and a valid access token is specified in the request header, only content playable in the country associated with the user account is returned.
- Users can view the country that is associated with their account in the account settings. A user must grant access to the user-read-private scope prior to when the access token is issued.

limit
Optional.
Maximum number of results to return.
Default: 20
Minimum: 1
Maximum: 50

offset
Optional.
The index of the first result to return.
Default: 0 (the first result).
Maximum offset (including limit): 10,000.
Use with limit to get the next page of search results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

Value

A data frame with the label information of the artists.

Examples

# \donttest{
get_label_artists('brainfeeder')

#> $q
#> [1] "label:\"brainfeeder\""
#> 
#> $type
#> [1] "artist"
#> 
#> $market
#> NULL
#> 
#> $limit
#> [1] 20
#> 
#> $offset
#> [1] 0
#> 
#> $access_token
#> [1] "BQBR9FchXGlS6UPC6B0Rpw_TnQGWqviwXBWvA4oeAxqRUe5GzDXieicC6Tr8aO30BCaryE40gOJh5wUcIew"
#> 

#> # A tibble: 20 x 12
#>    genres  href    id     images name  popularity type  uri    external_urls.sp~
#>    <list>  <chr>   <chr>  <list> <chr>      <int> <chr> <chr>  <chr>            
#>  1 <chr [~ https:~ 4frXp~ <df [~ Thun~         72 arti~ spoti~ https://open.spo~
#>  2 <chr [~ https:~ 6EHS9~ <df [~ Sala~         40 arti~ spoti~ https://open.spo~
#>  3 <chr [~ https:~ 1ybIN~ <df [~ Smino         73 arti~ spoti~ https://open.spo~
#>  4 <chr [~ https:~ 46Ce0~ <df [~ Lapa~         46 arti~ spoti~ https://open.spo~
#>  5 <chr [~ https:~ 29XOe~ <df [~ Flyi~         66 arti~ spoti~ https://open.spo~
#>  6 <chr [~ https:~ 2L2un~ <df [~ Teebs         49 arti~ spoti~ https://open.spo~
#>  7 <chr [~ https:~ 3Y3xI~ <df [~ Kenn~         71 arti~ spoti~ https://open.spo~
#>  8 <chr [~ https:~ 1tP3R~ <df [~ Arth~         50 arti~ spoti~ https://open.spo~
#>  9 <chr [~ https:~ 43Jlw~ <df [~ Hiat~         62 arti~ spoti~ https://open.spo~
#> 10 <chr [~ https:~ 7LCDn~ <df [~ Iglo~         46 arti~ spoti~ https://open.spo~
#> 11 <chr [~ https:~ 6HQYn~ <df [~ Kama~         61 arti~ spoti~ https://open.spo~
#> 12 <chr [~ https:~ 1Ma3p~ <df [~ Ross~         53 arti~ spoti~ https://open.spo~
#> 13 <chr [~ https:~ 5cM1P~ <df [~ Robe~         66 arti~ spoti~ https://open.spo~
#> 14 <chr [~ https:~ 3VwKS~ <df [~ TOKi~         60 arti~ spoti~ https://open.spo~
#> 15 <chr [~ https:~ 1PmVy~ <df [~ Lorn          58 arti~ spoti~ https://open.spo~
#> 16 <chr [~ https:~ 0NcPK~ <df [~ Guap~         66 arti~ spoti~ https://open.spo~
#> 17 <chr [~ https:~ 2Njfa~ <df [~ Aziz~         59 arti~ spoti~ https://open.spo~
#> 18 <chr [~ https:~ 6jlWj~ <df [~ Emil~         56 arti~ spoti~ https://open.spo~
#> 19 <chr [~ https:~ 6uIst~ <df [~ Loui~         48 arti~ spoti~ https://open.spo~
#> 20 <chr [~ https:~ 65dGL~ <df [~ BADB~         64 arti~ spoti~ https://open.spo~
#> # ... with 3 more variables: followers.href <lgl>, followers.total <int>,
#> #   label <chr>

# }