Get New Releases

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country
Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases.
See https://developer.spotify.com/documentation/web-api/reference/#category-browse for more information.

Examples

# \donttest{
## Get new Swedish music
get_new_releases(country = 'SE', limit = 5)

#>   album_type
#> 1     single
#> 2     single
#> 3     single
#> 4      album
#> 5     single
#>                                                                                                                                                                                                                                                                                                                                                                                                            artists
#> 1 https://api.spotify.com/v1/artists/1eX2QZGKv4RFxpDKUGxgm0, https://api.spotify.com/v1/artists/3H7Ez7cwaYw4L3ELy4v3Lc, 1eX2QZGKv4RFxpDKUGxgm0, 3H7Ez7cwaYw4L3ELy4v3Lc, Aden x Asme, Håkan Hellström, artist, artist, spotify:artist:1eX2QZGKv4RFxpDKUGxgm0, spotify:artist:3H7Ez7cwaYw4L3ELy4v3Lc, https://open.spotify.com/artist/1eX2QZGKv4RFxpDKUGxgm0, https://open.spotify.com/artist/3H7Ez7cwaYw4L3ELy4v3Lc
#> 2                                                                                                                                                                                                                  https://api.spotify.com/v1/artists/62QZPjYQMoo5g56FP9Webq, 62QZPjYQMoo5g56FP9Webq, Laleh, artist, spotify:artist:62QZPjYQMoo5g56FP9Webq, https://open.spotify.com/artist/62QZPjYQMoo5g56FP9Webq
#> 3         https://api.spotify.com/v1/artists/7vk5e3vY1uw9plTHJAMwjN, https://api.spotify.com/v1/artists/5rGrDvrLOV2VV8SCFVGWlj, 7vk5e3vY1uw9plTHJAMwjN, 5rGrDvrLOV2VV8SCFVGWlj, Alan Walker, Imanbek, artist, artist, spotify:artist:7vk5e3vY1uw9plTHJAMwjN, spotify:artist:5rGrDvrLOV2VV8SCFVGWlj, https://open.spotify.com/artist/7vk5e3vY1uw9plTHJAMwjN, https://open.spotify.com/artist/5rGrDvrLOV2VV8SCFVGWlj
#> 4                                                                                                                                                                                                       https://api.spotify.com/v1/artists/4C86wrjOUfXQDlt6A4oJOC, 4C86wrjOUfXQDlt6A4oJOC, Daniel Adams-Ray, artist, spotify:artist:4C86wrjOUfXQDlt6A4oJOC, https://open.spotify.com/artist/4C86wrjOUfXQDlt6A4oJOC
#> 5                                                                                                                                                                                                             https://api.spotify.com/v1/artists/4zGH6gaO4NohcfCXZ31bQG, 4zGH6gaO4NohcfCXZ31bQG, Eric Saade, artist, spotify:artist:4zGH6gaO4NohcfCXZ31bQG, https://open.spotify.com/artist/4zGH6gaO4NohcfCXZ31bQG
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 AM, AZ, BY, GE, KG, KZ, MD, RU, UA, UZ
#> 4 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 5 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                       href
#> 1 https://api.spotify.com/v1/albums/5G8bQMGvwxtTIJGuni3jzL
#> 2 https://api.spotify.com/v1/albums/2hxKPuoKlnH6ZMInQ3Xg9x
#> 3 https://api.spotify.com/v1/albums/7ihIr7mrNqtqqqSUG5Ts7I
#> 4 https://api.spotify.com/v1/albums/0XwIMEBisaLYj4SWj6wjjS
#> 5 https://api.spotify.com/v1/albums/2yWlDwkU9Es8E5QhNxuZzf
#>                       id
#> 1 5G8bQMGvwxtTIJGuni3jzL
#> 2 2hxKPuoKlnH6ZMInQ3Xg9x
#> 3 7ihIr7mrNqtqqqSUG5Ts7I
#> 4 0XwIMEBisaLYj4SWj6wjjS
#> 5 2yWlDwkU9Es8E5QhNxuZzf
#>                                                                                                                                                                                                                             images
#> 1 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e042c88090a1f831deb85803, https://i.scdn.co/image/ab67616d00001e02e042c88090a1f831deb85803, https://i.scdn.co/image/ab67616d00004851e042c88090a1f831deb85803, 640, 300, 64
#> 2 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a9474c447d0782a6342253d0, https://i.scdn.co/image/ab67616d00001e02a9474c447d0782a6342253d0, https://i.scdn.co/image/ab67616d00004851a9474c447d0782a6342253d0, 640, 300, 64
#> 3 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735ad3d2beded6316a3668dd07, https://i.scdn.co/image/ab67616d00001e025ad3d2beded6316a3668dd07, https://i.scdn.co/image/ab67616d000048515ad3d2beded6316a3668dd07, 640, 300, 64
#> 4 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27331c4e5b5afa4caae278f5170, https://i.scdn.co/image/ab67616d00001e0231c4e5b5afa4caae278f5170, https://i.scdn.co/image/ab67616d0000485131c4e5b5afa4caae278f5170, 640, 300, 64
#> 5 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27302b7fae77367b8750572cdab, https://i.scdn.co/image/ab67616d00001e0202b7fae77367b8750572cdab, https://i.scdn.co/image/ab67616d0000485102b7fae77367b8750572cdab, 640, 300, 64
#>                               name release_date release_date_precision
#> 1 Känn ingen sorg för mig Göteborg   2021-06-11                    day
#> 2                           Change   2021-06-11                    day
#> 3                     Sweet Dreams   2021-06-11                    day
#> 4        Döda hjärtan kan slå igen   2021-06-11                    day
#> 5                   Like U Used To   2021-06-11                    day
#>   total_tracks  type                                  uri
#> 1            1 album spotify:album:5G8bQMGvwxtTIJGuni3jzL
#> 2            1 album spotify:album:2hxKPuoKlnH6ZMInQ3Xg9x
#> 3            1 album spotify:album:7ihIr7mrNqtqqqSUG5Ts7I
#> 4           12 album spotify:album:0XwIMEBisaLYj4SWj6wjjS
#> 5            3 album spotify:album:2yWlDwkU9Es8E5QhNxuZzf
#>                                   external_urls.spotify
#> 1 https://open.spotify.com/album/5G8bQMGvwxtTIJGuni3jzL
#> 2 https://open.spotify.com/album/2hxKPuoKlnH6ZMInQ3Xg9x
#> 3 https://open.spotify.com/album/7ihIr7mrNqtqqqSUG5Ts7I
#> 4 https://open.spotify.com/album/0XwIMEBisaLYj4SWj6wjjS
#> 5 https://open.spotify.com/album/2yWlDwkU9Es8E5QhNxuZzf

# }