Get new releases

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country

Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit

Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset

Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info

Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases.

See https://developer.spotify.com/documentation/web-api/reference/#category-browse for more information.

Examples

# \donttest{
## Get new Swedish music
get_new_releases(country = 'SE', limit = 5)
#>   album_type
#> 1     single
#> 2     single
#> 3     single
#> 4     single
#> 5     single
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     artists
#> 1                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/60d24wfXkVzDSfLS6hyCjZ, https://api.spotify.com/v1/artists/164Uj4eKjl6zTBKfJLFKKK, 60d24wfXkVzDSfLS6hyCjZ, 164Uj4eKjl6zTBKfJLFKKK, Martin Garrix, JVKE, artist, artist, spotify:artist:60d24wfXkVzDSfLS6hyCjZ, spotify:artist:164Uj4eKjl6zTBKfJLFKKK, https://open.spotify.com/artist/60d24wfXkVzDSfLS6hyCjZ, https://open.spotify.com/artist/164Uj4eKjl6zTBKfJLFKKK
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                    https://api.spotify.com/v1/artists/2W8yFh0Ga6Yf3jiayVxwkE, 2W8yFh0Ga6Yf3jiayVxwkE, Dove Cameron, artist, spotify:artist:2W8yFh0Ga6Yf3jiayVxwkE, https://open.spotify.com/artist/2W8yFh0Ga6Yf3jiayVxwkE
#> 3 https://api.spotify.com/v1/artists/7f5Zgnp2spUuuzKplmRkt7, https://api.spotify.com/v1/artists/67MNhiAICFY6Pwc2YxCO0K, https://api.spotify.com/v1/artists/3NPpFNZtSTHheNBaWC82rB, 7f5Zgnp2spUuuzKplmRkt7, 67MNhiAICFY6Pwc2YxCO0K, 3NPpFNZtSTHheNBaWC82rB, Lost Frequencies, Elley Duhé, X Ambassadors, artist, artist, artist, spotify:artist:7f5Zgnp2spUuuzKplmRkt7, spotify:artist:67MNhiAICFY6Pwc2YxCO0K, spotify:artist:3NPpFNZtSTHheNBaWC82rB, https://open.spotify.com/artist/7f5Zgnp2spUuuzKplmRkt7, https://open.spotify.com/artist/67MNhiAICFY6Pwc2YxCO0K, https://open.spotify.com/artist/3NPpFNZtSTHheNBaWC82rB
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                           https://api.spotify.com/v1/artists/1rKFeRryEci6cxNkdvHzNr, 1rKFeRryEci6cxNkdvHzNr, Darin, artist, spotify:artist:1rKFeRryEci6cxNkdvHzNr, https://open.spotify.com/artist/1rKFeRryEci6cxNkdvHzNr
#> 5                                                                                                                                                                                                                         https://api.spotify.com/v1/artists/2Dor6diK1zw9BEluKBOdoA, https://api.spotify.com/v1/artists/6eWjyO1r0QlvsIuIIcgGxI, 2Dor6diK1zw9BEluKBOdoA, 6eWjyO1r0QlvsIuIIcgGxI, 23, C.Gambino, artist, artist, spotify:artist:2Dor6diK1zw9BEluKBOdoA, spotify:artist:6eWjyO1r0QlvsIuIIcgGxI, https://open.spotify.com/artist/2Dor6diK1zw9BEluKBOdoA, https://open.spotify.com/artist/6eWjyO1r0QlvsIuIIcgGxI
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#> 3 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         SE
#> 5 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                       href
#> 1 https://api.spotify.com/v1/albums/1D8XFqGY27IpYFAKB61h8v
#> 2 https://api.spotify.com/v1/albums/5ieBMOA3cYk5kofLMCSnMv
#> 3 https://api.spotify.com/v1/albums/1wntuGFGZSdkDL3pOTcdpr
#> 4 https://api.spotify.com/v1/albums/0SfWzA3ntIDxTDhGulOxGm
#> 5 https://api.spotify.com/v1/albums/3gfzgUuXhBJC2KMjvWgGQ7
#>                       id
#> 1 1D8XFqGY27IpYFAKB61h8v
#> 2 5ieBMOA3cYk5kofLMCSnMv
#> 3 1wntuGFGZSdkDL3pOTcdpr
#> 4 0SfWzA3ntIDxTDhGulOxGm
#> 5 3gfzgUuXhBJC2KMjvWgGQ7
#>                                                                                                                                                                                                                             images
#> 1 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732aa989be7ba10bb7beee57a7, https://i.scdn.co/image/ab67616d00001e022aa989be7ba10bb7beee57a7, https://i.scdn.co/image/ab67616d000048512aa989be7ba10bb7beee57a7, 640, 300, 64
#> 2 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735d0f9fb963040e3280885e8f, https://i.scdn.co/image/ab67616d00001e025d0f9fb963040e3280885e8f, https://i.scdn.co/image/ab67616d000048515d0f9fb963040e3280885e8f, 640, 300, 64
#> 3 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737a1b61448054bd2f97406822, https://i.scdn.co/image/ab67616d00001e027a1b61448054bd2f97406822, https://i.scdn.co/image/ab67616d000048517a1b61448054bd2f97406822, 640, 300, 64
#> 4 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736e9a63d82258a824b7b27dfd, https://i.scdn.co/image/ab67616d00001e026e9a63d82258a824b7b27dfd, https://i.scdn.co/image/ab67616d000048516e9a63d82258a824b7b27dfd, 640, 300, 64
#> 5 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737b8f93913bffccef6c72c322, https://i.scdn.co/image/ab67616d00001e027b8f93913bffccef6c72c322, https://i.scdn.co/image/ab67616d000048517b8f93913bffccef6c72c322, 640, 300, 64
#>               name release_date release_date_precision total_tracks  type
#> 1             Hero   2022-12-08                    day            1 album
#> 2     Girl Like Me   2022-12-09                    day            1 album
#> 3      Back To You   2022-12-09                    day            1 album
#> 4 My Purple Clouds   2022-12-09                    day            5 album
#> 5            M.O.B   2022-12-09                    day            1 album
#>                                    uri
#> 1 spotify:album:1D8XFqGY27IpYFAKB61h8v
#> 2 spotify:album:5ieBMOA3cYk5kofLMCSnMv
#> 3 spotify:album:1wntuGFGZSdkDL3pOTcdpr
#> 4 spotify:album:0SfWzA3ntIDxTDhGulOxGm
#> 5 spotify:album:3gfzgUuXhBJC2KMjvWgGQ7
#>                                   external_urls.spotify
#> 1 https://open.spotify.com/album/1D8XFqGY27IpYFAKB61h8v
#> 2 https://open.spotify.com/album/5ieBMOA3cYk5kofLMCSnMv
#> 3 https://open.spotify.com/album/1wntuGFGZSdkDL3pOTcdpr
#> 4 https://open.spotify.com/album/0SfWzA3ntIDxTDhGulOxGm
#> 5 https://open.spotify.com/album/3gfzgUuXhBJC2KMjvWgGQ7
# }