Get new releases

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country

Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit

Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset

Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info

Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases.

See https://developer.spotify.com/documentation/web-api/reference/#category-browse for more information.

Examples

if (FALSE) {
## Get new Swedish music
get_new_releases(country = 'SE', limit = 5)
}