Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country
Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
## Get new Swedish music
get_new_releases(country = 'SE')

#>    album_type
#> 1      single
#> 2      single
#> 3      single
#> 4       album
#> 5      single
#> 6      single
#> 7      single
#> 8      single
#> 9      single
#> 10     single
#> 11     single
#> 12     single
#> 13     single
#> 14     single
#> 15     single
#> 16     single
#> 17      album
#> 18     single
#> 19     single
#> 20     single
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      artists
#> 1                                                                                                                                                                                           https://api.spotify.com/v1/artists/1eX2QZGKv4RFxpDKUGxgm0, https://api.spotify.com/v1/artists/3H7Ez7cwaYw4L3ELy4v3Lc, 1eX2QZGKv4RFxpDKUGxgm0, 3H7Ez7cwaYw4L3ELy4v3Lc, Aden x Asme, Håkan Hellström, artist, artist, spotify:artist:1eX2QZGKv4RFxpDKUGxgm0, spotify:artist:3H7Ez7cwaYw4L3ELy4v3Lc, https://open.spotify.com/artist/1eX2QZGKv4RFxpDKUGxgm0, https://open.spotify.com/artist/3H7Ez7cwaYw4L3ELy4v3Lc
#> 2                                                                                                                                                                                                                                                                                                                                                                                                            https://api.spotify.com/v1/artists/62QZPjYQMoo5g56FP9Webq, 62QZPjYQMoo5g56FP9Webq, Laleh, artist, spotify:artist:62QZPjYQMoo5g56FP9Webq, https://open.spotify.com/artist/62QZPjYQMoo5g56FP9Webq
#> 3                                                                                                                                                                                                   https://api.spotify.com/v1/artists/7vk5e3vY1uw9plTHJAMwjN, https://api.spotify.com/v1/artists/5rGrDvrLOV2VV8SCFVGWlj, 7vk5e3vY1uw9plTHJAMwjN, 5rGrDvrLOV2VV8SCFVGWlj, Alan Walker, Imanbek, artist, artist, spotify:artist:7vk5e3vY1uw9plTHJAMwjN, spotify:artist:5rGrDvrLOV2VV8SCFVGWlj, https://open.spotify.com/artist/7vk5e3vY1uw9plTHJAMwjN, https://open.spotify.com/artist/5rGrDvrLOV2VV8SCFVGWlj
#> 4                                                                                                                                                                                                                                                                                                                                                                                                 https://api.spotify.com/v1/artists/4C86wrjOUfXQDlt6A4oJOC, 4C86wrjOUfXQDlt6A4oJOC, Daniel Adams-Ray, artist, spotify:artist:4C86wrjOUfXQDlt6A4oJOC, https://open.spotify.com/artist/4C86wrjOUfXQDlt6A4oJOC
#> 5                                                                                                                                                                                                                                                                                                                                                                                                       https://api.spotify.com/v1/artists/4zGH6gaO4NohcfCXZ31bQG, 4zGH6gaO4NohcfCXZ31bQG, Eric Saade, artist, spotify:artist:4zGH6gaO4NohcfCXZ31bQG, https://open.spotify.com/artist/4zGH6gaO4NohcfCXZ31bQG
#> 6                                                                                                                                                                                                                                                                                                                                                                                                       https://api.spotify.com/v1/artists/1J0C99LuAwjCC6IQOmJEDE, 1J0C99LuAwjCC6IQOmJEDE, Alex Järvi, artist, spotify:artist:1J0C99LuAwjCC6IQOmJEDE, https://open.spotify.com/artist/1J0C99LuAwjCC6IQOmJEDE
#> 7                                                                                                                                                                                                         https://api.spotify.com/v1/artists/0H39MdGGX6dbnnQPt6NQkZ, https://api.spotify.com/v1/artists/7tYKF4w9nC0nq9CsPZTHyP, 0H39MdGGX6dbnnQPt6NQkZ, 7tYKF4w9nC0nq9CsPZTHyP, SAINt JHN, SZA, artist, artist, spotify:artist:0H39MdGGX6dbnnQPt6NQkZ, spotify:artist:7tYKF4w9nC0nq9CsPZTHyP, https://open.spotify.com/artist/0H39MdGGX6dbnnQPt6NQkZ, https://open.spotify.com/artist/7tYKF4w9nC0nq9CsPZTHyP
#> 8                                                                                                                                                                                                                                                                                                                                                                                                  https://api.spotify.com/v1/artists/3wjsrpfO6odEphTZWx45RQ, 3wjsrpfO6odEphTZWx45RQ, Hearts & Colors, artist, spotify:artist:3wjsrpfO6odEphTZWx45RQ, https://open.spotify.com/artist/3wjsrpfO6odEphTZWx45RQ
#> 9                                                                                                                                                                                                                                                                                                                                                                                                       https://api.spotify.com/v1/artists/6GpcBKNmZDIrRzYkPJu7Wd, 6GpcBKNmZDIrRzYkPJu7Wd, flora cash, artist, spotify:artist:6GpcBKNmZDIrRzYkPJu7Wd, https://open.spotify.com/artist/6GpcBKNmZDIrRzYkPJu7Wd
#> 10                                                                                                                                                                                                                                                                                                                                                                                                             https://api.spotify.com/v1/artists/4ErLqbv0gQ49ED1AKtvrBr, 4ErLqbv0gQ49ED1AKtvrBr, HON, artist, spotify:artist:4ErLqbv0gQ49ED1AKtvrBr, https://open.spotify.com/artist/4ErLqbv0gQ49ED1AKtvrBr
#> 11                                                                                                                                                                                                                                                                                                                                                                                                         https://api.spotify.com/v1/artists/3Rjud126EhtcSDte9lORR7, 3Rjud126EhtcSDte9lORR7, Movits!, artist, spotify:artist:3Rjud126EhtcSDte9lORR7, https://open.spotify.com/artist/3Rjud126EhtcSDte9lORR7
#> 12                                                                                                                                                                                                                                                                                                                                                                                                 https://api.spotify.com/v1/artists/3WnmyxLdTHKN83h75tcb8P, 3WnmyxLdTHKN83h75tcb8P, Anis Don Demina, artist, spotify:artist:3WnmyxLdTHKN83h75tcb8P, https://open.spotify.com/artist/3WnmyxLdTHKN83h75tcb8P
#> 13                                                                                                                                                                                                                                                                                                                                                                                                  https://api.spotify.com/v1/artists/6vXTefBL93Dj5IqAWq6OTv, 6vXTefBL93Dj5IqAWq6OTv, French Montana, artist, spotify:artist:6vXTefBL93Dj5IqAWq6OTv, https://open.spotify.com/artist/6vXTefBL93Dj5IqAWq6OTv
#> 14                                                                                                                                                                                                                                                                                                                                                                                                        https://api.spotify.com/v1/artists/5Ys0Puc1F5jOw7EpmhMSTK, 5Ys0Puc1F5jOw7EpmhMSTK, Sam Ezeh, artist, spotify:artist:5Ys0Puc1F5jOw7EpmhMSTK, https://open.spotify.com/artist/5Ys0Puc1F5jOw7EpmhMSTK
#> 15                                                                                                                                                                                                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/6xrCU6zdcSTsG2hLrojpmI, 6xrCU6zdcSTsG2hLrojpmI, José González, artist, spotify:artist:6xrCU6zdcSTsG2hLrojpmI, https://open.spotify.com/artist/6xrCU6zdcSTsG2hLrojpmI
#> 16                                                                                                                                                                                                     https://api.spotify.com/v1/artists/3oKLD3PuOc20nNDoSCKDY7, https://api.spotify.com/v1/artists/15k33gArN2NU6cEYyQC3xZ, 3oKLD3PuOc20nNDoSCKDY7, 15k33gArN2NU6cEYyQC3xZ, Niello, Lisa Ajax, artist, artist, spotify:artist:3oKLD3PuOc20nNDoSCKDY7, spotify:artist:15k33gArN2NU6cEYyQC3xZ, https://open.spotify.com/artist/3oKLD3PuOc20nNDoSCKDY7, https://open.spotify.com/artist/15k33gArN2NU6cEYyQC3xZ
#> 17                                                                                                                                                                                                                                                                                                                                                                                                        https://api.spotify.com/v1/artists/3gbaHBYDy62irTs0ZsHNmi, 3gbaHBYDy62irTs0ZsHNmi, Wiktoria, artist, spotify:artist:3gbaHBYDy62irTs0ZsHNmi, https://open.spotify.com/artist/3gbaHBYDy62irTs0ZsHNmi
#> 18 https://api.spotify.com/v1/artists/087cmrx63FD1wMFuYBN9gt, https://api.spotify.com/v1/artists/2u9ZIMtCrLs7ya2HfIpLKZ, https://api.spotify.com/v1/artists/62oPQvptWCPp2UCsqYwcIf, 087cmrx63FD1wMFuYBN9gt, 2u9ZIMtCrLs7ya2HfIpLKZ, 62oPQvptWCPp2UCsqYwcIf, Mona Masrour, Jelassi, L1NA, artist, artist, artist, spotify:artist:087cmrx63FD1wMFuYBN9gt, spotify:artist:2u9ZIMtCrLs7ya2HfIpLKZ, spotify:artist:62oPQvptWCPp2UCsqYwcIf, https://open.spotify.com/artist/087cmrx63FD1wMFuYBN9gt, https://open.spotify.com/artist/2u9ZIMtCrLs7ya2HfIpLKZ, https://open.spotify.com/artist/62oPQvptWCPp2UCsqYwcIf
#> 19                                                                                                                                                                                        https://api.spotify.com/v1/artists/6rgEPiKjowlMKZC1DF6W75, https://api.spotify.com/v1/artists/15xvsJMf8phaNa1LYvL9Qv, 6rgEPiKjowlMKZC1DF6W75, 15xvsJMf8phaNa1LYvL9Qv, Rebecca & Fiona, John Dahlbäck, artist, artist, spotify:artist:6rgEPiKjowlMKZC1DF6W75, spotify:artist:15xvsJMf8phaNa1LYvL9Qv, https://open.spotify.com/artist/6rgEPiKjowlMKZC1DF6W75, https://open.spotify.com/artist/15xvsJMf8phaNa1LYvL9Qv
#> 20                                                                                                                                                                                                                                                                                                                                                                                                       https://api.spotify.com/v1/artists/3GtCKJA4SkOxN65MWjm82R, 3GtCKJA4SkOxN65MWjm82R, Hot Shade, artist, spotify:artist:3GtCKJA4SkOxN65MWjm82R, https://open.spotify.com/artist/3GtCKJA4SkOxN65MWjm82R
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         available_markets
#> 1  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 2  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  AM, AZ, BY, GE, KG, KZ, MD, RU, UA, UZ
#> 4  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 5  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 6  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 7  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 8  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 9                      AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DM, DO, DZ, EC, EE, EG, ES, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 10 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 11         AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 12 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 13 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 14 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     AE, AM, AZ, BD, BH, BN, BT, HK, ID, IN, JO, JP, KG, KH, KR, KW, LA, LK, MN, MO, MV, MY, NP, OM, PH, PK, QA, SA, SG, TH, TW, UZ, VN
#> 16 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 17 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 18 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 19 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 20 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                        href
#> 1  https://api.spotify.com/v1/albums/5G8bQMGvwxtTIJGuni3jzL
#> 2  https://api.spotify.com/v1/albums/2hxKPuoKlnH6ZMInQ3Xg9x
#> 3  https://api.spotify.com/v1/albums/7ihIr7mrNqtqqqSUG5Ts7I
#> 4  https://api.spotify.com/v1/albums/0XwIMEBisaLYj4SWj6wjjS
#> 5  https://api.spotify.com/v1/albums/2yWlDwkU9Es8E5QhNxuZzf
#> 6  https://api.spotify.com/v1/albums/7H7RKFP7bX25Y0hH7b8otX
#> 7  https://api.spotify.com/v1/albums/0uXCtaJMMAAgaB1AavNLSh
#> 8  https://api.spotify.com/v1/albums/44bRq4YcqgX5EtK8NY88lB
#> 9  https://api.spotify.com/v1/albums/5GGnarWGlNGFEGOo5ztQH2
#> 10 https://api.spotify.com/v1/albums/2BwwgpxslWjsJDd3mU3ha9
#> 11 https://api.spotify.com/v1/albums/4p5G46BCdT6STXz5fYH7xM
#> 12 https://api.spotify.com/v1/albums/2bAxn4ls6OmXfdcsbUNFhI
#> 13 https://api.spotify.com/v1/albums/58zbc1nYIQmLd4rxnH1Blt
#> 14 https://api.spotify.com/v1/albums/43skczn9Io3vAXh97760qM
#> 15 https://api.spotify.com/v1/albums/2vAB6q0kGH2OIU2mBaEqPr
#> 16 https://api.spotify.com/v1/albums/0y0auZt6Q8rpgWWItA4nFq
#> 17 https://api.spotify.com/v1/albums/4Qpyeo0oQZ69tjG2SqoJ2H
#> 18 https://api.spotify.com/v1/albums/3Iq3AmMNBAn1Stw3IcGudF
#> 19 https://api.spotify.com/v1/albums/2wNDj0IjDtseiKPvaM1lp5
#> 20 https://api.spotify.com/v1/albums/4AF2p0HZ3Dxy5SSMDHtBRj
#>                        id
#> 1  5G8bQMGvwxtTIJGuni3jzL
#> 2  2hxKPuoKlnH6ZMInQ3Xg9x
#> 3  7ihIr7mrNqtqqqSUG5Ts7I
#> 4  0XwIMEBisaLYj4SWj6wjjS
#> 5  2yWlDwkU9Es8E5QhNxuZzf
#> 6  7H7RKFP7bX25Y0hH7b8otX
#> 7  0uXCtaJMMAAgaB1AavNLSh
#> 8  44bRq4YcqgX5EtK8NY88lB
#> 9  5GGnarWGlNGFEGOo5ztQH2
#> 10 2BwwgpxslWjsJDd3mU3ha9
#> 11 4p5G46BCdT6STXz5fYH7xM
#> 12 2bAxn4ls6OmXfdcsbUNFhI
#> 13 58zbc1nYIQmLd4rxnH1Blt
#> 14 43skczn9Io3vAXh97760qM
#> 15 2vAB6q0kGH2OIU2mBaEqPr
#> 16 0y0auZt6Q8rpgWWItA4nFq
#> 17 4Qpyeo0oQZ69tjG2SqoJ2H
#> 18 3Iq3AmMNBAn1Stw3IcGudF
#> 19 2wNDj0IjDtseiKPvaM1lp5
#> 20 4AF2p0HZ3Dxy5SSMDHtBRj
#>                                                                                                                                                                                                                              images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e042c88090a1f831deb85803, https://i.scdn.co/image/ab67616d00001e02e042c88090a1f831deb85803, https://i.scdn.co/image/ab67616d00004851e042c88090a1f831deb85803, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a9474c447d0782a6342253d0, https://i.scdn.co/image/ab67616d00001e02a9474c447d0782a6342253d0, https://i.scdn.co/image/ab67616d00004851a9474c447d0782a6342253d0, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735ad3d2beded6316a3668dd07, https://i.scdn.co/image/ab67616d00001e025ad3d2beded6316a3668dd07, https://i.scdn.co/image/ab67616d000048515ad3d2beded6316a3668dd07, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27331c4e5b5afa4caae278f5170, https://i.scdn.co/image/ab67616d00001e0231c4e5b5afa4caae278f5170, https://i.scdn.co/image/ab67616d0000485131c4e5b5afa4caae278f5170, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27302b7fae77367b8750572cdab, https://i.scdn.co/image/ab67616d00001e0202b7fae77367b8750572cdab, https://i.scdn.co/image/ab67616d0000485102b7fae77367b8750572cdab, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a44ffe84ee29531636361e3f, https://i.scdn.co/image/ab67616d00001e02a44ffe84ee29531636361e3f, https://i.scdn.co/image/ab67616d00004851a44ffe84ee29531636361e3f, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d0126e53352bb15e3e232643, https://i.scdn.co/image/ab67616d00001e02d0126e53352bb15e3e232643, https://i.scdn.co/image/ab67616d00004851d0126e53352bb15e3e232643, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a2133d68d11a1eb8ff73a8e7, https://i.scdn.co/image/ab67616d00001e02a2133d68d11a1eb8ff73a8e7, https://i.scdn.co/image/ab67616d00004851a2133d68d11a1eb8ff73a8e7, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273740322e4a9648cedd5e440e2, https://i.scdn.co/image/ab67616d00001e02740322e4a9648cedd5e440e2, https://i.scdn.co/image/ab67616d00004851740322e4a9648cedd5e440e2, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c40f14ec130a1240f3fc895e, https://i.scdn.co/image/ab67616d00001e02c40f14ec130a1240f3fc895e, https://i.scdn.co/image/ab67616d00004851c40f14ec130a1240f3fc895e, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732534c9fa25ef5efbf449b27f, https://i.scdn.co/image/ab67616d00001e022534c9fa25ef5efbf449b27f, https://i.scdn.co/image/ab67616d000048512534c9fa25ef5efbf449b27f, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f76db709f2bf7a168fb75e4b, https://i.scdn.co/image/ab67616d00001e02f76db709f2bf7a168fb75e4b, https://i.scdn.co/image/ab67616d00004851f76db709f2bf7a168fb75e4b, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2730cbe63f4412808ec15a645a7, https://i.scdn.co/image/ab67616d00001e020cbe63f4412808ec15a645a7, https://i.scdn.co/image/ab67616d000048510cbe63f4412808ec15a645a7, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27381190ce304834cfdabead41f, https://i.scdn.co/image/ab67616d00001e0281190ce304834cfdabead41f, https://i.scdn.co/image/ab67616d0000485181190ce304834cfdabead41f, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273328ade64828ccdc3dc590a1d, https://i.scdn.co/image/ab67616d00001e02328ade64828ccdc3dc590a1d, https://i.scdn.co/image/ab67616d00004851328ade64828ccdc3dc590a1d, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732556ac6ba69f5659b815f6fb, https://i.scdn.co/image/ab67616d00001e022556ac6ba69f5659b815f6fb, https://i.scdn.co/image/ab67616d000048512556ac6ba69f5659b815f6fb, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273be4b6df6969ddbfa12eb39a6, https://i.scdn.co/image/ab67616d00001e02be4b6df6969ddbfa12eb39a6, https://i.scdn.co/image/ab67616d00004851be4b6df6969ddbfa12eb39a6, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27309fe4d58b2a5acee234f28f0, https://i.scdn.co/image/ab67616d00001e0209fe4d58b2a5acee234f28f0, https://i.scdn.co/image/ab67616d0000485109fe4d58b2a5acee234f28f0, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d25bbf8e90d374fca3af02aa, https://i.scdn.co/image/ab67616d00001e02d25bbf8e90d374fca3af02aa, https://i.scdn.co/image/ab67616d00004851d25bbf8e90d374fca3af02aa, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e080753c79a18c4d0c59fc0c, https://i.scdn.co/image/ab67616d00001e02e080753c79a18c4d0c59fc0c, https://i.scdn.co/image/ab67616d00004851e080753c79a18c4d0c59fc0c, 640, 300, 64
#>                                     name release_date release_date_precision
#> 1       Känn ingen sorg för mig Göteborg   2021-06-11                    day
#> 2                                 Change   2021-06-11                    day
#> 3                           Sweet Dreams   2021-06-11                    day
#> 4              Döda hjärtan kan slå igen   2021-06-11                    day
#> 5                         Like U Used To   2021-06-11                    day
#> 6                               Ruta ett   2021-06-11                    day
#> 7  Just For Me (Space Jam: A New Legacy)   2021-06-11                    day
#> 8                                Need Me   2021-06-11                    day
#> 9          Don’t You Look At Me That Way   2021-06-11                    day
#> 10                  När allt rinner över   2021-06-11                    day
#> 11                               Blixtar   2021-06-11                    day
#> 12                             Underbarn   2021-06-11                    day
#> 13                                FWMGAB   2021-06-11                    day
#> 14                                 Relay   2021-06-11                    day
#> 15                               Head On   2021-06-09                    day
#> 16                           Ingen annan   2021-06-11                    day
#> 17                               EXPOSED   2021-06-11                    day
#> 18                        2 STEG (REMIX)   2021-06-11                    day
#> 19                                Replay   2021-06-11                    day
#> 20                             Good Life   2021-06-11                    day
#>    total_tracks  type                                  uri
#> 1             1 album spotify:album:5G8bQMGvwxtTIJGuni3jzL
#> 2             1 album spotify:album:2hxKPuoKlnH6ZMInQ3Xg9x
#> 3             1 album spotify:album:7ihIr7mrNqtqqqSUG5Ts7I
#> 4            12 album spotify:album:0XwIMEBisaLYj4SWj6wjjS
#> 5             3 album spotify:album:2yWlDwkU9Es8E5QhNxuZzf
#> 6             1 album spotify:album:7H7RKFP7bX25Y0hH7b8otX
#> 7             1 album spotify:album:0uXCtaJMMAAgaB1AavNLSh
#> 8             1 album spotify:album:44bRq4YcqgX5EtK8NY88lB
#> 9             4 album spotify:album:5GGnarWGlNGFEGOo5ztQH2
#> 10            5 album spotify:album:2BwwgpxslWjsJDd3mU3ha9
#> 11            2 album spotify:album:4p5G46BCdT6STXz5fYH7xM
#> 12            5 album spotify:album:2bAxn4ls6OmXfdcsbUNFhI
#> 13            1 album spotify:album:58zbc1nYIQmLd4rxnH1Blt
#> 14            1 album spotify:album:43skczn9Io3vAXh97760qM
#> 15            1 album spotify:album:2vAB6q0kGH2OIU2mBaEqPr
#> 16            1 album spotify:album:0y0auZt6Q8rpgWWItA4nFq
#> 17           10 album spotify:album:4Qpyeo0oQZ69tjG2SqoJ2H
#> 18            1 album spotify:album:3Iq3AmMNBAn1Stw3IcGudF
#> 19            1 album spotify:album:2wNDj0IjDtseiKPvaM1lp5
#> 20            1 album spotify:album:4AF2p0HZ3Dxy5SSMDHtBRj
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/album/5G8bQMGvwxtTIJGuni3jzL
#> 2  https://open.spotify.com/album/2hxKPuoKlnH6ZMInQ3Xg9x
#> 3  https://open.spotify.com/album/7ihIr7mrNqtqqqSUG5Ts7I
#> 4  https://open.spotify.com/album/0XwIMEBisaLYj4SWj6wjjS
#> 5  https://open.spotify.com/album/2yWlDwkU9Es8E5QhNxuZzf
#> 6  https://open.spotify.com/album/7H7RKFP7bX25Y0hH7b8otX
#> 7  https://open.spotify.com/album/0uXCtaJMMAAgaB1AavNLSh
#> 8  https://open.spotify.com/album/44bRq4YcqgX5EtK8NY88lB
#> 9  https://open.spotify.com/album/5GGnarWGlNGFEGOo5ztQH2
#> 10 https://open.spotify.com/album/2BwwgpxslWjsJDd3mU3ha9
#> 11 https://open.spotify.com/album/4p5G46BCdT6STXz5fYH7xM
#> 12 https://open.spotify.com/album/2bAxn4ls6OmXfdcsbUNFhI
#> 13 https://open.spotify.com/album/58zbc1nYIQmLd4rxnH1Blt
#> 14 https://open.spotify.com/album/43skczn9Io3vAXh97760qM
#> 15 https://open.spotify.com/album/2vAB6q0kGH2OIU2mBaEqPr
#> 16 https://open.spotify.com/album/0y0auZt6Q8rpgWWItA4nFq
#> 17 https://open.spotify.com/album/4Qpyeo0oQZ69tjG2SqoJ2H
#> 18 https://open.spotify.com/album/3Iq3AmMNBAn1Stw3IcGudF
#> 19 https://open.spotify.com/album/2wNDj0IjDtseiKPvaM1lp5
#> 20 https://open.spotify.com/album/4AF2p0HZ3Dxy5SSMDHtBRj

# }