Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

get_new_releases(
  country = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token(),
  include_meta_info = FALSE
)

Arguments

country
Optional. A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

limit
Optional. The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset
Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_meta_info
Optional. Boolean indicating whether to include full result, with meta information such as "country", "offset", and "limit". Defaults to FALSE.

Value

Returns a data frame of results containing new releases. See https://developer.spotify.com/documentation/web-api/reference/users-profile/get-current-users-profile/ for more information.

Examples

# \donttest{
## Get new Swedish music
get_new_releases(country = 'SE')

#>    album_type
#> 1      single
#> 2      single
#> 3      single
#> 4      single
#> 5      single
#> 6      single
#> 7       album
#> 8      single
#> 9       album
#> 10     single
#> 11     single
#> 12      album
#> 13     single
#> 14     single
#> 15     single
#> 16      album
#> 17      album
#> 18     single
#> 19     single
#> 20     single
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               artists
#> 1                                                                                                                                                                                                            https://api.spotify.com/v1/artists/7vk5e3vY1uw9plTHJAMwjN, https://api.spotify.com/v1/artists/5rGrDvrLOV2VV8SCFVGWlj, 7vk5e3vY1uw9plTHJAMwjN, 5rGrDvrLOV2VV8SCFVGWlj, Alan Walker, Imanbek, artist, artist, spotify:artist:7vk5e3vY1uw9plTHJAMwjN, spotify:artist:5rGrDvrLOV2VV8SCFVGWlj, https://open.spotify.com/artist/7vk5e3vY1uw9plTHJAMwjN, https://open.spotify.com/artist/5rGrDvrLOV2VV8SCFVGWlj
#> 2                                                                                                                                                                                                                                                                                                                                                                                                              https://api.spotify.com/v1/artists/4IWBUUAFIplrNtaOHcJPRM, 4IWBUUAFIplrNtaOHcJPRM, James Arthur, artist, spotify:artist:4IWBUUAFIplrNtaOHcJPRM, https://open.spotify.com/artist/4IWBUUAFIplrNtaOHcJPRM
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                      https://api.spotify.com/v1/artists/5KKpBU5eC2tJDzf0wmlRp2, 5KKpBU5eC2tJDzf0wmlRp2, RAYE, artist, spotify:artist:5KKpBU5eC2tJDzf0wmlRp2, https://open.spotify.com/artist/5KKpBU5eC2tJDzf0wmlRp2
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/4npEfmQ6YuiwW1GpUmaq3F, 4npEfmQ6YuiwW1GpUmaq3F, Ava Max, artist, spotify:artist:4npEfmQ6YuiwW1GpUmaq3F, https://open.spotify.com/artist/4npEfmQ6YuiwW1GpUmaq3F
#> 5                                                                                                                                                                                                                                                                                                                                                                                                               https://api.spotify.com/v1/artists/2NjfBq1NflQcKSeiDooVjY, 2NjfBq1NflQcKSeiDooVjY, Tones And I, artist, spotify:artist:2NjfBq1NflQcKSeiDooVjY, https://open.spotify.com/artist/2NjfBq1NflQcKSeiDooVjY
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                 https://api.spotify.com/v1/artists/3GxKJzJK4LpsYGXQrw77wz, 3GxKJzJK4LpsYGXQrw77wz, Mimi Webb, artist, spotify:artist:3GxKJzJK4LpsYGXQrw77wz, https://open.spotify.com/artist/3GxKJzJK4LpsYGXQrw77wz
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                  https://api.spotify.com/v1/artists/04gDigrS5kc9YWfZHwBETP, 04gDigrS5kc9YWfZHwBETP, Maroon 5, artist, spotify:artist:04gDigrS5kc9YWfZHwBETP, https://open.spotify.com/artist/04gDigrS5kc9YWfZHwBETP
#> 8                                                                                                                                                                                                                                                                                                                                                                                                       https://api.spotify.com/v1/artists/181bsRPaVXVlUKXrxwZfHK, 181bsRPaVXVlUKXrxwZfHK, Megan Thee Stallion, artist, spotify:artist:181bsRPaVXVlUKXrxwZfHK, https://open.spotify.com/artist/181bsRPaVXVlUKXrxwZfHK
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                     https://api.spotify.com/v1/artists/6oMuImdp5ZcFhWP0ESe6mG, 6oMuImdp5ZcFhWP0ESe6mG, Migos, artist, spotify:artist:6oMuImdp5ZcFhWP0ESe6mG, https://open.spotify.com/artist/6oMuImdp5ZcFhWP0ESe6mG
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                 https://api.spotify.com/v1/artists/5cj0lLjcoR7YOSnhnX0Po5, 5cj0lLjcoR7YOSnhnX0Po5, Doja Cat, artist, spotify:artist:5cj0lLjcoR7YOSnhnX0Po5, https://open.spotify.com/artist/5cj0lLjcoR7YOSnhnX0Po5
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                    https://api.spotify.com/v1/artists/163tK9Wjr9P9DmM0AVK7lm, 163tK9Wjr9P9DmM0AVK7lm, Lorde, artist, spotify:artist:163tK9Wjr9P9DmM0AVK7lm, https://open.spotify.com/artist/163tK9Wjr9P9DmM0AVK7lm
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/6AgTAQt8XS6jRWi4sX7w49, 6AgTAQt8XS6jRWi4sX7w49, Polo G, artist, spotify:artist:6AgTAQt8XS6jRWi4sX7w49, https://open.spotify.com/artist/6AgTAQt8XS6jRWi4sX7w49
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                     https://api.spotify.com/v1/artists/68dW5BU6sdVNf099EylxEt, 68dW5BU6sdVNf099EylxEt, Hov1, artist, spotify:artist:68dW5BU6sdVNf099EylxEt, https://open.spotify.com/artist/68dW5BU6sdVNf099EylxEt
#> 14 https://api.spotify.com/v1/artists/7c0XG5cIJTrrAgEC3ULPiq, https://api.spotify.com/v1/artists/2LIk90788K0zvyj2JJVwkJ, https://api.spotify.com/v1/artists/6fWVd57NKTalqvmjRd2t8Z, 7c0XG5cIJTrrAgEC3ULPiq, 2LIk90788K0zvyj2JJVwkJ, 6fWVd57NKTalqvmjRd2t8Z, Ty Dolla $ign, Jack Harlow, 24kGoldn, artist, artist, artist, spotify:artist:7c0XG5cIJTrrAgEC3ULPiq, spotify:artist:2LIk90788K0zvyj2JJVwkJ, spotify:artist:6fWVd57NKTalqvmjRd2t8Z, https://open.spotify.com/artist/7c0XG5cIJTrrAgEC3ULPiq, https://open.spotify.com/artist/2LIk90788K0zvyj2JJVwkJ, https://open.spotify.com/artist/6fWVd57NKTalqvmjRd2t8Z
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                      https://api.spotify.com/v1/artists/5Wg2b4Mp42gicxEeDNawf7, 5Wg2b4Mp42gicxEeDNawf7, A7S, artist, spotify:artist:5Wg2b4Mp42gicxEeDNawf7, https://open.spotify.com/artist/5Wg2b4Mp42gicxEeDNawf7
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/4bZ30evEi5vMG1A3xdW8U2, 4bZ30evEi5vMG1A3xdW8U2, Fricky, artist, spotify:artist:4bZ30evEi5vMG1A3xdW8U2, https://open.spotify.com/artist/4bZ30evEi5vMG1A3xdW8U2
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                  https://api.spotify.com/v1/artists/59E9dYtwLTUAGuAppwFKAW, 59E9dYtwLTUAGuAppwFKAW, Cherrie, artist, spotify:artist:59E9dYtwLTUAGuAppwFKAW, https://open.spotify.com/artist/59E9dYtwLTUAGuAppwFKAW
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                   https://api.spotify.com/v1/artists/5A0Bu9azuFEnud3q7t0V2r, 5A0Bu9azuFEnud3q7t0V2r, Petter, artist, spotify:artist:5A0Bu9azuFEnud3q7t0V2r, https://open.spotify.com/artist/5A0Bu9azuFEnud3q7t0V2r
#> 19                                                                                                                                                                                                                                                                                                                                                                                                              https://api.spotify.com/v1/artists/757aE44tKEUQEqRuT6GnEB, 757aE44tKEUQEqRuT6GnEB, Roddy Ricch, artist, spotify:artist:757aE44tKEUQEqRuT6GnEB, https://open.spotify.com/artist/757aE44tKEUQEqRuT6GnEB
#> 20                                                                                                                                                                                                                                                                                                                                                                                                                 https://api.spotify.com/v1/artists/4bOG1sx3QHFbOUVLNmMpPe, 4bOG1sx3QHFbOUVLNmMpPe, Timbuktu, artist, spotify:artist:4bOG1sx3QHFbOUVLNmMpPe, https://open.spotify.com/artist/4bOG1sx3QHFbOUVLNmMpPe
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         available_markets
#> 1                                              AE, AU, BD, BG, BH, BN, BT, CY, DJ, EE, FI, FJ, FM, GR, HK, ID, IL, IN, JO, JP, KE, KH, KI, KM, KR, KW, LA, LB, LK, LT, LV, MG, MH, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UG, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 2  AE, AM, AU, AZ, BD, BG, BH, BN, BT, BY, CY, DJ, EE, FI, FJ, FM, GE, GR, HK, ID, IL, IN, JO, JP, KE, KG, KH, KI, KM, KR, KW, KZ, LA, LB, LK, LT, LV, MD, MG, MH, MN, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, RU, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UA, UG, UZ, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 3  TO, WS, FJ, KI, MH, NR, NZ, TV, FM, SB, VU, AU, PG, JP, KR, PW, TL, BN, HK, MN, MO, MY, PH, SG, TW, ID, KH, LA, TH, VN, BD, BT, KG, KZ, NP, IN, LK, MV, PK, UZ, AE, AM, AZ, GE, MU, OM, SC, BG, BH, BY, CY, DJ, EE, FI, GR, IL, JO, KE, KM, KW, LB, LT, LV, MD, MG, PS, QA, RO, RU, SA, TR, TZ, UA, UG, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 4  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 5          AD, AE, AG, AL, AM, AO, AR, AT, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 6  AE, AM, AU, AZ, BD, BG, BH, BN, BT, BY, CY, DJ, EE, FI, FJ, FM, GE, GR, HK, ID, IL, IN, JO, JP, KE, KG, KH, KI, KM, KR, KW, KZ, LA, LB, LK, LT, LV, MD, MG, MH, MN, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, RU, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UA, UG, UZ, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 7  AE, AM, AU, AZ, BD, BG, BH, BN, BT, BY, CY, DJ, EE, FI, FJ, FM, GE, GR, HK, ID, IL, IN, JO, JP, KE, KG, KH, KI, KM, KR, KW, KZ, LA, LB, LK, LT, LV, MD, MG, MH, MN, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, RU, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UA, UG, UZ, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 8  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 9  AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 10 AE, AM, AU, AZ, BD, BG, BH, BN, BT, BY, CY, DJ, EE, FI, FJ, FM, GE, GR, HK, ID, IL, IN, JO, JP, KE, KG, KH, KI, KM, KR, KW, KZ, LA, LB, LK, LT, LV, MD, MG, MH, MN, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, RU, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UA, UG, UZ, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 11                 AD, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 12 AE, AM, AU, AZ, BD, BG, BH, BN, BT, BY, CY, DJ, EE, FI, FJ, FM, GE, GR, HK, ID, IL, IN, JO, JP, KE, KG, KH, KI, KM, KR, KW, KZ, LA, LB, LK, LT, LV, MD, MG, MH, MN, MO, MU, MV, MY, NP, NR, NZ, OM, PG, PH, PK, PS, PW, QA, RO, RU, SA, SB, SC, SG, TH, TL, TO, TR, TV, TW, TZ, UA, UG, UZ, VN, VU, WS, AD, AL, AT, BA, BE, BI, BW, CH, CZ, DE, DK, EG, ES, FR, HR, HU, IT, LI, LS, LU, MC, ME, MK, MT, MW, MZ, NA, NL, NO, PL, RS, RW, SE, SI, SK, SM, SZ, XK, ZA, ZM, ZW, AO, BJ, CM, DZ, GA, GB, GQ, IE, MA, NE, NG, PT, TD, TN, BF, CI, GH, GM, GN, GW, IS, LR, ML, MR, SL, SN, ST, TG, CV, AR, BR, SR, UY, AG, BB, BO, BS, CA, CL, CW, DM, DO, GD, GY, HT, KN, LC, PY, TT, US, VC, CO, EC, JM, MX, PA, PE, BZ, CR, GT, HN, NI, SV
#> 13 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 CA, US
#> 15 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 16 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 17 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 18 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 19 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 20 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                        href
#> 1  https://api.spotify.com/v1/albums/5WV1czZI9hnO6REVS6yKHR
#> 2  https://api.spotify.com/v1/albums/1gHwbB7Db9yJ7kASKJFSnc
#> 3  https://api.spotify.com/v1/albums/4l78jaWeylIINTLaAdIHJy
#> 4  https://api.spotify.com/v1/albums/5W79aGcuoBYk0Mb2QL2Jcw
#> 5  https://api.spotify.com/v1/albums/3IbmFFjHMuAkhfk1zFDSow
#> 6  https://api.spotify.com/v1/albums/0xU5lh9KbThw5Y7b50Ilm8
#> 7  https://api.spotify.com/v1/albums/1pCA38N6MkLlthXtAOvZTU
#> 8  https://api.spotify.com/v1/albums/7gEGuvkVs79hY4tG7OYnzx
#> 9  https://api.spotify.com/v1/albums/3uyOwJu4r3yroAkFywNFM3
#> 10 https://api.spotify.com/v1/albums/7N3SiwnBMR1C8hcA4W1GtH
#> 11 https://api.spotify.com/v1/albums/0pAF4Jr6MPQ2u0vqG2VOjd
#> 12 https://api.spotify.com/v1/albums/7KSf6p0G0mZd7j60etVTwT
#> 13 https://api.spotify.com/v1/albums/4NQguSrcgRiDznaKoowh8u
#> 14 https://api.spotify.com/v1/albums/01RtA0qCjfGy0FY4ON23nD
#> 15 https://api.spotify.com/v1/albums/4VrGw3dww2O6U0hZTXOHae
#> 16 https://api.spotify.com/v1/albums/2v5eP5PoTRTy3yOxX5pfAi
#> 17 https://api.spotify.com/v1/albums/1UTctXDMnBfSheaCnpT7O5
#> 18 https://api.spotify.com/v1/albums/5Bw1AqtLIewl5hUQ1Lj8Rv
#> 19 https://api.spotify.com/v1/albums/2WFC0XUBROo4X2vxLwtB4u
#> 20 https://api.spotify.com/v1/albums/6690jqPTlPzzo3xoQwLSwh
#>                        id
#> 1  5WV1czZI9hnO6REVS6yKHR
#> 2  1gHwbB7Db9yJ7kASKJFSnc
#> 3  4l78jaWeylIINTLaAdIHJy
#> 4  5W79aGcuoBYk0Mb2QL2Jcw
#> 5  3IbmFFjHMuAkhfk1zFDSow
#> 6  0xU5lh9KbThw5Y7b50Ilm8
#> 7  1pCA38N6MkLlthXtAOvZTU
#> 8  7gEGuvkVs79hY4tG7OYnzx
#> 9  3uyOwJu4r3yroAkFywNFM3
#> 10 7N3SiwnBMR1C8hcA4W1GtH
#> 11 0pAF4Jr6MPQ2u0vqG2VOjd
#> 12 7KSf6p0G0mZd7j60etVTwT
#> 13 4NQguSrcgRiDznaKoowh8u
#> 14 01RtA0qCjfGy0FY4ON23nD
#> 15 4VrGw3dww2O6U0hZTXOHae
#> 16 2v5eP5PoTRTy3yOxX5pfAi
#> 17 1UTctXDMnBfSheaCnpT7O5
#> 18 5Bw1AqtLIewl5hUQ1Lj8Rv
#> 19 2WFC0XUBROo4X2vxLwtB4u
#> 20 6690jqPTlPzzo3xoQwLSwh
#>                                                                                                                                                                                                                              images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737327806c243a77cde4502d6a, https://i.scdn.co/image/ab67616d00001e027327806c243a77cde4502d6a, https://i.scdn.co/image/ab67616d000048517327806c243a77cde4502d6a, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c1dd95aa29782af7e152d0eb, https://i.scdn.co/image/ab67616d00001e02c1dd95aa29782af7e152d0eb, https://i.scdn.co/image/ab67616d00004851c1dd95aa29782af7e152d0eb, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27314713ec2e658f8cd66d0446f, https://i.scdn.co/image/ab67616d00001e0214713ec2e658f8cd66d0446f, https://i.scdn.co/image/ab67616d0000485114713ec2e658f8cd66d0446f, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27393fc5ac4fab909bf00efb02e, https://i.scdn.co/image/ab67616d00001e0293fc5ac4fab909bf00efb02e, https://i.scdn.co/image/ab67616d0000485193fc5ac4fab909bf00efb02e, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273532f46ac33cbb831767c5934, https://i.scdn.co/image/ab67616d00001e02532f46ac33cbb831767c5934, https://i.scdn.co/image/ab67616d00004851532f46ac33cbb831767c5934, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e910ee190e44404aafe9f886, https://i.scdn.co/image/ab67616d00001e02e910ee190e44404aafe9f886, https://i.scdn.co/image/ab67616d00004851e910ee190e44404aafe9f886, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2738dde90e4275716e423bf0ba2, https://i.scdn.co/image/ab67616d00001e028dde90e4275716e423bf0ba2, https://i.scdn.co/image/ab67616d000048518dde90e4275716e423bf0ba2, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273ac1c01aec658e90fcd23f857, https://i.scdn.co/image/ab67616d00001e02ac1c01aec658e90fcd23f857, https://i.scdn.co/image/ab67616d00004851ac1c01aec658e90fcd23f857, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2739590780c28b0bf5df966ee43, https://i.scdn.co/image/ab67616d00001e029590780c28b0bf5df966ee43, https://i.scdn.co/image/ab67616d000048519590780c28b0bf5df966ee43, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d948a70dd076a2a2a10a795b, https://i.scdn.co/image/ab67616d00001e02d948a70dd076a2a2a10a795b, https://i.scdn.co/image/ab67616d00004851d948a70dd076a2a2a10a795b, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d08ee7bfe4aaf52168247de7, https://i.scdn.co/image/ab67616d00001e02d08ee7bfe4aaf52168247de7, https://i.scdn.co/image/ab67616d00004851d08ee7bfe4aaf52168247de7, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a493e05c99d8ec5e8020ff2b, https://i.scdn.co/image/ab67616d00001e02a493e05c99d8ec5e8020ff2b, https://i.scdn.co/image/ab67616d00004851a493e05c99d8ec5e8020ff2b, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2731b3a5bd79d90213fe5f5e0df, https://i.scdn.co/image/ab67616d00001e021b3a5bd79d90213fe5f5e0df, https://i.scdn.co/image/ab67616d000048511b3a5bd79d90213fe5f5e0df, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273fb9baf4194eda393b587c75d, https://i.scdn.co/image/ab67616d00001e02fb9baf4194eda393b587c75d, https://i.scdn.co/image/ab67616d00004851fb9baf4194eda393b587c75d, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2730ed9de57ca9e59ca05e0f04b, https://i.scdn.co/image/ab67616d00001e020ed9de57ca9e59ca05e0f04b, https://i.scdn.co/image/ab67616d000048510ed9de57ca9e59ca05e0f04b, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732a5eb5b119eca221fbe68728, https://i.scdn.co/image/ab67616d00001e022a5eb5b119eca221fbe68728, https://i.scdn.co/image/ab67616d000048512a5eb5b119eca221fbe68728, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273424db9c5126cb63d5e3f4ab9, https://i.scdn.co/image/ab67616d00001e02424db9c5126cb63d5e3f4ab9, https://i.scdn.co/image/ab67616d00004851424db9c5126cb63d5e3f4ab9, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273520845b97c0ebbee3608e794, https://i.scdn.co/image/ab67616d00001e02520845b97c0ebbee3608e794, https://i.scdn.co/image/ab67616d00004851520845b97c0ebbee3608e794, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273294d49871cc7625ed4603f59, https://i.scdn.co/image/ab67616d00001e02294d49871cc7625ed4603f59, https://i.scdn.co/image/ab67616d00004851294d49871cc7625ed4603f59, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737c29525e4e1787a061a74b10, https://i.scdn.co/image/ab67616d00001e027c29525e4e1787a061a74b10, https://i.scdn.co/image/ab67616d000048517c29525e4e1787a061a74b10, 640, 300, 64
#>                  name release_date release_date_precision total_tracks  type
#> 1        Sweet Dreams   2021-06-11                    day            1 album
#> 2           September   2021-06-11                    day            1 album
#> 3          Call On Me   2021-06-11                    day            1 album
#> 4     EveryTime I Cry   2021-06-08                    day            1 album
#> 5          Cloudy Day   2021-06-10                    day            3 album
#> 6           Dumb Love   2021-06-11                    day            1 album
#> 7      JORDI (Deluxe)   2021-06-11                    day           14 album
#> 8           Thot Shit   2021-06-11                    day            1 album
#> 9         Culture III   2021-06-11                    day           19 album
#> 10       Need To Know   2021-06-11                    day            1 album
#> 11        Solar Power   2021-06-10                    day            1 album
#> 12       Hall of Fame   2021-06-11                    day           20 album
#> 13                Blå   2021-06-04                    day            1 album
#> 14              I Won   2021-06-04                    day            1 album
#> 15            Nirvana   2021-06-04                    day            1 album
#> 16          Fricktion   2021-06-03                    day           10 album
#> 17          Naag Nool   2021-06-04                    day           10 album
#> 18        Förutom dig   2021-06-04                    day            1 album
#> 19      Late At Night   2021-06-04                    day            1 album
#> 20 Comme Ci, Comme Ça   2021-06-04                    day            1 album
#>                                     uri
#> 1  spotify:album:5WV1czZI9hnO6REVS6yKHR
#> 2  spotify:album:1gHwbB7Db9yJ7kASKJFSnc
#> 3  spotify:album:4l78jaWeylIINTLaAdIHJy
#> 4  spotify:album:5W79aGcuoBYk0Mb2QL2Jcw
#> 5  spotify:album:3IbmFFjHMuAkhfk1zFDSow
#> 6  spotify:album:0xU5lh9KbThw5Y7b50Ilm8
#> 7  spotify:album:1pCA38N6MkLlthXtAOvZTU
#> 8  spotify:album:7gEGuvkVs79hY4tG7OYnzx
#> 9  spotify:album:3uyOwJu4r3yroAkFywNFM3
#> 10 spotify:album:7N3SiwnBMR1C8hcA4W1GtH
#> 11 spotify:album:0pAF4Jr6MPQ2u0vqG2VOjd
#> 12 spotify:album:7KSf6p0G0mZd7j60etVTwT
#> 13 spotify:album:4NQguSrcgRiDznaKoowh8u
#> 14 spotify:album:01RtA0qCjfGy0FY4ON23nD
#> 15 spotify:album:4VrGw3dww2O6U0hZTXOHae
#> 16 spotify:album:2v5eP5PoTRTy3yOxX5pfAi
#> 17 spotify:album:1UTctXDMnBfSheaCnpT7O5
#> 18 spotify:album:5Bw1AqtLIewl5hUQ1Lj8Rv
#> 19 spotify:album:2WFC0XUBROo4X2vxLwtB4u
#> 20 spotify:album:6690jqPTlPzzo3xoQwLSwh
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/album/5WV1czZI9hnO6REVS6yKHR
#> 2  https://open.spotify.com/album/1gHwbB7Db9yJ7kASKJFSnc
#> 3  https://open.spotify.com/album/4l78jaWeylIINTLaAdIHJy
#> 4  https://open.spotify.com/album/5W79aGcuoBYk0Mb2QL2Jcw
#> 5  https://open.spotify.com/album/3IbmFFjHMuAkhfk1zFDSow
#> 6  https://open.spotify.com/album/0xU5lh9KbThw5Y7b50Ilm8
#> 7  https://open.spotify.com/album/1pCA38N6MkLlthXtAOvZTU
#> 8  https://open.spotify.com/album/7gEGuvkVs79hY4tG7OYnzx
#> 9  https://open.spotify.com/album/3uyOwJu4r3yroAkFywNFM3
#> 10 https://open.spotify.com/album/7N3SiwnBMR1C8hcA4W1GtH
#> 11 https://open.spotify.com/album/0pAF4Jr6MPQ2u0vqG2VOjd
#> 12 https://open.spotify.com/album/7KSf6p0G0mZd7j60etVTwT
#> 13 https://open.spotify.com/album/4NQguSrcgRiDznaKoowh8u
#> 14 https://open.spotify.com/album/01RtA0qCjfGy0FY4ON23nD
#> 15 https://open.spotify.com/album/4VrGw3dww2O6U0hZTXOHae
#> 16 https://open.spotify.com/album/2v5eP5PoTRTy3yOxX5pfAi
#> 17 https://open.spotify.com/album/1UTctXDMnBfSheaCnpT7O5
#> 18 https://open.spotify.com/album/5Bw1AqtLIewl5hUQ1Lj8Rv
#> 19 https://open.spotify.com/album/2WFC0XUBROo4X2vxLwtB4u
#> 20 https://open.spotify.com/album/6690jqPTlPzzo3xoQwLSwh

# }