Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit
Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market
Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists
A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres
A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks
A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms
Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key
Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode
Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity
Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo
Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature
Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms
Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key
Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode
Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity
Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo
Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature
Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms
Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key
Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode
Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity
Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo
Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature
Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response
Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

# \donttest{
get_recommendations(market = 'SE', seed_genres = 'rock')

#>                                                                                                                                                                                                                   artists
#> 1                   https://api.spotify.com/v1/artists/7243txmysJ4KbRmH8UAMKO, 7243txmysJ4KbRmH8UAMKO, The Amazons, artist, spotify:artist:7243txmysJ4KbRmH8UAMKO, https://open.spotify.com/artist/7243txmysJ4KbRmH8UAMKO
#> 2                      https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#> 3              https://api.spotify.com/v1/artists/2xiIXseIJcq3nG7C8fHeBj, 2xiIXseIJcq3nG7C8fHeBj, Three Days Grace, artist, spotify:artist:2xiIXseIJcq3nG7C8fHeBj, https://open.spotify.com/artist/2xiIXseIJcq3nG7C8fHeBj
#> 4                        https://api.spotify.com/v1/artists/3NChzMpu9exTlNPiqUQ2DE, 3NChzMpu9exTlNPiqUQ2DE, Thrice, artist, spotify:artist:3NChzMpu9exTlNPiqUQ2DE, https://open.spotify.com/artist/3NChzMpu9exTlNPiqUQ2DE
#> 5            https://api.spotify.com/v1/artists/22bE4uQ6baNwSHPVcDxLCe, 22bE4uQ6baNwSHPVcDxLCe, The Rolling Stones, artist, spotify:artist:22bE4uQ6baNwSHPVcDxLCe, https://open.spotify.com/artist/22bE4uQ6baNwSHPVcDxLCe
#> 6                        https://api.spotify.com/v1/artists/6KO6G41BBLTDNYOLefWTMU, 6KO6G41BBLTDNYOLefWTMU, P.O.D., artist, spotify:artist:6KO6G41BBLTDNYOLefWTMU, https://open.spotify.com/artist/6KO6G41BBLTDNYOLefWTMU
#> 7                      https://api.spotify.com/v1/artists/05fG473iIaoy82BF1aGhL8, 05fG473iIaoy82BF1aGhL8, Slipknot, artist, spotify:artist:05fG473iIaoy82BF1aGhL8, https://open.spotify.com/artist/05fG473iIaoy82BF1aGhL8
#> 8  https://api.spotify.com/v1/artists/3IYUhFvPQItj6xySrBmZkd, 3IYUhFvPQItj6xySrBmZkd, Creedence Clearwater Revival, artist, spotify:artist:3IYUhFvPQItj6xySrBmZkd, https://open.spotify.com/artist/3IYUhFvPQItj6xySrBmZkd
#> 9                   https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 10              https://api.spotify.com/v1/artists/53XhwfbYqKCa1cC15pYq2q, 53XhwfbYqKCa1cC15pYq2q, Imagine Dragons, artist, spotify:artist:53XhwfbYqKCa1cC15pYq2q, https://open.spotify.com/artist/53XhwfbYqKCa1cC15pYq2q
#> 11                   https://api.spotify.com/v1/artists/2ziB7fzrXBoh1HUPS6sVFn, 2ziB7fzrXBoh1HUPS6sVFn, Audioslave, artist, spotify:artist:2ziB7fzrXBoh1HUPS6sVFn, https://open.spotify.com/artist/2ziB7fzrXBoh1HUPS6sVFn
#> 12                       https://api.spotify.com/v1/artists/5KDIH2gF0VpelTqyQS7udb, 5KDIH2gF0VpelTqyQS7udb, Staind, artist, spotify:artist:5KDIH2gF0VpelTqyQS7udb, https://open.spotify.com/artist/5KDIH2gF0VpelTqyQS7udb
#> 13                   https://api.spotify.com/v1/artists/2o18h28enlHxj887tATc58, 2o18h28enlHxj887tATc58, HammerFall, artist, spotify:artist:2o18h28enlHxj887tATc58, https://open.spotify.com/artist/2o18h28enlHxj887tATc58
#> 14                    https://api.spotify.com/v1/artists/3TOqt5oJwL9BE2NG9MEwDa, 3TOqt5oJwL9BE2NG9MEwDa, Disturbed, artist, spotify:artist:3TOqt5oJwL9BE2NG9MEwDa, https://open.spotify.com/artist/3TOqt5oJwL9BE2NG9MEwDa
#> 15                        https://api.spotify.com/v1/artists/6oFs3qk4VepIVFdoD4jmsy, 6oFs3qk4VepIVFdoD4jmsy, PVRIS, artist, spotify:artist:6oFs3qk4VepIVFdoD4jmsy, https://open.spotify.com/artist/6oFs3qk4VepIVFdoD4jmsy
#> 16                   https://api.spotify.com/v1/artists/7lzordPuZEXxwt9aoVZYmG, 7lzordPuZEXxwt9aoVZYmG, Billy Idol, artist, spotify:artist:7lzordPuZEXxwt9aoVZYmG, https://open.spotify.com/artist/7lzordPuZEXxwt9aoVZYmG
#> 17                         https://api.spotify.com/v1/artists/0ofhRkasSdhp6LUs7tFcF7, 0ofhRkasSdhp6LUs7tFcF7, WSTR, artist, spotify:artist:0ofhRkasSdhp6LUs7tFcF7, https://open.spotify.com/artist/0ofhRkasSdhp6LUs7tFcF7
#> 18               https://api.spotify.com/v1/artists/4MVyzYMgTwdP7Z49wAZHx0, 4MVyzYMgTwdP7Z49wAZHx0, Lynyrd Skynyrd, artist, spotify:artist:4MVyzYMgTwdP7Z49wAZHx0, https://open.spotify.com/artist/4MVyzYMgTwdP7Z49wAZHx0
#> 19                       https://api.spotify.com/v1/artists/0qT79UgT5tY4yudH9VfsdT, 0qT79UgT5tY4yudH9VfsdT, Sum 41, artist, spotify:artist:0qT79UgT5tY4yudH9VfsdT, https://open.spotify.com/artist/0qT79UgT5tY4yudH9VfsdT
#> 20                         https://api.spotify.com/v1/artists/3RNrq3jvMZxD9ZyoOZbQOD, 3RNrq3jvMZxD9ZyoOZbQOD, Korn, artist, spotify:artist:3RNrq3jvMZxD9ZyoOZbQOD, https://open.spotify.com/artist/3RNrq3jvMZxD9ZyoOZbQOD
#>    disc_number duration_ms explicit
#> 1            1      271560    FALSE
#> 2            1      266773    FALSE
#> 3            1      231480    FALSE
#> 4            1      239450    FALSE
#> 5            1      213066    FALSE
#> 6            1      256240    FALSE
#> 7            1      278600    FALSE
#> 8            1      140773    FALSE
#> 9            1      366136    FALSE
#> 10           1      232906    FALSE
#> 11           1      334906    FALSE
#> 12           1      264706     TRUE
#> 13           1      231386    FALSE
#> 14           1      204826    FALSE
#> 15           1      254358    FALSE
#> 16           1      286840    FALSE
#> 17           1      208179    FALSE
#> 18           1      547106    FALSE
#> 19           1      198493     TRUE
#> 20           1      255733     TRUE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/243b3t0P416vtdmxRllHEV
#> 2  https://api.spotify.com/v1/tracks/3AJwUDP919kvQ9QcozQPxg
#> 3  https://api.spotify.com/v1/tracks/0M955bMOoilikPXwKLYpoi
#> 4  https://api.spotify.com/v1/tracks/5OeY5LebIOzD9Zvat1Dikf
#> 5  https://api.spotify.com/v1/tracks/7HKez549fwJQDzx3zLjHKC
#> 6  https://api.spotify.com/v1/tracks/0UrWr7Jnu1heq1o99ZwUd0
#> 7  https://api.spotify.com/v1/tracks/6wqJeItl3Vc3az4ZicSQAB
#> 8  https://api.spotify.com/v1/tracks/4BP3uh0hFLFRb5cjsgLqDh
#> 9  https://api.spotify.com/v1/tracks/3uMmllZo1AfoEnVT4ENCD3
#> 10 https://api.spotify.com/v1/tracks/1y3bE5i57eUm3hfLAc7h08
#> 11 https://api.spotify.com/v1/tracks/3bwAS8cN7Xtzvv5UurJOOX
#> 12 https://api.spotify.com/v1/tracks/1BExV8JprNJLvLeDE9qdEW
#> 13 https://api.spotify.com/v1/tracks/0eV3PB3T0OxW4feG1DlOjQ
#> 14 https://api.spotify.com/v1/tracks/2TNPagu8Z4zvDOpmnktF00
#> 15 https://api.spotify.com/v1/tracks/7eN3lGwu6SSmwYGH9PlpYa
#> 16 https://api.spotify.com/v1/tracks/2ZTIw0fZhFp3nnvF41nvVc
#> 17 https://api.spotify.com/v1/tracks/4Jl3syKPuB2YFZZhsSeniv
#> 18 https://api.spotify.com/v1/tracks/5EWPGh7jbTNO2wakv8LjUI
#> 19 https://api.spotify.com/v1/tracks/6hqt1z34Oz0OZtSfy62OFD
#> 20 https://api.spotify.com/v1/tracks/6W21LNLz9Sw7sUSNWMSHRu
#>                        id is_local is_playable                           name
#> 1  243b3t0P416vtdmxRllHEV    FALSE        TRUE                    Black Magic
#> 2  3AJwUDP919kvQ9QcozQPxg    FALSE        TRUE                         Yellow
#> 3  0M955bMOoilikPXwKLYpoi    FALSE        TRUE    I Hate Everything About You
#> 4  5OeY5LebIOzD9Zvat1Dikf    FALSE        TRUE                    Black Honey
#> 5  7HKez549fwJQDzx3zLjHKC    FALSE        TRUE  Start Me Up - Remastered 2009
#> 6  0UrWr7Jnu1heq1o99ZwUd0    FALSE        TRUE            Youth of the Nation
#> 7  6wqJeItl3Vc3az4ZicSQAB    FALSE        TRUE                Before I Forget
#> 8  4BP3uh0hFLFRb5cjsgLqDh    FALSE        TRUE                  Fortunate Son
#> 9  3uMmllZo1AfoEnVT4ENCD3    FALSE        TRUE Highway Star - Remastered 2012
#> 10 1y3bE5i57eUm3hfLAc7h08    FALSE        TRUE               Walking The Wire
#> 11 3bwAS8cN7Xtzvv5UurJOOX    FALSE        TRUE               I Am the Highway
#> 12 1BExV8JprNJLvLeDE9qdEW    FALSE        TRUE               It's Been Awhile
#> 13 0eV3PB3T0OxW4feG1DlOjQ    FALSE        TRUE                 Hearts on Fire
#> 14 2TNPagu8Z4zvDOpmnktF00    FALSE        TRUE                        Warrior
#> 15 7eN3lGwu6SSmwYGH9PlpYa    FALSE        TRUE                         Heaven
#> 16 2ZTIw0fZhFp3nnvF41nvVc    FALSE        TRUE        Rebel Yell - Remastered
#> 17 4Jl3syKPuB2YFZZhsSeniv    FALSE        TRUE               Eastbound & Down
#> 18 5EWPGh7jbTNO2wakv8LjUI    FALSE        TRUE                      Free Bird
#> 19 6hqt1z34Oz0OZtSfy62OFD    FALSE        TRUE                  The Hell Song
#> 20 6W21LNLz9Sw7sUSNWMSHRu    FALSE        TRUE               Freak On a Leash
#>    popularity
#> 1          55
#> 2          86
#> 3          72
#> 4          60
#> 5          77
#> 6          71
#> 7          75
#> 8          82
#> 9          68
#> 10         69
#> 11         66
#> 12         66
#> 13         62
#> 14         63
#> 15         46
#> 16         64
#> 17         48
#> 18         75
#> 19         65
#> 20         75
#>                                                                                                    preview_url
#> 1                                                                                                         <NA>
#> 2  https://p.scdn.co/mp3-preview/26ad91af2f13c2b7729cbc70856ce43f4c790a7b?cid=125a10da5fa94687b3d0913168bc3698
#> 3  https://p.scdn.co/mp3-preview/8893669447625edf76339cdd67a338dad034a7ea?cid=125a10da5fa94687b3d0913168bc3698
#> 4  https://p.scdn.co/mp3-preview/5597144cc3b18d706d3cd7ee9e1802af6dcaddd4?cid=125a10da5fa94687b3d0913168bc3698
#> 5                                                                                                         <NA>
#> 6  https://p.scdn.co/mp3-preview/e539c574dfe89ea3cdd6aaa4aa93d89595e6e379?cid=125a10da5fa94687b3d0913168bc3698
#> 7  https://p.scdn.co/mp3-preview/835a766a2ae7f61480c38e5b97443520fecc325d?cid=125a10da5fa94687b3d0913168bc3698
#> 8                                                                                                         <NA>
#> 9                                                                                                         <NA>
#> 10                                                                                                        <NA>
#> 11 https://p.scdn.co/mp3-preview/eb5a71933e12afb5e357f3d2f22db28d780f08cf?cid=125a10da5fa94687b3d0913168bc3698
#> 12 https://p.scdn.co/mp3-preview/7c98137672d37068f9081aff2c5031d91dff772d?cid=125a10da5fa94687b3d0913168bc3698
#> 13 https://p.scdn.co/mp3-preview/6bb57fa117d04489b514cb89166f4fd9f0d7a47e?cid=125a10da5fa94687b3d0913168bc3698
#> 14 https://p.scdn.co/mp3-preview/f6a24824d47a9b7831a6c90a91e9776283bf668b?cid=125a10da5fa94687b3d0913168bc3698
#> 15 https://p.scdn.co/mp3-preview/a382b9a8bbb418d0b0245a2c2c7910d1269687f4?cid=125a10da5fa94687b3d0913168bc3698
#> 16                                                                                                        <NA>
#> 17 https://p.scdn.co/mp3-preview/53f484b80f067ff7f0c880afea2ca08382e35e7f?cid=125a10da5fa94687b3d0913168bc3698
#> 18                                                                                                        <NA>
#> 19                                                                                                        <NA>
#> 20 https://p.scdn.co/mp3-preview/14933695bf1a6f25b2c602d9bf4d774e5dcb2d6c?cid=125a10da5fa94687b3d0913168bc3698
#>    track_number  type                                  uri album.album_type
#> 1             6 track spotify:track:243b3t0P416vtdmxRllHEV            ALBUM
#> 2             5 track spotify:track:3AJwUDP919kvQ9QcozQPxg            ALBUM
#> 3             3 track spotify:track:0M955bMOoilikPXwKLYpoi            ALBUM
#> 4             7 track spotify:track:5OeY5LebIOzD9Zvat1Dikf            ALBUM
#> 5             1 track spotify:track:7HKez549fwJQDzx3zLjHKC            ALBUM
#> 6             4 track spotify:track:0UrWr7Jnu1heq1o99ZwUd0            ALBUM
#> 7            10 track spotify:track:6wqJeItl3Vc3az4ZicSQAB            ALBUM
#> 8             6 track spotify:track:4BP3uh0hFLFRb5cjsgLqDh            ALBUM
#> 9             1 track spotify:track:3uMmllZo1AfoEnVT4ENCD3            ALBUM
#> 10            5 track spotify:track:1y3bE5i57eUm3hfLAc7h08            ALBUM
#> 11            8 track spotify:track:3bwAS8cN7Xtzvv5UurJOOX            ALBUM
#> 12            4 track spotify:track:1BExV8JprNJLvLeDE9qdEW            ALBUM
#> 13            2 track spotify:track:0eV3PB3T0OxW4feG1DlOjQ            ALBUM
#> 14            4 track spotify:track:2TNPagu8Z4zvDOpmnktF00            ALBUM
#> 15            1 track spotify:track:7eN3lGwu6SSmwYGH9PlpYa            ALBUM
#> 16            5 track spotify:track:2ZTIw0fZhFp3nnvF41nvVc      COMPILATION
#> 17            6 track spotify:track:4Jl3syKPuB2YFZZhsSeniv            ALBUM
#> 18            8 track spotify:track:5EWPGh7jbTNO2wakv8LjUI            ALBUM
#> 19            1 track spotify:track:6hqt1z34Oz0OZtSfy62OFD            ALBUM
#> 20            2 track spotify:track:6W21LNLz9Sw7sUSNWMSHRu            ALBUM
#>                                                                                                                                                                                                             album.artists
#> 1                   https://api.spotify.com/v1/artists/7243txmysJ4KbRmH8UAMKO, 7243txmysJ4KbRmH8UAMKO, The Amazons, artist, spotify:artist:7243txmysJ4KbRmH8UAMKO, https://open.spotify.com/artist/7243txmysJ4KbRmH8UAMKO
#> 2                      https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#> 3              https://api.spotify.com/v1/artists/2xiIXseIJcq3nG7C8fHeBj, 2xiIXseIJcq3nG7C8fHeBj, Three Days Grace, artist, spotify:artist:2xiIXseIJcq3nG7C8fHeBj, https://open.spotify.com/artist/2xiIXseIJcq3nG7C8fHeBj
#> 4                        https://api.spotify.com/v1/artists/3NChzMpu9exTlNPiqUQ2DE, 3NChzMpu9exTlNPiqUQ2DE, Thrice, artist, spotify:artist:3NChzMpu9exTlNPiqUQ2DE, https://open.spotify.com/artist/3NChzMpu9exTlNPiqUQ2DE
#> 5            https://api.spotify.com/v1/artists/22bE4uQ6baNwSHPVcDxLCe, 22bE4uQ6baNwSHPVcDxLCe, The Rolling Stones, artist, spotify:artist:22bE4uQ6baNwSHPVcDxLCe, https://open.spotify.com/artist/22bE4uQ6baNwSHPVcDxLCe
#> 6                        https://api.spotify.com/v1/artists/6KO6G41BBLTDNYOLefWTMU, 6KO6G41BBLTDNYOLefWTMU, P.O.D., artist, spotify:artist:6KO6G41BBLTDNYOLefWTMU, https://open.spotify.com/artist/6KO6G41BBLTDNYOLefWTMU
#> 7                      https://api.spotify.com/v1/artists/05fG473iIaoy82BF1aGhL8, 05fG473iIaoy82BF1aGhL8, Slipknot, artist, spotify:artist:05fG473iIaoy82BF1aGhL8, https://open.spotify.com/artist/05fG473iIaoy82BF1aGhL8
#> 8  https://api.spotify.com/v1/artists/3IYUhFvPQItj6xySrBmZkd, 3IYUhFvPQItj6xySrBmZkd, Creedence Clearwater Revival, artist, spotify:artist:3IYUhFvPQItj6xySrBmZkd, https://open.spotify.com/artist/3IYUhFvPQItj6xySrBmZkd
#> 9                   https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 10              https://api.spotify.com/v1/artists/53XhwfbYqKCa1cC15pYq2q, 53XhwfbYqKCa1cC15pYq2q, Imagine Dragons, artist, spotify:artist:53XhwfbYqKCa1cC15pYq2q, https://open.spotify.com/artist/53XhwfbYqKCa1cC15pYq2q
#> 11                   https://api.spotify.com/v1/artists/2ziB7fzrXBoh1HUPS6sVFn, 2ziB7fzrXBoh1HUPS6sVFn, Audioslave, artist, spotify:artist:2ziB7fzrXBoh1HUPS6sVFn, https://open.spotify.com/artist/2ziB7fzrXBoh1HUPS6sVFn
#> 12                       https://api.spotify.com/v1/artists/5KDIH2gF0VpelTqyQS7udb, 5KDIH2gF0VpelTqyQS7udb, Staind, artist, spotify:artist:5KDIH2gF0VpelTqyQS7udb, https://open.spotify.com/artist/5KDIH2gF0VpelTqyQS7udb
#> 13                   https://api.spotify.com/v1/artists/2o18h28enlHxj887tATc58, 2o18h28enlHxj887tATc58, HammerFall, artist, spotify:artist:2o18h28enlHxj887tATc58, https://open.spotify.com/artist/2o18h28enlHxj887tATc58
#> 14                    https://api.spotify.com/v1/artists/3TOqt5oJwL9BE2NG9MEwDa, 3TOqt5oJwL9BE2NG9MEwDa, Disturbed, artist, spotify:artist:3TOqt5oJwL9BE2NG9MEwDa, https://open.spotify.com/artist/3TOqt5oJwL9BE2NG9MEwDa
#> 15                        https://api.spotify.com/v1/artists/6oFs3qk4VepIVFdoD4jmsy, 6oFs3qk4VepIVFdoD4jmsy, PVRIS, artist, spotify:artist:6oFs3qk4VepIVFdoD4jmsy, https://open.spotify.com/artist/6oFs3qk4VepIVFdoD4jmsy
#> 16                   https://api.spotify.com/v1/artists/7lzordPuZEXxwt9aoVZYmG, 7lzordPuZEXxwt9aoVZYmG, Billy Idol, artist, spotify:artist:7lzordPuZEXxwt9aoVZYmG, https://open.spotify.com/artist/7lzordPuZEXxwt9aoVZYmG
#> 17                         https://api.spotify.com/v1/artists/0ofhRkasSdhp6LUs7tFcF7, 0ofhRkasSdhp6LUs7tFcF7, WSTR, artist, spotify:artist:0ofhRkasSdhp6LUs7tFcF7, https://open.spotify.com/artist/0ofhRkasSdhp6LUs7tFcF7
#> 18               https://api.spotify.com/v1/artists/4MVyzYMgTwdP7Z49wAZHx0, 4MVyzYMgTwdP7Z49wAZHx0, Lynyrd Skynyrd, artist, spotify:artist:4MVyzYMgTwdP7Z49wAZHx0, https://open.spotify.com/artist/4MVyzYMgTwdP7Z49wAZHx0
#> 19                       https://api.spotify.com/v1/artists/0qT79UgT5tY4yudH9VfsdT, 0qT79UgT5tY4yudH9VfsdT, Sum 41, artist, spotify:artist:0qT79UgT5tY4yudH9VfsdT, https://open.spotify.com/artist/0qT79UgT5tY4yudH9VfsdT
#> 20                         https://api.spotify.com/v1/artists/3RNrq3jvMZxD9ZyoOZbQOD, 3RNrq3jvMZxD9ZyoOZbQOD, Korn, artist, spotify:artist:3RNrq3jvMZxD9ZyoOZbQOD, https://open.spotify.com/artist/3RNrq3jvMZxD9ZyoOZbQOD
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/2vsawVLAVvyQ8Oqu190Qbo
#> 2  https://api.spotify.com/v1/albums/6ZG5lRT77aJ3btmArcykra
#> 3  https://api.spotify.com/v1/albums/13topfW33NjnACjnRiZBX7
#> 4  https://api.spotify.com/v1/albums/4hFDcpvIU9F4OleRAR9nNZ
#> 5  https://api.spotify.com/v1/albums/15XNBzVWARPMlu0sEbfBjJ
#> 6  https://api.spotify.com/v1/albums/4eUmncMYQRg0Ot29KKqbuk
#> 7  https://api.spotify.com/v1/albums/4ZDBQSIDIZRUBOG2OHcN3T
#> 8  https://api.spotify.com/v1/albums/31q47gQszFt0CddSyMksgO
#> 9  https://api.spotify.com/v1/albums/32NQ56VZDTXSH3SMv4XSGN
#> 10 https://api.spotify.com/v1/albums/33pt9HBdGlAbRGBHQgsZsU
#> 11 https://api.spotify.com/v1/albums/78guAsers0klWl6RwzgDLd
#> 12 https://api.spotify.com/v1/albums/0HogGfosAvj4tF3GTv2KIE
#> 13 https://api.spotify.com/v1/albums/2uehY2zRsh0h6ZmJt8LLQi
#> 14 https://api.spotify.com/v1/albums/23Kf2cl3kz7pFNKhl4iMp9
#> 15 https://api.spotify.com/v1/albums/0jLafOidtMogAbdIjOXXIe
#> 16 https://api.spotify.com/v1/albums/1gTvJG5YnrTiwr0uDuzaoA
#> 17 https://api.spotify.com/v1/albums/4gw6oQ9Yk4m6mreJi6oGAV
#> 18 https://api.spotify.com/v1/albums/6DExt1eX4lflLacVjHHbOs
#> 19 https://api.spotify.com/v1/albums/2iOJEUCKJaYseov1qA8ot8
#> 20 https://api.spotify.com/v1/albums/0gsiszk6JWYwAyGvaTTud4
#>                  album.id
#> 1  2vsawVLAVvyQ8Oqu190Qbo
#> 2  6ZG5lRT77aJ3btmArcykra
#> 3  13topfW33NjnACjnRiZBX7
#> 4  4hFDcpvIU9F4OleRAR9nNZ
#> 5  15XNBzVWARPMlu0sEbfBjJ
#> 6  4eUmncMYQRg0Ot29KKqbuk
#> 7  4ZDBQSIDIZRUBOG2OHcN3T
#> 8  31q47gQszFt0CddSyMksgO
#> 9  32NQ56VZDTXSH3SMv4XSGN
#> 10 33pt9HBdGlAbRGBHQgsZsU
#> 11 78guAsers0klWl6RwzgDLd
#> 12 0HogGfosAvj4tF3GTv2KIE
#> 13 2uehY2zRsh0h6ZmJt8LLQi
#> 14 23Kf2cl3kz7pFNKhl4iMp9
#> 15 0jLafOidtMogAbdIjOXXIe
#> 16 1gTvJG5YnrTiwr0uDuzaoA
#> 17 4gw6oQ9Yk4m6mreJi6oGAV
#> 18 6DExt1eX4lflLacVjHHbOs
#> 19 2iOJEUCKJaYseov1qA8ot8
#> 20 0gsiszk6JWYwAyGvaTTud4
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27384e5a436b4b16efdeee3cb2a, https://i.scdn.co/image/ab67616d00001e0284e5a436b4b16efdeee3cb2a, https://i.scdn.co/image/ab67616d0000485184e5a436b4b16efdeee3cb2a, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2733d92b2ad5af9fbc8637425f0, https://i.scdn.co/image/ab67616d00001e023d92b2ad5af9fbc8637425f0, https://i.scdn.co/image/ab67616d000048513d92b2ad5af9fbc8637425f0, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27377922eaa071fb388270f203f, https://i.scdn.co/image/ab67616d00001e0277922eaa071fb388270f203f, https://i.scdn.co/image/ab67616d0000485177922eaa071fb388270f203f, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a65c9844d454375a10178690, https://i.scdn.co/image/ab67616d00001e02a65c9844d454375a10178690, https://i.scdn.co/image/ab67616d00004851a65c9844d454375a10178690, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27308fc42e575043a753f60d675, https://i.scdn.co/image/ab67616d00001e0208fc42e575043a753f60d675, https://i.scdn.co/image/ab67616d0000485108fc42e575043a753f60d675, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27395741aa186187b05e0653245, https://i.scdn.co/image/ab67616d00001e0295741aa186187b05e0653245, https://i.scdn.co/image/ab67616d0000485195741aa186187b05e0653245, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736b3463e7160d333ada4b175a, https://i.scdn.co/image/ab67616d00001e026b3463e7160d333ada4b175a, https://i.scdn.co/image/ab67616d000048516b3463e7160d333ada4b175a, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2739f39192f9f8ca1c90847b3e5, https://i.scdn.co/image/ab67616d00001e029f39192f9f8ca1c90847b3e5, https://i.scdn.co/image/ab67616d000048519f39192f9f8ca1c90847b3e5, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00001e02a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00004851a729c9c3dec04b99d889c66f, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735675e83f707f1d7271e5cf8a, https://i.scdn.co/image/ab67616d00001e025675e83f707f1d7271e5cf8a, https://i.scdn.co/image/ab67616d000048515675e83f707f1d7271e5cf8a, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a7292b6863258e889b78d787, https://i.scdn.co/image/ab67616d00001e02a7292b6863258e889b78d787, https://i.scdn.co/image/ab67616d00004851a7292b6863258e889b78d787, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27300e9c78a5d11dd6cf8e8f43e, https://i.scdn.co/image/ab67616d00001e0200e9c78a5d11dd6cf8e8f43e, https://i.scdn.co/image/ab67616d0000485100e9c78a5d11dd6cf8e8f43e, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273bceefa541bb181ec0b8d32e4, https://i.scdn.co/image/ab67616d00001e02bceefa541bb181ec0b8d32e4, https://i.scdn.co/image/ab67616d00004851bceefa541bb181ec0b8d32e4, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273fa6551bf5f8a821baa48a1e8, https://i.scdn.co/image/ab67616d00001e02fa6551bf5f8a821baa48a1e8, https://i.scdn.co/image/ab67616d00004851fa6551bf5f8a821baa48a1e8, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736d81fe74bb715685e4219113, https://i.scdn.co/image/ab67616d00001e026d81fe74bb715685e4219113, https://i.scdn.co/image/ab67616d000048516d81fe74bb715685e4219113, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735425de975a56867550d2dd65, https://i.scdn.co/image/ab67616d00001e025425de975a56867550d2dd65, https://i.scdn.co/image/ab67616d000048515425de975a56867550d2dd65, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736965cb863cb6fa5f93459805, https://i.scdn.co/image/ab67616d00001e026965cb863cb6fa5f93459805, https://i.scdn.co/image/ab67616d000048516965cb863cb6fa5f93459805, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273128450651c9f0442780d8eb8, https://i.scdn.co/image/ab67616d00001e02128450651c9f0442780d8eb8, https://i.scdn.co/image/ab67616d00004851128450651c9f0442780d8eb8, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d6176a711e031d6522840c52, https://i.scdn.co/image/ab67616d00001e02d6176a711e031d6522840c52, https://i.scdn.co/image/ab67616d00004851d6176a711e031d6522840c52, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27350d216aebaf98e8ac9947fd5, https://i.scdn.co/image/ab67616d00001e0250d216aebaf98e8ac9947fd5, https://i.scdn.co/image/ab67616d0000485150d216aebaf98e8ac9947fd5, 640, 300, 64
#>                                    album.name album.release_date
#> 1                        The Amazons (Deluxe)         2017-05-26
#> 2                                  Parachutes         2000-07-10
#> 3                            Three Days Grace         2003-04-19
#> 4           To Be Everywhere Is To Be Nowhere         2016-05-27
#> 5               Tattoo You (2009 Re-Mastered)         1981-08-24
#> 6  Satellite (U.S. Version; WEA Distribution)         2001-09-11
#> 7               Vol. 3: The Subliminal Verses               2004
#> 8                     Willy And The Poor Boys         1969-11-02
#> 9                   Machine Head (Remastered)         1972-03-25
#> 10                                     Evolve         2017-06-23
#> 11                                 Audioslave         2002-11-17
#> 12                            Break the Cycle         2001-05-22
#> 13                            Crimson Thunder         2002-10-28
#> 14                    Asylum (Deluxe Edition)         2010-08-20
#> 15 All We Know Of Heaven, All We Need Of Hell         2017-08-25
#> 16                              Greatest Hits         2001-01-01
#> 17                    Red, Green Or Inbetween         2017-01-20
#> 18          Pronounced' Leh-'Nerd 'Skin-'Nerd         1973-01-01
#> 19                   Does This Look Infected?         2002-01-01
#> 20                          Follow The Leader         1998-08-18
#>    album.release_date_precision album.total_tracks album.type
#> 1                           day                 17      album
#> 2                           day                 10      album
#> 3                           day                 12      album
#> 4                           day                 11      album
#> 5                           day                 11      album
#> 6                           day                 15      album
#> 7                          year                 22      album
#> 8                           day                 13      album
#> 9                           day                  8      album
#> 10                          day                 12      album
#> 11                          day                 14      album
#> 12                          day                 14      album
#> 13                          day                 12      album
#> 14                          day                 15      album
#> 15                          day                 10      album
#> 16                          day                 16      album
#> 17                          day                 11      album
#> 18                          day                  8      album
#> 19                          day                 12      album
#> 20                          day                 14      album
#>                               album.uri
#> 1  spotify:album:2vsawVLAVvyQ8Oqu190Qbo
#> 2  spotify:album:6ZG5lRT77aJ3btmArcykra
#> 3  spotify:album:13topfW33NjnACjnRiZBX7
#> 4  spotify:album:4hFDcpvIU9F4OleRAR9nNZ
#> 5  spotify:album:15XNBzVWARPMlu0sEbfBjJ
#> 6  spotify:album:4eUmncMYQRg0Ot29KKqbuk
#> 7  spotify:album:4ZDBQSIDIZRUBOG2OHcN3T
#> 8  spotify:album:31q47gQszFt0CddSyMksgO
#> 9  spotify:album:32NQ56VZDTXSH3SMv4XSGN
#> 10 spotify:album:33pt9HBdGlAbRGBHQgsZsU
#> 11 spotify:album:78guAsers0klWl6RwzgDLd
#> 12 spotify:album:0HogGfosAvj4tF3GTv2KIE
#> 13 spotify:album:2uehY2zRsh0h6ZmJt8LLQi
#> 14 spotify:album:23Kf2cl3kz7pFNKhl4iMp9
#> 15 spotify:album:0jLafOidtMogAbdIjOXXIe
#> 16 spotify:album:1gTvJG5YnrTiwr0uDuzaoA
#> 17 spotify:album:4gw6oQ9Yk4m6mreJi6oGAV
#> 18 spotify:album:6DExt1eX4lflLacVjHHbOs
#> 19 spotify:album:2iOJEUCKJaYseov1qA8ot8
#> 20 spotify:album:0gsiszk6JWYwAyGvaTTud4
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/2vsawVLAVvyQ8Oqu190Qbo      GBUM71604648
#> 2  https://open.spotify.com/album/6ZG5lRT77aJ3btmArcykra      GBAYE0000267
#> 3  https://open.spotify.com/album/13topfW33NjnACjnRiZBX7      USJI10300075
#> 4  https://open.spotify.com/album/4hFDcpvIU9F4OleRAR9nNZ      QMRSZ1600157
#> 5  https://open.spotify.com/album/15XNBzVWARPMlu0sEbfBjJ      GBUM70909474
#> 6  https://open.spotify.com/album/4eUmncMYQRg0Ot29KKqbuk      USAT20106877
#> 7  https://open.spotify.com/album/4ZDBQSIDIZRUBOG2OHcN3T      NLA320420824
#> 8  https://open.spotify.com/album/31q47gQszFt0CddSyMksgO      USC4R0817618
#> 9  https://open.spotify.com/album/32NQ56VZDTXSH3SMv4XSGN      GBDXG1200002
#> 10 https://open.spotify.com/album/33pt9HBdGlAbRGBHQgsZsU      USUM71704648
#> 11 https://open.spotify.com/album/78guAsers0klWl6RwzgDLd      USSM10211592
#> 12 https://open.spotify.com/album/0HogGfosAvj4tF3GTv2KIE      USEE10100288
#> 13 https://open.spotify.com/album/2uehY2zRsh0h6ZmJt8LLQi      DED830264702
#> 14 https://open.spotify.com/album/23Kf2cl3kz7pFNKhl4iMp9      USRE11000555
#> 15 https://open.spotify.com/album/0jLafOidtMogAbdIjOXXIe      QMRSZ1701194
#> 16 https://open.spotify.com/album/1gTvJG5YnrTiwr0uDuzaoA      USCH39900054
#> 17 https://open.spotify.com/album/4gw6oQ9Yk4m6mreJi6oGAV      USZZ81610031
#> 18 https://open.spotify.com/album/6DExt1eX4lflLacVjHHbOs      USMC17301722
#> 19 https://open.spotify.com/album/2iOJEUCKJaYseov1qA8ot8      USIR20200791
#> 20 https://open.spotify.com/album/0gsiszk6JWYwAyGvaTTud4      USSM19801755
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/243b3t0P416vtdmxRllHEV
#> 2  https://open.spotify.com/track/3AJwUDP919kvQ9QcozQPxg
#> 3  https://open.spotify.com/track/0M955bMOoilikPXwKLYpoi
#> 4  https://open.spotify.com/track/5OeY5LebIOzD9Zvat1Dikf
#> 5  https://open.spotify.com/track/7HKez549fwJQDzx3zLjHKC
#> 6  https://open.spotify.com/track/0UrWr7Jnu1heq1o99ZwUd0
#> 7  https://open.spotify.com/track/6wqJeItl3Vc3az4ZicSQAB
#> 8  https://open.spotify.com/track/4BP3uh0hFLFRb5cjsgLqDh
#> 9  https://open.spotify.com/track/3uMmllZo1AfoEnVT4ENCD3
#> 10 https://open.spotify.com/track/1y3bE5i57eUm3hfLAc7h08
#> 11 https://open.spotify.com/track/3bwAS8cN7Xtzvv5UurJOOX
#> 12 https://open.spotify.com/track/1BExV8JprNJLvLeDE9qdEW
#> 13 https://open.spotify.com/track/0eV3PB3T0OxW4feG1DlOjQ
#> 14 https://open.spotify.com/track/2TNPagu8Z4zvDOpmnktF00
#> 15 https://open.spotify.com/track/7eN3lGwu6SSmwYGH9PlpYa
#> 16 https://open.spotify.com/track/2ZTIw0fZhFp3nnvF41nvVc
#> 17 https://open.spotify.com/track/4Jl3syKPuB2YFZZhsSeniv
#> 18 https://open.spotify.com/track/5EWPGh7jbTNO2wakv8LjUI
#> 19 https://open.spotify.com/track/6hqt1z34Oz0OZtSfy62OFD
#> 20 https://open.spotify.com/track/6W21LNLz9Sw7sUSNWMSHRu
#>                                            linked_from.href
#> 1                                                      <NA>
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4                                                      <NA>
#> 5                                                      <NA>
#> 6                                                      <NA>
#> 7                                                      <NA>
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#> 11 https://api.spotify.com/v1/tracks/4Im33JWLEobPtMTTAw7GUF
#> 12                                                     <NA>
#> 13 https://api.spotify.com/v1/tracks/4EFtG09zBCVqItRg3aSTrJ
#> 14                                                     <NA>
#> 15                                                     <NA>
#> 16 https://api.spotify.com/v1/tracks/4TIJ7zSBNejpoIPaWpWRKc
#> 17 https://api.spotify.com/v1/tracks/4NNzO1RQqdYzQXTISmHZDe
#> 18                                                     <NA>
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1                    <NA>             <NA>                                 <NA>
#> 2                    <NA>             <NA>                                 <NA>
#> 3                    <NA>             <NA>                                 <NA>
#> 4                    <NA>             <NA>                                 <NA>
#> 5                    <NA>             <NA>                                 <NA>
#> 6                    <NA>             <NA>                                 <NA>
#> 7                    <NA>             <NA>                                 <NA>
#> 8                    <NA>             <NA>                                 <NA>
#> 9                    <NA>             <NA>                                 <NA>
#> 10                   <NA>             <NA>                                 <NA>
#> 11 4Im33JWLEobPtMTTAw7GUF            track spotify:track:4Im33JWLEobPtMTTAw7GUF
#> 12                   <NA>             <NA>                                 <NA>
#> 13 4EFtG09zBCVqItRg3aSTrJ            track spotify:track:4EFtG09zBCVqItRg3aSTrJ
#> 14                   <NA>             <NA>                                 <NA>
#> 15                   <NA>             <NA>                                 <NA>
#> 16 4TIJ7zSBNejpoIPaWpWRKc            track spotify:track:4TIJ7zSBNejpoIPaWpWRKc
#> 17 4NNzO1RQqdYzQXTISmHZDe            track spotify:track:4NNzO1RQqdYzQXTISmHZDe
#> 18                   <NA>             <NA>                                 <NA>
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1                                                   <NA>
#> 2                                                   <NA>
#> 3                                                   <NA>
#> 4                                                   <NA>
#> 5                                                   <NA>
#> 6                                                   <NA>
#> 7                                                   <NA>
#> 8                                                   <NA>
#> 9                                                   <NA>
#> 10                                                  <NA>
#> 11 https://open.spotify.com/track/4Im33JWLEobPtMTTAw7GUF
#> 12                                                  <NA>
#> 13 https://open.spotify.com/track/4EFtG09zBCVqItRg3aSTrJ
#> 14                                                  <NA>
#> 15                                                  <NA>
#> 16 https://open.spotify.com/track/4TIJ7zSBNejpoIPaWpWRKc
#> 17 https://open.spotify.com/track/4NNzO1RQqdYzQXTISmHZDe
#> 18                                                  <NA>
#> 19                                                  <NA>
#> 20                                                  <NA>

# }