Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit

Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market

Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists

A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres

A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks

A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms

Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key

Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode

Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity

Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo

Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature

Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence

Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms

Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key

Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode

Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity

Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo

Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature

Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence

Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms

Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key

Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode

Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity

Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo

Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature

Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence

Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization

Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response

Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

# \donttest{
get_recommendations(market = 'SE', seed_genres = 'rock')
#>                                                                                                                                                                                                                   artists
#> 1             https://api.spotify.com/v1/artists/3YQKmKGau1PzlVlkL1iodx, 3YQKmKGau1PzlVlkL1iodx, Twenty One Pilots, artist, spotify:artist:3YQKmKGau1PzlVlkL1iodx, https://open.spotify.com/artist/3YQKmKGau1PzlVlkL1iodx
#> 2  https://api.spotify.com/v1/artists/3IYUhFvPQItj6xySrBmZkd, 3IYUhFvPQItj6xySrBmZkd, Creedence Clearwater Revival, artist, spotify:artist:3IYUhFvPQItj6xySrBmZkd, https://open.spotify.com/artist/3IYUhFvPQItj6xySrBmZkd
#> 3                     https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 4                      https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#> 5                  https://api.spotify.com/v1/artists/36QJpDe2go2KgaRleHCDTp, 36QJpDe2go2KgaRleHCDTp, Led Zeppelin, artist, spotify:artist:36QJpDe2go2KgaRleHCDTp, https://open.spotify.com/artist/36QJpDe2go2KgaRleHCDTp
#> 6                https://api.spotify.com/v1/artists/55vs7NT1KxcFjbMC4y202E, 55vs7NT1KxcFjbMC4y202E, REO Speedwagon, artist, spotify:artist:55vs7NT1KxcFjbMC4y202E, https://open.spotify.com/artist/55vs7NT1KxcFjbMC4y202E
#> 7                 https://api.spotify.com/v1/artists/3qm84nBOXUEQ2vnTfUTTFC, 3qm84nBOXUEQ2vnTfUTTFC, Guns N' Roses, artist, spotify:artist:3qm84nBOXUEQ2vnTfUTTFC, https://open.spotify.com/artist/3qm84nBOXUEQ2vnTfUTTFC
#> 8                       https://api.spotify.com/v1/artists/6w7j5wQ5AI5OQYlcM15s2L, 6w7j5wQ5AI5OQYlcM15s2L, Extreme, artist, spotify:artist:6w7j5wQ5AI5OQYlcM15s2L, https://open.spotify.com/artist/6w7j5wQ5AI5OQYlcM15s2L
#> 9                     https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 10          https://api.spotify.com/v1/artists/20JZFwl6HVl6yg8a4H3ZqK, 20JZFwl6HVl6yg8a4H3ZqK, Panic! At The Disco, artist, spotify:artist:20JZFwl6HVl6yg8a4H3ZqK, https://open.spotify.com/artist/20JZFwl6HVl6yg8a4H3ZqK
#> 11                   https://api.spotify.com/v1/artists/2ziB7fzrXBoh1HUPS6sVFn, 2ziB7fzrXBoh1HUPS6sVFn, Audioslave, artist, spotify:artist:2ziB7fzrXBoh1HUPS6sVFn, https://open.spotify.com/artist/2ziB7fzrXBoh1HUPS6sVFn
#> 12                        https://api.spotify.com/v1/artists/43sZBwHjahUvgbx1WNIkIz, 43sZBwHjahUvgbx1WNIkIz, Creed, artist, spotify:artist:43sZBwHjahUvgbx1WNIkIz, https://open.spotify.com/artist/43sZBwHjahUvgbx1WNIkIz
#> 13            https://api.spotify.com/v1/artists/4NiJW4q9ichVqL1aUsgGAN, 4NiJW4q9ichVqL1aUsgGAN, A Day To Remember, artist, spotify:artist:4NiJW4q9ichVqL1aUsgGAN, https://open.spotify.com/artist/4NiJW4q9ichVqL1aUsgGAN
#> 14                        https://api.spotify.com/v1/artists/6oFs3qk4VepIVFdoD4jmsy, 6oFs3qk4VepIVFdoD4jmsy, PVRIS, artist, spotify:artist:6oFs3qk4VepIVFdoD4jmsy, https://open.spotify.com/artist/6oFs3qk4VepIVFdoD4jmsy
#> 15               https://api.spotify.com/v1/artists/2vd2HnNh4pdYa9gDVHFjEu, 2vd2HnNh4pdYa9gDVHFjEu, As I Lay Dying, artist, spotify:artist:2vd2HnNh4pdYa9gDVHFjEu, https://open.spotify.com/artist/2vd2HnNh4pdYa9gDVHFjEu
#> 16                    https://api.spotify.com/v1/artists/6vwjIs0tbIiseJMR3pqwiL, 6vwjIs0tbIiseJMR3pqwiL, Beartooth, artist, spotify:artist:6vwjIs0tbIiseJMR3pqwiL, https://open.spotify.com/artist/6vwjIs0tbIiseJMR3pqwiL
#> 17      https://api.spotify.com/v1/artists/5t28BP42x2axFnqOOMg3CM, 5t28BP42x2axFnqOOMg3CM, Five Finger Death Punch, artist, spotify:artist:5t28BP42x2axFnqOOMg3CM, https://open.spotify.com/artist/5t28BP42x2axFnqOOMg3CM
#> 18                  https://api.spotify.com/v1/artists/165ZgPlLkK7bf5bDoFc6Sb, 165ZgPlLkK7bf5bDoFc6Sb, Limp Bizkit, artist, spotify:artist:165ZgPlLkK7bf5bDoFc6Sb, https://open.spotify.com/artist/165ZgPlLkK7bf5bDoFc6Sb
#> 19                 https://api.spotify.com/v1/artists/7jy3rLJdDQY21OgRLCZ9sD, 7jy3rLJdDQY21OgRLCZ9sD, Foo Fighters, artist, spotify:artist:7jy3rLJdDQY21OgRLCZ9sD, https://open.spotify.com/artist/7jy3rLJdDQY21OgRLCZ9sD
#> 20                      https://api.spotify.com/v1/artists/0rvjqX7ttXeg3mTy8Xscbt, 0rvjqX7ttXeg3mTy8Xscbt, Journey, artist, spotify:artist:0rvjqX7ttXeg3mTy8Xscbt, https://open.spotify.com/artist/0rvjqX7ttXeg3mTy8Xscbt
#>    disc_number duration_ms explicit
#> 1            1      195920    FALSE
#> 2            1      152466    FALSE
#> 3            1      233986    FALSE
#> 4            1      307879    FALSE
#> 5            1      517125    FALSE
#> 6            1      239680    FALSE
#> 7            1      283733    FALSE
#> 8            1      334106    FALSE
#> 9            1      224596    FALSE
#> 10           1      213093    FALSE
#> 11           1      334906    FALSE
#> 12           1      242000    FALSE
#> 13           1      202640    FALSE
#> 14           1      254358    FALSE
#> 15           1      253317    FALSE
#> 16           1      213576    FALSE
#> 17           1      275212    FALSE
#> 18           1      272973     TRUE
#> 19           1      323373    FALSE
#> 20           1      201693    FALSE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/6i0V12jOa3mr6uu4WYhUBr
#> 2  https://api.spotify.com/v1/tracks/1YqJSpIxdavjQ8BxAttU0r
#> 3  https://api.spotify.com/v1/tracks/1TRr8Pp2uOqL24fMHpJyV3
#> 4  https://api.spotify.com/v1/tracks/0BCPKOYdS2jbQ8iyB56Zns
#> 5  https://api.spotify.com/v1/tracks/6Vjk8MNXpQpi0F4BefdTyq
#> 6  https://api.spotify.com/v1/tracks/5gys5nzVQIYhgHIfiOJYva
#> 7  https://api.spotify.com/v1/tracks/2N2yrmodOnVF10mKvItC9P
#> 8  https://api.spotify.com/v1/tracks/1gVgkQFOKa8Wc1HYsJtPdH
#> 9  https://api.spotify.com/v1/tracks/4LFBVDJV0xpayUvquQVzhS
#> 10 https://api.spotify.com/v1/tracks/2fh3bZ8jZhMxOcfESE9nQY
#> 11 https://api.spotify.com/v1/tracks/3bwAS8cN7Xtzvv5UurJOOX
#> 12 https://api.spotify.com/v1/tracks/0eFklulxFnNCypNSrywzSB
#> 13 https://api.spotify.com/v1/tracks/4JsDHMv5PVO8N07DbDq33r
#> 14 https://api.spotify.com/v1/tracks/7eN3lGwu6SSmwYGH9PlpYa
#> 15 https://api.spotify.com/v1/tracks/0CcqWuAEJC93K8cBMbAjgI
#> 16 https://api.spotify.com/v1/tracks/03jdl5GTBhKsgB1Yw8mVPh
#> 17 https://api.spotify.com/v1/tracks/0h8VX60lat4433GnbP6vQs
#> 18 https://api.spotify.com/v1/tracks/2gSVKxPDww9Eep5rdvtdem
#> 19 https://api.spotify.com/v1/tracks/1wLQwg0mloy3yXjL0jPE0N
#> 20 https://api.spotify.com/v1/tracks/71SvEDmsOwIWw1IozsZoMA
#>                        id is_local is_playable
#> 1  6i0V12jOa3mr6uu4WYhUBr    FALSE        TRUE
#> 2  1YqJSpIxdavjQ8BxAttU0r    FALSE        TRUE
#> 3  1TRr8Pp2uOqL24fMHpJyV3    FALSE        TRUE
#> 4  0BCPKOYdS2jbQ8iyB56Zns    FALSE        TRUE
#> 5  6Vjk8MNXpQpi0F4BefdTyq    FALSE        TRUE
#> 6  5gys5nzVQIYhgHIfiOJYva    FALSE        TRUE
#> 7  2N2yrmodOnVF10mKvItC9P    FALSE        TRUE
#> 8  1gVgkQFOKa8Wc1HYsJtPdH    FALSE        TRUE
#> 9  4LFBVDJV0xpayUvquQVzhS    FALSE        TRUE
#> 10 2fh3bZ8jZhMxOcfESE9nQY    FALSE        TRUE
#> 11 3bwAS8cN7Xtzvv5UurJOOX    FALSE        TRUE
#> 12 0eFklulxFnNCypNSrywzSB    FALSE        TRUE
#> 13 4JsDHMv5PVO8N07DbDq33r    FALSE        TRUE
#> 14 7eN3lGwu6SSmwYGH9PlpYa    FALSE        TRUE
#> 15 0CcqWuAEJC93K8cBMbAjgI    FALSE        TRUE
#> 16 03jdl5GTBhKsgB1Yw8mVPh    FALSE        TRUE
#> 17 0h8VX60lat4433GnbP6vQs    FALSE        TRUE
#> 18 2gSVKxPDww9Eep5rdvtdem    FALSE        TRUE
#> 19 1wLQwg0mloy3yXjL0jPE0N    FALSE        TRUE
#> 20 71SvEDmsOwIWw1IozsZoMA    FALSE        TRUE
#>                                  name popularity
#> 1                            Heathens         81
#> 2            Lookin' Out My Back Door         68
#> 3                    Sound of Madness         44
#> 4                              Clocks         81
#> 5                  Kashmir - Remaster         70
#> 6                  Take It On the Run         67
#> 7                Don't Cry (Original)         79
#> 8                     More Than Words         75
#> 9                        Cut the Cord         67
#> 10 Don't Threaten Me with a Good Time         68
#> 11                   I Am the Highway         65
#> 12    One Last Breath - Radio Version         48
#> 13                         All I Want         68
#> 14                             Heaven         43
#> 15                       My Own Grave         60
#> 16                         In Between         66
#> 17                          Gone Away         50
#> 18                             My Way         72
#> 19                                Run         59
#> 20                Any Way You Want It         73
#>                                                                                                    preview_url
#> 1  https://p.scdn.co/mp3-preview/1eec0afcc9b808702506c14bb5a03f084d1051f8?cid=0e73e430c7814968988cd21c104bf113
#> 2                                                                                                         <NA>
#> 3  https://p.scdn.co/mp3-preview/9e9834df030dd8dcf0921de92143e8228f2ac8a7?cid=0e73e430c7814968988cd21c104bf113
#> 4  https://p.scdn.co/mp3-preview/566503d79384d4c80e9879cbc5f75880062b6c12?cid=0e73e430c7814968988cd21c104bf113
#> 5  https://p.scdn.co/mp3-preview/64908f0ca685e8727a2f07615ee147d28adbc450?cid=0e73e430c7814968988cd21c104bf113
#> 6  https://p.scdn.co/mp3-preview/9923235617087b4c488aa0fd6dcb43c9477e9800?cid=0e73e430c7814968988cd21c104bf113
#> 7                                                                                                         <NA>
#> 8                                                                                                         <NA>
#> 9  https://p.scdn.co/mp3-preview/d5a95e18dbe0eb0d59e407df37bf60ec5160c915?cid=0e73e430c7814968988cd21c104bf113
#> 10 https://p.scdn.co/mp3-preview/748decf82eaea4ca43bf8b501e4d836f3e7867f7?cid=0e73e430c7814968988cd21c104bf113
#> 11 https://p.scdn.co/mp3-preview/094459dc6af24f7c59dc1f09779a396eb58d17e6?cid=0e73e430c7814968988cd21c104bf113
#> 12                                                                                                        <NA>
#> 13                                                                                                        <NA>
#> 14 https://p.scdn.co/mp3-preview/9cc0dab191ceb5bef8a5c32ad9a56da1fadf83a1?cid=0e73e430c7814968988cd21c104bf113
#> 15 https://p.scdn.co/mp3-preview/68767f360616afcaa01eb7ec54ba631994379c2a?cid=0e73e430c7814968988cd21c104bf113
#> 16 https://p.scdn.co/mp3-preview/79b891286d1dc958d6cf7f45726d502da31b421a?cid=0e73e430c7814968988cd21c104bf113
#> 17 https://p.scdn.co/mp3-preview/c79f780133fe3ad26f8c5d278cc50e1780473a45?cid=0e73e430c7814968988cd21c104bf113
#> 18                                                                                                        <NA>
#> 19 https://p.scdn.co/mp3-preview/40ce1210ad84fc1025205a0b9dd7a1759c94a408?cid=0e73e430c7814968988cd21c104bf113
#> 20 https://p.scdn.co/mp3-preview/17d711fcd526a218c0ba9d0c7c957733afa91c2c?cid=0e73e430c7814968988cd21c104bf113
#>    track_number  type                                  uri album.album_type
#> 1             1 track spotify:track:6i0V12jOa3mr6uu4WYhUBr           SINGLE
#> 2             5 track spotify:track:1YqJSpIxdavjQ8BxAttU0r            ALBUM
#> 3             2 track spotify:track:1TRr8Pp2uOqL24fMHpJyV3            ALBUM
#> 4             5 track spotify:track:0BCPKOYdS2jbQ8iyB56Zns            ALBUM
#> 5             6 track spotify:track:6Vjk8MNXpQpi0F4BefdTyq            ALBUM
#> 6             5 track spotify:track:5gys5nzVQIYhgHIfiOJYva            ALBUM
#> 7             4 track spotify:track:2N2yrmodOnVF10mKvItC9P            ALBUM
#> 8             5 track spotify:track:1gVgkQFOKa8Wc1HYsJtPdH            ALBUM
#> 9             2 track spotify:track:4LFBVDJV0xpayUvquQVzhS            ALBUM
#> 10            2 track spotify:track:2fh3bZ8jZhMxOcfESE9nQY            ALBUM
#> 11            8 track spotify:track:3bwAS8cN7Xtzvv5UurJOOX            ALBUM
#> 12           11 track spotify:track:0eFklulxFnNCypNSrywzSB            ALBUM
#> 13            2 track spotify:track:4JsDHMv5PVO8N07DbDq33r            ALBUM
#> 14            1 track spotify:track:7eN3lGwu6SSmwYGH9PlpYa            ALBUM
#> 15            8 track spotify:track:0CcqWuAEJC93K8cBMbAjgI           SINGLE
#> 16            4 track spotify:track:03jdl5GTBhKsgB1Yw8mVPh            ALBUM
#> 17            2 track spotify:track:0h8VX60lat4433GnbP6vQs      COMPILATION
#> 18            5 track spotify:track:2gSVKxPDww9Eep5rdvtdem            ALBUM
#> 19            2 track spotify:track:1wLQwg0mloy3yXjL0jPE0N            ALBUM
#> 20            1 track spotify:track:71SvEDmsOwIWw1IozsZoMA            ALBUM
#>                                                                                                                                                                                                             album.artists
#> 1             https://api.spotify.com/v1/artists/3YQKmKGau1PzlVlkL1iodx, 3YQKmKGau1PzlVlkL1iodx, Twenty One Pilots, artist, spotify:artist:3YQKmKGau1PzlVlkL1iodx, https://open.spotify.com/artist/3YQKmKGau1PzlVlkL1iodx
#> 2  https://api.spotify.com/v1/artists/3IYUhFvPQItj6xySrBmZkd, 3IYUhFvPQItj6xySrBmZkd, Creedence Clearwater Revival, artist, spotify:artist:3IYUhFvPQItj6xySrBmZkd, https://open.spotify.com/artist/3IYUhFvPQItj6xySrBmZkd
#> 3                     https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 4                      https://api.spotify.com/v1/artists/4gzpq5DPGxSnKTe4SA8HAU, 4gzpq5DPGxSnKTe4SA8HAU, Coldplay, artist, spotify:artist:4gzpq5DPGxSnKTe4SA8HAU, https://open.spotify.com/artist/4gzpq5DPGxSnKTe4SA8HAU
#> 5                  https://api.spotify.com/v1/artists/36QJpDe2go2KgaRleHCDTp, 36QJpDe2go2KgaRleHCDTp, Led Zeppelin, artist, spotify:artist:36QJpDe2go2KgaRleHCDTp, https://open.spotify.com/artist/36QJpDe2go2KgaRleHCDTp
#> 6                https://api.spotify.com/v1/artists/55vs7NT1KxcFjbMC4y202E, 55vs7NT1KxcFjbMC4y202E, REO Speedwagon, artist, spotify:artist:55vs7NT1KxcFjbMC4y202E, https://open.spotify.com/artist/55vs7NT1KxcFjbMC4y202E
#> 7                 https://api.spotify.com/v1/artists/3qm84nBOXUEQ2vnTfUTTFC, 3qm84nBOXUEQ2vnTfUTTFC, Guns N' Roses, artist, spotify:artist:3qm84nBOXUEQ2vnTfUTTFC, https://open.spotify.com/artist/3qm84nBOXUEQ2vnTfUTTFC
#> 8                       https://api.spotify.com/v1/artists/6w7j5wQ5AI5OQYlcM15s2L, 6w7j5wQ5AI5OQYlcM15s2L, Extreme, artist, spotify:artist:6w7j5wQ5AI5OQYlcM15s2L, https://open.spotify.com/artist/6w7j5wQ5AI5OQYlcM15s2L
#> 9                     https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 10          https://api.spotify.com/v1/artists/20JZFwl6HVl6yg8a4H3ZqK, 20JZFwl6HVl6yg8a4H3ZqK, Panic! At The Disco, artist, spotify:artist:20JZFwl6HVl6yg8a4H3ZqK, https://open.spotify.com/artist/20JZFwl6HVl6yg8a4H3ZqK
#> 11                   https://api.spotify.com/v1/artists/2ziB7fzrXBoh1HUPS6sVFn, 2ziB7fzrXBoh1HUPS6sVFn, Audioslave, artist, spotify:artist:2ziB7fzrXBoh1HUPS6sVFn, https://open.spotify.com/artist/2ziB7fzrXBoh1HUPS6sVFn
#> 12                        https://api.spotify.com/v1/artists/43sZBwHjahUvgbx1WNIkIz, 43sZBwHjahUvgbx1WNIkIz, Creed, artist, spotify:artist:43sZBwHjahUvgbx1WNIkIz, https://open.spotify.com/artist/43sZBwHjahUvgbx1WNIkIz
#> 13            https://api.spotify.com/v1/artists/4NiJW4q9ichVqL1aUsgGAN, 4NiJW4q9ichVqL1aUsgGAN, A Day To Remember, artist, spotify:artist:4NiJW4q9ichVqL1aUsgGAN, https://open.spotify.com/artist/4NiJW4q9ichVqL1aUsgGAN
#> 14                        https://api.spotify.com/v1/artists/6oFs3qk4VepIVFdoD4jmsy, 6oFs3qk4VepIVFdoD4jmsy, PVRIS, artist, spotify:artist:6oFs3qk4VepIVFdoD4jmsy, https://open.spotify.com/artist/6oFs3qk4VepIVFdoD4jmsy
#> 15               https://api.spotify.com/v1/artists/2vd2HnNh4pdYa9gDVHFjEu, 2vd2HnNh4pdYa9gDVHFjEu, As I Lay Dying, artist, spotify:artist:2vd2HnNh4pdYa9gDVHFjEu, https://open.spotify.com/artist/2vd2HnNh4pdYa9gDVHFjEu
#> 16                    https://api.spotify.com/v1/artists/6vwjIs0tbIiseJMR3pqwiL, 6vwjIs0tbIiseJMR3pqwiL, Beartooth, artist, spotify:artist:6vwjIs0tbIiseJMR3pqwiL, https://open.spotify.com/artist/6vwjIs0tbIiseJMR3pqwiL
#> 17      https://api.spotify.com/v1/artists/5t28BP42x2axFnqOOMg3CM, 5t28BP42x2axFnqOOMg3CM, Five Finger Death Punch, artist, spotify:artist:5t28BP42x2axFnqOOMg3CM, https://open.spotify.com/artist/5t28BP42x2axFnqOOMg3CM
#> 18                  https://api.spotify.com/v1/artists/165ZgPlLkK7bf5bDoFc6Sb, 165ZgPlLkK7bf5bDoFc6Sb, Limp Bizkit, artist, spotify:artist:165ZgPlLkK7bf5bDoFc6Sb, https://open.spotify.com/artist/165ZgPlLkK7bf5bDoFc6Sb
#> 19                 https://api.spotify.com/v1/artists/7jy3rLJdDQY21OgRLCZ9sD, 7jy3rLJdDQY21OgRLCZ9sD, Foo Fighters, artist, spotify:artist:7jy3rLJdDQY21OgRLCZ9sD, https://open.spotify.com/artist/7jy3rLJdDQY21OgRLCZ9sD
#> 20                      https://api.spotify.com/v1/artists/0rvjqX7ttXeg3mTy8Xscbt, 0rvjqX7ttXeg3mTy8Xscbt, Journey, artist, spotify:artist:0rvjqX7ttXeg3mTy8Xscbt, https://open.spotify.com/artist/0rvjqX7ttXeg3mTy8Xscbt
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/3J8W9AOjQhnBLCX33m3atT
#> 2  https://api.spotify.com/v1/albums/6vuApmTBQPuqB31Kg8hIao
#> 3  https://api.spotify.com/v1/albums/661dSFpYtPulSNDZZLkokq
#> 4  https://api.spotify.com/v1/albums/0RHX9XECH8IVI3LNgWDpmQ
#> 5  https://api.spotify.com/v1/albums/1lZahjeu4AhPkg9JARZr5F
#> 6  https://api.spotify.com/v1/albums/0X4ZNTZw7SYgrp5rlBQC3N
#> 7  https://api.spotify.com/v1/albums/0CxPbTRARqKUYighiEY9Sz
#> 8  https://api.spotify.com/v1/albums/7DKHQxJTI32UyCdDdGwvRC
#> 9  https://api.spotify.com/v1/albums/7969zwJB8p021EBcCUjsv9
#> 10 https://api.spotify.com/v1/albums/6twKQ0EsUJHVlAr6XBylrj
#> 11 https://api.spotify.com/v1/albums/7ysDeat0kDelEE7M6MUwXZ
#> 12 https://api.spotify.com/v1/albums/2ENlJXygQX6en4iziijet6
#> 13 https://api.spotify.com/v1/albums/2kllnMabPw1hYID0MhQRK0
#> 14 https://api.spotify.com/v1/albums/0jLafOidtMogAbdIjOXXIe
#> 15 https://api.spotify.com/v1/albums/4zSeVXasb7coGxo7ymD3EM
#> 16 https://api.spotify.com/v1/albums/0cq8sFikYGr4vv5ZVDr9qg
#> 17 https://api.spotify.com/v1/albums/6Q6erxuA6kqfBobtkmO5pH
#> 18 https://api.spotify.com/v1/albums/5mi7FKaWE5CtcOjdyxScA7
#> 19 https://api.spotify.com/v1/albums/6KMkuqIwKkwUhUYRPL6dUc
#> 20 https://api.spotify.com/v1/albums/2OyVtIEp7O7a6o82DF4Ba5
#>                  album.id
#> 1  3J8W9AOjQhnBLCX33m3atT
#> 2  6vuApmTBQPuqB31Kg8hIao
#> 3  661dSFpYtPulSNDZZLkokq
#> 4  0RHX9XECH8IVI3LNgWDpmQ
#> 5  1lZahjeu4AhPkg9JARZr5F
#> 6  0X4ZNTZw7SYgrp5rlBQC3N
#> 7  0CxPbTRARqKUYighiEY9Sz
#> 8  7DKHQxJTI32UyCdDdGwvRC
#> 9  7969zwJB8p021EBcCUjsv9
#> 10 6twKQ0EsUJHVlAr6XBylrj
#> 11 7ysDeat0kDelEE7M6MUwXZ
#> 12 2ENlJXygQX6en4iziijet6
#> 13 2kllnMabPw1hYID0MhQRK0
#> 14 0jLafOidtMogAbdIjOXXIe
#> 15 4zSeVXasb7coGxo7ymD3EM
#> 16 0cq8sFikYGr4vv5ZVDr9qg
#> 17 6Q6erxuA6kqfBobtkmO5pH
#> 18 5mi7FKaWE5CtcOjdyxScA7
#> 19 6KMkuqIwKkwUhUYRPL6dUc
#> 20 2OyVtIEp7O7a6o82DF4Ba5
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732ca3ba8f334ca5a5f0312efb, https://i.scdn.co/image/ab67616d00001e022ca3ba8f334ca5a5f0312efb, https://i.scdn.co/image/ab67616d000048512ca3ba8f334ca5a5f0312efb, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e5656849a50efc0552cc0dc9, https://i.scdn.co/image/ab67616d00001e02e5656849a50efc0552cc0dc9, https://i.scdn.co/image/ab67616d00004851e5656849a50efc0552cc0dc9, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a24befd5cbb5a31f0ac9eacb, https://i.scdn.co/image/ab67616d00001e02a24befd5cbb5a31f0ac9eacb, https://i.scdn.co/image/ab67616d00004851a24befd5cbb5a31f0ac9eacb, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273de09e02aa7febf30b7c02d82, https://i.scdn.co/image/ab67616d00001e02de09e02aa7febf30b7c02d82, https://i.scdn.co/image/ab67616d00004851de09e02aa7febf30b7c02d82, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732abc2d86a442efb6cc631d0a, https://i.scdn.co/image/ab67616d00001e022abc2d86a442efb6cc631d0a, https://i.scdn.co/image/ab67616d000048512abc2d86a442efb6cc631d0a, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d514470784e3d02ee0bcdb80, https://i.scdn.co/image/ab67616d00001e02d514470784e3d02ee0bcdb80, https://i.scdn.co/image/ab67616d00004851d514470784e3d02ee0bcdb80, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e44963b8bb127552ac761873, https://i.scdn.co/image/ab67616d00001e02e44963b8bb127552ac761873, https://i.scdn.co/image/ab67616d00004851e44963b8bb127552ac761873, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273fb16b9b69eb891027654bcb1, https://i.scdn.co/image/ab67616d00001e02fb16b9b69eb891027654bcb1, https://i.scdn.co/image/ab67616d00004851fb16b9b69eb891027654bcb1, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737b6fda92d5912fef92f373de, https://i.scdn.co/image/ab67616d00001e027b6fda92d5912fef92f373de, https://i.scdn.co/image/ab67616d000048517b6fda92d5912fef92f373de, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27323152d9337d6c57b116ed13a, https://i.scdn.co/image/ab67616d00001e0223152d9337d6c57b116ed13a, https://i.scdn.co/image/ab67616d0000485123152d9337d6c57b116ed13a, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734d94e1251b7e8c5e247de39c, https://i.scdn.co/image/ab67616d00001e024d94e1251b7e8c5e247de39c, https://i.scdn.co/image/ab67616d000048514d94e1251b7e8c5e247de39c, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a32f2ae4dade83e0d7a294f4, https://i.scdn.co/image/ab67616d00001e02a32f2ae4dade83e0d7a294f4, https://i.scdn.co/image/ab67616d00004851a32f2ae4dade83e0d7a294f4, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d2477f14fcb48c2a3043afdb, https://i.scdn.co/image/ab67616d00001e02d2477f14fcb48c2a3043afdb, https://i.scdn.co/image/ab67616d00004851d2477f14fcb48c2a3043afdb, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736d81fe74bb715685e4219113, https://i.scdn.co/image/ab67616d00001e026d81fe74bb715685e4219113, https://i.scdn.co/image/ab67616d000048516d81fe74bb715685e4219113, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273ab876e767fcb1cace4048db4, https://i.scdn.co/image/ab67616d00001e02ab876e767fcb1cace4048db4, https://i.scdn.co/image/ab67616d00004851ab876e767fcb1cace4048db4, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2738aed8d6bb25e579e088ebbfc, https://i.scdn.co/image/ab67616d00001e028aed8d6bb25e579e088ebbfc, https://i.scdn.co/image/ab67616d000048518aed8d6bb25e579e088ebbfc, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b641fc33c6288c6d613bc552, https://i.scdn.co/image/ab67616d00001e02b641fc33c6288c6d613bc552, https://i.scdn.co/image/ab67616d00004851b641fc33c6288c6d613bc552, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734a31b146c7cf07705d912efe, https://i.scdn.co/image/ab67616d00001e024a31b146c7cf07705d912efe, https://i.scdn.co/image/ab67616d000048514a31b146c7cf07705d912efe, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273bc02d6e90da76957899eee36, https://i.scdn.co/image/ab67616d00001e02bc02d6e90da76957899eee36, https://i.scdn.co/image/ab67616d00004851bc02d6e90da76957899eee36, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737e8045e318486885fe243817, https://i.scdn.co/image/ab67616d00001e027e8045e318486885fe243817, https://i.scdn.co/image/ab67616d000048517e8045e318486885fe243817, 640, 300, 64
#>                                           album.name album.release_date
#> 1                                           Heathens         2016-06-16
#> 2                                    Cosmo's Factory         1970-07-25
#> 3                               The Sound of Madness         2008-06-23
#> 4                        A Rush of Blood to the Head         2002-08-08
#> 5                  Physical Graffiti (1994 Remaster)         1975-02-24
#> 6           Hi Infidelity (30th Anniversary Edition)               1980
#> 7                                Use Your Illusion I         1991-09-17
#> 8                        Extreme II - Pornograffitti         1990-01-01
#> 9                                 Threat to Survival         2015-09-18
#> 10                               Death of a Bachelor         2016-01-15
#> 11                                        Audioslave               2002
#> 12                                         Weathered         2001-01-01
#> 13                        What Separates Me From You               2010
#> 14        All We Know Of Heaven, All We Need Of Hell         2017-08-25
#> 15                                      My Own Grave         2018-06-08
#> 16                                        Disgusting         2014-06-10
#> 17                           A Decade Of Destruction         2017-12-01
#> 18 Chocolate Starfish And The Hot Dog Flavored Water         2000-10-17
#> 19                                 Concrete and Gold         2017-09-15
#> 20                                         Departure               1980
#>    album.release_date_precision album.total_tracks album.type
#> 1                           day                  1      album
#> 2                           day                 11      album
#> 3                           day                 11      album
#> 4                           day                 11      album
#> 5                           day                 15      album
#> 6                          year                 19      album
#> 7                           day                 16      album
#> 8                           day                 13      album
#> 9                           day                 11      album
#> 10                          day                 11      album
#> 11                         year                 14      album
#> 12                          day                 11      album
#> 13                         year                 10      album
#> 14                          day                 10      album
#> 15                          day                  1      album
#> 16                          day                 12      album
#> 17                          day                 16      album
#> 18                          day                 15      album
#> 19                          day                 11      album
#> 20                         year                 14      album
#>                               album.uri
#> 1  spotify:album:3J8W9AOjQhnBLCX33m3atT
#> 2  spotify:album:6vuApmTBQPuqB31Kg8hIao
#> 3  spotify:album:661dSFpYtPulSNDZZLkokq
#> 4  spotify:album:0RHX9XECH8IVI3LNgWDpmQ
#> 5  spotify:album:1lZahjeu4AhPkg9JARZr5F
#> 6  spotify:album:0X4ZNTZw7SYgrp5rlBQC3N
#> 7  spotify:album:0CxPbTRARqKUYighiEY9Sz
#> 8  spotify:album:7DKHQxJTI32UyCdDdGwvRC
#> 9  spotify:album:7969zwJB8p021EBcCUjsv9
#> 10 spotify:album:6twKQ0EsUJHVlAr6XBylrj
#> 11 spotify:album:7ysDeat0kDelEE7M6MUwXZ
#> 12 spotify:album:2ENlJXygQX6en4iziijet6
#> 13 spotify:album:2kllnMabPw1hYID0MhQRK0
#> 14 spotify:album:0jLafOidtMogAbdIjOXXIe
#> 15 spotify:album:4zSeVXasb7coGxo7ymD3EM
#> 16 spotify:album:0cq8sFikYGr4vv5ZVDr9qg
#> 17 spotify:album:6Q6erxuA6kqfBobtkmO5pH
#> 18 spotify:album:5mi7FKaWE5CtcOjdyxScA7
#> 19 spotify:album:6KMkuqIwKkwUhUYRPL6dUc
#> 20 spotify:album:2OyVtIEp7O7a6o82DF4Ba5
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/3J8W9AOjQhnBLCX33m3atT      USAT21601930
#> 2  https://open.spotify.com/album/6vuApmTBQPuqB31Kg8hIao      USFI87000056
#> 3  https://open.spotify.com/album/661dSFpYtPulSNDZZLkokq      USAT20802280
#> 4  https://open.spotify.com/album/0RHX9XECH8IVI3LNgWDpmQ      GBAYE0200771
#> 5  https://open.spotify.com/album/1lZahjeu4AhPkg9JARZr5F      USAT21300969
#> 6  https://open.spotify.com/album/0X4ZNTZw7SYgrp5rlBQC3N      USSM11102620
#> 7  https://open.spotify.com/album/0CxPbTRARqKUYighiEY9Sz      USGF19141504
#> 8  https://open.spotify.com/album/7DKHQxJTI32UyCdDdGwvRC      USAM19090005
#> 9  https://open.spotify.com/album/7969zwJB8p021EBcCUjsv9      USAT21501541
#> 10 https://open.spotify.com/album/6twKQ0EsUJHVlAr6XBylrj      USAT21503695
#> 11 https://open.spotify.com/album/7ysDeat0kDelEE7M6MUwXZ      USSM10211592
#> 12 https://open.spotify.com/album/2ENlJXygQX6en4iziijet6      USWU30200001
#> 13 https://open.spotify.com/album/2kllnMabPw1hYID0MhQRK0      USVIC1060302
#> 14 https://open.spotify.com/album/0jLafOidtMogAbdIjOXXIe      QMRSZ1701194
#> 15 https://open.spotify.com/album/4zSeVXasb7coGxo7ymD3EM      QM4TX1859424
#> 16 https://open.spotify.com/album/0cq8sFikYGr4vv5ZVDr9qg      USP6L1400182
#> 17 https://open.spotify.com/album/6Q6erxuA6kqfBobtkmO5pH      TCADH1751905
#> 18 https://open.spotify.com/album/5mi7FKaWE5CtcOjdyxScA7      USIR10001177
#> 19 https://open.spotify.com/album/6KMkuqIwKkwUhUYRPL6dUc      USRW31700013
#> 20 https://open.spotify.com/album/2OyVtIEp7O7a6o82DF4Ba5      USSM18100114
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/6i0V12jOa3mr6uu4WYhUBr
#> 2  https://open.spotify.com/track/1YqJSpIxdavjQ8BxAttU0r
#> 3  https://open.spotify.com/track/1TRr8Pp2uOqL24fMHpJyV3
#> 4  https://open.spotify.com/track/0BCPKOYdS2jbQ8iyB56Zns
#> 5  https://open.spotify.com/track/6Vjk8MNXpQpi0F4BefdTyq
#> 6  https://open.spotify.com/track/5gys5nzVQIYhgHIfiOJYva
#> 7  https://open.spotify.com/track/2N2yrmodOnVF10mKvItC9P
#> 8  https://open.spotify.com/track/1gVgkQFOKa8Wc1HYsJtPdH
#> 9  https://open.spotify.com/track/4LFBVDJV0xpayUvquQVzhS
#> 10 https://open.spotify.com/track/2fh3bZ8jZhMxOcfESE9nQY
#> 11 https://open.spotify.com/track/3bwAS8cN7Xtzvv5UurJOOX
#> 12 https://open.spotify.com/track/0eFklulxFnNCypNSrywzSB
#> 13 https://open.spotify.com/track/4JsDHMv5PVO8N07DbDq33r
#> 14 https://open.spotify.com/track/7eN3lGwu6SSmwYGH9PlpYa
#> 15 https://open.spotify.com/track/0CcqWuAEJC93K8cBMbAjgI
#> 16 https://open.spotify.com/track/03jdl5GTBhKsgB1Yw8mVPh
#> 17 https://open.spotify.com/track/0h8VX60lat4433GnbP6vQs
#> 18 https://open.spotify.com/track/2gSVKxPDww9Eep5rdvtdem
#> 19 https://open.spotify.com/track/1wLQwg0mloy3yXjL0jPE0N
#> 20 https://open.spotify.com/track/71SvEDmsOwIWw1IozsZoMA
#>                                            linked_from.href
#> 1                                                      <NA>
#> 2  https://api.spotify.com/v1/tracks/3GqeZK1VfO6z9MS3qmCmJk
#> 3                                                      <NA>
#> 4                                                      <NA>
#> 5  https://api.spotify.com/v1/tracks/2nVHqZbOGkKWzlcy1aMbE7
#> 6                                                      <NA>
#> 7                                                      <NA>
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#> 11 https://api.spotify.com/v1/tracks/4Im33JWLEobPtMTTAw7GUF
#> 12 https://api.spotify.com/v1/tracks/42T2QQv3xgBlpQxaSP7lnK
#> 13 https://api.spotify.com/v1/tracks/3imITUqEhkebNbOoS1yrWm
#> 14                                                     <NA>
#> 15 https://api.spotify.com/v1/tracks/2QcHIZFd5Meeroa1HB1L90
#> 16 https://api.spotify.com/v1/tracks/58jVSXqb2cypfC6fwxrulm
#> 17                                                     <NA>
#> 18                                                     <NA>
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1                    <NA>             <NA>                                 <NA>
#> 2  3GqeZK1VfO6z9MS3qmCmJk            track spotify:track:3GqeZK1VfO6z9MS3qmCmJk
#> 3                    <NA>             <NA>                                 <NA>
#> 4                    <NA>             <NA>                                 <NA>
#> 5  2nVHqZbOGkKWzlcy1aMbE7            track spotify:track:2nVHqZbOGkKWzlcy1aMbE7
#> 6                    <NA>             <NA>                                 <NA>
#> 7                    <NA>             <NA>                                 <NA>
#> 8                    <NA>             <NA>                                 <NA>
#> 9                    <NA>             <NA>                                 <NA>
#> 10                   <NA>             <NA>                                 <NA>
#> 11 4Im33JWLEobPtMTTAw7GUF            track spotify:track:4Im33JWLEobPtMTTAw7GUF
#> 12 42T2QQv3xgBlpQxaSP7lnK            track spotify:track:42T2QQv3xgBlpQxaSP7lnK
#> 13 3imITUqEhkebNbOoS1yrWm            track spotify:track:3imITUqEhkebNbOoS1yrWm
#> 14                   <NA>             <NA>                                 <NA>
#> 15 2QcHIZFd5Meeroa1HB1L90            track spotify:track:2QcHIZFd5Meeroa1HB1L90
#> 16 58jVSXqb2cypfC6fwxrulm            track spotify:track:58jVSXqb2cypfC6fwxrulm
#> 17                   <NA>             <NA>                                 <NA>
#> 18                   <NA>             <NA>                                 <NA>
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1                                                   <NA>
#> 2  https://open.spotify.com/track/3GqeZK1VfO6z9MS3qmCmJk
#> 3                                                   <NA>
#> 4                                                   <NA>
#> 5  https://open.spotify.com/track/2nVHqZbOGkKWzlcy1aMbE7
#> 6                                                   <NA>
#> 7                                                   <NA>
#> 8                                                   <NA>
#> 9                                                   <NA>
#> 10                                                  <NA>
#> 11 https://open.spotify.com/track/4Im33JWLEobPtMTTAw7GUF
#> 12 https://open.spotify.com/track/42T2QQv3xgBlpQxaSP7lnK
#> 13 https://open.spotify.com/track/3imITUqEhkebNbOoS1yrWm
#> 14                                                  <NA>
#> 15 https://open.spotify.com/track/2QcHIZFd5Meeroa1HB1L90
#> 16 https://open.spotify.com/track/58jVSXqb2cypfC6fwxrulm
#> 17                                                  <NA>
#> 18                                                  <NA>
#> 19                                                  <NA>
#> 20                                                  <NA>
# }