Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit
Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market
Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists
A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres
A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks
A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms
Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key
Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode
Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity
Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo
Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature
Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms
Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key
Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode
Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity
Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo
Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature
Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms
Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key
Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode
Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity
Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo
Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature
Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response
Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

 get_recommendations(market = 'SE', seed_genres = 'rock') 

#>                                                                                                                                                                                                                                                                                                                                                                                                    artists
#> 1                                                                                                                                                                                                    https://api.spotify.com/v1/artists/6kACVPfCOnqzgfEF5ryl0x, 6kACVPfCOnqzgfEF5ryl0x, Johnny Cash, artist, spotify:artist:6kACVPfCOnqzgfEF5ryl0x, https://open.spotify.com/artist/6kACVPfCOnqzgfEF5ryl0x
#> 2                                                                                                                                                                                                    https://api.spotify.com/v1/artists/6XyY86QOPPrYVGvF9ch6wz, 6XyY86QOPPrYVGvF9ch6wz, Linkin Park, artist, spotify:artist:6XyY86QOPPrYVGvF9ch6wz, https://open.spotify.com/artist/6XyY86QOPPrYVGvF9ch6wz
#> 3                                                                                                                                                                                                           https://api.spotify.com/v1/artists/3RNrq3jvMZxD9ZyoOZbQOD, 3RNrq3jvMZxD9ZyoOZbQOD, Korn, artist, spotify:artist:3RNrq3jvMZxD9ZyoOZbQOD, https://open.spotify.com/artist/3RNrq3jvMZxD9ZyoOZbQOD
#> 4                                                                                                                                                                                                          https://api.spotify.com/v1/artists/6FQqZYVfTNQ1pCqfkwVFEa, 6FQqZYVfTNQ1pCqfkwVFEa, Foals, artist, spotify:artist:6FQqZYVfTNQ1pCqfkwVFEa, https://open.spotify.com/artist/6FQqZYVfTNQ1pCqfkwVFEa
#> 5                                                                                                                                                                                          https://api.spotify.com/v1/artists/18H0sAptzdwid08XGg1Lcj, 18H0sAptzdwid08XGg1Lcj, Death From Above 1979, artist, spotify:artist:18H0sAptzdwid08XGg1Lcj, https://open.spotify.com/artist/18H0sAptzdwid08XGg1Lcj
#> 6                                                                                                                                                                                                        https://api.spotify.com/v1/artists/0EdvGhlC1FkGItLOWQzG4J, 0EdvGhlC1FkGItLOWQzG4J, Sublime, artist, spotify:artist:0EdvGhlC1FkGItLOWQzG4J, https://open.spotify.com/artist/0EdvGhlC1FkGItLOWQzG4J
#> 7                                                                                                                                                                                                 https://api.spotify.com/v1/artists/2pqd3HsfsvcBGtHvPOg6eg, 2pqd3HsfsvcBGtHvPOg6eg, Highly Suspect, artist, spotify:artist:2pqd3HsfsvcBGtHvPOg6eg, https://open.spotify.com/artist/2pqd3HsfsvcBGtHvPOg6eg
#> 8                                                                                                                                                                                                   https://api.spotify.com/v1/artists/1yAwtBaoHLEDWAnWR87hBT, 1yAwtBaoHLEDWAnWR87hBT, Modest Mouse, artist, spotify:artist:1yAwtBaoHLEDWAnWR87hBT, https://open.spotify.com/artist/1yAwtBaoHLEDWAnWR87hBT
#> 9                                                                                                                                                                                       https://api.spotify.com/v1/artists/4rjlerN21ygkIhmUv55irs, 4rjlerN21ygkIhmUv55irs, The Jesus and Mary Chain, artist, spotify:artist:4rjlerN21ygkIhmUv55irs, https://open.spotify.com/artist/4rjlerN21ygkIhmUv55irs
#> 10                                                                                                                                                                                                            https://api.spotify.com/v1/artists/51Blml2LZPmy7TTiAg47vQ, 51Blml2LZPmy7TTiAg47vQ, U2, artist, spotify:artist:51Blml2LZPmy7TTiAg47vQ, https://open.spotify.com/artist/51Blml2LZPmy7TTiAg47vQ
#> 11                                                                                                                                                                                               https://api.spotify.com/v1/artists/3lPQ2Fk5JOwGWAF3ORFCqH, 3lPQ2Fk5JOwGWAF3ORFCqH, John Mellencamp, artist, spotify:artist:3lPQ2Fk5JOwGWAF3ORFCqH, https://open.spotify.com/artist/3lPQ2Fk5JOwGWAF3ORFCqH
#> 12                                                                                                                                                                                                     https://api.spotify.com/v1/artists/7Ey4PD4MYsKc5I2dolUwbH, 7Ey4PD4MYsKc5I2dolUwbH, Aerosmith, artist, spotify:artist:7Ey4PD4MYsKc5I2dolUwbH, https://open.spotify.com/artist/7Ey4PD4MYsKc5I2dolUwbH
#> 13                                                                                                                                                                                                      https://api.spotify.com/v1/artists/11wRdbnoYqRddKBrpHt4Ue, 11wRdbnoYqRddKBrpHt4Ue, Kasabian, artist, spotify:artist:11wRdbnoYqRddKBrpHt4Ue, https://open.spotify.com/artist/11wRdbnoYqRddKBrpHt4Ue
#> 14 https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, https://api.spotify.com/v1/artists/3aBkeBhwadnWMWoVJ2CxJC, 6GI52t8N5F02MxU0g5U69P, 3aBkeBhwadnWMWoVJ2CxJC, Santana, Rob Thomas, artist, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, spotify:artist:3aBkeBhwadnWMWoVJ2CxJC, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/3aBkeBhwadnWMWoVJ2CxJC
#> 15                                                                                                                                                                                               https://api.spotify.com/v1/artists/4NpFxQe2UvRCAjto3JqlSl, 4NpFxQe2UvRCAjto3JqlSl, Greta Van Fleet, artist, spotify:artist:4NpFxQe2UvRCAjto3JqlSl, https://open.spotify.com/artist/4NpFxQe2UvRCAjto3JqlSl
#> 16                                                                                                                                                                                                        https://api.spotify.com/v1/artists/3jOstUTkEu2JkjvRdBA5Gu, 3jOstUTkEu2JkjvRdBA5Gu, Weezer, artist, spotify:artist:3jOstUTkEu2JkjvRdBA5Gu, https://open.spotify.com/artist/3jOstUTkEu2JkjvRdBA5Gu
#> 17                                                                                                                                                                                             https://api.spotify.com/v1/artists/2sil8z5kiy4r76CRTXxBCA, 2sil8z5kiy4r76CRTXxBCA, The Goo Goo Dolls, artist, spotify:artist:2sil8z5kiy4r76CRTXxBCA, https://open.spotify.com/artist/2sil8z5kiy4r76CRTXxBCA
#> 18                                                                                                                                                                                                        https://api.spotify.com/v1/artists/6KO6G41BBLTDNYOLefWTMU, 6KO6G41BBLTDNYOLefWTMU, P.O.D., artist, spotify:artist:6KO6G41BBLTDNYOLefWTMU, https://open.spotify.com/artist/6KO6G41BBLTDNYOLefWTMU
#> 19                                                                                                                                                                                                https://api.spotify.com/v1/artists/4MVyzYMgTwdP7Z49wAZHx0, 4MVyzYMgTwdP7Z49wAZHx0, Lynyrd Skynyrd, artist, spotify:artist:4MVyzYMgTwdP7Z49wAZHx0, https://open.spotify.com/artist/4MVyzYMgTwdP7Z49wAZHx0
#> 20                                                                                                                                                                                                        https://api.spotify.com/v1/artists/4KWTAlx2RvbpseOGMEmROg, 4KWTAlx2RvbpseOGMEmROg, R.E.M., artist, spotify:artist:4KWTAlx2RvbpseOGMEmROg, https://open.spotify.com/artist/4KWTAlx2RvbpseOGMEmROg
#>    disc_number duration_ms explicit
#> 1            1      216533    FALSE
#> 2            1      255066    FALSE
#> 3            1      255733     TRUE
#> 4            1      300720    FALSE
#> 5            1      198081    FALSE
#> 6            1      171066     TRUE
#> 7            1      416266     TRUE
#> 8            1      208466    FALSE
#> 9            1      203933    FALSE
#> 10           1      277476    FALSE
#> 11           1      254493    FALSE
#> 12           1      308333    FALSE
#> 13           1      215693    FALSE
#> 14           1      294986    FALSE
#> 15           1      267333    FALSE
#> 16           1      195893    FALSE
#> 17           1      289533    FALSE
#> 18           1      205026    FALSE
#> 19           1      450226    FALSE
#> 20           1      246933    FALSE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/28cnXtME493VX9NOw9cIUh
#> 2  https://api.spotify.com/v1/tracks/3xXBsjrbG1xQIm1xv1cKOt
#> 3  https://api.spotify.com/v1/tracks/6W21LNLz9Sw7sUSNWMSHRu
#> 4  https://api.spotify.com/v1/tracks/78tgXRq9Q6tPNP9hKCpgwB
#> 5  https://api.spotify.com/v1/tracks/50jXnyrzW4mcBBqFTMEoAC
#> 6  https://api.spotify.com/v1/tracks/5nkYHBWXu7KCgcdwP6jcB1
#> 7  https://api.spotify.com/v1/tracks/2S7R4trL4HITPIfvKUxcA4
#> 8  https://api.spotify.com/v1/tracks/2lwwrWVKdf3LR9lbbhnr6R
#> 9  https://api.spotify.com/v1/tracks/4OnhAQQnmzMMD3mogOAfQC
#> 10 https://api.spotify.com/v1/tracks/6wpGqhRvJGNNXwWlPmkMyO
#> 11 https://api.spotify.com/v1/tracks/43btz2xjMKpcmjkuRsvxyg
#> 12 https://api.spotify.com/v1/tracks/0NJC0FDCODpPUntRTTQq97
#> 13 https://api.spotify.com/v1/tracks/0AJhcuRl3i1FfPNr88ZScv
#> 14 https://api.spotify.com/v1/tracks/0n2SEXB2qoRQg171q7XqeW
#> 15 https://api.spotify.com/v1/tracks/4f9UzrnxHsWKIheG1TRrEG
#> 16 https://api.spotify.com/v1/tracks/2jz1bw1p0WQj0PDnVDP0uY
#> 17 https://api.spotify.com/v1/tracks/6Qyc6fS4DsZjB2mRW9DsQs
#> 18 https://api.spotify.com/v1/tracks/1X4Ntw6Lbaa1ACgilCqMpr
#> 19 https://api.spotify.com/v1/tracks/64sn7wgJfeQKY3Jisnl59w
#> 20 https://api.spotify.com/v1/tracks/2oSpQ7QtIKTNFfA08Cy0ku
#>                        id is_local is_playable
#> 1  28cnXtME493VX9NOw9cIUh    FALSE        TRUE
#> 2  3xXBsjrbG1xQIm1xv1cKOt    FALSE        TRUE
#> 3  6W21LNLz9Sw7sUSNWMSHRu    FALSE        TRUE
#> 4  78tgXRq9Q6tPNP9hKCpgwB    FALSE        TRUE
#> 5  50jXnyrzW4mcBBqFTMEoAC    FALSE        TRUE
#> 6  5nkYHBWXu7KCgcdwP6jcB1    FALSE        TRUE
#> 7  2S7R4trL4HITPIfvKUxcA4    FALSE        TRUE
#> 8  2lwwrWVKdf3LR9lbbhnr6R    FALSE        TRUE
#> 9  4OnhAQQnmzMMD3mogOAfQC    FALSE        TRUE
#> 10 6wpGqhRvJGNNXwWlPmkMyO    FALSE        TRUE
#> 11 43btz2xjMKpcmjkuRsvxyg    FALSE        TRUE
#> 12 0NJC0FDCODpPUntRTTQq97    FALSE        TRUE
#> 13 0AJhcuRl3i1FfPNr88ZScv    FALSE        TRUE
#> 14 0n2SEXB2qoRQg171q7XqeW    FALSE        TRUE
#> 15 4f9UzrnxHsWKIheG1TRrEG    FALSE        TRUE
#> 16 2jz1bw1p0WQj0PDnVDP0uY    FALSE        TRUE
#> 17 6Qyc6fS4DsZjB2mRW9DsQs    FALSE        TRUE
#> 18 1X4Ntw6Lbaa1ACgilCqMpr    FALSE        TRUE
#> 19 64sn7wgJfeQKY3Jisnl59w    FALSE        TRUE
#> 20 2oSpQ7QtIKTNFfA08Cy0ku    FALSE        TRUE
#>                                                         name popularity
#> 1                                                       Hurt         75
#> 2                                             One More Light         71
#> 3                                           Freak On a Leash         75
#> 4                                             What Went Down         57
#> 5                                                  Freeze Me         60
#> 6                                                 What I Got         50
#> 7                                                       Wolf         49
#> 8                                                   Float On         72
#> 9                                                 Amputation         34
#> 10                I Still Haven't Found What I'm Looking For         79
#> 11                                              Jack & Diane         75
#> 12                                                    Cryin'         72
#> 13                              You're in Love with a Psycho         60
#> 14                                 Smooth (feat. Rob Thomas)         70
#> 15                                             Watching Over         54
#> 16                                         Feels Like Summer         58
#> 17                                                      Iris         79
#> 18                               Alive - Chris Lord-Alge Mix         63
#> 19                                            Tuesday's Gone         65
#> 20 It's The End Of The World As We Know It (And I Feel Fine)         69
#>                                                                                                    preview_url
#> 1                                                                                                         <NA>
#> 2  https://p.scdn.co/mp3-preview/7ea4eef2ed25e30cd90bbef3dc5ef3503ac1afeb?cid=125a10da5fa94687b3d0913168bc3698
#> 3  https://p.scdn.co/mp3-preview/14933695bf1a6f25b2c602d9bf4d774e5dcb2d6c?cid=125a10da5fa94687b3d0913168bc3698
#> 4  https://p.scdn.co/mp3-preview/50c56f59a12d26bd77b8071466bca0baaeca4981?cid=125a10da5fa94687b3d0913168bc3698
#> 5  https://p.scdn.co/mp3-preview/80f76475f28744ee4823b6905d5f6dbef0545885?cid=125a10da5fa94687b3d0913168bc3698
#> 6                                                                                                         <NA>
#> 7  https://p.scdn.co/mp3-preview/38c4d01cc0ac0d1c7de312a66ff989da3f5fac07?cid=125a10da5fa94687b3d0913168bc3698
#> 8  https://p.scdn.co/mp3-preview/0d741d740ea0e449057d326439a47e974447e86d?cid=125a10da5fa94687b3d0913168bc3698
#> 9  https://p.scdn.co/mp3-preview/e7502dfdc7d8f084450f597df332542473dcc486?cid=125a10da5fa94687b3d0913168bc3698
#> 10                                                                                                        <NA>
#> 11                                                                                                        <NA>
#> 12                                                                                                        <NA>
#> 13 https://p.scdn.co/mp3-preview/7607e607e030341b0c34813ea0adb916eff66e92?cid=125a10da5fa94687b3d0913168bc3698
#> 14 https://p.scdn.co/mp3-preview/9f977ebb2e695502c8b1d3bd678ef706c9ea4ccb?cid=125a10da5fa94687b3d0913168bc3698
#> 15                                                                                                        <NA>
#> 16 https://p.scdn.co/mp3-preview/530a0fa1eaf7bc59c46b4695610be07e7ce80d12?cid=125a10da5fa94687b3d0913168bc3698
#> 17 https://p.scdn.co/mp3-preview/0e87ade9fa869e31b3c32115c2c10ee8fa522c81?cid=125a10da5fa94687b3d0913168bc3698
#> 18 https://p.scdn.co/mp3-preview/c3564adb4b6c07a51767b2d339f1bf512269ade5?cid=125a10da5fa94687b3d0913168bc3698
#> 19                                                                                                        <NA>
#> 20                                                                                                        <NA>
#>    track_number  type                                  uri album.album_type
#> 1             2 track spotify:track:28cnXtME493VX9NOw9cIUh            ALBUM
#> 2             9 track spotify:track:3xXBsjrbG1xQIm1xv1cKOt            ALBUM
#> 3             2 track spotify:track:6W21LNLz9Sw7sUSNWMSHRu            ALBUM
#> 4             1 track spotify:track:78tgXRq9Q6tPNP9hKCpgwB            ALBUM
#> 5             2 track spotify:track:50jXnyrzW4mcBBqFTMEoAC            ALBUM
#> 6             2 track spotify:track:5nkYHBWXu7KCgcdwP6jcB1            ALBUM
#> 7            11 track spotify:track:2S7R4trL4HITPIfvKUxcA4            ALBUM
#> 8             3 track spotify:track:2lwwrWVKdf3LR9lbbhnr6R            ALBUM
#> 9             1 track spotify:track:4OnhAQQnmzMMD3mogOAfQC            ALBUM
#> 10            2 track spotify:track:6wpGqhRvJGNNXwWlPmkMyO            ALBUM
#> 11            2 track spotify:track:43btz2xjMKpcmjkuRsvxyg            ALBUM
#> 12            9 track spotify:track:0NJC0FDCODpPUntRTTQq97            ALBUM
#> 13            2 track spotify:track:0AJhcuRl3i1FfPNr88ZScv            ALBUM
#> 14            5 track spotify:track:0n2SEXB2qoRQg171q7XqeW            ALBUM
#> 15            4 track spotify:track:4f9UzrnxHsWKIheG1TRrEG            ALBUM
#> 16            3 track spotify:track:2jz1bw1p0WQj0PDnVDP0uY            ALBUM
#> 17           11 track spotify:track:6Qyc6fS4DsZjB2mRW9DsQs            ALBUM
#> 18            2 track spotify:track:1X4Ntw6Lbaa1ACgilCqMpr            ALBUM
#> 19            2 track spotify:track:64sn7wgJfeQKY3Jisnl59w            ALBUM
#> 20            6 track spotify:track:2oSpQ7QtIKTNFfA08Cy0ku            ALBUM
#>                                                                                                                                                                                                         album.artists
#> 1               https://api.spotify.com/v1/artists/6kACVPfCOnqzgfEF5ryl0x, 6kACVPfCOnqzgfEF5ryl0x, Johnny Cash, artist, spotify:artist:6kACVPfCOnqzgfEF5ryl0x, https://open.spotify.com/artist/6kACVPfCOnqzgfEF5ryl0x
#> 2               https://api.spotify.com/v1/artists/6XyY86QOPPrYVGvF9ch6wz, 6XyY86QOPPrYVGvF9ch6wz, Linkin Park, artist, spotify:artist:6XyY86QOPPrYVGvF9ch6wz, https://open.spotify.com/artist/6XyY86QOPPrYVGvF9ch6wz
#> 3                      https://api.spotify.com/v1/artists/3RNrq3jvMZxD9ZyoOZbQOD, 3RNrq3jvMZxD9ZyoOZbQOD, Korn, artist, spotify:artist:3RNrq3jvMZxD9ZyoOZbQOD, https://open.spotify.com/artist/3RNrq3jvMZxD9ZyoOZbQOD
#> 4                     https://api.spotify.com/v1/artists/6FQqZYVfTNQ1pCqfkwVFEa, 6FQqZYVfTNQ1pCqfkwVFEa, Foals, artist, spotify:artist:6FQqZYVfTNQ1pCqfkwVFEa, https://open.spotify.com/artist/6FQqZYVfTNQ1pCqfkwVFEa
#> 5     https://api.spotify.com/v1/artists/18H0sAptzdwid08XGg1Lcj, 18H0sAptzdwid08XGg1Lcj, Death From Above 1979, artist, spotify:artist:18H0sAptzdwid08XGg1Lcj, https://open.spotify.com/artist/18H0sAptzdwid08XGg1Lcj
#> 6                   https://api.spotify.com/v1/artists/0EdvGhlC1FkGItLOWQzG4J, 0EdvGhlC1FkGItLOWQzG4J, Sublime, artist, spotify:artist:0EdvGhlC1FkGItLOWQzG4J, https://open.spotify.com/artist/0EdvGhlC1FkGItLOWQzG4J
#> 7            https://api.spotify.com/v1/artists/2pqd3HsfsvcBGtHvPOg6eg, 2pqd3HsfsvcBGtHvPOg6eg, Highly Suspect, artist, spotify:artist:2pqd3HsfsvcBGtHvPOg6eg, https://open.spotify.com/artist/2pqd3HsfsvcBGtHvPOg6eg
#> 8              https://api.spotify.com/v1/artists/1yAwtBaoHLEDWAnWR87hBT, 1yAwtBaoHLEDWAnWR87hBT, Modest Mouse, artist, spotify:artist:1yAwtBaoHLEDWAnWR87hBT, https://open.spotify.com/artist/1yAwtBaoHLEDWAnWR87hBT
#> 9  https://api.spotify.com/v1/artists/4rjlerN21ygkIhmUv55irs, 4rjlerN21ygkIhmUv55irs, The Jesus and Mary Chain, artist, spotify:artist:4rjlerN21ygkIhmUv55irs, https://open.spotify.com/artist/4rjlerN21ygkIhmUv55irs
#> 10                       https://api.spotify.com/v1/artists/51Blml2LZPmy7TTiAg47vQ, 51Blml2LZPmy7TTiAg47vQ, U2, artist, spotify:artist:51Blml2LZPmy7TTiAg47vQ, https://open.spotify.com/artist/51Blml2LZPmy7TTiAg47vQ
#> 11          https://api.spotify.com/v1/artists/3lPQ2Fk5JOwGWAF3ORFCqH, 3lPQ2Fk5JOwGWAF3ORFCqH, John Mellencamp, artist, spotify:artist:3lPQ2Fk5JOwGWAF3ORFCqH, https://open.spotify.com/artist/3lPQ2Fk5JOwGWAF3ORFCqH
#> 12                https://api.spotify.com/v1/artists/7Ey4PD4MYsKc5I2dolUwbH, 7Ey4PD4MYsKc5I2dolUwbH, Aerosmith, artist, spotify:artist:7Ey4PD4MYsKc5I2dolUwbH, https://open.spotify.com/artist/7Ey4PD4MYsKc5I2dolUwbH
#> 13                 https://api.spotify.com/v1/artists/11wRdbnoYqRddKBrpHt4Ue, 11wRdbnoYqRddKBrpHt4Ue, Kasabian, artist, spotify:artist:11wRdbnoYqRddKBrpHt4Ue, https://open.spotify.com/artist/11wRdbnoYqRddKBrpHt4Ue
#> 14                  https://api.spotify.com/v1/artists/6GI52t8N5F02MxU0g5U69P, 6GI52t8N5F02MxU0g5U69P, Santana, artist, spotify:artist:6GI52t8N5F02MxU0g5U69P, https://open.spotify.com/artist/6GI52t8N5F02MxU0g5U69P
#> 15          https://api.spotify.com/v1/artists/4NpFxQe2UvRCAjto3JqlSl, 4NpFxQe2UvRCAjto3JqlSl, Greta Van Fleet, artist, spotify:artist:4NpFxQe2UvRCAjto3JqlSl, https://open.spotify.com/artist/4NpFxQe2UvRCAjto3JqlSl
#> 16                   https://api.spotify.com/v1/artists/3jOstUTkEu2JkjvRdBA5Gu, 3jOstUTkEu2JkjvRdBA5Gu, Weezer, artist, spotify:artist:3jOstUTkEu2JkjvRdBA5Gu, https://open.spotify.com/artist/3jOstUTkEu2JkjvRdBA5Gu
#> 17        https://api.spotify.com/v1/artists/2sil8z5kiy4r76CRTXxBCA, 2sil8z5kiy4r76CRTXxBCA, The Goo Goo Dolls, artist, spotify:artist:2sil8z5kiy4r76CRTXxBCA, https://open.spotify.com/artist/2sil8z5kiy4r76CRTXxBCA
#> 18                   https://api.spotify.com/v1/artists/6KO6G41BBLTDNYOLefWTMU, 6KO6G41BBLTDNYOLefWTMU, P.O.D., artist, spotify:artist:6KO6G41BBLTDNYOLefWTMU, https://open.spotify.com/artist/6KO6G41BBLTDNYOLefWTMU
#> 19           https://api.spotify.com/v1/artists/4MVyzYMgTwdP7Z49wAZHx0, 4MVyzYMgTwdP7Z49wAZHx0, Lynyrd Skynyrd, artist, spotify:artist:4MVyzYMgTwdP7Z49wAZHx0, https://open.spotify.com/artist/4MVyzYMgTwdP7Z49wAZHx0
#> 20                   https://api.spotify.com/v1/artists/4KWTAlx2RvbpseOGMEmROg, 4KWTAlx2RvbpseOGMEmROg, R.E.M., artist, spotify:artist:4KWTAlx2RvbpseOGMEmROg, https://open.spotify.com/artist/4KWTAlx2RvbpseOGMEmROg
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/2BlL4Gv2DLPu8p58Wcmlm9
#> 2  https://api.spotify.com/v1/albums/5Eevxp2BCbWq25ZdiXRwYd
#> 3  https://api.spotify.com/v1/albums/0gsiszk6JWYwAyGvaTTud4
#> 4  https://api.spotify.com/v1/albums/0RyCpIKlCV0kgEuzrmp73O
#> 5  https://api.spotify.com/v1/albums/4kKKH6cFBtNd9etSJzkMjZ
#> 6  https://api.spotify.com/v1/albums/4R3tXoorBpHji6Jdms8a4Q
#> 7  https://api.spotify.com/v1/albums/3TC40H9dIJArFzy0rWnWCg
#> 8  https://api.spotify.com/v1/albums/0TGTGuc2vXv6ZECoAf52N0
#> 9  https://api.spotify.com/v1/albums/4ew0FYIDjOmGQVWmaWfvsb
#> 10 https://api.spotify.com/v1/albums/5y6wlw1LnqFnQFruMeiwGU
#> 11 https://api.spotify.com/v1/albums/4gouGcdQn9OvjX42xnWrF0
#> 12 https://api.spotify.com/v1/albums/1tuM8yBePaekEruGsH2J79
#> 13 https://api.spotify.com/v1/albums/0ymsF6XO1mZM1VKqI680jU
#> 14 https://api.spotify.com/v1/albums/10aiDpdFGyfCFEcqpx6XTq
#> 15 https://api.spotify.com/v1/albums/7zeCZY6rQRufc8IHGKyXGX
#> 16 https://api.spotify.com/v1/albums/3EwfQtjvyRAXsPWAKO5FDP
#> 17 https://api.spotify.com/v1/albums/4UMjBXcRqIgMZ1XumU2x5T
#> 18 https://api.spotify.com/v1/albums/4eUmncMYQRg0Ot29KKqbuk
#> 19 https://api.spotify.com/v1/albums/6DExt1eX4lflLacVjHHbOs
#> 20 https://api.spotify.com/v1/albums/6gMv3MgFlieOM6Uz5GZBzy
#>                  album.id
#> 1  2BlL4Gv2DLPu8p58Wcmlm9
#> 2  5Eevxp2BCbWq25ZdiXRwYd
#> 3  0gsiszk6JWYwAyGvaTTud4
#> 4  0RyCpIKlCV0kgEuzrmp73O
#> 5  4kKKH6cFBtNd9etSJzkMjZ
#> 6  4R3tXoorBpHji6Jdms8a4Q
#> 7  3TC40H9dIJArFzy0rWnWCg
#> 8  0TGTGuc2vXv6ZECoAf52N0
#> 9  4ew0FYIDjOmGQVWmaWfvsb
#> 10 5y6wlw1LnqFnQFruMeiwGU
#> 11 4gouGcdQn9OvjX42xnWrF0
#> 12 1tuM8yBePaekEruGsH2J79
#> 13 0ymsF6XO1mZM1VKqI680jU
#> 14 10aiDpdFGyfCFEcqpx6XTq
#> 15 7zeCZY6rQRufc8IHGKyXGX
#> 16 3EwfQtjvyRAXsPWAKO5FDP
#> 17 4UMjBXcRqIgMZ1XumU2x5T
#> 18 4eUmncMYQRg0Ot29KKqbuk
#> 19 6DExt1eX4lflLacVjHHbOs
#> 20 6gMv3MgFlieOM6Uz5GZBzy
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736f4f62da3d811b6501a69ffa, https://i.scdn.co/image/ab67616d00001e026f4f62da3d811b6501a69ffa, https://i.scdn.co/image/ab67616d000048516f4f62da3d811b6501a69ffa, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273145e1cf12538f5666511237d, https://i.scdn.co/image/ab67616d00001e02145e1cf12538f5666511237d, https://i.scdn.co/image/ab67616d00004851145e1cf12538f5666511237d, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27350d216aebaf98e8ac9947fd5, https://i.scdn.co/image/ab67616d00001e0250d216aebaf98e8ac9947fd5, https://i.scdn.co/image/ab67616d0000485150d216aebaf98e8ac9947fd5, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b036f68e97ce9f5372bfb350, https://i.scdn.co/image/ab67616d00001e02b036f68e97ce9f5372bfb350, https://i.scdn.co/image/ab67616d00004851b036f68e97ce9f5372bfb350, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734fc70ffe01a56e966101b140, https://i.scdn.co/image/ab67616d00001e024fc70ffe01a56e966101b140, https://i.scdn.co/image/ab67616d000048514fc70ffe01a56e966101b140, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c4cb3ab94558b3893b1ccbd9, https://i.scdn.co/image/ab67616d00001e02c4cb3ab94558b3893b1ccbd9, https://i.scdn.co/image/ab67616d00004851c4cb3ab94558b3893b1ccbd9, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2730412a482b98670e4d98a55fd, https://i.scdn.co/image/ab67616d00001e020412a482b98670e4d98a55fd, https://i.scdn.co/image/ab67616d000048510412a482b98670e4d98a55fd, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273cc68329bfbf34037df965dc1, https://i.scdn.co/image/ab67616d00001e02cc68329bfbf34037df965dc1, https://i.scdn.co/image/ab67616d00004851cc68329bfbf34037df965dc1, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734133a0db3e6a30222e1e099d, https://i.scdn.co/image/ab67616d00001e024133a0db3e6a30222e1e099d, https://i.scdn.co/image/ab67616d000048514133a0db3e6a30222e1e099d, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b7bea3d01f04e6d0408d2afe, https://i.scdn.co/image/ab67616d00001e02b7bea3d01f04e6d0408d2afe, https://i.scdn.co/image/ab67616d00004851b7bea3d01f04e6d0408d2afe, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b1945b44ce8868740f132f8e, https://i.scdn.co/image/ab67616d00001e02b1945b44ce8868740f132f8e, https://i.scdn.co/image/ab67616d00004851b1945b44ce8868740f132f8e, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273587b26d49596162397b29ec8, https://i.scdn.co/image/ab67616d00001e02587b26d49596162397b29ec8, https://i.scdn.co/image/ab67616d00004851587b26d49596162397b29ec8, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a3070377fc047ec055351d3b, https://i.scdn.co/image/ab67616d00001e02a3070377fc047ec055351d3b, https://i.scdn.co/image/ab67616d00004851a3070377fc047ec055351d3b, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27347eb3ea5a92904c19e102e54, https://i.scdn.co/image/ab67616d00001e0247eb3ea5a92904c19e102e54, https://i.scdn.co/image/ab67616d0000485147eb3ea5a92904c19e102e54, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2737b9e5a9d697bcb8bf86a83b4, https://i.scdn.co/image/ab67616d00001e027b9e5a9d697bcb8bf86a83b4, https://i.scdn.co/image/ab67616d000048517b9e5a9d697bcb8bf86a83b4, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f04ed0e9d5af8b1b5f011b2f, https://i.scdn.co/image/ab67616d00001e02f04ed0e9d5af8b1b5f011b2f, https://i.scdn.co/image/ab67616d00004851f04ed0e9d5af8b1b5f011b2f, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273eda9478c39a21e1cdc6609ca, https://i.scdn.co/image/ab67616d00001e02eda9478c39a21e1cdc6609ca, https://i.scdn.co/image/ab67616d00004851eda9478c39a21e1cdc6609ca, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27395741aa186187b05e0653245, https://i.scdn.co/image/ab67616d00001e0295741aa186187b05e0653245, https://i.scdn.co/image/ab67616d0000485195741aa186187b05e0653245, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273128450651c9f0442780d8eb8, https://i.scdn.co/image/ab67616d00001e02128450651c9f0442780d8eb8, https://i.scdn.co/image/ab67616d00004851128450651c9f0442780d8eb8, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a7c7f94cfccd5ab6de233916, https://i.scdn.co/image/ab67616d00001e02a7c7f94cfccd5ab6de233916, https://i.scdn.co/image/ab67616d00004851a7c7f94cfccd5ab6de233916, 640, 300, 64
#>                                    album.name album.release_date
#> 1           American IV: The Man Comes Around         2002-01-01
#> 2                              One More Light         2017-05-19
#> 3                           Follow The Leader         1998-08-18
#> 4                              What Went Down         2015-08-28
#> 5                             Outrage! Is Now         2017-09-08
#> 6                                     Sublime         1996-07-30
#> 7                       The Boy Who Died Wolf         2016-11-18
#> 8      Good News For People Who Love Bad News         2004-04-05
#> 9                              Damage And Joy         2017-04-04
#> 10             The Joshua Tree (Super Deluxe)         1987-03-03
#> 11                              American Fool               1982
#> 12                                 Get A Grip         1993-01-01
#> 13               For Crying Out Loud (Deluxe)         2017-05-05
#> 14                  Supernatural (Remastered)         1999-06-15
#> 15                Anthem Of The Peaceful Army         2018-10-19
#> 16                           Pacific Daydream         2017-10-27
#> 17                          Dizzy up the Girl         1998-09-11
#> 18 Satellite (U.S. Version; WEA Distribution)         2001-09-11
#> 19          Pronounced' Leh-'Nerd 'Skin-'Nerd         1973-01-01
#> 20                    Document (R.E.M. No. 5)         1987-09-01
#>    album.release_date_precision album.total_tracks album.type
#> 1                           day                 15      album
#> 2                           day                 10      album
#> 3                           day                 14      album
#> 4                           day                 18      album
#> 5                           day                 10      album
#> 6                           day                 23      album
#> 7                           day                 11      album
#> 8                           day                 16      album
#> 9                           day                 14      album
#> 10                          day                 49      album
#> 11                         year                 10      album
#> 12                          day                 14      album
#> 13                          day                 27      album
#> 14                          day                 14      album
#> 15                          day                 11      album
#> 16                          day                 10      album
#> 17                          day                 13      album
#> 18                          day                 15      album
#> 19                          day                  8      album
#> 20                          day                 11      album
#>                               album.uri
#> 1  spotify:album:2BlL4Gv2DLPu8p58Wcmlm9
#> 2  spotify:album:5Eevxp2BCbWq25ZdiXRwYd
#> 3  spotify:album:0gsiszk6JWYwAyGvaTTud4
#> 4  spotify:album:0RyCpIKlCV0kgEuzrmp73O
#> 5  spotify:album:4kKKH6cFBtNd9etSJzkMjZ
#> 6  spotify:album:4R3tXoorBpHji6Jdms8a4Q
#> 7  spotify:album:3TC40H9dIJArFzy0rWnWCg
#> 8  spotify:album:0TGTGuc2vXv6ZECoAf52N0
#> 9  spotify:album:4ew0FYIDjOmGQVWmaWfvsb
#> 10 spotify:album:5y6wlw1LnqFnQFruMeiwGU
#> 11 spotify:album:4gouGcdQn9OvjX42xnWrF0
#> 12 spotify:album:1tuM8yBePaekEruGsH2J79
#> 13 spotify:album:0ymsF6XO1mZM1VKqI680jU
#> 14 spotify:album:10aiDpdFGyfCFEcqpx6XTq
#> 15 spotify:album:7zeCZY6rQRufc8IHGKyXGX
#> 16 spotify:album:3EwfQtjvyRAXsPWAKO5FDP
#> 17 spotify:album:4UMjBXcRqIgMZ1XumU2x5T
#> 18 spotify:album:4eUmncMYQRg0Ot29KKqbuk
#> 19 spotify:album:6DExt1eX4lflLacVjHHbOs
#> 20 spotify:album:6gMv3MgFlieOM6Uz5GZBzy
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/2BlL4Gv2DLPu8p58Wcmlm9      USDJ20200650
#> 2  https://open.spotify.com/album/5Eevxp2BCbWq25ZdiXRwYd      USWB11700225
#> 3  https://open.spotify.com/album/0gsiszk6JWYwAyGvaTTud4      USSM19801755
#> 4  https://open.spotify.com/album/0RyCpIKlCV0kgEuzrmp73O      GBAHT1500379
#> 5  https://open.spotify.com/album/4kKKH6cFBtNd9etSJzkMjZ      CAL451785502
#> 6  https://open.spotify.com/album/4R3tXoorBpHji6Jdms8a4Q      USGA19649251
#> 7  https://open.spotify.com/album/3TC40H9dIJArFzy0rWnWCg      QMCE31600728
#> 8  https://open.spotify.com/album/0TGTGuc2vXv6ZECoAf52N0      USSM10315494
#> 9  https://open.spotify.com/album/4ew0FYIDjOmGQVWmaWfvsb      UKHNA1600001
#> 10 https://open.spotify.com/album/5y6wlw1LnqFnQFruMeiwGU      GBUM70709783
#> 11 https://open.spotify.com/album/4gouGcdQn9OvjX42xnWrF0      USJM18200001
#> 12 https://open.spotify.com/album/1tuM8yBePaekEruGsH2J79      USIR10000430
#> 13 https://open.spotify.com/album/0ymsF6XO1mZM1VKqI680jU      GBARL1700369
#> 14 https://open.spotify.com/album/10aiDpdFGyfCFEcqpx6XTq      USAR19900033
#> 15 https://open.spotify.com/album/7zeCZY6rQRufc8IHGKyXGX      USUM71811320
#> 16 https://open.spotify.com/album/3EwfQtjvyRAXsPWAKO5FDP      USAT21700603
#> 17 https://open.spotify.com/album/4UMjBXcRqIgMZ1XumU2x5T      USWB19800160
#> 18 https://open.spotify.com/album/4eUmncMYQRg0Ot29KKqbuk      USAT20106836
#> 19 https://open.spotify.com/album/6DExt1eX4lflLacVjHHbOs      USMC17347089
#> 20 https://open.spotify.com/album/6gMv3MgFlieOM6Uz5GZBzy      USIR38700037
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/28cnXtME493VX9NOw9cIUh
#> 2  https://open.spotify.com/track/3xXBsjrbG1xQIm1xv1cKOt
#> 3  https://open.spotify.com/track/6W21LNLz9Sw7sUSNWMSHRu
#> 4  https://open.spotify.com/track/78tgXRq9Q6tPNP9hKCpgwB
#> 5  https://open.spotify.com/track/50jXnyrzW4mcBBqFTMEoAC
#> 6  https://open.spotify.com/track/5nkYHBWXu7KCgcdwP6jcB1
#> 7  https://open.spotify.com/track/2S7R4trL4HITPIfvKUxcA4
#> 8  https://open.spotify.com/track/2lwwrWVKdf3LR9lbbhnr6R
#> 9  https://open.spotify.com/track/4OnhAQQnmzMMD3mogOAfQC
#> 10 https://open.spotify.com/track/6wpGqhRvJGNNXwWlPmkMyO
#> 11 https://open.spotify.com/track/43btz2xjMKpcmjkuRsvxyg
#> 12 https://open.spotify.com/track/0NJC0FDCODpPUntRTTQq97
#> 13 https://open.spotify.com/track/0AJhcuRl3i1FfPNr88ZScv
#> 14 https://open.spotify.com/track/0n2SEXB2qoRQg171q7XqeW
#> 15 https://open.spotify.com/track/4f9UzrnxHsWKIheG1TRrEG
#> 16 https://open.spotify.com/track/2jz1bw1p0WQj0PDnVDP0uY
#> 17 https://open.spotify.com/track/6Qyc6fS4DsZjB2mRW9DsQs
#> 18 https://open.spotify.com/track/1X4Ntw6Lbaa1ACgilCqMpr
#> 19 https://open.spotify.com/track/64sn7wgJfeQKY3Jisnl59w
#> 20 https://open.spotify.com/track/2oSpQ7QtIKTNFfA08Cy0ku
#>                                            linked_from.href
#> 1                                                      <NA>
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4  https://api.spotify.com/v1/tracks/5h27GYpKZWWhFov8fOunF6
#> 5  https://api.spotify.com/v1/tracks/7FeeQVg6yID8laNUH0LlX8
#> 6                                                      <NA>
#> 7                                                      <NA>
#> 8  https://api.spotify.com/v1/tracks/3b00LSFIRqnHvkoDEs5fkg
#> 9  https://api.spotify.com/v1/tracks/67L36FEtnci1yguR5s514G
#> 10                                                     <NA>
#> 11                                                     <NA>
#> 12                                                     <NA>
#> 13                                                     <NA>
#> 14                                                     <NA>
#> 15                                                     <NA>
#> 16                                                     <NA>
#> 17                                                     <NA>
#> 18                                                     <NA>
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1                    <NA>             <NA>                                 <NA>
#> 2                    <NA>             <NA>                                 <NA>
#> 3                    <NA>             <NA>                                 <NA>
#> 4  5h27GYpKZWWhFov8fOunF6            track spotify:track:5h27GYpKZWWhFov8fOunF6
#> 5  7FeeQVg6yID8laNUH0LlX8            track spotify:track:7FeeQVg6yID8laNUH0LlX8
#> 6                    <NA>             <NA>                                 <NA>
#> 7                    <NA>             <NA>                                 <NA>
#> 8  3b00LSFIRqnHvkoDEs5fkg            track spotify:track:3b00LSFIRqnHvkoDEs5fkg
#> 9  67L36FEtnci1yguR5s514G            track spotify:track:67L36FEtnci1yguR5s514G
#> 10                   <NA>             <NA>                                 <NA>
#> 11                   <NA>             <NA>                                 <NA>
#> 12                   <NA>             <NA>                                 <NA>
#> 13                   <NA>             <NA>                                 <NA>
#> 14                   <NA>             <NA>                                 <NA>
#> 15                   <NA>             <NA>                                 <NA>
#> 16                   <NA>             <NA>                                 <NA>
#> 17                   <NA>             <NA>                                 <NA>
#> 18                   <NA>             <NA>                                 <NA>
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1                                                   <NA>
#> 2                                                   <NA>
#> 3                                                   <NA>
#> 4  https://open.spotify.com/track/5h27GYpKZWWhFov8fOunF6
#> 5  https://open.spotify.com/track/7FeeQVg6yID8laNUH0LlX8
#> 6                                                   <NA>
#> 7                                                   <NA>
#> 8  https://open.spotify.com/track/3b00LSFIRqnHvkoDEs5fkg
#> 9  https://open.spotify.com/track/67L36FEtnci1yguR5s514G
#> 10                                                  <NA>
#> 11                                                  <NA>
#> 12                                                  <NA>
#> 13                                                  <NA>
#> 14                                                  <NA>
#> 15                                                  <NA>
#> 16                                                  <NA>
#> 17                                                  <NA>
#> 18                                                  <NA>
#> 19                                                  <NA>
#> 20                                                  <NA>