Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit
Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market
Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists
A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres
A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks
A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms
Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key
Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode
Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity
Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo
Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature
Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms
Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key
Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode
Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity
Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo
Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature
Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms
Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key
Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode
Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity
Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo
Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature
Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response
Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

# \donttest{
get_recommendations(market = 'SE', seed_genres = 'rock')

#>                                                                                                                                                                                                            artists
#> 1                  https://api.spotify.com/v1/artists/74oJ4qxwOZvX6oSsu1DGnw, 74oJ4qxwOZvX6oSsu1DGnw, Cream, artist, spotify:artist:74oJ4qxwOZvX6oSsu1DGnw, https://open.spotify.com/artist/74oJ4qxwOZvX6oSsu1DGnw
#> 2               https://api.spotify.com/v1/artists/6gZq1Q6bdOxsUPUG1TaFbF, 6gZq1Q6bdOxsUPUG1TaFbF, Godsmack, artist, spotify:artist:6gZq1Q6bdOxsUPUG1TaFbF, https://open.spotify.com/artist/6gZq1Q6bdOxsUPUG1TaFbF
#> 3  https://api.spotify.com/v1/artists/40Yq4vzPs9VNUrIBG5Jr2i, 40Yq4vzPs9VNUrIBG5Jr2i, The Smashing Pumpkins, artist, spotify:artist:40Yq4vzPs9VNUrIBG5Jr2i, https://open.spotify.com/artist/40Yq4vzPs9VNUrIBG5Jr2i
#> 4       https://api.spotify.com/v1/artists/25Maank76ry2Tmbi2Ql1SF, 25Maank76ry2Tmbi2Ql1SF, The Interrupters, artist, spotify:artist:25Maank76ry2Tmbi2Ql1SF, https://open.spotify.com/artist/25Maank76ry2Tmbi2Ql1SF
#> 5      https://api.spotify.com/v1/artists/0u3d5PM2FuEuG5QuUdt8mT, 0u3d5PM2FuEuG5QuUdt8mT, Hands Like Houses, artist, spotify:artist:0u3d5PM2FuEuG5QuUdt8mT, https://open.spotify.com/artist/0u3d5PM2FuEuG5QuUdt8mT
#> 6                     https://api.spotify.com/v1/artists/51Blml2LZPmy7TTiAg47vQ, 51Blml2LZPmy7TTiAg47vQ, U2, artist, spotify:artist:51Blml2LZPmy7TTiAg47vQ, https://open.spotify.com/artist/51Blml2LZPmy7TTiAg47vQ
#> 7    https://api.spotify.com/v1/artists/2UazAtjfzqBF0Nho2awK4z, 2UazAtjfzqBF0Nho2awK4z, Stone Temple Pilots, artist, spotify:artist:2UazAtjfzqBF0Nho2awK4z, https://open.spotify.com/artist/2UazAtjfzqBF0Nho2awK4z
#> 8         https://api.spotify.com/v1/artists/00YTqRClk82aMchQQpYMd5, 00YTqRClk82aMchQQpYMd5, Our Last Night, artist, spotify:artist:00YTqRClk82aMchQQpYMd5, https://open.spotify.com/artist/00YTqRClk82aMchQQpYMd5
#> 9          https://api.spotify.com/v1/artists/2otRZcOVNcmdkZ5nEoogy6, 2otRZcOVNcmdkZ5nEoogy6, Dead by April, artist, spotify:artist:2otRZcOVNcmdkZ5nEoogy6, https://open.spotify.com/artist/2otRZcOVNcmdkZ5nEoogy6
#> 10              https://api.spotify.com/v1/artists/22RISwgVJyZu9lpqAcv1F5, 22RISwgVJyZu9lpqAcv1F5, Blossoms, artist, spotify:artist:22RISwgVJyZu9lpqAcv1F5, https://open.spotify.com/artist/22RISwgVJyZu9lpqAcv1F5
#> 11             https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 12           https://api.spotify.com/v1/artists/5xUf6j4upBrXZPg6AI4MRK, 5xUf6j4upBrXZPg6AI4MRK, Soundgarden, artist, spotify:artist:5xUf6j4upBrXZPg6AI4MRK, https://open.spotify.com/artist/5xUf6j4upBrXZPg6AI4MRK
#> 13              https://api.spotify.com/v1/artists/2kO6mP0olFJGGh6kvUdNC8, 2kO6mP0olFJGGh6kvUdNC8, Des Rocs, artist, spotify:artist:2kO6mP0olFJGGh6kvUdNC8, https://open.spotify.com/artist/2kO6mP0olFJGGh6kvUdNC8
#> 14             https://api.spotify.com/v1/artists/4o0pNHbyj36LPvukNqEug0, 4o0pNHbyj36LPvukNqEug0, The Maine, artist, spotify:artist:4o0pNHbyj36LPvukNqEug0, https://open.spotify.com/artist/4o0pNHbyj36LPvukNqEug0
#> 15      https://api.spotify.com/v1/artists/2xiIXseIJcq3nG7C8fHeBj, 2xiIXseIJcq3nG7C8fHeBj, Three Days Grace, artist, spotify:artist:2xiIXseIJcq3nG7C8fHeBj, https://open.spotify.com/artist/2xiIXseIJcq3nG7C8fHeBj
#> 16     https://api.spotify.com/v1/artists/3eqjTLE0HfPfh78zjh6TqT, 3eqjTLE0HfPfh78zjh6TqT, Bruce Springsteen, artist, spotify:artist:3eqjTLE0HfPfh78zjh6TqT, https://open.spotify.com/artist/3eqjTLE0HfPfh78zjh6TqT
#> 17                 https://api.spotify.com/v1/artists/1nTJnNxTJmEiZSYf8ScLQ8, 1nTJnNxTJmEiZSYf8ScLQ8, Smile, artist, spotify:artist:1nTJnNxTJmEiZSYf8ScLQ8, https://open.spotify.com/artist/1nTJnNxTJmEiZSYf8ScLQ8
#> 18                  https://api.spotify.com/v1/artists/0ofhRkasSdhp6LUs7tFcF7, 0ofhRkasSdhp6LUs7tFcF7, WSTR, artist, spotify:artist:0ofhRkasSdhp6LUs7tFcF7, https://open.spotify.com/artist/0ofhRkasSdhp6LUs7tFcF7
#> 19       https://api.spotify.com/v1/artists/64tNsm6TnZe2zpcMVMOoHL, 64tNsm6TnZe2zpcMVMOoHL, Alice In Chains, artist, spotify:artist:64tNsm6TnZe2zpcMVMOoHL, https://open.spotify.com/artist/64tNsm6TnZe2zpcMVMOoHL
#> 20             https://api.spotify.com/v1/artists/1w5Kfo2jwwIPruYS2UWh56, 1w5Kfo2jwwIPruYS2UWh56, Pearl Jam, artist, spotify:artist:1w5Kfo2jwwIPruYS2UWh56, https://open.spotify.com/artist/1w5Kfo2jwwIPruYS2UWh56
#>    disc_number duration_ms explicit
#> 1            2      258466    FALSE
#> 2            1      282653    FALSE
#> 3            2      266200    FALSE
#> 4            1      170453    FALSE
#> 5            1      207200    FALSE
#> 6            1      295515    FALSE
#> 7            1      176546    FALSE
#> 8            1      238181    FALSE
#> 9            1      265226    FALSE
#> 10           1      222488    FALSE
#> 11           1      233986    FALSE
#> 12           1      283026    FALSE
#> 13           1      207141    FALSE
#> 14           1      192825    FALSE
#> 15           1      231480    FALSE
#> 16           1      155880    FALSE
#> 17           1      196840    FALSE
#> 18           1      208179    FALSE
#> 19           1      149466    FALSE
#> 20           1      318226    FALSE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/6PUabSMXmPnZna361Wwmf7
#> 2  https://api.spotify.com/v1/tracks/5uGsG0LfotfWDq6hql4h53
#> 3  https://api.spotify.com/v1/tracks/5QLHGv0DfpeXLNFo7SFEy1
#> 4  https://api.spotify.com/v1/tracks/60FJeNTK5edr72G2xtCWuP
#> 5  https://api.spotify.com/v1/tracks/2optm7JeCh2iEM502qFLUX
#> 6  https://api.spotify.com/v1/tracks/6ADSaE87h8Y3lccZlBJdXH
#> 7  https://api.spotify.com/v1/tracks/4fjsOdrWMaKf70x21h5XbX
#> 8  https://api.spotify.com/v1/tracks/506YO37JuV40cnjeVuwlGD
#> 9  https://api.spotify.com/v1/tracks/4cfEA7zQhaQpkhUFonoOPw
#> 10 https://api.spotify.com/v1/tracks/5pVbLYTdihSKd8Kylfhwa7
#> 11 https://api.spotify.com/v1/tracks/1TRr8Pp2uOqL24fMHpJyV3
#> 12 https://api.spotify.com/v1/tracks/1wlKw9NAWVWpx7OIAyg9EA
#> 13 https://api.spotify.com/v1/tracks/481idAnpIOHEjR4YPsbA0y
#> 14 https://api.spotify.com/v1/tracks/47ojH5LeQhl1ZltSAHBEFF
#> 15 https://api.spotify.com/v1/tracks/0M955bMOoilikPXwKLYpoi
#> 16 https://api.spotify.com/v1/tracks/3PzsbWSQdLCKDLxn7YZfkM
#> 17 https://api.spotify.com/v1/tracks/69VE5kmfqG4dJSkrB1OtCI
#> 18 https://api.spotify.com/v1/tracks/4Jl3syKPuB2YFZZhsSeniv
#> 19 https://api.spotify.com/v1/tracks/4A065x9kJt955eGVqf813g
#> 20 https://api.spotify.com/v1/tracks/62nQ8UZVqR2RMvkJHkcO2o
#>                        id is_local is_playable
#> 1  6PUabSMXmPnZna361Wwmf7    FALSE        TRUE
#> 2  5uGsG0LfotfWDq6hql4h53    FALSE        TRUE
#> 3  5QLHGv0DfpeXLNFo7SFEy1    FALSE        TRUE
#> 4  60FJeNTK5edr72G2xtCWuP    FALSE        TRUE
#> 5  2optm7JeCh2iEM502qFLUX    FALSE        TRUE
#> 6  6ADSaE87h8Y3lccZlBJdXH    FALSE        TRUE
#> 7  4fjsOdrWMaKf70x21h5XbX    FALSE        TRUE
#> 8  506YO37JuV40cnjeVuwlGD    FALSE        TRUE
#> 9  4cfEA7zQhaQpkhUFonoOPw    FALSE        TRUE
#> 10 5pVbLYTdihSKd8Kylfhwa7    FALSE        TRUE
#> 11 1TRr8Pp2uOqL24fMHpJyV3    FALSE        TRUE
#> 12 1wlKw9NAWVWpx7OIAyg9EA    FALSE        TRUE
#> 13 481idAnpIOHEjR4YPsbA0y    FALSE        TRUE
#> 14 47ojH5LeQhl1ZltSAHBEFF    FALSE        TRUE
#> 15 0M955bMOoilikPXwKLYpoi    FALSE        TRUE
#> 16 3PzsbWSQdLCKDLxn7YZfkM    FALSE        TRUE
#> 17 69VE5kmfqG4dJSkrB1OtCI    FALSE        TRUE
#> 18 4Jl3syKPuB2YFZZhsSeniv    FALSE        TRUE
#> 19 4A065x9kJt955eGVqf813g    FALSE        TRUE
#> 20 62nQ8UZVqR2RMvkJHkcO2o    FALSE        TRUE
#>                                                  name popularity
#> 1                                   Crossroads - Live         57
#> 2                                              Voodoo         66
#> 3                              1979 - Remastered 2012         76
#> 4                                      She's Kerosene         60
#> 5                                             Monster         48
#> 6                    With Or Without You - Remastered         79
#> 7                            Vasoline - 2019 Remaster         62
#> 8                                       Common Ground         51
#> 9                                     Crying Over You         55
#> 10 There's A Reason Why (I Never Returned Your Calls)         57
#> 11                                   Sound of Madness         42
#> 12                                 Fell On Black Days         67
#> 13                           Let Me Live / Let Me Die         64
#> 14                                    Don't Come Down         55
#> 15                        I Hate Everything About You         72
#> 16                                        I'm On Fire         75
#> 17                     Doing All Right - ...Revisited         55
#> 18                                   Eastbound & Down         48
#> 19                                         Them Bones         67
#> 20                                             Jeremy         72
#>                                                                                                    preview_url
#> 1                                                                                                         <NA>
#> 2                                                                                                         <NA>
#> 3                                                                                                         <NA>
#> 4  https://p.scdn.co/mp3-preview/b3211d53964127b425ed63b29effe0347e139101?cid=125a10da5fa94687b3d0913168bc3698
#> 5  https://p.scdn.co/mp3-preview/e5f9a1c79b0a9cee5cdbb2e19fd95a6c71d6e9aa?cid=125a10da5fa94687b3d0913168bc3698
#> 6                                                                                                         <NA>
#> 7  https://p.scdn.co/mp3-preview/a8e4886ecfbf049afcbe90c74d5313d15f7c5e53?cid=125a10da5fa94687b3d0913168bc3698
#> 8  https://p.scdn.co/mp3-preview/dba3fc2354613666047a632d23854a17751f03ec?cid=125a10da5fa94687b3d0913168bc3698
#> 9                                                                                                         <NA>
#> 10                                                                                                        <NA>
#> 11 https://p.scdn.co/mp3-preview/26186c5152e1f647a1b47148fef1799019c8d563?cid=125a10da5fa94687b3d0913168bc3698
#> 12                                                                                                        <NA>
#> 13 https://p.scdn.co/mp3-preview/b3dc80e33929b9b9b351bb77ebd6bdced8eb1784?cid=125a10da5fa94687b3d0913168bc3698
#> 14 https://p.scdn.co/mp3-preview/e0363988e76f3a7daeeca603535ae60924c545a7?cid=125a10da5fa94687b3d0913168bc3698
#> 15 https://p.scdn.co/mp3-preview/8893669447625edf76339cdd67a338dad034a7ea?cid=125a10da5fa94687b3d0913168bc3698
#> 16 https://p.scdn.co/mp3-preview/f3bae9928744d8a6840dc8bf43a5ed1416adadff?cid=125a10da5fa94687b3d0913168bc3698
#> 17                                                                                                        <NA>
#> 18 https://p.scdn.co/mp3-preview/53f484b80f067ff7f0c880afea2ca08382e35e7f?cid=125a10da5fa94687b3d0913168bc3698
#> 19 https://p.scdn.co/mp3-preview/2d7ef6644acd2336ddb7176b56c35f4935106416?cid=125a10da5fa94687b3d0913168bc3698
#> 20 https://p.scdn.co/mp3-preview/c5742cc09643dc0704b397d938b425c575f89b57?cid=125a10da5fa94687b3d0913168bc3698
#>    track_number  type                                  uri album.album_type
#> 1             1 track spotify:track:6PUabSMXmPnZna361Wwmf7            ALBUM
#> 2            12 track spotify:track:5uGsG0LfotfWDq6hql4h53            ALBUM
#> 3             5 track spotify:track:5QLHGv0DfpeXLNFo7SFEy1            ALBUM
#> 4             3 track spotify:track:60FJeNTK5edr72G2xtCWuP            ALBUM
#> 5             2 track spotify:track:2optm7JeCh2iEM502qFLUX            ALBUM
#> 6             3 track spotify:track:6ADSaE87h8Y3lccZlBJdXH            ALBUM
#> 7             2 track spotify:track:4fjsOdrWMaKf70x21h5XbX            ALBUM
#> 8             6 track spotify:track:506YO37JuV40cnjeVuwlGD            ALBUM
#> 9             1 track spotify:track:4cfEA7zQhaQpkhUFonoOPw            ALBUM
#> 10            1 track spotify:track:5pVbLYTdihSKd8Kylfhwa7            ALBUM
#> 11            2 track spotify:track:1TRr8Pp2uOqL24fMHpJyV3            ALBUM
#> 12            3 track spotify:track:1wlKw9NAWVWpx7OIAyg9EA            ALBUM
#> 13            1 track spotify:track:481idAnpIOHEjR4YPsbA0y           SINGLE
#> 14            1 track spotify:track:47ojH5LeQhl1ZltSAHBEFF            ALBUM
#> 15            3 track spotify:track:0M955bMOoilikPXwKLYpoi            ALBUM
#> 16            6 track spotify:track:3PzsbWSQdLCKDLxn7YZfkM            ALBUM
#> 17            3 track spotify:track:69VE5kmfqG4dJSkrB1OtCI            ALBUM
#> 18            6 track spotify:track:4Jl3syKPuB2YFZZhsSeniv            ALBUM
#> 19            1 track spotify:track:4A065x9kJt955eGVqf813g            ALBUM
#> 20            6 track spotify:track:62nQ8UZVqR2RMvkJHkcO2o            ALBUM
#>                                                                                                                                                                                                      album.artists
#> 1                  https://api.spotify.com/v1/artists/74oJ4qxwOZvX6oSsu1DGnw, 74oJ4qxwOZvX6oSsu1DGnw, Cream, artist, spotify:artist:74oJ4qxwOZvX6oSsu1DGnw, https://open.spotify.com/artist/74oJ4qxwOZvX6oSsu1DGnw
#> 2               https://api.spotify.com/v1/artists/6gZq1Q6bdOxsUPUG1TaFbF, 6gZq1Q6bdOxsUPUG1TaFbF, Godsmack, artist, spotify:artist:6gZq1Q6bdOxsUPUG1TaFbF, https://open.spotify.com/artist/6gZq1Q6bdOxsUPUG1TaFbF
#> 3  https://api.spotify.com/v1/artists/40Yq4vzPs9VNUrIBG5Jr2i, 40Yq4vzPs9VNUrIBG5Jr2i, The Smashing Pumpkins, artist, spotify:artist:40Yq4vzPs9VNUrIBG5Jr2i, https://open.spotify.com/artist/40Yq4vzPs9VNUrIBG5Jr2i
#> 4       https://api.spotify.com/v1/artists/25Maank76ry2Tmbi2Ql1SF, 25Maank76ry2Tmbi2Ql1SF, The Interrupters, artist, spotify:artist:25Maank76ry2Tmbi2Ql1SF, https://open.spotify.com/artist/25Maank76ry2Tmbi2Ql1SF
#> 5      https://api.spotify.com/v1/artists/0u3d5PM2FuEuG5QuUdt8mT, 0u3d5PM2FuEuG5QuUdt8mT, Hands Like Houses, artist, spotify:artist:0u3d5PM2FuEuG5QuUdt8mT, https://open.spotify.com/artist/0u3d5PM2FuEuG5QuUdt8mT
#> 6                     https://api.spotify.com/v1/artists/51Blml2LZPmy7TTiAg47vQ, 51Blml2LZPmy7TTiAg47vQ, U2, artist, spotify:artist:51Blml2LZPmy7TTiAg47vQ, https://open.spotify.com/artist/51Blml2LZPmy7TTiAg47vQ
#> 7    https://api.spotify.com/v1/artists/2UazAtjfzqBF0Nho2awK4z, 2UazAtjfzqBF0Nho2awK4z, Stone Temple Pilots, artist, spotify:artist:2UazAtjfzqBF0Nho2awK4z, https://open.spotify.com/artist/2UazAtjfzqBF0Nho2awK4z
#> 8         https://api.spotify.com/v1/artists/00YTqRClk82aMchQQpYMd5, 00YTqRClk82aMchQQpYMd5, Our Last Night, artist, spotify:artist:00YTqRClk82aMchQQpYMd5, https://open.spotify.com/artist/00YTqRClk82aMchQQpYMd5
#> 9          https://api.spotify.com/v1/artists/2otRZcOVNcmdkZ5nEoogy6, 2otRZcOVNcmdkZ5nEoogy6, Dead by April, artist, spotify:artist:2otRZcOVNcmdkZ5nEoogy6, https://open.spotify.com/artist/2otRZcOVNcmdkZ5nEoogy6
#> 10              https://api.spotify.com/v1/artists/22RISwgVJyZu9lpqAcv1F5, 22RISwgVJyZu9lpqAcv1F5, Blossoms, artist, spotify:artist:22RISwgVJyZu9lpqAcv1F5, https://open.spotify.com/artist/22RISwgVJyZu9lpqAcv1F5
#> 11             https://api.spotify.com/v1/artists/70BYFdaZbEKbeauJ670ysI, 70BYFdaZbEKbeauJ670ysI, Shinedown, artist, spotify:artist:70BYFdaZbEKbeauJ670ysI, https://open.spotify.com/artist/70BYFdaZbEKbeauJ670ysI
#> 12           https://api.spotify.com/v1/artists/5xUf6j4upBrXZPg6AI4MRK, 5xUf6j4upBrXZPg6AI4MRK, Soundgarden, artist, spotify:artist:5xUf6j4upBrXZPg6AI4MRK, https://open.spotify.com/artist/5xUf6j4upBrXZPg6AI4MRK
#> 13              https://api.spotify.com/v1/artists/2kO6mP0olFJGGh6kvUdNC8, 2kO6mP0olFJGGh6kvUdNC8, Des Rocs, artist, spotify:artist:2kO6mP0olFJGGh6kvUdNC8, https://open.spotify.com/artist/2kO6mP0olFJGGh6kvUdNC8
#> 14             https://api.spotify.com/v1/artists/4o0pNHbyj36LPvukNqEug0, 4o0pNHbyj36LPvukNqEug0, The Maine, artist, spotify:artist:4o0pNHbyj36LPvukNqEug0, https://open.spotify.com/artist/4o0pNHbyj36LPvukNqEug0
#> 15      https://api.spotify.com/v1/artists/2xiIXseIJcq3nG7C8fHeBj, 2xiIXseIJcq3nG7C8fHeBj, Three Days Grace, artist, spotify:artist:2xiIXseIJcq3nG7C8fHeBj, https://open.spotify.com/artist/2xiIXseIJcq3nG7C8fHeBj
#> 16     https://api.spotify.com/v1/artists/3eqjTLE0HfPfh78zjh6TqT, 3eqjTLE0HfPfh78zjh6TqT, Bruce Springsteen, artist, spotify:artist:3eqjTLE0HfPfh78zjh6TqT, https://open.spotify.com/artist/3eqjTLE0HfPfh78zjh6TqT
#> 17                 https://api.spotify.com/v1/artists/1dfeR4HaWDbWqFHLkxsg1d, 1dfeR4HaWDbWqFHLkxsg1d, Queen, artist, spotify:artist:1dfeR4HaWDbWqFHLkxsg1d, https://open.spotify.com/artist/1dfeR4HaWDbWqFHLkxsg1d
#> 18                  https://api.spotify.com/v1/artists/0ofhRkasSdhp6LUs7tFcF7, 0ofhRkasSdhp6LUs7tFcF7, WSTR, artist, spotify:artist:0ofhRkasSdhp6LUs7tFcF7, https://open.spotify.com/artist/0ofhRkasSdhp6LUs7tFcF7
#> 19       https://api.spotify.com/v1/artists/64tNsm6TnZe2zpcMVMOoHL, 64tNsm6TnZe2zpcMVMOoHL, Alice In Chains, artist, spotify:artist:64tNsm6TnZe2zpcMVMOoHL, https://open.spotify.com/artist/64tNsm6TnZe2zpcMVMOoHL
#> 20             https://api.spotify.com/v1/artists/1w5Kfo2jwwIPruYS2UWh56, 1w5Kfo2jwwIPruYS2UWh56, Pearl Jam, artist, spotify:artist:1w5Kfo2jwwIPruYS2UWh56, https://open.spotify.com/artist/1w5Kfo2jwwIPruYS2UWh56
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/0zrtTZC7yY2TOEhnbJzSb9
#> 2  https://api.spotify.com/v1/albums/1w7vC8hjYXhK1fS5cM2fUM
#> 3  https://api.spotify.com/v1/albums/55RhFRyQFihIyGf61MgcfV
#> 4  https://api.spotify.com/v1/albums/6AoWkAFBSNlX63z1fJVNIL
#> 5  https://api.spotify.com/v1/albums/2qtLF5Z8OxMV919Knbi3oo
#> 6  https://api.spotify.com/v1/albums/5y6wlw1LnqFnQFruMeiwGU
#> 7  https://api.spotify.com/v1/albums/57lcTrUlYgfMIPvBUsVU6h
#> 8  https://api.spotify.com/v1/albums/482Zolwz9ekvPD4kfj8BGo
#> 9  https://api.spotify.com/v1/albums/3rgz4IjijsFvJwqicCqle3
#> 10 https://api.spotify.com/v1/albums/0yO5Mwu2oDdGbeRFpAaBGa
#> 11 https://api.spotify.com/v1/albums/661dSFpYtPulSNDZZLkokq
#> 12 https://api.spotify.com/v1/albums/29sTacnS0qA9xri6YS8xLA
#> 13 https://api.spotify.com/v1/albums/0fnAV8la7b1AZjAfY1oyXZ
#> 14 https://api.spotify.com/v1/albums/2jStuYeaRty1cwxjoxfo9K
#> 15 https://api.spotify.com/v1/albums/13topfW33NjnACjnRiZBX7
#> 16 https://api.spotify.com/v1/albums/0PMasrHdpaoIRuHuhHp72O
#> 17 https://api.spotify.com/v1/albums/6i6folBtxKV28WX3msQ4FE
#> 18 https://api.spotify.com/v1/albums/4gw6oQ9Yk4m6mreJi6oGAV
#> 19 https://api.spotify.com/v1/albums/58NXIEYqmq5dQHg9nV9duM
#> 20 https://api.spotify.com/v1/albums/5B4PYA7wNN4WdEXdIJu58a
#>                  album.id
#> 1  0zrtTZC7yY2TOEhnbJzSb9
#> 2  1w7vC8hjYXhK1fS5cM2fUM
#> 3  55RhFRyQFihIyGf61MgcfV
#> 4  6AoWkAFBSNlX63z1fJVNIL
#> 5  2qtLF5Z8OxMV919Knbi3oo
#> 6  5y6wlw1LnqFnQFruMeiwGU
#> 7  57lcTrUlYgfMIPvBUsVU6h
#> 8  482Zolwz9ekvPD4kfj8BGo
#> 9  3rgz4IjijsFvJwqicCqle3
#> 10 0yO5Mwu2oDdGbeRFpAaBGa
#> 11 661dSFpYtPulSNDZZLkokq
#> 12 29sTacnS0qA9xri6YS8xLA
#> 13 0fnAV8la7b1AZjAfY1oyXZ
#> 14 2jStuYeaRty1cwxjoxfo9K
#> 15 13topfW33NjnACjnRiZBX7
#> 16 0PMasrHdpaoIRuHuhHp72O
#> 17 6i6folBtxKV28WX3msQ4FE
#> 18 4gw6oQ9Yk4m6mreJi6oGAV
#> 19 58NXIEYqmq5dQHg9nV9duM
#> 20 5B4PYA7wNN4WdEXdIJu58a
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2735c7731f5acdcb2d02d78b7ee, https://i.scdn.co/image/ab67616d00001e025c7731f5acdcb2d02d78b7ee, https://i.scdn.co/image/ab67616d000048515c7731f5acdcb2d02d78b7ee, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b27382a891641608e2f2d6c3a8b8, https://i.scdn.co/image/ab67616d00001e0282a891641608e2f2d6c3a8b8, https://i.scdn.co/image/ab67616d0000485182a891641608e2f2d6c3a8b8, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273431ac6e6f393acf475730ec6, https://i.scdn.co/image/ab67616d00001e02431ac6e6f393acf475730ec6, https://i.scdn.co/image/ab67616d00004851431ac6e6f393acf475730ec6, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c9c292d93a9d27c4762cb559, https://i.scdn.co/image/ab67616d00001e02c9c292d93a9d27c4762cb559, https://i.scdn.co/image/ab67616d00004851c9c292d93a9d27c4762cb559, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f9bb3fff4e660ec5ba66af40, https://i.scdn.co/image/ab67616d00001e02f9bb3fff4e660ec5ba66af40, https://i.scdn.co/image/ab67616d00004851f9bb3fff4e660ec5ba66af40, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273b7bea3d01f04e6d0408d2afe, https://i.scdn.co/image/ab67616d00001e02b7bea3d01f04e6d0408d2afe, https://i.scdn.co/image/ab67616d00004851b7bea3d01f04e6d0408d2afe, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736db0b75182c54680392aaa20, https://i.scdn.co/image/ab67616d00001e026db0b75182c54680392aaa20, https://i.scdn.co/image/ab67616d000048516db0b75182c54680392aaa20, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c3becb63afa2b5f9a1c7623d, https://i.scdn.co/image/ab67616d00001e02c3becb63afa2b5f9a1c7623d, https://i.scdn.co/image/ab67616d00004851c3becb63afa2b5f9a1c7623d, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273387e3fe93a12844cf969399e, https://i.scdn.co/image/ab67616d00001e02387e3fe93a12844cf969399e, https://i.scdn.co/image/ab67616d00004851387e3fe93a12844cf969399e, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732889f5b28a2acdf603bfcd8a, https://i.scdn.co/image/ab67616d00001e022889f5b28a2acdf603bfcd8a, https://i.scdn.co/image/ab67616d000048512889f5b28a2acdf603bfcd8a, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a24befd5cbb5a31f0ac9eacb, https://i.scdn.co/image/ab67616d00001e02a24befd5cbb5a31f0ac9eacb, https://i.scdn.co/image/ab67616d00004851a24befd5cbb5a31f0ac9eacb, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273ae6d8e36136353d550b2587d, https://i.scdn.co/image/ab67616d00001e02ae6d8e36136353d550b2587d, https://i.scdn.co/image/ab67616d00004851ae6d8e36136353d550b2587d, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732eb4ba21c445087266c23823, https://i.scdn.co/image/ab67616d00001e022eb4ba21c445087266c23823, https://i.scdn.co/image/ab67616d000048512eb4ba21c445087266c23823, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273dc0cb47f6cee440cf6221a76, https://i.scdn.co/image/ab67616d00001e02dc0cb47f6cee440cf6221a76, https://i.scdn.co/image/ab67616d00004851dc0cb47f6cee440cf6221a76, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27377922eaa071fb388270f203f, https://i.scdn.co/image/ab67616d00001e0277922eaa071fb388270f203f, https://i.scdn.co/image/ab67616d0000485177922eaa071fb388270f203f, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a7865e686c36a4adda6c9978, https://i.scdn.co/image/ab67616d00001e02a7865e686c36a4adda6c9978, https://i.scdn.co/image/ab67616d00004851a7865e686c36a4adda6c9978, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e8b066f70c206551210d902b, https://i.scdn.co/image/ab67616d00001e02e8b066f70c206551210d902b, https://i.scdn.co/image/ab67616d00004851e8b066f70c206551210d902b, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736965cb863cb6fa5f93459805, https://i.scdn.co/image/ab67616d00001e026965cb863cb6fa5f93459805, https://i.scdn.co/image/ab67616d000048516965cb863cb6fa5f93459805, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27331ea2c36c5da3979177127b1, https://i.scdn.co/image/ab67616d00001e0231ea2c36c5da3979177127b1, https://i.scdn.co/image/ab67616d0000485131ea2c36c5da3979177127b1, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273d400d27cba05bb0545533864, https://i.scdn.co/image/ab67616d00001e02d400d27cba05bb0545533864, https://i.scdn.co/image/ab67616d00004851d400d27cba05bb0545533864, 640, 300, 64
#>                                                 album.name album.release_date
#> 1                                           Wheels Of Fire         1968-07-01
#> 2                                                 Godsmack         1998-01-01
#> 3  Mellon Collie And The Infinite Sadness (Deluxe Edition)               1995
#> 4                                     Fight the Good Fight         2018-06-29
#> 5                                                    Anon.         2018-10-12
#> 6                           The Joshua Tree (Super Deluxe)         1987-03-03
#> 7                                   Purple (2019 Remaster)         1994-06-07
#> 8                                        Selective Hearing         2017-06-09
#> 9                                           Worlds Collide         2017-04-07
#> 10                                  Cool Like You (Deluxe)         2018-04-27
#> 11                                    The Sound of Madness         2008-06-23
#> 12                           Superunknown (Deluxe Edition)         1994-03-09
#> 13                                     Let the Vultures In         2018-11-02
#> 14                                    Lovely Little Lonely         2017-04-07
#> 15                                        Three Days Grace         2003-04-19
#> 16                                      Born In The U.S.A.         1984-06-04
#> 17             Bohemian Rhapsody (The Original Soundtrack)         2018-10-19
#> 18                                 Red, Green Or Inbetween         2017-01-20
#> 19                                                    Dirt         1992-09-29
#> 20                                                     Ten         1991-08-27
#>    album.release_date_precision album.total_tracks album.type
#> 1                           day                 13      album
#> 2                           day                 12      album
#> 3                          year                 92      album
#> 4                           day                 12      album
#> 5                           day                 10      album
#> 6                           day                 49      album
#> 7                           day                 11      album
#> 8                           day                  7      album
#> 9                           day                 11      album
#> 10                          day                 22      album
#> 11                          day                 11      album
#> 12                          day                 32      album
#> 13                          day                  5      album
#> 14                          day                 12      album
#> 15                          day                 12      album
#> 16                          day                 12      album
#> 17                          day                 22      album
#> 18                          day                 11      album
#> 19                          day                 13      album
#> 20                          day                 11      album
#>                               album.uri
#> 1  spotify:album:0zrtTZC7yY2TOEhnbJzSb9
#> 2  spotify:album:1w7vC8hjYXhK1fS5cM2fUM
#> 3  spotify:album:55RhFRyQFihIyGf61MgcfV
#> 4  spotify:album:6AoWkAFBSNlX63z1fJVNIL
#> 5  spotify:album:2qtLF5Z8OxMV919Knbi3oo
#> 6  spotify:album:5y6wlw1LnqFnQFruMeiwGU
#> 7  spotify:album:57lcTrUlYgfMIPvBUsVU6h
#> 8  spotify:album:482Zolwz9ekvPD4kfj8BGo
#> 9  spotify:album:3rgz4IjijsFvJwqicCqle3
#> 10 spotify:album:0yO5Mwu2oDdGbeRFpAaBGa
#> 11 spotify:album:661dSFpYtPulSNDZZLkokq
#> 12 spotify:album:29sTacnS0qA9xri6YS8xLA
#> 13 spotify:album:0fnAV8la7b1AZjAfY1oyXZ
#> 14 spotify:album:2jStuYeaRty1cwxjoxfo9K
#> 15 spotify:album:13topfW33NjnACjnRiZBX7
#> 16 spotify:album:0PMasrHdpaoIRuHuhHp72O
#> 17 spotify:album:6i6folBtxKV28WX3msQ4FE
#> 18 spotify:album:4gw6oQ9Yk4m6mreJi6oGAV
#> 19 spotify:album:58NXIEYqmq5dQHg9nV9duM
#> 20 spotify:album:5B4PYA7wNN4WdEXdIJu58a
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/0zrtTZC7yY2TOEhnbJzSb9      GBA076800010
#> 2  https://open.spotify.com/album/1w7vC8hjYXhK1fS5cM2fUM      USUR19980188
#> 3  https://open.spotify.com/album/55RhFRyQFihIyGf61MgcfV      USVI21200910
#> 4  https://open.spotify.com/album/6AoWkAFBSNlX63z1fJVNIL      USHEL1801003
#> 5  https://open.spotify.com/album/2qtLF5Z8OxMV919Knbi3oo      USHR21811202
#> 6  https://open.spotify.com/album/5y6wlw1LnqFnQFruMeiwGU      GBUM70709792
#> 7  https://open.spotify.com/album/57lcTrUlYgfMIPvBUsVU6h      USRH11901675
#> 8  https://open.spotify.com/album/482Zolwz9ekvPD4kfj8BGo      GBDMT1600259
#> 9  https://open.spotify.com/album/3rgz4IjijsFvJwqicCqle3      SEUM71601951
#> 10 https://open.spotify.com/album/0yO5Mwu2oDdGbeRFpAaBGa      GBUM71706179
#> 11 https://open.spotify.com/album/661dSFpYtPulSNDZZLkokq      USAT20802280
#> 12 https://open.spotify.com/album/29sTacnS0qA9xri6YS8xLA      USAM19400003
#> 13 https://open.spotify.com/album/0fnAV8la7b1AZjAfY1oyXZ      GBKPL1812549
#> 14 https://open.spotify.com/album/2jStuYeaRty1cwxjoxfo9K      QMEU31700194
#> 15 https://open.spotify.com/album/13topfW33NjnACjnRiZBX7      USJI10300075
#> 16 https://open.spotify.com/album/0PMasrHdpaoIRuHuhHp72O      USSM18400411
#> 17 https://open.spotify.com/album/6i6folBtxKV28WX3msQ4FE      GBUM71805971
#> 18 https://open.spotify.com/album/4gw6oQ9Yk4m6mreJi6oGAV      USZZ81610031
#> 19 https://open.spotify.com/album/58NXIEYqmq5dQHg9nV9duM      USSM19200799
#> 20 https://open.spotify.com/album/5B4PYA7wNN4WdEXdIJu58a      USSM19100674
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/6PUabSMXmPnZna361Wwmf7
#> 2  https://open.spotify.com/track/5uGsG0LfotfWDq6hql4h53
#> 3  https://open.spotify.com/track/5QLHGv0DfpeXLNFo7SFEy1
#> 4  https://open.spotify.com/track/60FJeNTK5edr72G2xtCWuP
#> 5  https://open.spotify.com/track/2optm7JeCh2iEM502qFLUX
#> 6  https://open.spotify.com/track/6ADSaE87h8Y3lccZlBJdXH
#> 7  https://open.spotify.com/track/4fjsOdrWMaKf70x21h5XbX
#> 8  https://open.spotify.com/track/506YO37JuV40cnjeVuwlGD
#> 9  https://open.spotify.com/track/4cfEA7zQhaQpkhUFonoOPw
#> 10 https://open.spotify.com/track/5pVbLYTdihSKd8Kylfhwa7
#> 11 https://open.spotify.com/track/1TRr8Pp2uOqL24fMHpJyV3
#> 12 https://open.spotify.com/track/1wlKw9NAWVWpx7OIAyg9EA
#> 13 https://open.spotify.com/track/481idAnpIOHEjR4YPsbA0y
#> 14 https://open.spotify.com/track/47ojH5LeQhl1ZltSAHBEFF
#> 15 https://open.spotify.com/track/0M955bMOoilikPXwKLYpoi
#> 16 https://open.spotify.com/track/3PzsbWSQdLCKDLxn7YZfkM
#> 17 https://open.spotify.com/track/69VE5kmfqG4dJSkrB1OtCI
#> 18 https://open.spotify.com/track/4Jl3syKPuB2YFZZhsSeniv
#> 19 https://open.spotify.com/track/4A065x9kJt955eGVqf813g
#> 20 https://open.spotify.com/track/62nQ8UZVqR2RMvkJHkcO2o
#>                                            linked_from.href
#> 1                                                      <NA>
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4                                                      <NA>
#> 5  https://api.spotify.com/v1/tracks/5apUdi1b0lvZw574hq1Fsr
#> 6  https://api.spotify.com/v1/tracks/4ha9ajL4Kz98uldAfuos6p
#> 7  https://api.spotify.com/v1/tracks/5vCE0xRLIEG1Zej2tgWFDb
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#> 11                                                     <NA>
#> 12 https://api.spotify.com/v1/tracks/6B4JGMaqe1LaEMlsVzNdNM
#> 13 https://api.spotify.com/v1/tracks/2UKMqNCVsHrAYzBZMA0rxL
#> 14                                                     <NA>
#> 15                                                     <NA>
#> 16                                                     <NA>
#> 17                                                     <NA>
#> 18 https://api.spotify.com/v1/tracks/4NNzO1RQqdYzQXTISmHZDe
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1                    <NA>             <NA>                                 <NA>
#> 2                    <NA>             <NA>                                 <NA>
#> 3                    <NA>             <NA>                                 <NA>
#> 4                    <NA>             <NA>                                 <NA>
#> 5  5apUdi1b0lvZw574hq1Fsr            track spotify:track:5apUdi1b0lvZw574hq1Fsr
#> 6  4ha9ajL4Kz98uldAfuos6p            track spotify:track:4ha9ajL4Kz98uldAfuos6p
#> 7  5vCE0xRLIEG1Zej2tgWFDb            track spotify:track:5vCE0xRLIEG1Zej2tgWFDb
#> 8                    <NA>             <NA>                                 <NA>
#> 9                    <NA>             <NA>                                 <NA>
#> 10                   <NA>             <NA>                                 <NA>
#> 11                   <NA>             <NA>                                 <NA>
#> 12 6B4JGMaqe1LaEMlsVzNdNM            track spotify:track:6B4JGMaqe1LaEMlsVzNdNM
#> 13 2UKMqNCVsHrAYzBZMA0rxL            track spotify:track:2UKMqNCVsHrAYzBZMA0rxL
#> 14                   <NA>             <NA>                                 <NA>
#> 15                   <NA>             <NA>                                 <NA>
#> 16                   <NA>             <NA>                                 <NA>
#> 17                   <NA>             <NA>                                 <NA>
#> 18 4NNzO1RQqdYzQXTISmHZDe            track spotify:track:4NNzO1RQqdYzQXTISmHZDe
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1                                                   <NA>
#> 2                                                   <NA>
#> 3                                                   <NA>
#> 4                                                   <NA>
#> 5  https://open.spotify.com/track/5apUdi1b0lvZw574hq1Fsr
#> 6  https://open.spotify.com/track/4ha9ajL4Kz98uldAfuos6p
#> 7  https://open.spotify.com/track/5vCE0xRLIEG1Zej2tgWFDb
#> 8                                                   <NA>
#> 9                                                   <NA>
#> 10                                                  <NA>
#> 11                                                  <NA>
#> 12 https://open.spotify.com/track/6B4JGMaqe1LaEMlsVzNdNM
#> 13 https://open.spotify.com/track/2UKMqNCVsHrAYzBZMA0rxL
#> 14                                                  <NA>
#> 15                                                  <NA>
#> 16                                                  <NA>
#> 17                                                  <NA>
#> 18 https://open.spotify.com/track/4NNzO1RQqdYzQXTISmHZDe
#> 19                                                  <NA>
#> 20                                                  <NA>

# }