Create a playlist-style listening experience based on seed artists, tracks and genres.

All parameters are optional, but at least one of seed_artists, seed_tracks and seed_genres must be given.

get_recommendations(
  limit = 20,
  market = NULL,
  seed_artists = NULL,
  seed_genres = NULL,
  seed_tracks = NULL,
  max_acousticness = NULL,
  max_danceability = NULL,
  max_duration_ms = NULL,
  max_energy = NULL,
  max_instrumentalness = NULL,
  max_key = NULL,
  max_liveness = NULL,
  max_loudness = NULL,
  max_mode = NULL,
  max_popularity = NULL,
  max_speechiness = NULL,
  max_tempo = NULL,
  max_time_signature = NULL,
  max_valence = NULL,
  min_acousticness = NULL,
  min_danceability = NULL,
  min_duration_ms = NULL,
  min_energy = NULL,
  min_instrumentalness = NULL,
  min_key = NULL,
  min_liveness = NULL,
  min_loudness = NULL,
  min_mode = NULL,
  min_popularity = NULL,
  min_speechiness = NULL,
  min_tempo = NULL,
  min_time_signature = NULL,
  min_valence = NULL,
  target_acousticness = NULL,
  target_danceability = NULL,
  target_duration_ms = NULL,
  target_energy = NULL,
  target_instrumentalness = NULL,
  target_key = NULL,
  target_liveness = NULL,
  target_loudness = NULL,
  target_mode = NULL,
  target_popularity = NULL,
  target_speechiness = NULL,
  target_tempo = NULL,
  target_time_signature = NULL,
  target_valence = NULL,
  authorization = get_spotify_access_token(),
  include_seeds_in_response = FALSE
)

Arguments

limit
Optional. The target size of the list of recommended tracks. For seeds with unusually small pools or when highly restrictive filtering is applied, it may be impossible to generate the requested number of recommended tracks. Debugging information for such cases is available in the response. Default: 20. Minimum: 1. Maximum: 100.

market
Optional. An ISO 3166-1 alpha-2 country code or the string from_token. Provide this parameter if you want to apply Track Relinking. Because min_*, max_* and target_* are applied to pools before relinking, the generated results may not precisely match the filters applied. Original, non-relinked tracks are available via the linked_from attribute of the relinked track response.

seed_artists
A character vector of Spotify IDs for seed artists. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_genres
A character vector of any genres in the set of available genre seeds. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

seed_tracks
A character vector of Spotify IDs for a seed track. Up to 5 seed values may be provided in any combination of seed_artists, seed_tracks and seed_genres.

max_acousticness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' acousticness.

max_danceability
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' danceability.

max_duration_ms
Optional. Integer indicating a hard ceiling on recommended tracks' duration in milliseconds.

max_energy
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' energy.

max_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' instrumentalness.

max_key
Optional. Integer ranging from 1 to 12 indicating a hard ceiling on recommended tracks' key.

max_liveness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' liveness.

max_loudness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' loudness.

max_mode
Optional. Integer ranging from 0 to 1 indicating a hard ceiling on recommended tracks' mode.

max_popularity
Optional. Integer ranging between 0 and 100 indicating a hard ceiling on recommended tracks' popularity.

max_speechiness
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' speechiness.

max_tempo
Optional. Numeric value indicating a hard ceiling on recommended tracks' tempo.

max_time_signature
Optional. Integer indicating a hard ceiling on recommended tracks' time signature.

max_valence
Optional. Numeric value between 0 and 1 indicating a hard ceiling on recommended tracks' valence.

min_acousticness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' acousticness.

min_danceability
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' danceability.

min_duration_ms
Optional. Integer indicating a hard floor on recommended tracks' duration in milliseconds.

min_energy
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' energy.

min_instrumentalness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' instrumentalness.

min_key
Optional. Integer ranging from 1 to 12 indicating a hard floor on recommended tracks' key.

min_liveness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' liveness.

min_loudness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' loudness.

min_mode
Optional. Integer ranging from 0 to 1 indicating a hard floor on recommended tracks' mode.

min_popularity
Optional. Integer ranging between 0 and 100 indicating a hard floor on recommended tracks' popularity.

min_speechiness
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' speechiness.

min_tempo
Optional. Numeric value indicating a hard floor on recommended tracks' tempo.

min_time_signature
Optional. Integer indicating a hard floor on recommended tracks' time signature.

min_valence
Optional. Numeric value between 0 and 1 indicating a hard floor on recommended tracks' valence.

target_acousticness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' acousticness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_danceability
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' danceability. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_duration_ms
Optional. Integer indicating a target value for recommended tracks' duration in milliseconds. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_energy
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' energy. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_instrumentalness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' instrumentalness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_key
Optional. Integer ranging from 1 to 12 indicating a target value for recommended tracks' key. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_liveness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' liveness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_loudness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' loudness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_mode
Optional. Integer ranging from 0 to 1 indicating a target value for recommended tracks' mode. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_popularity
Optional. Integer ranging between 0 and 100 indicating a target value for recommended tracks' popularity. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_speechiness
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' speechiness. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_tempo
Optional. Numeric value indicating a target value for recommended tracks' tempo. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_time_signature
Optional. Integer indicating a target value for recommended tracks' time signature. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

target_valence
Optional. Numeric value between 0 and 1 indicating a target value for recommended tracks' valence. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request target_energy = 0.6 and target_danceability = 0.8. All target values will be weighed equally in ranking results.

authorization
Required. A valid access token from the Spotify Accounts service. See the Web API authorization guide for more details. Defaults to spotifyr::get_spotify_access_token()

include_seeds_in_response
Optional. Boolean for whether to include seed object in response. Defaults to FALSE.

Value

Returns a data frame of results recommendations. See the official Spotify Web API documentation for more information.

See also

Examples

 get_recommendations(market = 'SE', seed_genres = 'rock') 

#>                                                                                                                                                                                                              artists
#> 1                     https://api.spotify.com/v1/artists/0PFtn5NtBbbUNbU9EAmIWF, 0PFtn5NtBbbUNbU9EAmIWF, TOTO, artist, spotify:artist:0PFtn5NtBbbUNbU9EAmIWF, https://open.spotify.com/artist/0PFtn5NtBbbUNbU9EAmIWF
#> 2           https://api.spotify.com/v1/artists/00YTqRClk82aMchQQpYMd5, 00YTqRClk82aMchQQpYMd5, Our Last Night, artist, spotify:artist:00YTqRClk82aMchQQpYMd5, https://open.spotify.com/artist/00YTqRClk82aMchQQpYMd5
#> 3                   https://api.spotify.com/v1/artists/3jOstUTkEu2JkjvRdBA5Gu, 3jOstUTkEu2JkjvRdBA5Gu, Weezer, artist, spotify:artist:3jOstUTkEu2JkjvRdBA5Gu, https://open.spotify.com/artist/3jOstUTkEu2JkjvRdBA5Gu
#> 4         https://api.spotify.com/v1/artists/00tVTdpEhQQw1bqdu8RCx2, 00tVTdpEhQQw1bqdu8RCx2, Blue Öyster Cult, artist, spotify:artist:00tVTdpEhQQw1bqdu8RCx2, https://open.spotify.com/artist/00tVTdpEhQQw1bqdu8RCx2
#> 5             https://api.spotify.com/v1/artists/776Uo845nYHJpNaStv1Ds4, 776Uo845nYHJpNaStv1Ds4, Jimi Hendrix, artist, spotify:artist:776Uo845nYHJpNaStv1Ds4, https://open.spotify.com/artist/776Uo845nYHJpNaStv1Ds4
#> 6                https://api.spotify.com/v1/artists/7oPftvlwr6VrsViSDV7fJY, 7oPftvlwr6VrsViSDV7fJY, Green Day, artist, spotify:artist:7oPftvlwr6VrsViSDV7fJY, https://open.spotify.com/artist/7oPftvlwr6VrsViSDV7fJY
#> 7              https://api.spotify.com/v1/artists/165ZgPlLkK7bf5bDoFc6Sb, 165ZgPlLkK7bf5bDoFc6Sb, Limp Bizkit, artist, spotify:artist:165ZgPlLkK7bf5bDoFc6Sb, https://open.spotify.com/artist/165ZgPlLkK7bf5bDoFc6Sb
#> 8              https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 9                 https://api.spotify.com/v1/artists/6Ghvu1VvMGScGpOUJBAHNH, 6Ghvu1VvMGScGpOUJBAHNH, Deftones, artist, spotify:artist:6Ghvu1VvMGScGpOUJBAHNH, https://open.spotify.com/artist/6Ghvu1VvMGScGpOUJBAHNH
#> 10                   https://api.spotify.com/v1/artists/3JaAGmSTpJK35DqWrDUzBz, 3JaAGmSTpJK35DqWrDUzBz, Sweet, artist, spotify:artist:3JaAGmSTpJK35DqWrDUzBz, https://open.spotify.com/artist/3JaAGmSTpJK35DqWrDUzBz
#> 11             https://api.spotify.com/v1/artists/0cc6vw3VN8YlIcvr1v7tBL, 0cc6vw3VN8YlIcvr1v7tBL, Mötley Crüe, artist, spotify:artist:0cc6vw3VN8YlIcvr1v7tBL, https://open.spotify.com/artist/0cc6vw3VN8YlIcvr1v7tBL
#> 12               https://api.spotify.com/v1/artists/2cnMpRsOVqtPMfq7YiFE6K, 2cnMpRsOVqtPMfq7YiFE6K, Van Halen, artist, spotify:artist:2cnMpRsOVqtPMfq7YiFE6K, https://open.spotify.com/artist/2cnMpRsOVqtPMfq7YiFE6K
#> 13       https://api.spotify.com/v1/artists/4NiJW4q9ichVqL1aUsgGAN, 4NiJW4q9ichVqL1aUsgGAN, A Day To Remember, artist, spotify:artist:4NiJW4q9ichVqL1aUsgGAN, https://open.spotify.com/artist/4NiJW4q9ichVqL1aUsgGAN
#> 14 https://api.spotify.com/v1/artists/5t28BP42x2axFnqOOMg3CM, 5t28BP42x2axFnqOOMg3CM, Five Finger Death Punch, artist, spotify:artist:5t28BP42x2axFnqOOMg3CM, https://open.spotify.com/artist/5t28BP42x2axFnqOOMg3CM
#> 15               https://api.spotify.com/v1/artists/3TOqt5oJwL9BE2NG9MEwDa, 3TOqt5oJwL9BE2NG9MEwDa, Disturbed, artist, spotify:artist:3TOqt5oJwL9BE2NG9MEwDa, https://open.spotify.com/artist/3TOqt5oJwL9BE2NG9MEwDa
#> 16           https://api.spotify.com/v1/artists/08GQAI4eElDnROBrJRGE0X, 08GQAI4eElDnROBrJRGE0X, Fleetwood Mac, artist, spotify:artist:08GQAI4eElDnROBrJRGE0X, https://open.spotify.com/artist/08GQAI4eElDnROBrJRGE0X
#> 17           https://api.spotify.com/v1/artists/2otRZcOVNcmdkZ5nEoogy6, 2otRZcOVNcmdkZ5nEoogy6, Dead by April, artist, spotify:artist:2otRZcOVNcmdkZ5nEoogy6, https://open.spotify.com/artist/2otRZcOVNcmdkZ5nEoogy6
#> 18       https://api.spotify.com/v1/artists/4F84IBURUo98rz4r61KF70, 4F84IBURUo98rz4r61KF70, The White Stripes, artist, spotify:artist:4F84IBURUo98rz4r61KF70, https://open.spotify.com/artist/4F84IBURUo98rz4r61KF70
#> 19     https://api.spotify.com/v1/artists/74eX4C98E4FCrAMl39qRsJ, 74eX4C98E4FCrAMl39qRsJ, Theory of a Deadman, artist, spotify:artist:74eX4C98E4FCrAMl39qRsJ, https://open.spotify.com/artist/74eX4C98E4FCrAMl39qRsJ
#> 20            https://api.spotify.com/v1/artists/10Z7WzKMeIdNBKexi1YarP, 10Z7WzKMeIdNBKexi1YarP, Fame on Fire, artist, spotify:artist:10Z7WzKMeIdNBKexi1YarP, https://open.spotify.com/artist/10Z7WzKMeIdNBKexi1YarP
#>    disc_number duration_ms explicit
#> 1            1      235546    FALSE
#> 2            1      238181    FALSE
#> 3            1      195893    FALSE
#> 4            1      271000    FALSE
#> 5            1      146573    FALSE
#> 6            1      224760    FALSE
#> 7            1      272973     TRUE
#> 8            1      366136    FALSE
#> 9            1      214760    FALSE
#> 10           1      241746    FALSE
#> 11           1      281960    FALSE
#> 12           1      227693    FALSE
#> 13           1      202640    FALSE
#> 14           1      209703     TRUE
#> 15           1      252002    FALSE
#> 16           1      257800    FALSE
#> 17           1      265226    FALSE
#> 18           1      231733    FALSE
#> 19           1      219360     TRUE
#> 20           1      236250    FALSE
#>                                                        href
#> 1  https://api.spotify.com/v1/tracks/4aVuWgvD0X63hcOCnZtNFA
#> 2  https://api.spotify.com/v1/tracks/506YO37JuV40cnjeVuwlGD
#> 3  https://api.spotify.com/v1/tracks/2jz1bw1p0WQj0PDnVDP0uY
#> 4  https://api.spotify.com/v1/tracks/3fkPMWQ6cBNBLuFcPyMS8s
#> 5  https://api.spotify.com/v1/tracks/1ntxpzIUbSsizvuAy6lTYY
#> 6  https://api.spotify.com/v1/tracks/19YmvsVCetCBeVj6O2mljR
#> 7  https://api.spotify.com/v1/tracks/2gSVKxPDww9Eep5rdvtdem
#> 8  https://api.spotify.com/v1/tracks/3uMmllZo1AfoEnVT4ENCD3
#> 9  https://api.spotify.com/v1/tracks/1158ckiB5S4cpsdYHDB9IF
#> 10 https://api.spotify.com/v1/tracks/61ZVP6ytLwYnbLonbJtzR6
#> 11 https://api.spotify.com/v1/tracks/41wWrTueWew7qUU0Ru87Mp
#> 12 https://api.spotify.com/v1/tracks/0TaT50ZZxT4ytZxuqkE3A9
#> 13 https://api.spotify.com/v1/tracks/4JsDHMv5PVO8N07DbDq33r
#> 14 https://api.spotify.com/v1/tracks/0nqIqABZcujg05mkixJV60
#> 15 https://api.spotify.com/v1/tracks/3jjU4Pky1ja5J1onU6ei4T
#> 16 https://api.spotify.com/v1/tracks/0ofHAoxe9vBkTCp2UQIavz
#> 17 https://api.spotify.com/v1/tracks/4cfEA7zQhaQpkhUFonoOPw
#> 18 https://api.spotify.com/v1/tracks/3dPQuX8Gs42Y7b454ybpMR
#> 19 https://api.spotify.com/v1/tracks/6kp8F9FpHXzjga2MeRDSG8
#> 20 https://api.spotify.com/v1/tracks/54SMw8TnDcuieolVRXBmni
#>                        id is_local is_playable                           name
#> 1  4aVuWgvD0X63hcOCnZtNFA    FALSE        TRUE                  Hold the Line
#> 2  506YO37JuV40cnjeVuwlGD    FALSE        TRUE                  Common Ground
#> 3  2jz1bw1p0WQj0PDnVDP0uY    FALSE        TRUE              Feels Like Summer
#> 4  3fkPMWQ6cBNBLuFcPyMS8s    FALSE        TRUE                Burnin' for You
#> 5  1ntxpzIUbSsizvuAy6lTYY    FALSE        TRUE              Crosstown Traffic
#> 6  19YmvsVCetCBeVj6O2mljR    FALSE        TRUE                Still Breathing
#> 7  2gSVKxPDww9Eep5rdvtdem    FALSE        TRUE                         My Way
#> 8  3uMmllZo1AfoEnVT4ENCD3    FALSE        TRUE Highway Star - Remastered 2012
#> 9  1158ckiB5S4cpsdYHDB9IF    FALSE        TRUE       My Own Summer (Shove It)
#> 10 61ZVP6ytLwYnbLonbJtzR6    FALSE        TRUE             The Ballroom Blitz
#> 11 41wWrTueWew7qUU0Ru87Mp    FALSE        TRUE                      Wild Side
#> 12 0TaT50ZZxT4ytZxuqkE3A9    FALSE        TRUE         Why Can't This Be Love
#> 13 4JsDHMv5PVO8N07DbDq33r    FALSE        TRUE                     All I Want
#> 14 0nqIqABZcujg05mkixJV60    FALSE        TRUE                      Sham Pain
#> 15 3jjU4Pky1ja5J1onU6ei4T    FALSE        TRUE               The Vengeful One
#> 16 0ofHAoxe9vBkTCp2UQIavz    FALSE        TRUE         Dreams - 2004 Remaster
#> 17 4cfEA7zQhaQpkhUFonoOPw    FALSE        TRUE                Crying Over You
#> 18 3dPQuX8Gs42Y7b454ybpMR    FALSE        TRUE              Seven Nation Army
#> 19 6kp8F9FpHXzjga2MeRDSG8    FALSE        TRUE                Bitch Came Back
#> 20 54SMw8TnDcuieolVRXBmni    FALSE        TRUE                   Shape of You
#>    popularity
#> 1          77
#> 2          51
#> 3          58
#> 4          68
#> 5          60
#> 6          62
#> 7          71
#> 8          68
#> 9          70
#> 10         45
#> 11         66
#> 12         66
#> 13         65
#> 14         59
#> 15         67
#> 16         85
#> 17         55
#> 18         83
#> 19         63
#> 20         55
#>                                                                                                    preview_url
#> 1  https://p.scdn.co/mp3-preview/ae5b5990efb6105af7ccf2d5b2690c1bdb08cdf3?cid=125a10da5fa94687b3d0913168bc3698
#> 2  https://p.scdn.co/mp3-preview/dba3fc2354613666047a632d23854a17751f03ec?cid=125a10da5fa94687b3d0913168bc3698
#> 3  https://p.scdn.co/mp3-preview/530a0fa1eaf7bc59c46b4695610be07e7ce80d12?cid=125a10da5fa94687b3d0913168bc3698
#> 4  https://p.scdn.co/mp3-preview/7df1a3c5045bdde3deed8f62414e5bfe086feda8?cid=125a10da5fa94687b3d0913168bc3698
#> 5  https://p.scdn.co/mp3-preview/1497a8c3e9977b13b43f41a83ac33b3884b3ce82?cid=125a10da5fa94687b3d0913168bc3698
#> 6  https://p.scdn.co/mp3-preview/43872cebf1362055bf4fc1997ccb1992e2996b0d?cid=125a10da5fa94687b3d0913168bc3698
#> 7                                                                                                         <NA>
#> 8                                                                                                         <NA>
#> 9  https://p.scdn.co/mp3-preview/8d311acfa67d329ce5eec2dbcf336c330c70533a?cid=125a10da5fa94687b3d0913168bc3698
#> 10 https://p.scdn.co/mp3-preview/8a8bd2a0ca2db5a0637e01fa80a0d8bf3cf7d6e3?cid=125a10da5fa94687b3d0913168bc3698
#> 11 https://p.scdn.co/mp3-preview/d00f10e517b0d9b47ee4f14abfbe375adbc40d9c?cid=125a10da5fa94687b3d0913168bc3698
#> 12 https://p.scdn.co/mp3-preview/618e555b21c1a4f3819a86c6fd838e92e022be72?cid=125a10da5fa94687b3d0913168bc3698
#> 13                                                                                                        <NA>
#> 14 https://p.scdn.co/mp3-preview/795470dd9a181e2aa86ad517856e532e67b84971?cid=125a10da5fa94687b3d0913168bc3698
#> 15 https://p.scdn.co/mp3-preview/5a2d57f9cc6e6d4fde4e618dd47666535f7c1a55?cid=125a10da5fa94687b3d0913168bc3698
#> 16 https://p.scdn.co/mp3-preview/40d1f6d82f16b05b62aae47e8e130de04b407fc8?cid=125a10da5fa94687b3d0913168bc3698
#> 17                                                                                                        <NA>
#> 18 https://p.scdn.co/mp3-preview/9ca99ae29be0130196f0f2257b6d88ede85f7099?cid=125a10da5fa94687b3d0913168bc3698
#> 19 https://p.scdn.co/mp3-preview/3a3dd902b854edb960023224574ab781740c43c9?cid=125a10da5fa94687b3d0913168bc3698
#> 20 https://p.scdn.co/mp3-preview/752d58f11d845a4462a7d2878ce35757be8906b9?cid=125a10da5fa94687b3d0913168bc3698
#>    track_number  type                                  uri album.album_type
#> 1             9 track spotify:track:4aVuWgvD0X63hcOCnZtNFA            ALBUM
#> 2             6 track spotify:track:506YO37JuV40cnjeVuwlGD            ALBUM
#> 3             3 track spotify:track:2jz1bw1p0WQj0PDnVDP0uY            ALBUM
#> 4             2 track spotify:track:3fkPMWQ6cBNBLuFcPyMS8s            ALBUM
#> 5             3 track spotify:track:1ntxpzIUbSsizvuAy6lTYY            ALBUM
#> 6             7 track spotify:track:19YmvsVCetCBeVj6O2mljR            ALBUM
#> 7             5 track spotify:track:2gSVKxPDww9Eep5rdvtdem            ALBUM
#> 8             1 track spotify:track:3uMmllZo1AfoEnVT4ENCD3            ALBUM
#> 9             1 track spotify:track:1158ckiB5S4cpsdYHDB9IF            ALBUM
#> 10            9 track spotify:track:61ZVP6ytLwYnbLonbJtzR6      COMPILATION
#> 11            1 track spotify:track:41wWrTueWew7qUU0Ru87Mp            ALBUM
#> 12            2 track spotify:track:0TaT50ZZxT4ytZxuqkE3A9            ALBUM
#> 13            2 track spotify:track:4JsDHMv5PVO8N07DbDq33r            ALBUM
#> 14            4 track spotify:track:0nqIqABZcujg05mkixJV60            ALBUM
#> 15            3 track spotify:track:3jjU4Pky1ja5J1onU6ei4T            ALBUM
#> 16            2 track spotify:track:0ofHAoxe9vBkTCp2UQIavz            ALBUM
#> 17            1 track spotify:track:4cfEA7zQhaQpkhUFonoOPw            ALBUM
#> 18            1 track spotify:track:3dPQuX8Gs42Y7b454ybpMR            ALBUM
#> 19            2 track spotify:track:6kp8F9FpHXzjga2MeRDSG8            ALBUM
#> 20            1 track spotify:track:54SMw8TnDcuieolVRXBmni           SINGLE
#>                                                                                                                                                                                                        album.artists
#> 1                     https://api.spotify.com/v1/artists/0PFtn5NtBbbUNbU9EAmIWF, 0PFtn5NtBbbUNbU9EAmIWF, TOTO, artist, spotify:artist:0PFtn5NtBbbUNbU9EAmIWF, https://open.spotify.com/artist/0PFtn5NtBbbUNbU9EAmIWF
#> 2           https://api.spotify.com/v1/artists/00YTqRClk82aMchQQpYMd5, 00YTqRClk82aMchQQpYMd5, Our Last Night, artist, spotify:artist:00YTqRClk82aMchQQpYMd5, https://open.spotify.com/artist/00YTqRClk82aMchQQpYMd5
#> 3                   https://api.spotify.com/v1/artists/3jOstUTkEu2JkjvRdBA5Gu, 3jOstUTkEu2JkjvRdBA5Gu, Weezer, artist, spotify:artist:3jOstUTkEu2JkjvRdBA5Gu, https://open.spotify.com/artist/3jOstUTkEu2JkjvRdBA5Gu
#> 4         https://api.spotify.com/v1/artists/00tVTdpEhQQw1bqdu8RCx2, 00tVTdpEhQQw1bqdu8RCx2, Blue Öyster Cult, artist, spotify:artist:00tVTdpEhQQw1bqdu8RCx2, https://open.spotify.com/artist/00tVTdpEhQQw1bqdu8RCx2
#> 5             https://api.spotify.com/v1/artists/776Uo845nYHJpNaStv1Ds4, 776Uo845nYHJpNaStv1Ds4, Jimi Hendrix, artist, spotify:artist:776Uo845nYHJpNaStv1Ds4, https://open.spotify.com/artist/776Uo845nYHJpNaStv1Ds4
#> 6                https://api.spotify.com/v1/artists/7oPftvlwr6VrsViSDV7fJY, 7oPftvlwr6VrsViSDV7fJY, Green Day, artist, spotify:artist:7oPftvlwr6VrsViSDV7fJY, https://open.spotify.com/artist/7oPftvlwr6VrsViSDV7fJY
#> 7              https://api.spotify.com/v1/artists/165ZgPlLkK7bf5bDoFc6Sb, 165ZgPlLkK7bf5bDoFc6Sb, Limp Bizkit, artist, spotify:artist:165ZgPlLkK7bf5bDoFc6Sb, https://open.spotify.com/artist/165ZgPlLkK7bf5bDoFc6Sb
#> 8              https://api.spotify.com/v1/artists/568ZhdwyaiCyOGJRtNYhWf, 568ZhdwyaiCyOGJRtNYhWf, Deep Purple, artist, spotify:artist:568ZhdwyaiCyOGJRtNYhWf, https://open.spotify.com/artist/568ZhdwyaiCyOGJRtNYhWf
#> 9                 https://api.spotify.com/v1/artists/6Ghvu1VvMGScGpOUJBAHNH, 6Ghvu1VvMGScGpOUJBAHNH, Deftones, artist, spotify:artist:6Ghvu1VvMGScGpOUJBAHNH, https://open.spotify.com/artist/6Ghvu1VvMGScGpOUJBAHNH
#> 10                   https://api.spotify.com/v1/artists/3JaAGmSTpJK35DqWrDUzBz, 3JaAGmSTpJK35DqWrDUzBz, Sweet, artist, spotify:artist:3JaAGmSTpJK35DqWrDUzBz, https://open.spotify.com/artist/3JaAGmSTpJK35DqWrDUzBz
#> 11             https://api.spotify.com/v1/artists/0cc6vw3VN8YlIcvr1v7tBL, 0cc6vw3VN8YlIcvr1v7tBL, Mötley Crüe, artist, spotify:artist:0cc6vw3VN8YlIcvr1v7tBL, https://open.spotify.com/artist/0cc6vw3VN8YlIcvr1v7tBL
#> 12               https://api.spotify.com/v1/artists/2cnMpRsOVqtPMfq7YiFE6K, 2cnMpRsOVqtPMfq7YiFE6K, Van Halen, artist, spotify:artist:2cnMpRsOVqtPMfq7YiFE6K, https://open.spotify.com/artist/2cnMpRsOVqtPMfq7YiFE6K
#> 13       https://api.spotify.com/v1/artists/4NiJW4q9ichVqL1aUsgGAN, 4NiJW4q9ichVqL1aUsgGAN, A Day To Remember, artist, spotify:artist:4NiJW4q9ichVqL1aUsgGAN, https://open.spotify.com/artist/4NiJW4q9ichVqL1aUsgGAN
#> 14 https://api.spotify.com/v1/artists/5t28BP42x2axFnqOOMg3CM, 5t28BP42x2axFnqOOMg3CM, Five Finger Death Punch, artist, spotify:artist:5t28BP42x2axFnqOOMg3CM, https://open.spotify.com/artist/5t28BP42x2axFnqOOMg3CM
#> 15               https://api.spotify.com/v1/artists/3TOqt5oJwL9BE2NG9MEwDa, 3TOqt5oJwL9BE2NG9MEwDa, Disturbed, artist, spotify:artist:3TOqt5oJwL9BE2NG9MEwDa, https://open.spotify.com/artist/3TOqt5oJwL9BE2NG9MEwDa
#> 16           https://api.spotify.com/v1/artists/08GQAI4eElDnROBrJRGE0X, 08GQAI4eElDnROBrJRGE0X, Fleetwood Mac, artist, spotify:artist:08GQAI4eElDnROBrJRGE0X, https://open.spotify.com/artist/08GQAI4eElDnROBrJRGE0X
#> 17           https://api.spotify.com/v1/artists/2otRZcOVNcmdkZ5nEoogy6, 2otRZcOVNcmdkZ5nEoogy6, Dead by April, artist, spotify:artist:2otRZcOVNcmdkZ5nEoogy6, https://open.spotify.com/artist/2otRZcOVNcmdkZ5nEoogy6
#> 18       https://api.spotify.com/v1/artists/4F84IBURUo98rz4r61KF70, 4F84IBURUo98rz4r61KF70, The White Stripes, artist, spotify:artist:4F84IBURUo98rz4r61KF70, https://open.spotify.com/artist/4F84IBURUo98rz4r61KF70
#> 19     https://api.spotify.com/v1/artists/74eX4C98E4FCrAMl39qRsJ, 74eX4C98E4FCrAMl39qRsJ, Theory of a Deadman, artist, spotify:artist:74eX4C98E4FCrAMl39qRsJ, https://open.spotify.com/artist/74eX4C98E4FCrAMl39qRsJ
#> 20            https://api.spotify.com/v1/artists/10Z7WzKMeIdNBKexi1YarP, 10Z7WzKMeIdNBKexi1YarP, Fame on Fire, artist, spotify:artist:10Z7WzKMeIdNBKexi1YarP, https://open.spotify.com/artist/10Z7WzKMeIdNBKexi1YarP
#>                                                  album.href
#> 1  https://api.spotify.com/v1/albums/1mnu4hYvdwQgZXcNvtJ3D3
#> 2  https://api.spotify.com/v1/albums/482Zolwz9ekvPD4kfj8BGo
#> 3  https://api.spotify.com/v1/albums/3EwfQtjvyRAXsPWAKO5FDP
#> 4  https://api.spotify.com/v1/albums/7v4kEpVtppoMm80m43lGzt
#> 5  https://api.spotify.com/v1/albums/5z090LQztiqh13wYspQvKQ
#> 6  https://api.spotify.com/v1/albums/5a3LqvNt2nv1B4aRKXmgOV
#> 7  https://api.spotify.com/v1/albums/5mi7FKaWE5CtcOjdyxScA7
#> 8  https://api.spotify.com/v1/albums/32NQ56VZDTXSH3SMv4XSGN
#> 9  https://api.spotify.com/v1/albums/7o4UsmV37Sg5It2Eb7vHzu
#> 10 https://api.spotify.com/v1/albums/710mrF7sQJjTraiDVjVE49
#> 11 https://api.spotify.com/v1/albums/0Tnve4bDxyiBCRaj03qO8z
#> 12 https://api.spotify.com/v1/albums/5Ew5vOg3PYyT9QUPCrdIZq
#> 13 https://api.spotify.com/v1/albums/6I2SUr0MolRUiDEcZAbSWm
#> 14 https://api.spotify.com/v1/albums/2IA9d7atwNgtiS6Tpv9WKt
#> 15 https://api.spotify.com/v1/albums/1ROVEUk4lc2vD4a0IN1TWS
#> 16 https://api.spotify.com/v1/albums/0BwWUstDMUbgq2NYONRqlu
#> 17 https://api.spotify.com/v1/albums/3rgz4IjijsFvJwqicCqle3
#> 18 https://api.spotify.com/v1/albums/6D9urpsOWWKtYvF6PaorGE
#> 19 https://api.spotify.com/v1/albums/2JUpjRRxWKUZwVwY9sTXS1
#> 20 https://api.spotify.com/v1/albums/4eYjV7ET5e8iOtyxg5CLNq
#>                  album.id
#> 1  1mnu4hYvdwQgZXcNvtJ3D3
#> 2  482Zolwz9ekvPD4kfj8BGo
#> 3  3EwfQtjvyRAXsPWAKO5FDP
#> 4  7v4kEpVtppoMm80m43lGzt
#> 5  5z090LQztiqh13wYspQvKQ
#> 6  5a3LqvNt2nv1B4aRKXmgOV
#> 7  5mi7FKaWE5CtcOjdyxScA7
#> 8  32NQ56VZDTXSH3SMv4XSGN
#> 9  7o4UsmV37Sg5It2Eb7vHzu
#> 10 710mrF7sQJjTraiDVjVE49
#> 11 0Tnve4bDxyiBCRaj03qO8z
#> 12 5Ew5vOg3PYyT9QUPCrdIZq
#> 13 6I2SUr0MolRUiDEcZAbSWm
#> 14 2IA9d7atwNgtiS6Tpv9WKt
#> 15 1ROVEUk4lc2vD4a0IN1TWS
#> 16 0BwWUstDMUbgq2NYONRqlu
#> 17 3rgz4IjijsFvJwqicCqle3
#> 18 6D9urpsOWWKtYvF6PaorGE
#> 19 2JUpjRRxWKUZwVwY9sTXS1
#> 20 4eYjV7ET5e8iOtyxg5CLNq
#>                                                                                                                                                                                                                        album.images
#> 1  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f903e62767a0e22e33b7af83, https://i.scdn.co/image/ab67616d00001e02f903e62767a0e22e33b7af83, https://i.scdn.co/image/ab67616d00004851f903e62767a0e22e33b7af83, 640, 300, 64
#> 2  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273c3becb63afa2b5f9a1c7623d, https://i.scdn.co/image/ab67616d00001e02c3becb63afa2b5f9a1c7623d, https://i.scdn.co/image/ab67616d00004851c3becb63afa2b5f9a1c7623d, 640, 300, 64
#> 3  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273f04ed0e9d5af8b1b5f011b2f, https://i.scdn.co/image/ab67616d00001e02f04ed0e9d5af8b1b5f011b2f, https://i.scdn.co/image/ab67616d00004851f04ed0e9d5af8b1b5f011b2f, 640, 300, 64
#> 4  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2736895da04b237964b95c2a166, https://i.scdn.co/image/ab67616d00001e026895da04b237964b95c2a166, https://i.scdn.co/image/ab67616d000048516895da04b237964b95c2a166, 640, 300, 64
#> 5  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273522088789d49e216d9818292, https://i.scdn.co/image/ab67616d00001e02522088789d49e216d9818292, https://i.scdn.co/image/ab67616d00004851522088789d49e216d9818292, 640, 300, 64
#> 6  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273af71e889d130283bcd08b8a7, https://i.scdn.co/image/ab67616d00001e02af71e889d130283bcd08b8a7, https://i.scdn.co/image/ab67616d00004851af71e889d130283bcd08b8a7, 640, 300, 64
#> 7  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2734a31b146c7cf07705d912efe, https://i.scdn.co/image/ab67616d00001e024a31b146c7cf07705d912efe, https://i.scdn.co/image/ab67616d000048514a31b146c7cf07705d912efe, 640, 300, 64
#> 8  640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00001e02a729c9c3dec04b99d889c66f, https://i.scdn.co/image/ab67616d00004851a729c9c3dec04b99d889c66f, 640, 300, 64
#> 9  640, 300, 64, https://i.scdn.co/image/ab67616d0000b2730b1129853982ea17845d4eb6, https://i.scdn.co/image/ab67616d00001e020b1129853982ea17845d4eb6, https://i.scdn.co/image/ab67616d000048510b1129853982ea17845d4eb6, 640, 300, 64
#> 10 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a087b55c20710d687320737a, https://i.scdn.co/image/ab67616d00001e02a087b55c20710d687320737a, https://i.scdn.co/image/ab67616d00004851a087b55c20710d687320737a, 640, 300, 64
#> 11 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a226e31b10f88c83e00e828a, https://i.scdn.co/image/ab67616d00001e02a226e31b10f88c83e00e828a, https://i.scdn.co/image/ab67616d00004851a226e31b10f88c83e00e828a, 640, 300, 64
#> 12 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273def8994935a955331c21bc12, https://i.scdn.co/image/ab67616d00001e02def8994935a955331c21bc12, https://i.scdn.co/image/ab67616d00004851def8994935a955331c21bc12, 640, 300, 64
#> 13 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2739233270031da763169e5f96c, https://i.scdn.co/image/ab67616d00001e029233270031da763169e5f96c, https://i.scdn.co/image/ab67616d000048519233270031da763169e5f96c, 640, 300, 64
#> 14 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273fb94cdee49bcfbbefa8400ee, https://i.scdn.co/image/ab67616d00001e02fb94cdee49bcfbbefa8400ee, https://i.scdn.co/image/ab67616d00004851fb94cdee49bcfbbefa8400ee, 640, 300, 64
#> 15 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a58c62e75b82e9c3b47d9ef3, https://i.scdn.co/image/ab67616d00001e02a58c62e75b82e9c3b47d9ef3, https://i.scdn.co/image/ab67616d00004851a58c62e75b82e9c3b47d9ef3, 640, 300, 64
#> 16 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273e52a59a28efa4773dd2bfe1b, https://i.scdn.co/image/ab67616d00001e02e52a59a28efa4773dd2bfe1b, https://i.scdn.co/image/ab67616d00004851e52a59a28efa4773dd2bfe1b, 640, 300, 64
#> 17 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273387e3fe93a12844cf969399e, https://i.scdn.co/image/ab67616d00001e02387e3fe93a12844cf969399e, https://i.scdn.co/image/ab67616d00004851387e3fe93a12844cf969399e, 640, 300, 64
#> 18 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273a69f71a8794e2d867a52f98f, https://i.scdn.co/image/ab67616d00001e02a69f71a8794e2d867a52f98f, https://i.scdn.co/image/ab67616d00004851a69f71a8794e2d867a52f98f, 640, 300, 64
#> 19 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2731994d4fee1588a93a1c31745, https://i.scdn.co/image/ab67616d00001e021994d4fee1588a93a1c31745, https://i.scdn.co/image/ab67616d000048511994d4fee1588a93a1c31745, 640, 300, 64
#> 20 640, 300, 64, https://i.scdn.co/image/ab67616d0000b27336af4da408a79320bb873b49, https://i.scdn.co/image/ab67616d00001e0236af4da408a79320bb873b49, https://i.scdn.co/image/ab67616d0000485136af4da408a79320bb873b49, 640, 300, 64
#>                                           album.name album.release_date
#> 1                                               Toto         1978-10-10
#> 2                                  Selective Hearing         2017-06-09
#> 3                                   Pacific Daydream         2017-10-27
#> 4                             Fire of Unknown Origin         1981-10-23
#> 5                                  Electric Ladyland         1968-10-25
#> 6                                   Revolution Radio         2016-10-07
#> 7  Chocolate Starfish And The Hot Dog Flavored Water         2000-10-17
#> 8                          Machine Head (Remastered)         1972-03-25
#> 9                                     Around the Fur         1997-10-27
#> 10                                        Sweet Best               1976
#> 11                               Girls, Girls, Girls               1987
#> 12                                              5150         1986-03-24
#> 13                        What Separates Me From You         2010-11-16
#> 14                     And Justice for None (Deluxe)         2018-05-18
#> 15                     Immortalized (Deluxe Edition)         2015-08-15
#> 16                            Rumours (Super Deluxe)         1977-02-04
#> 17                                    Worlds Collide         2017-04-07
#> 18                                          Elephant         2003-04-01
#> 19                 The Truth Is... (Special Edition)         2011-07-08
#> 20                                      Shape of You         2017-01-29
#>    album.release_date_precision album.total_tracks album.type
#> 1                           day                 10      album
#> 2                           day                  7      album
#> 3                           day                 10      album
#> 4                           day                  9      album
#> 5                           day                 16      album
#> 6                           day                 12      album
#> 7                           day                 15      album
#> 8                           day                  8      album
#> 9                           day                 10      album
#> 10                         year                 22      album
#> 11                         year                 15      album
#> 12                          day                  9      album
#> 13                          day                 10      album
#> 14                          day                 16      album
#> 15                          day                 16      album
#> 16                          day                 58      album
#> 17                          day                 11      album
#> 18                          day                 14      album
#> 19                          day                 18      album
#> 20                          day                  1      album
#>                               album.uri
#> 1  spotify:album:1mnu4hYvdwQgZXcNvtJ3D3
#> 2  spotify:album:482Zolwz9ekvPD4kfj8BGo
#> 3  spotify:album:3EwfQtjvyRAXsPWAKO5FDP
#> 4  spotify:album:7v4kEpVtppoMm80m43lGzt
#> 5  spotify:album:5z090LQztiqh13wYspQvKQ
#> 6  spotify:album:5a3LqvNt2nv1B4aRKXmgOV
#> 7  spotify:album:5mi7FKaWE5CtcOjdyxScA7
#> 8  spotify:album:32NQ56VZDTXSH3SMv4XSGN
#> 9  spotify:album:7o4UsmV37Sg5It2Eb7vHzu
#> 10 spotify:album:710mrF7sQJjTraiDVjVE49
#> 11 spotify:album:0Tnve4bDxyiBCRaj03qO8z
#> 12 spotify:album:5Ew5vOg3PYyT9QUPCrdIZq
#> 13 spotify:album:6I2SUr0MolRUiDEcZAbSWm
#> 14 spotify:album:2IA9d7atwNgtiS6Tpv9WKt
#> 15 spotify:album:1ROVEUk4lc2vD4a0IN1TWS
#> 16 spotify:album:0BwWUstDMUbgq2NYONRqlu
#> 17 spotify:album:3rgz4IjijsFvJwqicCqle3
#> 18 spotify:album:6D9urpsOWWKtYvF6PaorGE
#> 19 spotify:album:2JUpjRRxWKUZwVwY9sTXS1
#> 20 spotify:album:4eYjV7ET5e8iOtyxg5CLNq
#>                              album.external_urls.spotify external_ids.isrc
#> 1  https://open.spotify.com/album/1mnu4hYvdwQgZXcNvtJ3D3      USSM17800444
#> 2  https://open.spotify.com/album/482Zolwz9ekvPD4kfj8BGo      GBDMT1600259
#> 3  https://open.spotify.com/album/3EwfQtjvyRAXsPWAKO5FDP      USAT21700603
#> 4  https://open.spotify.com/album/7v4kEpVtppoMm80m43lGzt      USSM10017652
#> 5  https://open.spotify.com/album/5z090LQztiqh13wYspQvKQ      USQX90900737
#> 6  https://open.spotify.com/album/5a3LqvNt2nv1B4aRKXmgOV      USRE11600432
#> 7  https://open.spotify.com/album/5mi7FKaWE5CtcOjdyxScA7      USIR10001177
#> 8  https://open.spotify.com/album/32NQ56VZDTXSH3SMv4XSGN      GBDXG1200002
#> 9  https://open.spotify.com/album/7o4UsmV37Sg5It2Eb7vHzu      USMV29700142
#> 10 https://open.spotify.com/album/710mrF7sQJjTraiDVjVE49      GBARL7300027
#> 11 https://open.spotify.com/album/0Tnve4bDxyiBCRaj03qO8z      USBY29900189
#> 12 https://open.spotify.com/album/5Ew5vOg3PYyT9QUPCrdIZq      USWB10902261
#> 13 https://open.spotify.com/album/6I2SUr0MolRUiDEcZAbSWm      USVIC1060302
#> 14 https://open.spotify.com/album/2IA9d7atwNgtiS6Tpv9WKt      US5Z51800005
#> 15 https://open.spotify.com/album/1ROVEUk4lc2vD4a0IN1TWS      USRE11500172
#> 16 https://open.spotify.com/album/0BwWUstDMUbgq2NYONRqlu      USWB11301111
#> 17 https://open.spotify.com/album/3rgz4IjijsFvJwqicCqle3      SEUM71601951
#> 18 https://open.spotify.com/album/6D9urpsOWWKtYvF6PaorGE      USVT10300001
#> 19 https://open.spotify.com/album/2JUpjRRxWKUZwVwY9sTXS1      NLA321191095
#> 20 https://open.spotify.com/album/4eYjV7ET5e8iOtyxg5CLNq      TCACX1729283
#>                                    external_urls.spotify
#> 1  https://open.spotify.com/track/4aVuWgvD0X63hcOCnZtNFA
#> 2  https://open.spotify.com/track/506YO37JuV40cnjeVuwlGD
#> 3  https://open.spotify.com/track/2jz1bw1p0WQj0PDnVDP0uY
#> 4  https://open.spotify.com/track/3fkPMWQ6cBNBLuFcPyMS8s
#> 5  https://open.spotify.com/track/1ntxpzIUbSsizvuAy6lTYY
#> 6  https://open.spotify.com/track/19YmvsVCetCBeVj6O2mljR
#> 7  https://open.spotify.com/track/2gSVKxPDww9Eep5rdvtdem
#> 8  https://open.spotify.com/track/3uMmllZo1AfoEnVT4ENCD3
#> 9  https://open.spotify.com/track/1158ckiB5S4cpsdYHDB9IF
#> 10 https://open.spotify.com/track/61ZVP6ytLwYnbLonbJtzR6
#> 11 https://open.spotify.com/track/41wWrTueWew7qUU0Ru87Mp
#> 12 https://open.spotify.com/track/0TaT50ZZxT4ytZxuqkE3A9
#> 13 https://open.spotify.com/track/4JsDHMv5PVO8N07DbDq33r
#> 14 https://open.spotify.com/track/0nqIqABZcujg05mkixJV60
#> 15 https://open.spotify.com/track/3jjU4Pky1ja5J1onU6ei4T
#> 16 https://open.spotify.com/track/0ofHAoxe9vBkTCp2UQIavz
#> 17 https://open.spotify.com/track/4cfEA7zQhaQpkhUFonoOPw
#> 18 https://open.spotify.com/track/3dPQuX8Gs42Y7b454ybpMR
#> 19 https://open.spotify.com/track/6kp8F9FpHXzjga2MeRDSG8
#> 20 https://open.spotify.com/track/54SMw8TnDcuieolVRXBmni
#>                                            linked_from.href
#> 1                                                      <NA>
#> 2                                                      <NA>
#> 3                                                      <NA>
#> 4                                                      <NA>
#> 5                                                      <NA>
#> 6                                                      <NA>
#> 7                                                      <NA>
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#> 11 https://api.spotify.com/v1/tracks/57ivEAWqHTUcTYeXc5lRvd
#> 12                                                     <NA>
#> 13 https://api.spotify.com/v1/tracks/3imITUqEhkebNbOoS1yrWm
#> 14                                                     <NA>
#> 15                                                     <NA>
#> 16                                                     <NA>
#> 17                                                     <NA>
#> 18 https://api.spotify.com/v1/tracks/1jNOi6m3Hn8nLEeHCp5Msr
#> 19                                                     <NA>
#> 20                                                     <NA>
#>            linked_from.id linked_from.type                      linked_from.uri
#> 1                    <NA>             <NA>                                 <NA>
#> 2                    <NA>             <NA>                                 <NA>
#> 3                    <NA>             <NA>                                 <NA>
#> 4                    <NA>             <NA>                                 <NA>
#> 5                    <NA>             <NA>                                 <NA>
#> 6                    <NA>             <NA>                                 <NA>
#> 7                    <NA>             <NA>                                 <NA>
#> 8                    <NA>             <NA>                                 <NA>
#> 9                    <NA>             <NA>                                 <NA>
#> 10                   <NA>             <NA>                                 <NA>
#> 11 57ivEAWqHTUcTYeXc5lRvd            track spotify:track:57ivEAWqHTUcTYeXc5lRvd
#> 12                   <NA>             <NA>                                 <NA>
#> 13 3imITUqEhkebNbOoS1yrWm            track spotify:track:3imITUqEhkebNbOoS1yrWm
#> 14                   <NA>             <NA>                                 <NA>
#> 15                   <NA>             <NA>                                 <NA>
#> 16                   <NA>             <NA>                                 <NA>
#> 17                   <NA>             <NA>                                 <NA>
#> 18 1jNOi6m3Hn8nLEeHCp5Msr            track spotify:track:1jNOi6m3Hn8nLEeHCp5Msr
#> 19                   <NA>             <NA>                                 <NA>
#> 20                   <NA>             <NA>                                 <NA>
#>                        linked_from.external_urls.spotify
#> 1                                                   <NA>
#> 2                                                   <NA>
#> 3                                                   <NA>
#> 4                                                   <NA>
#> 5                                                   <NA>
#> 6                                                   <NA>
#> 7                                                   <NA>
#> 8                                                   <NA>
#> 9                                                   <NA>
#> 10                                                  <NA>
#> 11 https://open.spotify.com/track/57ivEAWqHTUcTYeXc5lRvd
#> 12                                                  <NA>
#> 13 https://open.spotify.com/track/3imITUqEhkebNbOoS1yrWm
#> 14                                                  <NA>
#> 15                                                  <NA>
#> 16                                                  <NA>
#> 17                                                  <NA>
#> 18 https://open.spotify.com/track/1jNOi6m3Hn8nLEeHCp5Msr
#> 19                                                  <NA>
#> 20                                                  <NA>