Get recommendations for unlimited vector of track IDs

Get recommendations for a submitted vector of track IDs, with no limit on the number of seed tracks

get_recommendations_all(track_ids, valence = NULL)

Arguments

track_ids

A vector containing the IDs of the tracks you'd like recommendations for

valence

The target valence for the recommendations

Value

Returns a data frame containing binded recommendations from the Spotify API in a single data frame.

Details

This is a wrapper for the get_recommendations function, which provides a workaround for the limit of 5 seed tracks per recommendation call. The function splits a supplied vector of track IDs into subsets of length 5, then applies a get_recommendations call, 5 tracks at a time. This should generate a data frame of recommended tracks, with length equal to the supplied vector of track ids.

See also

Examples

# \donttest{
get_recommendations_all(
   track_ids = c("5VIpLopHgolKcSSj7JPCMA", "3QRGYDFFUVb4qneE4DX1gR")
   )
#>                                                                                                                                                                                                   artists
#> 1   https://api.spotify.com/v1/artists/3kjuyTCjPG1WMFCiyc5IuB, 3kjuyTCjPG1WMFCiyc5IuB, Arcade Fire, artist, spotify:artist:3kjuyTCjPG1WMFCiyc5IuB, https://open.spotify.com/artist/3kjuyTCjPG1WMFCiyc5IuB
#> 2 https://api.spotify.com/v1/artists/3KlPjpVKfm6vESPL46NDCh, 3KlPjpVKfm6vESPL46NDCh, Indigo Sparke, artist, spotify:artist:3KlPjpVKfm6vESPL46NDCh, https://open.spotify.com/artist/3KlPjpVKfm6vESPL46NDCh
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, ET, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#> 2     AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#>   disc_number duration_ms explicit
#> 1           1      287240    FALSE
#> 2           1      206950    FALSE
#>                                                       href
#> 1 https://api.spotify.com/v1/tracks/1rOlTL4pKQ9Y1fURua4AJR
#> 2 https://api.spotify.com/v1/tracks/0q0Vlb5g3DV8pysqIllbiR
#>                       id is_local                  name popularity
#> 1 1rOlTL4pKQ9Y1fURua4AJR    FALSE     My Body Is a Cage         61
#> 2 0q0Vlb5g3DV8pysqIllbiR    FALSE Everything Everything         35
#>                                                                                                   preview_url
#> 1 https://p.scdn.co/mp3-preview/cf1f881824d5c087480911530229051ef3519831?cid=0e73e430c7814968988cd21c104bf113
#> 2 https://p.scdn.co/mp3-preview/2b69522cf997bf8b58a153955ded73b4d9a98cc7?cid=0e73e430c7814968988cd21c104bf113
#>   track_number  type                                  uri album.album_type
#> 1           11 track spotify:track:1rOlTL4pKQ9Y1fURua4AJR            ALBUM
#> 2            9 track spotify:track:0q0Vlb5g3DV8pysqIllbiR            ALBUM
#>                                                                                                                                                                                             album.artists
#> 1   https://api.spotify.com/v1/artists/3kjuyTCjPG1WMFCiyc5IuB, 3kjuyTCjPG1WMFCiyc5IuB, Arcade Fire, artist, spotify:artist:3kjuyTCjPG1WMFCiyc5IuB, https://open.spotify.com/artist/3kjuyTCjPG1WMFCiyc5IuB
#> 2 https://api.spotify.com/v1/artists/3KlPjpVKfm6vESPL46NDCh, 3KlPjpVKfm6vESPL46NDCh, Indigo Sparke, artist, spotify:artist:3KlPjpVKfm6vESPL46NDCh, https://open.spotify.com/artist/3KlPjpVKfm6vESPL46NDCh
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          album.available_markets
#> 1     AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CD, CG, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, ET, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IQ, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, LY, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TJ, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VE, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                 album.href
#> 1 https://api.spotify.com/v1/albums/4GT6uZod4p5RiDMOVHOMme
#> 2 https://api.spotify.com/v1/albums/6akscmS833btWR5O7F5BMJ
#>                 album.id
#> 1 4GT6uZod4p5RiDMOVHOMme
#> 2 6akscmS833btWR5O7F5BMJ
#>                                                                                                                                                                                                                       album.images
#> 1 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2739e0e42aa866e5fa167b19608, https://i.scdn.co/image/ab67616d00001e029e0e42aa866e5fa167b19608, https://i.scdn.co/image/ab67616d000048519e0e42aa866e5fa167b19608, 640, 300, 64
#> 2 640, 300, 64, https://i.scdn.co/image/ab67616d0000b2732a8081249b6090274b9d4823, https://i.scdn.co/image/ab67616d00001e022a8081249b6090274b9d4823, https://i.scdn.co/image/ab67616d000048512a8081249b6090274b9d4823, 640, 300, 64
#>   album.name album.release_date album.release_date_precision album.total_tracks
#> 1 Neon Bible               2007                         year                 11
#> 2       Echo         2021-02-19                          day                  9
#>   album.type                            album.uri
#> 1      album spotify:album:4GT6uZod4p5RiDMOVHOMme
#> 2      album spotify:album:6akscmS833btWR5O7F5BMJ
#>                             album.external_urls.spotify external_ids.isrc
#> 1 https://open.spotify.com/album/4GT6uZod4p5RiDMOVHOMme      USMRG0728510
#> 2 https://open.spotify.com/album/6akscmS833btWR5O7F5BMJ      AUSXI1900009
#>                                   external_urls.spotify
#> 1 https://open.spotify.com/track/1rOlTL4pKQ9Y1fURua4AJR
#> 2 https://open.spotify.com/track/0q0Vlb5g3DV8pysqIllbiR
# }