Get recommendations for a submitted vector of track IDs, with no limit on the number of seed tracks

This is a wrapper for the get_recommendations function, which provides a workaround for the limit of 5 seed tracks per recommendation call. The function splits a supplied vector of track IDs into subsets of length 5, then applies a get_recommendations call, 5 tracks at a time. This should generate a data frame of recommended tracks, with length equal to the supplied vector of track ids.

get_recommendations_all(track_ids, valence = NULL)

Arguments

track_ids
A vector containing the IDs of the tracks you'd like recommendations for

valence
The target valence for the recommendations

Value

Returns a data frame containing recommendations from the Spotify API

See also

Examples

# \donttest{
get_recommendations_all(
   track_ids = c("5VIpLopHgolKcSSj7JPCMA", "3QRGYDFFUVb4qneE4DX1gR")
   )

#>                                                                                                                                                                                                   artists
#> 1          https://api.spotify.com/v1/artists/4kswme1Kl2NXRCJ326f14n, 4kswme1Kl2NXRCJ326f14n, WIES, artist, spotify:artist:4kswme1Kl2NXRCJ326f14n, https://open.spotify.com/artist/4kswme1Kl2NXRCJ326f14n
#> 2 https://api.spotify.com/v1/artists/02YLJJnWC7YQVixkjEBRn7, 02YLJJnWC7YQVixkjEBRn7, Billie Marten, artist, spotify:artist:02YLJJnWC7YQVixkjEBRn7, https://open.spotify.com/artist/02YLJJnWC7YQVixkjEBRn7
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>   disc_number duration_ms explicit
#> 1           1      186093    FALSE
#> 2           1      288493    FALSE
#>                                                       href
#> 1 https://api.spotify.com/v1/tracks/1tXLnT5iIyCxe2Kc3M9wH4
#> 2 https://api.spotify.com/v1/tracks/35ZHHlrMKr7CijbD239b0g
#>                       id is_local                       name popularity
#> 1 1tXLnT5iIyCxe2Kc3M9wH4    FALSE        Soms Is Het Te Laat         39
#> 2 35ZHHlrMKr7CijbD239b0g    FALSE Milk & Honey - Alt Version         51
#>                                                                                                   preview_url
#> 1 https://p.scdn.co/mp3-preview/66eb058f4f0a1502db6f1ecd83b2523e9caf1c96?cid=125a10da5fa94687b3d0913168bc3698
#> 2 https://p.scdn.co/mp3-preview/191491ef53ac90e04660c3c0c9f7c4a399f958da?cid=125a10da5fa94687b3d0913168bc3698
#>   track_number  type                                  uri album.album_type
#> 1            1 track spotify:track:1tXLnT5iIyCxe2Kc3M9wH4           SINGLE
#> 2           18 track spotify:track:35ZHHlrMKr7CijbD239b0g            ALBUM
#>                                                                                                                                                                                             album.artists
#> 1          https://api.spotify.com/v1/artists/4kswme1Kl2NXRCJ326f14n, 4kswme1Kl2NXRCJ326f14n, WIES, artist, spotify:artist:4kswme1Kl2NXRCJ326f14n, https://open.spotify.com/artist/4kswme1Kl2NXRCJ326f14n
#> 2 https://api.spotify.com/v1/artists/02YLJJnWC7YQVixkjEBRn7, 02YLJJnWC7YQVixkjEBRn7, Billie Marten, artist, spotify:artist:02YLJJnWC7YQVixkjEBRn7, https://open.spotify.com/artist/02YLJJnWC7YQVixkjEBRn7
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  album.available_markets
#> 1 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#> 2 AD, AE, AG, AL, AM, AO, AR, AT, AU, AZ, BA, BB, BD, BE, BF, BG, BH, BI, BJ, BN, BO, BR, BS, BT, BW, BY, BZ, CA, CH, CI, CL, CM, CO, CR, CV, CW, CY, CZ, DE, DJ, DK, DM, DO, DZ, EC, EE, EG, ES, FI, FJ, FM, FR, GA, GB, GD, GE, GH, GM, GN, GQ, GR, GT, GW, GY, HK, HN, HR, HT, HU, ID, IE, IL, IN, IS, IT, JM, JO, JP, KE, KG, KH, KI, KM, KN, KR, KW, KZ, LA, LB, LC, LI, LK, LR, LS, LT, LU, LV, MA, MC, MD, ME, MG, MH, MK, ML, MN, MO, MR, MT, MU, MV, MW, MX, MY, MZ, NA, NE, NG, NI, NL, NO, NP, NR, NZ, OM, PA, PE, PG, PH, PK, PL, PS, PT, PW, PY, QA, RO, RS, RU, RW, SA, SB, SC, SE, SG, SI, SK, SL, SM, SN, SR, ST, SV, SZ, TD, TG, TH, TL, TN, TO, TR, TT, TV, TW, TZ, UA, UG, US, UY, UZ, VC, VN, VU, WS, XK, ZA, ZM, ZW
#>                                                 album.href
#> 1 https://api.spotify.com/v1/albums/5lttYzHsiNIQAW0DUuGL5K
#> 2 https://api.spotify.com/v1/albums/3WHmv6ghPoM7oA0rP6jMkT
#>                 album.id
#> 1 5lttYzHsiNIQAW0DUuGL5K
#> 2 3WHmv6ghPoM7oA0rP6jMkT
#>                                                                                                                                                                                                                       album.images
#> 1 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273517502e40f7cd1cab9908889, https://i.scdn.co/image/ab67616d00001e02517502e40f7cd1cab9908889, https://i.scdn.co/image/ab67616d00004851517502e40f7cd1cab9908889, 640, 300, 64
#> 2 640, 300, 64, https://i.scdn.co/image/ab67616d0000b273792675f3474149244b0a65a2, https://i.scdn.co/image/ab67616d00001e02792675f3474149244b0a65a2, https://i.scdn.co/image/ab67616d00004851792675f3474149244b0a65a2, 640, 300, 64
#>                                      album.name album.release_date
#> 1                           Soms Is Het Te Laat         2019-09-13
#> 2 Writing of Blues and Yellows (Deluxe Version)         2016-09-23
#>   album.release_date_precision album.total_tracks album.type
#> 1                          day                  1      album
#> 2                          day                 18      album
#>                              album.uri
#> 1 spotify:album:5lttYzHsiNIQAW0DUuGL5K
#> 2 spotify:album:3WHmv6ghPoM7oA0rP6jMkT
#>                             album.external_urls.spotify external_ids.isrc
#> 1 https://open.spotify.com/album/5lttYzHsiNIQAW0DUuGL5K      NLSB71900160
#> 2 https://open.spotify.com/album/3WHmv6ghPoM7oA0rP6jMkT      GBARL1600841
#>                                   external_urls.spotify
#> 1 https://open.spotify.com/track/1tXLnT5iIyCxe2Kc3M9wH4
#> 2 https://open.spotify.com/track/35ZHHlrMKr7CijbD239b0g

# }